==> scripts/1.0_data_wrangling.Rmd <==
---
title: "Data Wrangling"
author: "Indra Boving & Joe Celebrezze"
date: "6/1/2022"
output: html_document
---

```{r, include=FALSE}
knitr::opts_chunk$set(echo = FALSE, warning = FALSE, message = FALSE)
```

# SETUP

```{r setup, include = FALSE}
#To load in necessary packages:
library(gapminder)
library(data.table)
library(purrr)
library(naniar)
library(tidyverse)
library(ggpubr)
library(simputation)
library(visdat)
library(here)
library(psych)
library(kableExtra)
library(rlang) # for {{}} in function writing
filter = dplyr::filter #correct filter problem (gets confused with stats::filter)
here = here::here
read_csv = readr::read_csv
select = dplyr::select
rename = dplyr::rename
```

This script deals with outliers and missing values in flammability metrics for tests conducted on C. megacarpus and A. fasciculatum outside Santa Barbara between 2016 and 2020 on both hotplate and epiradiator devices. 

In addition to measured metrics, it also computes other metrics from the data: flammability index (composite metric that incorporates flame height and time to ignition in a single score), proportion ignited (bins of 5 and 10)
```{r warning=FALSE}
#Read in data:
raw.data <- read.csv(here("raw-data", "flam.local.alldates.csv")) 
```

# 1. Initial Data Cleaning

### Filter out hot plate burns

For this analysis, we were only concerned with burn trials using the 
epiradiator design.

```{r}
raw.data.epi <- raw.data %>% 
  filter(model == "EPI") 
```

### NAs and non-ignitions

First, deal with values that don't make sense due to missingness (for
example, if the temp probe turned off and temp.max is listed as 0, or if
the weight is listed as 0, or if LFM is negative due to missing dry or
wet weight).

Do this to manipulated/controlled variables (weight, temp, mpa) as well
as with flammability metrics.

```{r}
#(need weird extra 0 types here for some reason?)
data.with.na <- raw.data.epi %>% 
  replace_with_na(replace = list(temp.max = c(0, 0.0))) %>% #for missing temp maxs, make NA
  replace_with_na(replace = list(ignition.temp = c(0, 0.0))) %>% #for missing temp at ignition, make NA (all 2016 & 2018 trials)
  replace_with_na(replace = list(start.temp = c(0, 0.0))) %>% #for missing starting temp, make na
  replace_with_na(replace = list(sample.wt = c(0, 0.0))) %>% #for sample weight...
  replace_with_na(replace = list(lfm = c(0.00000))) #for missing lfm 

#To make any odd values due to lack of ignition 'NA':
data.with.na[c("ttfg", "tti", "fd", "gd", "pfg", "fh", "mpa",  "prop.new", "water.wt", "fresh.wt", "dry.wt")][(data.with.na[c("ttfg", "tti", "fd", "gd", "pfg", "fh", "mpa", "prop.new", "water.wt", "fresh.wt", "dry.wt")] == 0)] <- NA
data.with.na[c("ttfg", "tti", "fd", "gd", "pfg", "gti",  "dry.wt", "water.wt", "fresh.wt", "lfm")][(data.with.na[c("ttfg", "tti", "fd", "gd", "pfg", "gti",  "dry.wt", "water.wt", "fresh.wt", "lfm")] < 0)] <- NA
data.with.na[c("ttfg", "tti", "fd", "gd", "pfg", "gti")][(data.with.na[c("ttfg", "tti", "fd", "gd", "pfg", "gti")] > 1000)] <- NA
```

### LFM Outliers

Next, visualize the impact of outliers in LFM with outlierKD function

(NOTE: outlierKD slightly changed here to NOT require yes/no input when
run. Instead, it is performed on a new column to retain unremoved data)

Details on OutlierKD: "To detect the outliers I use the command
boxplot.stats()\$out which use the Tukey's method to identify the
outliers ranged above and below the 1.5\*IQR."
(<https://www.r-bloggers.com/2016/04/identify-describe-plot-and-remove-the-outliers-from-the-dataset/>)

```{r}
##make column to take out outliers from lfm, while keeping column with original lfms too
data.with.na.outs <- data.with.na %>%
  mutate(lfm.outliers.out = lfm) %>%
  unite(year.month, c("year", "month"), remove = FALSE) %>%
  unite(year.month.spp, c("year", "month", "spp"), remove = FALSE)
```

#### OutlierKD2 Function
```{r}
source("http://goo.gl/UUyEzD") #outlier KD (original function)

    #The following function is derived from outlierKD (from above)
outlierKD2 <- function(dt, var) {
  var_name <- eval(substitute(var),eval(dt))
  tot <- sum(!is.na(var_name))
  na1 <- sum(is.na(var_name))
  m1 <- mean(var_name, na.rm = T)
  par(mfrow=c(2, 2), oma=c(0,0,3,0))
  boxplot(var_name, main="With outliers")
  hist(var_name, main="With outliers", xlab=NA, ylab=NA)
  outlier <- boxplot.stats(var_name)$out
  mo <- mean(outlier)
  var_name <- ifelse(var_name %in% outlier, NA, var_name)
  boxplot(var_name, main="Without outliers")
  hist(var_name, main="Without outliers", xlab=NA, ylab=NA)
  title("Outlier Check", outer=TRUE)
  na2 <- sum(is.na(var_name))
  message("Outliers identified: ", na2 - na1, " from ", tot, " observations")
  message("Proportion (%) of outliers: ", (na2 - na1) / tot*100)
  message("Mean of the outliers: ", mo)
  m2 <- mean(var_name, na.rm = T)
  message("Mean without removing outliers: ", m1)
  message("Mean if we remove outliers: ", m2)
    dt[as.character(substitute(var))] <- invisible(var_name)
    assign(as.character(as.list(match.call())$dt), dt, envir = .GlobalEnv)
    message("Outliers successfully removed", "\n")
    return(invisible(dt))
}
```

#### CEME

```{r}
data.with.na.ceme <- data.with.na.outs %>% 
  filter(spp == "CEME")

outlierKD2(data.with.na.ceme, lfm.outliers.out) #check for outliers and remove (happens in outliers.out column, lfm column still contains outliers)

data.with.na.ceme <- data.with.na.ceme %>% 
 mutate(lfm = ifelse(lfm > 400, NA, lfm)) #take out unreasonable values
```

#### ADFA

```{r}
data.with.na.adfa <- data.with.na.outs %>% 
  filter(spp == "ADFA")

outlierKD2(data.with.na.adfa, lfm.outliers.out) #check for outliers and remove (happens in outliers.out column, lfm column still contains outliers)

data.with.na.adfa <- data.with.na.adfa %>% 
 mutate(lfm = ifelse(lfm > 200, NA, lfm))
```

```{r}
data.with.na.new <- rbind(data.with.na.adfa, data.with.na.ceme)
```

### Sample weight outliers
```{r}
data.with.na.new <- data.with.na.new %>% 
  filter(sample.wt > 0.4 & sample.wt < 0.6) # filtering out outlier sample weights
```

### Sample ID adjustment
The sampling ID was slightly different for different sampling regimes (i.e. for more recent sampling, the identification number carries information about both the site and the individual in the format (site.individual or site.pod); however in certain years, only site 1 was sampled from and in these cases the identification number only carries information about the individual in either of the following formats: for individual 3, sample = 3.0 or sample = 3), so to make sure they are all given a sample ID that is similar, we will adjust sample IDs so they are in this format: site.pod.individual with the assumption that, upon each sampling effort, different individuals were sampled from (but from the same relative location (i.e., pod)), as sampling for flammability tests is rather invasive on the plants, so we typically avoided over-sampling if possible.

```{r}
data.with.na.new <- data.with.na.new %>% 
  mutate(year.month.num = case_when(
   year.month == '2016_December' ~ 1,
   year.month == '2018_January' ~ 2,
   year.month %in% c('2019_December','2020_January') ~ 3,
   year.month == '2020_September' ~ 4
  )) %>% 
  mutate(site.pod = case_when(
    sample == 1 ~ 1.1,
    sample == 2 ~ 1.2,
    sample == 3 ~ 1.3,
    sample == 4 ~ 1.4,
    sample == 5 ~ 1.5,
    sample == 6 ~ 1.6,
    sample == 7 ~ 1.7,
    sample == 8 ~ 1.8,
    sample == 9 ~ 1.9,
    sample %in% c(1.1, 1.2, 1.3, 1.4, 1.5, 1.6, 2.1, 2.2, 2.3, 2.4, 2.5, 2.6) ~ sample
  )) %>% 
  unite(sample, c("site.pod", "year.month.num"), sep = '.', remove = TRUE)
```


# 2. Impute missing values

### LFM
```{r}
#impute rlm for each spp. 
data <- data.with.na.new %>%
  mutate(lfm.NAs.imputed = lfm.outliers.out) %>%
  impute_rlm(lfm.NAs.imputed ~ mpa + dry.wt + fresh.wt | year.month.spp)
```

### Weights and Temps
```{r include=FALSE}
#impute values for missing weights:
data <- data %>%
  unite(individual, c("year", "month", "spp", "sample"), remove = FALSE) %>% #create 'individual' column
  impute_median_at(c("dry.wt", "fresh.wt", "water.wt", "sample.wt")) %>%
  ungroup() %>%
  impute_median_at("start.temp") %>%
  mutate(temp.change = temp.max - start.temp)
```

# 3. Add new variables:

Add some columns to the dataframe and manipulate some variables:

-   Add precip columns with 2 month previous precipitation (from SBBG
    daily precip. database, see Precip_SB_DATA.Rmd file for tidying that
    data).
-   Proportion ignite columns: for each increase in LFM of 10% or 5%, what
    proportion of the attempted burns ignited?

### Proportion Ignited 

#### LFM bins of 10
```{r ,include=FALSE}
data <- data %>% 
  #group_by(gr = cut(lfm.NAs.imputed, breaks= seq(0, 400, by = 10))) %>% #create 10 lfm.NAs.imputed segments, indicate these in new column called "gr"
  group_by(gr = cut(lfm, breaks= seq(0, 400, by = 10))) %>%
  ungroup() %>% #ungroup - this will stop issues in the future (we can always regroup again)
  group_by(year, month, model, spp, gr) %>% #create groups based on trial, model, species, and lfm segment
  add_tally %>% #column with n for each of these groups
  ungroup() %>% #ungroup so r doesnt get confused
  unite(model.spp.gr, c("year", "month", "model", "spp", "gr"), remove = FALSE) %>% #create column for each group as above, which will be associated with the n of each group based on the above add_tally step. 
  unite(model.spp.gr.sample, c("year", "month", "model", "spp", "gr", "sample"), remove = FALSE) %>% #create column with group for each indvidual sample (this is a unique ID)
  unite(individual, c('year','month','spp', 'sample'), remove = FALSE) %>% 
  #group for month and year
  mutate(precip.2month = year.month) %>%
 mutate(precip.2mo = recode(precip.2month, "2020_September" = 0.01, "2018_January" = 0.09, "2019_December" = 3.69, "2020_January" = 5.82, "2016_December" = 3.49)) %>%
  mutate(season = year.month) %>%
  mutate(season = recode(season, "2020_September" = "Dry", "2018_January" = "Dry", "2019_December" = "Wet", "2020_January" = "Wet", "2016_December" = "Wet" )) %>%  #group for month and year
  group_by(model.spp.gr)#group by year, model, spp, group column
 
setDT(data)[,total:=sum(ignition),by=model.spp.gr][] #add column called "total" with the total number of ignitions per group ("sum" works because we have 0s and 1s)

data <-  data %>%
  mutate(prop.ignite = paste0(round(100 * total/n))) %>% #for each group, divide total ignitions ("total" column) by total number burn attempts (n column), as a percentage
  ungroup() %>% #ungroup so r doesnt get confused
  dplyr::select(-n, -total, lfm) #remove columns we aren't interested in anymore

data$prop.ignite <- as.numeric(data$prop.ignite) #Since, otherwise, it would be a character vector
```

#### LFM bins of 5
```{r}
data.5bins <- data %>% 
  group_by(bins5lfm_gr = cut(lfm, breaks= seq(0, 400, by = 5))) %>% #create 10 lfm.NAs.imputed segments, indicate these in new column called "gr"
  ungroup() %>% #ungroup - this will stop issues in the future (we can always regroup again)
  group_by(year, month, model, spp, bins5lfm_gr) %>% #create groups based on trial, model, species, and lfm segment
  add_tally %>% #column with n for each of these groups
  ungroup() %>% #ungroup so r doesnt get confused
  unite(model.spp.bins5lfm_gr, c("year", "month", "model", "spp", "bins5lfm_gr"), remove = FALSE) %>% #create column for each group as above, which will be associated with the n of each group based on the above add_tally step. 
  unite(individual.bins5lfm_gr, c("year", "month", "model", "spp", "bins5lfm_gr", "sample"), remove = FALSE) %>% #create column with group for each indvidual sample (this is a unique ID)
  group_by(individual.bins5lfm_gr)#group by year, model, spp, group column
 
setDT(data.5bins)[,total_bins5lfm_gr:=sum(ignition),by=model.spp.bins5lfm_gr][] #add column called "total" with the total number of ignitions per group ("sum" works because we have 0s and 1s)

data.5bins <-  data.5bins %>%
  mutate(prop.ignite.bins5 = paste0(round(100 * total_bins5lfm_gr/n))) %>% #for each group, divide total ignitions ("total" column) by total number burn attempts (n column), as a percentage
  ungroup() %>% #ungroup so r doesnt get confused
  select(-n, -total_bins5lfm_gr) #remove columns we aren't interested in anymore

data.5bins$prop.ignite.bins5 <- as.numeric(data.5bins$prop.ignite.bins5) #Since, otherwise, it would be a character vector
```

Need to do the above, but without grouping by trial date
```{r}
data.5bins_nodate <- data.5bins %>% 
  #group_by(bins5lfm_gr = cut(lfm.NAs.imputed, breaks= seq(0, 400, by = 5))) %>% #create 10 lfm.NAs.imputed segments, indicate these in new column called "gr", did that already
  ungroup() %>% #ungroup - this will stop issues in the future (we can always regroup again)
  group_by(model, spp, bins5lfm_gr) %>% #create groups based on trial, model, species, and lfm segment
  add_tally %>% #column with n for each of these groups
  ungroup() %>% #ungroup so r doesnt get confused
  unite(model.spp.bins5lfm_gr_nodate, c("model", "spp", "bins5lfm_gr"), remove = FALSE) %>% #create column for each group as above, which will be associated with the n of each group based on the above add_tally step. 
  unite(model.spp.bins5lfm_gr.sample_nodate, c("model", "spp", "bins5lfm_gr", "sample"), remove = FALSE) %>% #create column with group for each indvidual sample (this is a unique ID)
  rename(individual.bins5lfm_gr_nodate = model.spp.bins5lfm_gr.sample_nodate) %>% 
  group_by(individual.bins5lfm_gr_nodate)#group by year, model, spp, group column
 
setDT(data.5bins_nodate)[,total_bins5lfm_gr_nodate:=sum(ignition),by=model.spp.bins5lfm_gr_nodate][] #add column called "total" with the total number of ignitions per group ("sum" works because we have 0s and 1s)

data.5bins_nodate <-  data.5bins_nodate %>%
  mutate(prop.ignite.bins5_nodate = paste0(round(100 * total_bins5lfm_gr_nodate/n))) %>% #for each group, divide total ignitions ("total" column) by total number burn attempts (n column), as a percentage
  ungroup() %>% #ungroup so r doesnt get confused
  select(-n, -total_bins5lfm_gr_nodate) #remove columns we aren't interested in anymore

data.5bins_nodate$prop.ignite.bins5_nodate <- as.numeric(data.5bins_nodate$prop.ignite.bins5_nodate) #Since, otherwise, it would be a character vector
```

```{r}
data.10bins_nodate <- data.5bins_nodate %>% 
  group_by(bins10lfm_gr = cut(lfm, breaks= seq(0, 400, by = 10))) %>% #create 10 lfm.NAs.imputed segments, indicate these in new column called "gr", did that already
  ungroup() %>% #ungroup - this will stop issues in the future (we can always regroup again)
  group_by(model, spp, bins10lfm_gr) %>% #create groups based on trial, model, species, and lfm segment
  add_tally %>% #column with n for each of these groups
  ungroup() %>% #ungroup so r doesnt get confused
  unite(model.spp.bins10lfm_gr_nodate, c("model", "spp", "bins10lfm_gr"), remove = FALSE) %>% #create column for each group as above, which will be associated with the n of each group based on the above add_tally step. 
  unite(model.spp.bins10lfm_gr.sample_nodate, c("model", "spp", "bins10lfm_gr", "sample"), remove = FALSE) %>% #create column with group for each indvidual sample (this is a unique ID)
  rename(individual.bins10lfm_gr_nodate = model.spp.bins10lfm_gr.sample_nodate) %>% 
  group_by(individual.bins10lfm_gr_nodate)#group by year, model, spp, group column
 
setDT(data.10bins_nodate)[,total_bins10lfm_gr_nodate:=sum(ignition),by=model.spp.bins10lfm_gr_nodate][] #add column called "total" with the total number of ignitions per group ("sum" works because we have 0s and 1s)

data.10bins_nodate <-  data.10bins_nodate %>%
  mutate(prop.ignite.bins10_nodate = paste0(round(100 * total_bins10lfm_gr_nodate/n))) %>% #for each group, divide total ignitions ("total" column) by total number burn attempts (n column), as a percentage
  ungroup() %>% #ungroup so r doesnt get confused
  select(-n, -total_bins10lfm_gr_nodate) #remove columns we aren't interested in anymore

data.10bins_nodate$prop.ignite.bins10_nodate <- as.numeric(data.10bins_nodate$prop.ignite.bins10_nodate) #Since, otherwise, it would be a character vector
```


```{r}
data <- data.10bins_nodate 
```
-   Categorical "dry" vs. "wet" vs. "moderate" column for season, based
    on prior precip and date (dry = <1 in rain, wet = >3 in rain)
-   Various categorical groups based on LFM, spp., models, date, etc.
    used in grouping later on or in creating the above columns.
-   Column labeled "hydration" with break at 60% LFM for dry, 60 - 90%
    for moderate, and \> 100% for hydrated
-   Make bins of varying lfm sizes (5, 10, 20)


```{r, include= FALSE}
#impute lfm based on relationship with Mpa, grouping by species:
data <- data %>%
  group_by(spp) %>%
#then create bins for LFM, in case we want to compare via bins
  mutate(hydration = cut(lfm, breaks = c(0, 60, 90, 400), labels = c("dry", "moderate", "hydrated"), ordered_result = TRUE)) %>% 
  ungroup() %>%
  group_by(bins5lfm_gr= cut(lfm, breaks= seq(0, 400, by = 5), ordered_result = TRUE)) %>% #create 10 LFM segments, indicate these in new column called "bins5lfm_gr"
  ungroup() %>% #ungroup - this will stop issues in the future (we can always regroup again)
  group_by(bins10lfm_gr= cut(lfm, breaks= seq(0, 400, by = 10), ordered_result = TRUE)) %>% #create 10 LFM segments, indicate these in new column called "bins10lfm_gr"
  ungroup() %>% #ungroup - this will stop issues in the future (we can always regroup again)
  group_by(bins20lfm_gr = cut(lfm, breaks= seq(0, 400, by = 20), ordered_result = TRUE)) %>% #create 20 LFM segments, indicate these in new column called "bins20lfm_gr"
  ungroup() 
```

### Dry Weight vs. Fresh Weight Metrics
```{r}
# Make dry weight columns, etc. for PV comparison: 

# calculate metrics for weight values (read . as "per", i.e. gdw.gfw is "gram dry weight per gram fresh weight"):
data <- data %>%
  dplyr::mutate(gdw.gfw = dry.wt/fresh.wt) %>%
  dplyr:::mutate(gww.gdw = lfm.NAs.imputed/100) %>%
  dplyr::mutate(gdw.gww = dry.wt/water.wt) %>%
  dplyr::mutate(dw.flam.sample = sample.wt * gdw.gfw) %>%
  dplyr::mutate(ww.flam.sample = sample.wt - dw.flam.sample, 
         ww.outs = ww.flam.sample,
         dw.outs = dw.flam.sample) %>%
  replace_with_na_at(.vars = c("ww.flam.sample"), condition = ~.x < 0) # oe was below 0 (due to missing lfm), so remove that one 
```

#### wet weights
```{r}
outlierKD2(data, ww.outs)
```

Remove very high ww
```{r}
data <- data %>% 
        filter(ww.flam.sample < .75)
```

#### dry weights

```{r}
outlierKD2(data, dw.outs)
```

```{r}
data %>% 
ggplot() +
  geom_point(aes(y = dw.flam.sample, x = mpa), color = "blue") +
  geom_point(aes(y = dw.outs, x = mpa), color = "goldenrod")
```
Remove very high: 

```{r}
data <- data %>% 
  filter(dw.flam.sample < 0.46)
```

------------------------------------------------------------------------

```{r}
data_flam_outs <- data %>% 
  mutate(tti.outs = tti, 
         gti.outs = gti, 
         gd.outs = gd, 
         fd.outs = fd, 
         pfg.outs = pfg, 
         fh.outs= fh, 
         temp.max.outs = temp.max, 
         ignition.temp.outs = ignition.temp
         )
```

# 4. Outliers: Flam. Metrics

## Outlier Visualization Function
```{r}
outlierVis <- function(df, metric, outlier, x.value = mpa) {
  x <- ggplot2::ggplot(data = df)+
                geom_point(aes(y = {{metric}}, x = {{x.value}}), color = "blue")+
                geom_point(aes(y = {{outlier}}, x = {{x.value}}), color = "goldenrod")+
                theme_bw()
  x
}
```


## CEME only:
```{r}
data.ceme <- data_flam_outs %>% 
  filter(spp == "CEME")

ignite.only.ceme <- data.ceme %>%
  filter(ignition == "1")
```

These run and remove outliers automatically, need to quiet the code if
that is not what we want.

#### tti
```{r}
outlierKD2(ignite.only.ceme, tti.outs)
outlierVis(ignite.only.ceme, tti, tti.outs)
```
Don't remove any 

#### gti

```{r}
outlierKD2(ignite.only.ceme, gti.outs)
outlierVis(ignite.only.ceme, gti, gti.outs)
```
Dont remove

#### gd
```{r}
outlierKD2(ignite.only.ceme, gd.outs)
outlierVis(ignite.only.ceme, gd, gd.outs)
```
Dont remove

#### fd
```{r}
outlierKD2(ignite.only.ceme, fd.outs)
outlierVis(ignite.only.ceme, fd, fd.outs)
```

Remove the point that is an obvious outlier at 0 mpa:
```{r}
ignite.only.ceme <- ignite.only.ceme %>% 
 mutate(fd = ifelse(fd > 30 | mpa > -1, NA, fd))
```

#### pfg
```{r}
outlierKD2(ignite.only.ceme, pfg.outs)
outlierVis(ignite.only.ceme, pfg, pfg.outs)
```
Dont remove

#### fh
```{r}
outlierKD2(ignite.only.ceme, fh.outs)
outlierVis(ignite.only.ceme, fh, fh.outs)
```

Remove high value at ~0 MPa:
```{r}
ignite.only.ceme <- ignite.only.ceme %>% 
  mutate(fh = ifelse(fh > 20 | mpa > -1, NA, fh))
```

#### max temp
```{r}
outlierKD2(ignite.only.ceme, temp.max.outs)
outlierVis(ignite.only.ceme, temp.max, temp.max.outs)
```
Dont remove

#### ignition temp
```{r}
outlierKD2(ignite.only.ceme, ignition.temp.outs)
outlierVis(ignite.only.ceme, ignition.temp, ignition.temp.outs)
```

Look at distributions of flammability metrics so we can think about how
to deal with NAs in those variables, and then deal with them.

## Distribution following imputation: 

```{r, include=FALSE}
### Subset data and perform imputations (median):
#str(data)
ignite.only.ceme %>%
  mutate(gti = as.numeric(gti)) 

ignite.only.ceme <- ignite.only.ceme %>%
  group_by(year.month) %>% 
  group_by(spp) %>% #this makes us take the median of based on the species, the model, and the lfm bins (of 10 lfm) the missing value is in. Preserves species differences, and lfm impact. 
  impute_median_at(c("ttfg", "gd", "gti", "fd", "pfg","temp.max", "start.temp", "ignition.temp", "prop.ignite", "fh", "tti")) 
```


## Look at flam characteristics to see if they make sense:

Here we are normalizing by the dry weight of the flammability sample,
which was calculated by multiplying the weight of the burned sample by
the ratio of dry weight to fresh weight in the samples used for lfm
measurements.

(Assumptions here: that the ratio of water to dry weight in burned and
lfm samples was the same)

```{r}
distributions.lfm <- ignite.only.ceme %>%
  ggplot() +
  geom_density(aes(x = sample.wt, color = spp)) 
distributions.lfm  

ignite.only.ceme <- ignite.only.ceme %>%  #Dealing with trials that ignited
  dplyr::mutate(mean.wt = mean(sample.wt, na.rm = TRUE)) %>% #take the mean of all sample weights for each model
  ungroup() %>% 
  mutate_if(is.integer, as.numeric) 

ignite.only.ceme$gti <- as.numeric(as.character(ignite.only.ceme$gti)) #not sure why this became a character...
#str(ignite.only.ceme)

distributions.lfm <- ignite.only.ceme %>%
  ggplot() +
  geom_density(aes(x = lfm.NAs.imputed, color = spp))
distributions.lfm 

distributions.lfm <- ignite.only.ceme %>%
  ggplot() +
  geom_density(aes(x = ww.flam.sample, color = spp))
distributions.lfm
```

```{r, include=FALSE}
ignite.only.ceme <- ignite.only.ceme %>%
  mutate(gti = replace(gti, gti == "0", "0.5")) %>% 
  mutate(gti = as.numeric(gti)) %>%
  mutate("dry.norm.fh" = fh/dw.flam.sample) %>% #create new columns, etc.
  mutate("dry.norm.gd" = gd/dw.flam.sample) %>%
  mutate("dry.norm.fd" = fd/dw.flam.sample) %>%
  mutate("dry.norm.pfg" = pfg/dw.flam.sample) %>%
  mutate("dry.norm.ttfg" = ttfg/dw.flam.sample) %>%
  mutate("dry.norm.tti" = tti/dw.flam.sample) %>% 
  mutate("dry.norm.gti" = gti/dw.flam.sample) 
#str(ignite.only.ceme)
```

## ADFA only:

```{r}
data.adfa <- data_flam_outs %>% 
  filter(spp == "ADFA")

ignite.only.adfa <- data.adfa %>%
  filter(ignition == "1") #only look at ignited samples
```

These run and remove outliers automatically, need to quiet the code if
that is not what we want.

#### tti

```{r}
outlierKD2(ignite.only.adfa, tti.outs)
outlierVis(ignite.only.adfa, tti, tti.outs)
```

#### gti

```{r}
outlierKD2(ignite.only.adfa, gti.outs)
outlierVis(ignite.only.adfa, gti, gti.outs)
```

#### gd

```{r}
outlierKD2(ignite.only.adfa, gd.outs)
outlierVis(ignite.only.adfa, gd, gd.outs)
```

#### fd
```{r}
outlierKD2(ignite.only.adfa, fd.outs)
outlierVis(ignite.only.adfa, fd, fd.outs)
```

Remove the point that is an obvious outlier at 0 mpa:
```{r}
ignite.only.adfa <- ignite.only.adfa %>% 
 mutate(fd = ifelse(fd > 30, NA, fd))
```

#### pfg
```{r}
outlierKD2(ignite.only.adfa, pfg.outs)
outlierVis(ignite.only.adfa, pfg, pfg.outs)
```
Dont remove

#### fh
```{r}
outlierKD2(ignite.only.adfa, fh.outs)
outlierVis(ignite.only.adfa, fh, fh.outs)
```

#### temp max

```{r}
outlierKD2(ignite.only.adfa, temp.max.outs)
outlierVis(ignite.only.adfa, temp.max, temp.max.outs)
```
Dont remove

#### ignition temp
```{r}
outlierKD2(ignite.only.adfa, ignition.temp.outs)
outlierVis(ignite.only.adfa, ignition.temp, ignition.temp.outs)
```

Look at distributions of flammability metrics so we can think about how
to deal with NAs in those variables, and then deal with them:

First, to visualize NA values in flam metrics: - Ignited samples only,
both models

### Subset data and perform imputations (median):

```{r, include=FALSE}
#str(data)
ignite.only.adfa %>%
  mutate(gti = as.numeric(gti)) 

ignite.only.adfa <- ignite.only.adfa %>%
  group_by(year.month) %>% 
  group_by(spp) %>% #this makes us take the median of based on the species, the model, and the lfm bins (of 10 lfm) the missing value is in. Preserves species differences, and lfm impact. 
  impute_median_at(c("ttfg", "gd", "gti", "fd", "pfg","temp.max", "start.temp", "ignition.temp", "prop.ignite", "fh", "tti")) 

```

Here we are normalizing by the dry weight of the flammability sample,
which was calculated by multiplying the weight of the burned sample by
the ratio of dry weight to fresh weight in the samples used for lfm
measurements.

(Assumptions here: that the ratio of water to dry weight in burned and
lfm samples was the same)

```{r}
ignite.only.adfa <- ignite.only.adfa %>%  #Dealing with trials that ignited (can change later)
  dplyr::mutate(mean.wt = mean(sample.wt, na.rm = TRUE)) %>% #take the mean of all sample weights for each model
  ungroup() %>% 
  mutate_if(is.integer, as.numeric) 

ignite.only.adfa$gti <- as.numeric(as.character(ignite.only.adfa$gti)) #not sure why this became a character...
```

```{r, include=FALSE}
ignite.only.adfa <- ignite.only.adfa %>%
  mutate(gti = replace(gti, gti == "0", "0.5")) %>% 
  mutate(gti = as.numeric(gti)) %>%
  mutate("dry.norm.fh" = fh/dw.flam.sample) %>% #create new columns, etc.
  mutate("dry.norm.gd" = gd/dw.flam.sample) %>%
  mutate("dry.norm.fd" = fd/dw.flam.sample) %>%
  mutate("dry.norm.pfg" = pfg/dw.flam.sample) %>%
  mutate("dry.norm.ttfg" = ttfg/dw.flam.sample) %>%
  mutate("dry.norm.tti" = tti/dw.flam.sample) %>% 
  mutate("dry.norm.gti" = gti/dw.flam.sample) 
```

------------------------------------------------------------------------

# Combine back into main datasets:

```{r}
no.ignites.ceme <- data.ceme %>%
  filter(ignition == "0") #only look at non-ignited samples

no.ignites.adfa <- data.adfa %>%
  filter(ignition == "0") #only look at non-ignited samples

no.ignites.all <- rbind(no.ignites.adfa, no.ignites.ceme)

data.local.ignites <- rbind(ignite.only.ceme, ignite.only.adfa)
```

## New Column: Flammability Index
Derived from Essaghi et. al. 2017, using maximum *tti* and maximum *fh*, an index of flammability as a scale of least (1) to most (20) flammable
```{r}
data.local.ignites <- data.local.ignites %>% 
  mutate(flam.index = 15.30412*((86-tti)/(8+tti))^((fh)/(fh + 27))^2)
```

```{r}
data.local <- bind_rows(data.local.ignites, no.ignites.all)

colSums(is.na(data.local))

write_csv(data.local, here("processed-data",  "local_flam_data_all.csv"))
```

------------------------------------------------------------------
# 5. Analysis-Specific
## Segmented Regressions
```{r}
mem.data.alldates <- read_csv(here("processed-data", "mem.data.local.epi.alldates.csv"), show_col_types = FALSE)

mem.data.raw <- mem.data.alldates %>%
  mutate(year = as.factor(year)) %>%
  mutate(month = as.factor(month)) %>% 
  filter(sample.wt > 0.4 & sample.wt < 0.6) # filtering out outlier sample weights (sorry for redundancy in code)

for(i in 1:length(mem.data.raw$mpa)) {
  if(mem.data.raw$mpa[i] > 0){
    mem.data.raw$mpa[i] <- mem.data.raw$mpa[i]*-1
  }
  else{mem.data.raw$mpa[i] <- mem.data.raw$mpa[i]}
}

clean.mem.data <- mem.data.raw[,c('PC1', 'PC2', 'PC3', 'PC4', 'spp', 'lfm.NAs.imputed', 'lfm.outliers.out', 'mpa', 'dw.flam.sample', 'year.month', 'individual', 'site','sample.wt', 'ww.flam.sample', 'start.temp',  'year', 'month', 'tti', 'fh', 'gd', 'fd', 'gti', 'pfg', 'ttfg', 'temp.max','flam.index', 'precip.2mo', 'prop.ignite.bins5_nodate', 'prop.ignite.bins10_nodate', 'gr')]
```

```{r}
seg.data.subset <- clean.mem.data %>% 
  mutate(lfm = lfm.NAs.imputed, 
         id = individual) %>% 
 dplyr::select(id, PC1, PC2, PC3, mpa, lfm, ww.flam.sample, dw.flam.sample, site, sample.wt, spp, tti, fh, fd, gd, gti, pfg, temp.max, year.month, prop.ignite.bins10_nodate, prop.ignite.bins5_nodate, ttfg, gr)

seg.data.subset<-seg.data.subset[order(seg.data.subset$id),]

write_csv(seg.data.subset, here("processed-data", "seg.reg.data.local.csv"))
```

## Mixed Effects Models
```{r}
mem.data.alldates <- read_csv(here("processed-data", "mem.data.local.epi.alldates.csv"), show_col_types = FALSE)

mem.data.alldates <- mem.data.alldates %>%
  mutate(mpa = case_when(
    mpa > 0 ~ (-1*mpa), 
    mpa < 0 ~ mpa)) %>% 
  mutate(year = as.factor(year)) %>%
  mutate(month = as.factor(month)) %>% 
  mutate(hydration = cut(lfm, breaks = c(0, 60, 90, 400), labels = 
        c("dry", "moderate", "hydrated"), ordered_result = TRUE)) %>% 
  mutate(excess_water.scaled = ww.flam.sample - dw.flam.sample) %>% 
  filter(sample.wt > 0.4 & sample.wt < 0.6) # filtering out outlier sample weights (sorry for redundancy in code)

clean.mem.data.alldates <- mem.data.alldates %>% 
  mutate(mpa.unscaled = mpa) %>% 
  mutate(lfm.unscaled = lfm)%>% 
  mutate(lfm.NAs.imputed.unscaled = lfm.NAs.imputed) %>% 
  mutate(dw.flam.sample.unscaled = dw.flam.sample) %>% 
  mutate(ww.flam.sample.unscaled = ww.flam.sample) %>% 
  mutate(flam.sample.wt.unscaled = sample.wt) %>% 
  mutate(excess_water.unscaled = excess_water.scaled)

#Parsing out the variables of interest for the mixed effects modeling including dependent variables and unscaled predictors
clean.mem.data.alldates <- clean.mem.data.alldates %>% 
  select(c('PC1', 'PC2', 'PC3', 'PC4', 'tti', 'fh', 'gd', 'fd', 'temp.max', 'pfg', 'gti', 'ttfg', 'flam.index', 'year', 'month', 'prop.ignite', 'prop.ignite.bins5', "ignition", 'spp', 'year.month', 'year', 'month', 'individual', 'site', 'start.temp', 'precip.2mo','mpa.unscaled', 'lfm.unscaled', 'dw.flam.sample.unscaled', 'lfm.NAs.imputed.unscaled', 'excess_water.unscaled', 'hydration', 'flam.sample.wt.unscaled', 'ww.flam.sample.unscaled')) %>% 
  arrange(individual)

#Now, splitting by species and then scaling predictors and then using rbind to put back into a single dataframe (scaled.mem.data.predictors.alldates)
scaled.mem.data.predictors.alldates.adfa <- mem.data.alldates %>%
  filter(spp == "ADFA") %>% 
  select(
    lfm.NAs.imputed,
    lfm.outliers.out,
    mpa,
    dw.flam.sample,
    sample.wt,
    ww.flam.sample,
    lfm,
    excess_water.scaled,
    individual
  ) %>%
  mutate (lfm.NAs.imputed = scale(lfm.NAs.imputed),
  lfm.outliers.out = scale(lfm.outliers.out),
  mpa.scaled = scale(mpa),
  dw.flam.sample = scale(dw.flam.sample),
  sample.wt = scale(sample.wt),
  ww.flam.sample = scale(ww.flam.sample),
  lfm = scale(lfm),
  excess_water.scaled = scale(excess_water.scaled)) %>% 
  dplyr::rename(individual.a = individual)
  
scaled.mem.data.predictors.alldates.ceme <- mem.data.alldates %>%
  filter(spp == "CEME") %>% 
  select(
    lfm.NAs.imputed,
    lfm.outliers.out,
    mpa,
    dw.flam.sample,
    sample.wt,
    ww.flam.sample,
    lfm,
    excess_water.scaled,
    individual
  ) %>% 
  mutate( lfm.NAs.imputed = scale(lfm.NAs.imputed),
    lfm.outliers.out = scale(lfm.outliers.out),
    mpa.scaled = scale(mpa),
    dw.flam.sample = scale(dw.flam.sample),
    sample.wt = scale(sample.wt),
    ww.flam.sample = scale(ww.flam.sample),
    lfm = scale(lfm),
    excess_water.scaled =
      scale(excess_water.scaled)
  ) %>% 
  dplyr::rename(individual.a = individual)
  
scaled.mem.data.predictors.alldates <- rbind(scaled.mem.data.predictors.alldates.adfa, scaled.mem.data.predictors.alldates.ceme) %>% 
  arrange(individual.a)

scaled.mem.data.alldates <- cbind(scaled.mem.data.predictors.alldates, 
                                   clean.mem.data.alldates)

scaled.mem.data.alldates <- scaled.mem.data.alldates %>%
  select(-individual.a) %>% 
  mutate(Species = case_when(
    spp == "ADFA" ~ "A. fasciculatum", 
    spp == "CEME" ~ "C. megacarpus"
  )) %>% 
  mutate(id = individual)  %>% 
  mutate(timing = case_when(
    year.month == "2016_December" ~ "Dec. 2016", 
    year.month == "2019_December" ~ "Dec. 2019",
    year.month == "2018_January" ~ "Jan. 2018",
    year.month == "2020_January" ~ "Jan. 2020",
    year.month == "2020_September" ~ "Sept. 2020"
  ))

scaled.mem.data.alldates.noNAs <- scaled.mem.data.alldates %>% 
  drop_na(mpa, lfm, PC1)

write.csv(scaled.mem.data.alldates.noNAs, here('processed-data', 'mem.data.subset.alldates.csv'))

```
