==> scripts/1.1_SBBG_precip.Rmd <==
---
title: "Precipitation Data"
author: "Indra Boving & Joe Celebrezze"
date: "2/26/2021"
output: html_document
---

```{r setup, include=FALSE}
knitr::opts_chunk$set(echo = TRUE)
```

```{r include=FALSE}
library(ggplot2)
library(gapminder)
library(data.table)
library(purrr) #for visualizing all variables in one big plot
library(naniar) #deal with NAs
library(dplyr)
library(tidyverse)
library(ggpubr)
library(here)
here=here::here
select = dplyr::select
filter = dplyr::filter #correct filter problem (gets confused with stats::filter)
```

Monthly data from SBBG's website: 

#2 months prior: 

```{r}
precip <- read.csv(here("raw-data","Precip_SBBG.csv")) %>% 
  filter(year > 2014) %>%
  unite(year.month, c("year", "month"), remove = FALSE) %>%
  group_by(year.month) %>%
  mutate(rainfall = sum(daily.rain))
precip 

Sept2020 <- precip %>%
  filter(year.month == "2020_8" | year.month == "2020_7") %>%
  group_by(year) %>%
  mutate(rainfall.2mo = sum(daily.rain), 
         timing = "Sept_2020")
Sept2020

Jan2020 <- precip %>%
  filter(year.month == "2020_01" | year.month == "2019_12") %>%
  group_by(year) %>%
  mutate(rainfall.2mo = sum(daily.rain), 
         timing = "Jan_2020")
Jan2020
#For Jan2020, uses 2020-01 and 2019-12

Dec20191 <- precip %>%
  filter(year.month == "2019_12") %>%
  filter(day < 10) %>%
  group_by(year) %>%
  mutate(rainfall.2mo = sum(daily.rain)) %>%
  ungroup()
Dec20191

Dec20192 <- precip %>%
  filter(year.month == "2019_11") %>%
  group_by(year) %>%
  mutate(rainfall.2mo = sum(daily.rain)) %>%
  ungroup()
Dec20192

Dec2019 <- rbind(Dec20191, Dec20192) %>%
  mutate(rainfall.2mo = sum(daily.rain),
         timing = "Dec_2019")
Dec2019
  
Jan2018 <- precip %>%
  filter(year.month == "2017_12" | year.month == "2017_11") %>%
  group_by(year) %>%
  mutate(rainfall.2mo = sum(daily.rain), 
         timing = "Jan_2018")
Jan2018
#but for jan2018, uses 2017-12 and 2017-11

Dec2016 <- precip %>%
  filter(year.month == "2016_12" | year.month == "2016_11") %>%
  group_by(year) %>%
  mutate(rainfall.2mo = sum(daily.rain), 
         timing = "Dec_2016")
Dec2016
# and then for dec2016, it uses 2015 data?

precip.filtered_2mo <- rbind(Dec2016, Dec2019, Jan2018, Dec2019, Sept2020, Jan2020) %>%
  select(water.year, year.month, day, daily.rain, rainfall.2mo, timing)
precip.filtered_2mo

write.csv(precip.filtered_2mo, here("processed-data", "precip.clean.csv"))
```

#4 months prior: 

```{r eval=FALSE, include=FALSE}
Sept2020 <- precip %>%
  filter(year.month == "2020_8" | year.month == "2020_7" | year.month == "2020_6" | year.month == "2020_5") %>%
  group_by(year) %>%
  mutate(rainfall.4mo = sum(daily.rain))
Sept2020

Jan2020 <- precip %>%
  filter(year.month == "2020_1" | year.month == "2019_12" | year.month == "2019_11" | year.month == "2019_10") %>%
  group_by(year) %>%
  mutate(rainfall.4mo = sum(daily.rain))
Jan2020

Dec20191 <- precip %>%
  filter(year.month == "2019_12") %>%
  filter(day < 10) 
Dec20191

Dec20192 <- precip %>%
  filter(year.month == "2019_11" | year.month == "2019_10" | year.month == "2019_9")
Dec20192

Dec2019 <- rbind(Dec20191, Dec20192) %>%
  group_by(year) %>%
  mutate(rainfall.4mo = sum(daily.rain))
Dec2019
  

Jan2018 <- precip %>%
  filter(year.month == "2017_12" | year.month == "2017_11" | year.month == "2017_10" | year.month == "2017_9") %>%
  group_by(water.year) %>%
  mutate(rainfall.4mo = sum(daily.rain))
Jan2018


Dec2016 <- precip %>%
  filter(year.month == "2016_12" | year.month == "2016_11" | year.month == "2016_10" | year.month == "2016_9") %>%
  group_by(year) %>%
  mutate(rainfall.4mo = sum(daily.rain))
Dec2016

precip.filtered <- rbind(Dec2016, Dec2019, Jan2018, Jan2020, Sept2020) %>%
  select(water.year, year.month, day, daily.rain, rainfall.4mo)
precip.filtered

write.csv(precip.filtered, here("processed-data", "precip.clean.csv"))
```
