==> scripts/2_PCA.Rmd <==
---
title: "PCA Figure and Table"
author: "Indra Boving & Joe Celebrezze"
date: "6/20/2022"
output: html_document
---

```{r setup, include=FALSE}
knitr::opts_chunk$set(echo = TRUE)
```

# Setup
Necessary packages:
```{r}
library(here)
here = here::here
library(tidyverse)
library(kableExtra)
library(sjPlot)
library(ggpubr)
rename = dplyr::rename
```

# 1.1. PCA Figure

Plot setup:
```{r warning=FALSE}
### Note: This dataset includes only Epiradiator and Ignited samples
epi.flamdata <- read_csv(here("processed-data", "local_flam_data_all.csv"), show_col_types = FALSE) %>%
  mutate(LFM = lfm.NAs.imputed) %>% 
  filter(ignition == 1, model == "EPI")

 # filter(year == 2020)

figure_epi_pca_quant <- (epi.flamdata[, c("fh", "ttfg", "tti", "fd", "gd", "gti", "pfg",
         "temp.max")]) %>%  
 transmute( "               Flame Duration" = fd, #using spaces to shift labels over instead of \t since \t is leading to little blank squares popping up on the plot.. spaces are just manual tabs anyways right?
           "                 Flame Height" = fh,
           "Post-flame Glow" = pfg,
          "Time to Ignition" = tti,
          "Max. Temp" = temp.max,
            "Glow Duration" = gd,
          "\nGlow to Ignition" = gti,
          "\nTime to First Glow" = ttfg
          )

figure_epi_pca_cat <- (epi.flamdata[c("hydration", "spp", "year.month", "sample")]) %>% 
  transmute("Hydration" = hydration, 
            spp = spp,
            "Timing" = year.month, 
            "Sample" = sample) %>% 
  mutate(Species = case_when(
    spp == "ADFA" ~ "A. fasciculatum", 
    spp == "CEME" ~ "C. megacarpus"
  ))

figure_prcomp <- prcomp(na.omit(figure_epi_pca_quant), center = TRUE, scale = TRUE, retx=TRUE)

#Setting up plotting data: 

#decisions for plotting:
choices = 1:2 
scale = 1
obs.scale = 1 - scale
var.scale = scale
ellipse.prob = 0.68
labels.size = 3
circle.prob = 0.69
choices = 1:2

#Run PCA
pcobj <- prcomp(na.omit(figure_epi_pca_quant), center = TRUE, scale = TRUE)

# extract PCA components: 
nobs.factor <- sqrt(nrow(pcobj$x) - 1)
    d <- pcobj$sdev
    u <- sweep(pcobj$x, 2, 1/(d * nobs.factor), FUN = "*")
    v <- pcobj$rotation

  choices <- pmin(choices, ncol(u))
  df.u <- as.data.frame(sweep(u[, choices], 2, d[choices]^obs.scale, 
                              FUN = "*"))
  v <- sweep(v, 2, d^var.scale, FUN = "*")
  df.v <- as.data.frame(v[, choices])
  names(df.u) <- c("PC1", "PC2")
  names(df.v) <- names(df.u)
  df.u <- df.u * nobs.factor

  r <- sqrt(qchisq(circle.prob, df = 2)) * prod(colMeans(df.u^2))^(1/4)
  
v.scale <- rowSums(v^2)
df.v <- r * df.v/sqrt(max(v.scale)) 
df.v <- df.v %>% mutate(PC1, PC2)
df.v$Variables <- rownames(v) 
PCAloadings = df.v
#df.v = dataset with loadings 

  #dataset with scores and categorical variables for plotting points: 
PCAvalues <- cbind(df.u, figure_epi_pca_cat)

#dataset with information for plotting circle: 
theta <- c(seq(-pi, pi, length = 50), seq(pi, -pi, length = 50))

circle <- data.frame(PC1 = r * cos(theta), PC2 = r * sin(theta)) 

#Group by flam. metric: 
PCA_combustability <- df.v %>% 
  filter(Variables %in% c("                 Flame Height", "Max. Temp","               Flame Duration"))
         
PCA_ignitability <- df.v %>% 
  filter(Variables %in% c("Time to Ignition", "\nGlow to Ignition"))

PCA_consumability <- df.v %>% 
  filter(Variables %in% c("Glow Duration", "Post-flame Glow", "\nTime to First Glow"))

# Calculate the angles and the label offset
df.v$Angle = ((180/pi) * atan(df.v$PC2/df.v$PC1))

df.v$Offset <- ((-2 * sign(df.v$PC1))/2)

#labels: 
u.axis.labs <- paste("PC", choices, sep = "")
u.axis.labs <- paste(u.axis.labs, sprintf("(%0.1f%% explained var.)", 
                                            100 * pcobj$sdev[choices]^2/sum(pcobj$sdev^2)))
```

Plot:
```{r}
ggplot(PCAvalues, aes(x = PC1, y = PC2)) +
  ylim(-4, 3) +
  geom_point(size = 1.75, alpha = .45, aes(shape = Species
                                       , color = Species)
                                       , data = PCAvalues) +
  guides(color = guide_legend(override.aes = list(size = 3))) +
  geom_segment(data = PCA_combustability, aes(x = 0, y = 0, 
                                              xend = PC1, yend = PC2),
     arrow = arrow(length = unit(1/2, "picas")), 
      size = .9, alpha = 0.9,
     color = "#64A400") +
  geom_segment(data = PCA_consumability, aes(x = 0, y = 0,
                                             xend = PC1, yend = PC2),
     arrow = arrow(length = unit(1/2, "picas")), 
     size = .9, alpha = 0.9,
     color = "#b08ba5") +
  geom_segment(data = PCA_ignitability, aes(x = 0, y = 0, 
                                             xend = PC1, yend = PC2),
     arrow = arrow(length = unit(1/2, "picas")), 
     size = .9, alpha = 0.9,
     color = '#669BBC') +
  annotate("text", 
           x = (PCA_combustability$PC1 * 0.7), 
          y = (PCA_combustability$PC2 * 1.15),
          label = PCA_combustability$Variables, size = 3.5,
          fontface = 'bold') +
  annotate("text",
           x = (PCA_ignitability$PC1 * 1.3), 
           y = (PCA_ignitability$PC2 * 1.45), 
           label = PCA_ignitability$Variables, size = 3.5,
           fontface = 'bold') +
  annotate("text",
           x = (PCA_consumability$PC1 * 1.3),
           y = (PCA_consumability$PC2 * 1.1), 
           label = PCA_consumability$Variables, size = 3.5,
           fontface = 'bold') + 
  geom_segment(aes(x = -3.1, xend = -3.1, y = -3.45, yend = -1.85), arrow = arrow(length = unit(1/2, 'picas')), size = 0.35, alpha = 0.85) +
  annotate('text', x = -3.25, y = -2.6, label = "Less Consumable", angle = 90, size = 3.5) +
  geom_segment(aes(x = -3.1, xend = -1.35, y = -3.45, yend = -3.45), arrow = arrow(length = unit(1/2, 'picas')), size = 0.35, alpha = 0.85) +
  annotate('text', x = -2.25, y = -3.68, label = "Larger & Hotter Flames, \n Faster Ignition", size = 3.5) +
  theme(panel.background = element_rect(fill='white', colour='black'), # Make background white and border black
          panel.grid.major = element_blank(),  # Hide major gridlines
          panel.grid.minor = element_blank(),
          legend.key = element_rect(fill = "white"),
          legend.position = c(0.2,0.9),
          legend.title = element_text(face = 'bold', size = 15),
          legend.text = element_text(face = 'italic', size = 13),
          axis.title = element_text(face = 'bold', size = 15)) +
  scale_color_manual(values = c("#9D0208","#F48C06")) + 
  geom_path(data = circle, color = "black",  size = 1/2, alpha = 1/3) + 
  xlab(u.axis.labs[1]) + 
   ylab(u.axis.labs[2]) +
  coord_equal()

ggsave(here("figures",'main-figures', "Fig1_pca.jpg"),
       height = 7, width = 9)
```

## Adding principal components into dataset
```{r warning=FALSE}
x.flam <-predict(pcobj)  ### generate scores for each row in epi.flamdata

pca.flam.data <- as.data.frame(x.flam) 

all.data.epi <- cbind(epi.flamdata, pca.flam.data)

# #because we wants instances where they didn't ignite for the logistic regression: 
epi.flamdata_noignitions <- read_csv(here("processed-data",  "local_flam_data_all.csv")) %>%
  mutate(LFM = lfm.NAs.imputed) %>%
  #filter(year == 2020) %>%
  filter(model == "EPI") %>%
  filter(ignition == 0)

all.data.withnoignitions <- bind_rows(all.data.epi, epi.flamdata_noignitions)

#dataset with the above columns, PCA using imputed NAs (medians), and flam metrics with NAs still in there:
mem.data <- all.data.withnoignitions

write_csv(mem.data, here("processed-data", "mem.data.local.epi.alldates.csv"))
```

# 1.2. PCA Table
```{r warning=FALSE}
figure_epi_pca_cat <- (epi.flamdata[c("hydration", "spp", "year.month", "sample")]) %>% 
  transmute("Hydration" = hydration, 
            spp = spp,
            "Timing" = year.month, 
            "Sample" = sample) %>% 
  mutate(Species = case_when(
    spp == "ADFA" ~ "A. fasciculatum", 
    spp == "CEME" ~ "C. megacarpus"
  ))

figure_varimax <- varimax(pcobj$rotation[,1:3])
figure_varimax

tab_pca(pcobj, 
        rotation = c("varimax"), 
        digits = 2,
        show.var = TRUE,
  string.pov = "Proportion of Variance",
  string.cpov = "Cumulative Proportion") 
```

## Nicer Table (kableExtra)
First, making dataframe with eigenvalues for principal components, proportion of variance explained and cumulative proportion:
```{r}
eig1 <- c(0.76, 0.62, 0.51, -0.72, -0.82, -0.11, 0.11, 0.01, 37.17, 37.17)
eig2 <- c(-0.17, 0.14, -0.16, -0.41, -0.39, -0.99, -0.93, -0.09, 23.23, 60.39)
eig3 <- c(-0.05, -0.31, 0.11, 0.45, -0.23, 0.04, 0.15, 0.98, 14.72, 75.11)
flam.metrics <- c("Flame Duration", "Flame Height", "Max. Temp.", "Time to Ignition", "Glow to Ignition", "Glow Duration", "Post-flame Glow", "Time to First Glow", "Proportion of Variance (%)", "Cumulative Proportion (%)")
pca.table.df <- data.frame(flam.metrics, eig1, eig2, eig3)
```
there was probably a more elegant way to do that, but since there aren't too many numbers to enter in by hand, I thought it would be quicker to just go ahead and do so.

Now, using kableExtra package to visualize how each component is weighted by each flam. metric
```{r}
pca.table.df %>%
  mutate(flam.metrics = cell_spec(flam.metrics, 'html', bold = ifelse(flam.metrics == c("Proportion of Variance (%)", "Cumulative Proportion (%)"), T, F), color = 'black')) %>% 
  mutate(eig1 = cell_spec(eig1, 'html', bold = ifelse(abs(eig1) > 0.5, T, F), color = 'black')) %>% 
  mutate(eig2 = cell_spec(eig2, 'html', bold = ifelse(abs(eig2) > 0.5, T, F), color = 'black')) %>% 
  mutate(eig3 = cell_spec(eig3, 'html', bold = ifelse(abs(eig3) > 0.5, T, F), color = 'black')) %>% 
  kable(format = 'html', escape = F, col.names = c(' ', 'Component 1', 'Component 2', 'Component 3')) %>% 
  kable_styling(bootstrap_options = c('hover', 'bordered', 'condensed'), fixed_thead = T, font_size = 16) %>% 
   row_spec(c(0), background = "#e5e5e5", bold = T, font_size = 16) %>% 
  #row_spec(c(0), background = "lightgrey", bold = T, font_size = 16) %>% 
  row_spec(c(9:10), background = '#D3D3D3') %>% 
  save_kable(here('figures', 'main-figures', 'Tab1_pca_table.html'))
```


# 2. Supplemental Figure
This figure shows two PCA's side-by-side, one for each species and will be included in the supplemental figures

## ADFA
Plot setup:
```{r warning=FALSE}
### Note: This dataset includes only Epiradiator and Ignited samples
epi.flamdata.adfa <- epi.flamdata %>% 
  filter(spp == "ADFA")

 # filter(year == 2020)

figure_epi_pca_quant.adfa <- (epi.flamdata.adfa[, c("fh", "ttfg", "tti", "fd", "gd", "gti", "pfg",
         "temp.max")]) %>%  
 transmute("\n\nFlame Duration" = fd,
           " Flame Height" = fh,
           "Post-flame Glow" = pfg,
           "\nTime to Ignition" = tti,
           "\nMax. Temp" = temp.max,
           "Glow Duration" = gd,
           "\nGlow to Ignition" = gti,
           "Time to First Glow\n" = ttfg
          )

figure_epi_pca_cat.adfa <- (epi.flamdata.adfa[c("hydration", "spp", "year.month", "sample")]) %>% 
  transmute("Hydration" = hydration, 
            spp = spp,
            "Timing" = year.month, 
            "Sample" = sample) %>% 
  mutate(Species = case_when(
    spp == "ADFA" ~ "Adenostoma fasciculatum"
  ))

figure_prcomp.adfa <- prcomp(na.omit(figure_epi_pca_quant.adfa), center = TRUE, scale = TRUE, retx=TRUE)

#Setting up plotting data: 

#decisions for plotting:
choices = 1:2 
scale = 1
obs.scale = 1 - scale
var.scale = scale
ellipse.prob = 0.68
labels.size = 3
circle.prob = 0.69
choices = 1:2

#Run PCA
pcobj <- prcomp(na.omit(figure_epi_pca_quant.adfa), center = TRUE, scale = TRUE)

# extract PCA components: 
nobs.factor <- sqrt(nrow(pcobj$x) - 1)
    d <- pcobj$sdev
    u <- sweep(pcobj$x, 2, 1/(d * nobs.factor), FUN = "*")
    v <- pcobj$rotation

  choices <- pmin(choices, ncol(u))
  df.u <- as.data.frame(sweep(u[, choices], 2, d[choices]^obs.scale, 
                              FUN = "*"))
  v <- sweep(v, 2, d^var.scale, FUN = "*")
  df.v <- as.data.frame(v[, choices])
  names(df.u) <- c("PC1", "PC2")
  names(df.v) <- names(df.u)
  df.u <- df.u * nobs.factor

  r <- sqrt(qchisq(circle.prob, df = 2)) * prod(colMeans(df.u^2))^(1/4)
  
v.scale <- rowSums(v^2)
df.v <- r * df.v/sqrt(max(v.scale)) 
df.v <- df.v %>% mutate(PC1, PC2)
df.v$Variables <- rownames(v) 
PCAloadings = df.v
#df.v = dataset with loadings 

  #dataset with scores and categorical variables for plotting points: 
PCAvalues.adfa <- cbind(df.u, figure_epi_pca_cat.adfa)

#dataset with information for plotting circle: 
theta <- c(seq(-pi, pi, length = 50), seq(pi, -pi, length = 50))

circle <- data.frame(PC1 = r * cos(theta), PC2 = r * sin(theta)) 

#Group by flam. metric: 
PCA_combustability <- df.v %>% 
  filter(Variables %in% c(" Flame Height", "\nMax. Temp",
                          "\n\nFlame Duration"))
         
PCA_ignitability <- df.v %>% 
  filter(Variables %in% c("\nTime to Ignition", "\nGlow to Ignition"))

PCA_consumability <- df.v %>% 
  filter(Variables %in% c("Glow Duration", "Post-flame Glow", "Time to First Glow\n"))

# Calculate the angles and the label offset
df.v$Angle = ((180/pi) * atan(df.v$PC2/df.v$PC1))

df.v$Offset <- ((-2 * sign(df.v$PC1))/2)

#labels: 
u.axis.labs <- paste("PC", choices, sep = "")
u.axis.labs <- paste(u.axis.labs, sprintf("(%0.1f%% explained var.)", 
                                            100 * pcobj$sdev[choices]^2/sum(pcobj$sdev^2)))
```

Plot:
```{r}
adfa.pca.plot <- ggplot(PCAvalues.adfa, aes(x = PC1, y = PC2)) +
  geom_point(size = 1.5, alpha = .4, color = "#9D0208", data = PCAvalues.adfa) +
  geom_segment(data = PCA_combustability, aes(x = 0, y = 0, 
                                              xend = PC1, yend = PC2),
     arrow = arrow(length = unit(1/2, "picas")), 
      size = .8, alpha = 0.85,
     color = "#64A400") +
  geom_segment(data = PCA_consumability, aes(x = 0, y = 0,
                                             xend = PC1, yend = PC2),
     arrow = arrow(length = unit(1/2, "picas")), 
     size = .8, alpha = 0.85,
     color = "#b08ba5") +
  geom_segment(data = PCA_ignitability, aes(x = 0, y = 0, 
                                             xend = PC1, yend = PC2),
     arrow = arrow(length = unit(1/2, "picas")), 
     size = .8, alpha = 0.85,
     color = '#669BBC') +
  annotate("text", 
           x = (PCA_combustability$PC1 * 0.95), 
          y = (PCA_combustability$PC2 * 1.325),
          label = PCA_combustability$Variables, size = 3.2,
          fontface = 'bold') +
  annotate("text",
           x = (PCA_ignitability$PC1 * 1.3), 
           y = (PCA_ignitability$PC2 * 0.95), 
           label = PCA_ignitability$Variables, size = 3.2,
           fontface = 'bold') +
  annotate("text",
           x = (PCA_consumability$PC1 * 1.3),
           y = (PCA_consumability$PC2 * 1.1), 
           label = PCA_consumability$Variables, size = 3.2,
           fontface = 'bold') + 
   geom_segment(aes(x = -3.1, xend = -3.1, y = -3.3, yend = -1.275), arrow = arrow(length = unit(1/2, 'picas')), size = 0.35, alpha = 0.85) +
  annotate('text', x = -3.25, y = -2.35, label = "More Consumable", angle = 90, size = 3.5) +
  geom_segment(aes(x = -3.1, xend = -1.4, y = -3.3, yend = -3.3), arrow = arrow(length = unit(1/2, 'picas')), size = 0.35, alpha = 0.85) +
  annotate('text', x = -2.25, y = -3.68, label = "Larger & Hotter Flames, \n Faster Ignition", size = 3.5) +
  annotate('text', x = -2.25, y = 4, label = "A. fasciculatum", size = 5, fontface = 4) +
  theme(panel.background = element_rect(fill='white', colour='black'), # Make background white and border black
          panel.grid.major = element_blank(),  # Hide major gridlines
          panel.grid.minor = element_blank(),
          legend.key = element_rect(fill = "white"),
          legend.title = element_text(face = 'bold', size = 12),
          legend.text = element_text(face = 'italic', size = 10),
          axis.title = element_text(face = 'bold', size = 12), 
          plot.title = element_text(face = 'italic'),
          aspect.ratio = 1/1) +
  geom_path(data = circle, color = "black",  size = 1/2, alpha = 1/3) +
  xlab(u.axis.labs[1]) + 
   ylab(u.axis.labs[2]) +
  coord_equal()
  
adfa.pca.plot
```

## CEME
Plot setup:
```{r warning=FALSE}
### Note: This dataset includes only Epiradiator and Ignited samples
epi.flamdata.ceme <- epi.flamdata %>% 
  filter(spp == "CEME")

 # filter(year == 2020)

figure_epi_pca_quant.ceme <- (epi.flamdata.ceme[, c("fh", "ttfg", "tti", "fd", "gd", "gti", "pfg",
         "temp.max")]) %>%  
 transmute("  Flame Duration" = fd,
           "     Flame Height" = fh,
           "Post-flame Glow" = pfg,
           "Time to Ignition" = tti,
           "Max. Temp" = temp.max,
           "Glow Duration" = gd,
           "Glow to Ignition" = gti,
           "\nTime to First Glow" = ttfg
          )

figure_epi_pca_cat.ceme <- (epi.flamdata.ceme[c("hydration", "spp", "year.month", "sample")]) %>% 
  transmute("Hydration" = hydration, 
            spp = spp,
            "Timing" = year.month, 
            "Sample" = sample) %>% 
  mutate(Species = case_when(
    spp == "CEME" ~ "Ceanothus megacarpus"
  ))

figure_prcomp.ceme <- prcomp(na.omit(figure_epi_pca_quant.ceme), center = TRUE, scale = TRUE, retx=TRUE)

#Setting up plotting data: 

#decisions for plotting:
choices = 1:2 
scale = 1
obs.scale = 1 - scale
var.scale = scale
ellipse.prob = 0.68
labels.size = 3
circle.prob = 0.69
choices = 1:2

#Run PCA
pcobj <- prcomp(na.omit(figure_epi_pca_quant.ceme), center = TRUE, scale = TRUE)

# extract PCA components: 
nobs.factor <- sqrt(nrow(pcobj$x) - 1)
    d <- pcobj$sdev
    u <- sweep(pcobj$x, 2, 1/(d * nobs.factor), FUN = "*")
    v <- pcobj$rotation

  choices <- pmin(choices, ncol(u))
  df.u <- as.data.frame(sweep(u[, choices], 2, d[choices]^obs.scale, 
                              FUN = "*"))
  v <- sweep(v, 2, d^var.scale, FUN = "*")
  df.v <- as.data.frame(v[, choices])
  names(df.u) <- c("PC1", "PC2")
  names(df.v) <- names(df.u)
  df.u <- df.u * nobs.factor

  r <- sqrt(qchisq(circle.prob, df = 2)) * prod(colMeans(df.u^2))^(1/4)
  
v.scale <- rowSums(v^2)
df.v <- r * df.v/sqrt(max(v.scale)) 
df.v <- df.v %>% mutate(PC1, PC2)
df.v$Variables <- rownames(v) 
PCAloadings = df.v
#df.v = dataset with loadings 

  #dataset with scores and categorical variables for plotting points: 
PCAvalues.ceme <- cbind(df.u, figure_epi_pca_cat.ceme)

#dataset with information for plotting circle: 
theta <- c(seq(-pi, pi, length = 50), seq(pi, -pi, length = 50))

circle <- data.frame(PC1 = r * cos(theta), PC2 = r * sin(theta)) 

#Group by flam. metric: 
PCA_combustability <- df.v %>% 
  filter(Variables %in% c("     Flame Height", "Max. Temp",
                          "  Flame Duration"))
         
PCA_ignitability <- df.v %>% 
  filter(Variables %in% c("Time to Ignition", "Glow to Ignition"))

PCA_consumability <- df.v %>% 
  filter(Variables %in% c("Glow Duration", "Post-flame Glow", "\nTime to First Glow"))

# Calculate the angles and the label offset
df.v$Angle = ((180/pi) * atan(df.v$PC2/df.v$PC1))

df.v$Offset <- ((-2 * sign(df.v$PC1))/2)

#labels: 
u.axis.labs <- paste("PC", choices, sep = "")
u.axis.labs <- paste(u.axis.labs, sprintf("(%0.1f%% explained var.)", 
                                            100 * pcobj$sdev[choices]^2/sum(pcobj$sdev^2)))
```

Plot:
```{r}
ceme.pca.plot <- ggplot(PCAvalues.ceme, aes(x = PC1, y = PC2)) +
  geom_point(size = 1.5, alpha = .5, shape = 17, color = "#F48C06", data = PCAvalues.ceme) +
  geom_segment(data = PCA_combustability, aes(x = 0, y = 0, 
                                              xend = PC1, yend = PC2),
     arrow = arrow(length = unit(1/2, "picas")), 
      size = .8, alpha = 0.85,
     color = "#64A400") +
  geom_segment(data = PCA_consumability, aes(x = 0, y = 0,
                                             xend = PC1, yend = PC2),
     arrow = arrow(length = unit(1/2, "picas")), 
     size = .8, alpha = 0.85,
     color = "#b08ba5") +
  geom_segment(data = PCA_ignitability, aes(x = 0, y = 0, 
                                             xend = PC1, yend = PC2),
     arrow = arrow(length = unit(1/2, "picas")), 
     size = .8, alpha = 0.85,
     color = '#669BBC') +
  annotate("text", 
           x = (PCA_combustability$PC1 * 1.5), 
          y = (PCA_combustability$PC2 * 1.1),
          label = PCA_combustability$Variables, size = 3.2,
          fontface = 'bold') +
  annotate("text",
           x = (PCA_ignitability$PC1 * 1.4), 
           y = (PCA_ignitability$PC2 * 1.05), 
           label = PCA_ignitability$Variables, size = 3.2,
           fontface = 'bold') +
  annotate("text",
           x = (PCA_consumability$PC1 * 1.3),
           y = (PCA_consumability$PC2 * 1.05), 
           label = PCA_consumability$Variables, size = 3.2,
           fontface = 'bold') + 
  geom_segment(aes(x = -3.1, xend = -3.1, y = -3.4, yend = -1.9), arrow = arrow(length = unit(1/2, 'picas')), size = 0.35, alpha = 0.85) +
  annotate('text', x = -3.25, y = -2.65, label = "Less Consumable", angle = 90, size = 3.5) +
  geom_segment(aes(x = -3.1, xend = -1.15, y = -3.4, yend = -3.4), arrow = arrow(length = unit(1/2, 'picas')), size = 0.35, alpha = 0.85) +
  annotate('text', x = -2.2, y = -3.68, label = "Larger & Hotter Flames, \n Faster Ignition", size = 3.5) +
  annotate('text', x = -2.2, y = 2, label = "C. megacarpus", size = 5, fontface = 4) +
  theme(panel.background = element_rect(fill='white', colour='black'), # Make background white and border black
          panel.grid.major = element_blank(),  # Hide major gridlines
          panel.grid.minor = element_blank(),
          legend.key = element_rect(fill = "white"),
          legend.title = element_text(face = 'bold', size = 12),
          legend.text = element_text(face = 'italic', size = 10),
          axis.title = element_text(face = 'bold', size = 12), 
          plot.title = element_text(face = 'italic'),
          aspect.ratio = 1/1) +
  geom_path(data = circle, color = "black",  size = 1/2, alpha = 1/3) +
  xlab(u.axis.labs[1]) + 
   ylab(u.axis.labs[2]) +
  coord_equal()
  
ceme.pca.plot
```

## Arranging Plots
```{r}
ggarrange(adfa.pca.plot,ceme.pca.plot)

ggsave(here('figures','supp-figures','FigS1_spp.pca.jpg'), height = 6, width = 12)
```
