==> scripts/3.0_mixed_effects_models.Rmd <==
---
title: "Mixed Effects Models"
author: "Indra Boving & Joe Celebrezze"
date: "6/21/2022"
output: html_document
csl: functional-ecology.csl
---

# Setup
```{r setup, include=FALSE}
knitr::opts_chunk$set(echo = TRUE)

#lots of extra here, but oh well... 
library(ggplot2)
library(gapminder)
library(data.table)
library(purrr) #for visualizing all variables in one big plot
library(naniar) #for dealing with NAs nicely 
library(tidyverse)
library(devtools)
library(ggfortify)
library(ggpubr)
library(jtools)
library(cowplot)
library(lmerTest)
library(ggeffects)  # install the package first if you haven't already, then load it
library(GGally) #for corr plots
require(kimisc) # has the nlist function to create a named list
require(AICcmodavg) # has the aictab function
library(psych)
devtools::install_github("strengejacke/strengejacke")
library(strengejacke)
library(sjPlot) # table functions
library(sjmisc) # sample data
library(lme4) # fitting models
library(here)
library(effects) #for plotting model effects
library(sjstats) #use for r2 functions
library(TMB)
library(glmmTMB)
library(lattice)
library(equatiomatic)
library("broom.mixed")
library(ggbiplot)
select = dplyr::select
here = here::here
library(MuMIn)
library(modelsummary)
#install_github("BlakeRMills/MetBrewer") 
#library("BlakeRMills/MetBrewer")
#library(memmented)
filter = dplyr::filter
mutate = dplyr::mutate
library(nlme)
library(MetBrewer)
#install.packages("memmented")
#devtools::install_github("hohenstein/remef")
library(remef)
library(kableExtra)

install.packages("gtsummary")
library(gtsummary)
library(grateful)
library(stats)
```

```{r}
citation("stats")
```


Set global options for color palettes and themes: 

```{r}
th <- theme(panel.background = element_rect(fill='white', colour='black'), # Make background white and border black
          panel.grid.major = element_blank(),  # Hide major gridlines
          panel.grid.minor = element_blank())

th.spp.wrap <- theme(legend.position = "bottom", 
        strip.text = element_text(face = "italic", size = 12) +
  scale_color_manual(values=met.brewer("Morgenstern", 7)))

Morgenstern = list(c("#7c668c", "#b08ba5", "#dfbbc8", "#ffc680", "#ffb178", "#db8872", "#a56457"), c(7, 5, 4, 6, 3, 2, 1), colorblind=TRUE)
```

#------------------------------------
# 1. Data wrangling
Reading in dataframe
```{r}
scaled.mem.data.alldates.noNAs <- read.csv(here('processed-data', 'mem.data.subset.alldates.csv'))
```

Seasonal mins and maxs from 2016 - 2020:

https://www.sbbg.org/about/onsite-weather-station-live-fuel-moisture 

```{r}
adfa_min <- 55
adfa_max <- 110

ceme_min <- 55
ceme_max <- 140
```

### All dates datasets
```{r, message=FALSE, warning=FALSE, include=FALSE}
mem.adfa.subset <- scaled.mem.data.alldates.noNAs  %>% 
  filter(spp == "ADFA") %>% 
  filter(ignition == "1")

mem.CEME.subset <- scaled.mem.data.alldates.noNAs  %>% 
  filter(spp == "CEME")%>% 
  filter(ignition == "1")

mem.adfa.subset.allignitions <- scaled.mem.data.alldates.noNAs  %>% 
  filter(spp == "ADFA")

mem.CEME.subset.allignitions <- scaled.mem.data.alldates.noNAs %>% 
  filter(spp == "CEME")

mem.adfa.subset.limited <- mem.adfa.subset %>% 
  filter(lfm.unscaled < adfa_max, lfm.unscaled > adfa_min)

mem.CEME.subset <- scaled.mem.data.alldates.noNAs %>% 
  filter(spp == "CEME")

mem.CEME.subset.limited <- mem.CEME.subset %>% 
  filter(lfm.unscaled < ceme_max, lfm.unscaled > ceme_min)
```

#------------------------------------
# 2.0 Predicting PC1 (part 1)

## 2.0.1 Figures of models

```{r}
mpa_mod_fig.mpa <- lmer(PC1 ~ mpa.unscaled + Species + year.month + (1 | individual), data = scaled.mem.data.alldates.noNAs)
summary(mpa_mod_fig.mpa)

#remove fixed effect of Species:
#     see what names are associated: 
term2coef(mpa_mod_fig.mpa, "Species")

r1_1 <- remef(mpa_mod_fig.mpa , fix = c("SpeciesC. megacarpus"))
summary(r1_1)
head(r1_1)

# remove fixed effect of 'Days' and the intercept
#r1_2 <- remef(fm1, fix = "Days", keep.intercept = FALSE)
```

```{r}

(pp <- plot_model(mpa_mod_fig.mpa,type="pred",
       terms=c("mpa.unscaled","year.month"),pred.type="re"))

(pp <- plot_model(mpa_mod_fig.mpa,type="pred",
       terms=c("mpa.unscaled","Species"),pred.type="re"))
```

#### Check assumptions

```{r}
plot(mpa_mod_fig.mpa)
#qqnorm(mpa_mod_fig.mpa, ~ranef(., level=2))
#https://stats.stackexchange.com/questions/77891/checking-assumptions-lmer-lme-mixed-models-in-r for above code - getting an error right now though
```

# ---------------------------------
# MODEL SELECTION 1
What covariates to include?

## Site, species, year.month, sample.wt
```{r}
mpa_mod_fig.mpa <- lmer(PC1 ~ mpa.unscaled + spp + site + year.month + sample.wt + (1 | individual), REML = F,  data = scaled.mem.data.alldates.noNAs)

performance::multicollinearity(mpa_mod_fig.mpa)
```

```{r}
max_model <- lmer(PC1 ~ spp + dw.flam.sample + ww.flam.sample + mpa + lfm.NAs.imputed + excess_water.scaled + site + year.month + sample.wt + (1 | individual), REML = F, data = scaled.mem.data.alldates.noNAs)

performance::multicollinearity(max_model)
```

```{r}

max_model_nolfm <- lmer(PC1 ~ spp + dw.flam.sample + ww.flam.sample + mpa + excess_water.scaled + site +  year.month + sample.wt + (1 | individual), REML = F,  data = scaled.mem.data.alldates.noNAs)

performance::multicollinearity(max_model_nolfm)
```

```{r}

max_model_nompa <- lmer(PC1 ~ spp + dw.flam.sample + ww.flam.sample + lfm.NAs.imputed + excess_water.scaled + site +  year.month + sample.wt + (1 | individual), REML = F,  data = scaled.mem.data.alldates.noNAs)

performance::multicollinearity(max_model_nompa)
```

```{r}

m3 <- lmer(PC1 ~ spp + ww.flam.sample + site + year.month + sample.wt + (1 | individual), REML = F,  data = scaled.mem.data.alldates.noNAs)

performance::multicollinearity(m3)
```

```{r}

m6 <- lmer(PC1 ~ dw.flam.sample + spp + site + year.month + sample.wt + (1 | individual), REML = F,  data = scaled.mem.data.alldates.noNAs)

performance::multicollinearity(m6)
```

```{r}

m6.5 <- lmer(PC1 ~ dw.flam.sample + ww.flam.sample + spp + site + year.month + sample.wt + (1 | individual), REML = F,  data = scaled.mem.data.alldates.noNAs)

performance::multicollinearity(m6.5)
```

```{r}
m7 <- lmer(PC1 ~ mpa + lfm.NAs.imputed + spp + site + year.month + sample.wt + (1 | individual), REML = F,  data = scaled.mem.data.alldates.noNAs)

performance::multicollinearity(m7)
```

```{r}
m10 <- lmer(PC1 ~ lfm.NAs.imputed + spp + site + year.month + sample.wt + (1 | individual), REML = F,  data = scaled.mem.data.alldates.noNAs)

performance::multicollinearity(m10)
```

```{r}
m11 <- lmer(PC1 ~ mpa*spp + site + year.month + sample.wt + (1 | individual), REML = F,  data = scaled.mem.data.alldates.noNAs)

performance::multicollinearity(m11)
```

```{r}
m12 <- lmer(PC1 ~ excess_water.scaled + spp + site + year.month + sample.wt + (1 | individual), REML = F,  data = scaled.mem.data.alldates.noNAs)

performance::multicollinearity(m12)
```

```{r}
m12.5 <- lmer(PC1 ~ excess_water.scaled + mpa + lfm.NAs.imputed + spp + site + year.month + sample.wt + (1 | individual), REML = F,  data = scaled.mem.data.alldates.noNAs)

performance::multicollinearity(m12.5) ##High correlation! Don't use
```

```{r}
m13 <- lmer(PC1 ~ excess_water.scaled + lfm.NAs.imputed + spp + site + year.month + sample.wt + (1 | individual), REML = F,  data = scaled.mem.data.alldates.noNAs)

performance::multicollinearity(m13) ##High correlation! Don't use
```

```{r}
m14 <- lmer(PC1 ~ excess_water.scaled + spp + site + year.month + sample.wt + (1 | individual), REML = F,  data = scaled.mem.data.alldates.noNAs)

performance::multicollinearity(m14)
```

```{r}
m15 <- lmer(PC1 ~ excess_water.scaled * spp + site + year.month + sample.wt + (1 | individual), REML = F,  data = scaled.mem.data.alldates.noNAs)

performance::multicollinearity(m15)
```

```{r}

tab_model(mpa_mod_fig.mpa, m11, m7, m10, m3, m6, m6.5, m12, m15,
          show.reflvl = TRUE, 
          digits = 3, 
          show.aic = TRUE, 
          show.ci = FALSE,
          show.icc = FALSE, 
          pred.labels = c('Intercept', 
                          'Water Potential (scaled)', 
                          'Species Effect (C. megacarpus)',
                          'Water Potential x Species Effect
                          (interaction)',
                          'LFM (%)',
                          'Water Weight (g)', 
                          'Dry Weight (g)',
                          'Site Effect (Site 1)',
                          'Excess Water (scaled)',
                          'Excess Water x Species Effect
                          (interaction)'),
         dv.labels = c("mpa_spp", 
                       "m11", 
                       "m7", 
                       "m7",
                       "m10",
                       "m3", 
                       "m6", 
                       "m6.5", 
                       "m12",
                       "m15"),
         string.pred = "Coeffcient", 
         title = "models with no collinearity",
  string.p = "P-Value", 
  p.style = "stars",
  file = here('figures', 'MEM.figures','site.yearmonth.samplewt.html')) # saving .html into figures folder
```

## Site, species, year.month
```{r}
mpa_mod_fig.mpa <- lmer(PC1 ~ mpa.unscaled + spp + site + year.month + (1 | individual), REML = F,  data = scaled.mem.data.alldates.noNAs)

qqnorm(resid(mpa_mod_fig.mpa))
r <- resid(mpa_mod_fig.mpa)
plot(fitted(mpa_mod_fig.mpa), r)

performance::multicollinearity(mpa_mod_fig.mpa)
```

```{r}
max_model <- lmer(PC1 ~ spp + dw.flam.sample + ww.flam.sample + mpa + lfm.NAs.imputed + excess_water.scaled + site + year.month + (1 | individual), REML = F,  data = scaled.mem.data.alldates.noNAs)

performance::multicollinearity(max_model)
```

```{r}
max_model_nolfm <- lmer(PC1 ~ spp + dw.flam.sample + ww.flam.sample + mpa + excess_water.scaled + site +  year.month + (1 | individual), REML = F,  data = scaled.mem.data.alldates.noNAs)

performance::multicollinearity(max_model_nolfm)
```

```{r}
max_model_nompa <- lmer(PC1 ~ spp + dw.flam.sample + ww.flam.sample + lfm.NAs.imputed + excess_water.scaled + site +  year.month + (1 | individual), REML = F,  data = scaled.mem.data.alldates.noNAs)

performance::multicollinearity(max_model_nompa)
```

```{r}
m3 <- lmer(PC1 ~ spp + ww.flam.sample + site + year.month + (1 | individual), REML = F,  data = scaled.mem.data.alldates.noNAs)

performance::multicollinearity(m3)
```

```{r}
m6 <- lmer(PC1 ~ dw.flam.sample + spp + site + year.month + (1 | individual), REML = F,  data = scaled.mem.data.alldates.noNAs)

qqnorm(resid(m6))
r <- resid(m6)
plot(fitted(m6), r)

performance::multicollinearity(m6)
```

```{r}

m6.5 <- lmer(PC1 ~ dw.flam.sample + ww.flam.sample + spp + site + year.month + (1 | individual), REML = F,  data = scaled.mem.data.alldates.noNAs)

performance::multicollinearity(m6.5)
```

```{r}

m7 <- lmer(PC1 ~ mpa + lfm.NAs.imputed + spp + site + year.month + (1 | individual), REML = F,  data = scaled.mem.data.alldates.noNAs)

performance::multicollinearity(m7)
```

```{r}

m10 <- lmer(PC1 ~ lfm.NAs.imputed + spp + site + year.month + (1 | individual), REML = F,  data = scaled.mem.data.alldates.noNAs)

qqnorm(resid(m10))
r <- resid(m10)
plot(fitted(m10), r)

performance::multicollinearity(m10)
```

```{r}
m11 <- lmer(PC1 ~ mpa*spp + site + year.month + (1 | individual), REML = F,  data = scaled.mem.data.alldates.noNAs)

performance::multicollinearity(m11)
```

```{r}
m12 <- lmer(PC1 ~ excess_water.scaled + spp + site + year.month + (1 | individual), REML = F,  data = scaled.mem.data.alldates.noNAs)

performance::multicollinearity(m12)
```

```{r}
m12.5 <- lmer(PC1 ~ dw.flam.sample + mpa + lfm.NAs.imputed + spp + site + year.month + (1 | individual), REML = F,  data = scaled.mem.data.alldates.noNAs)

qqnorm(resid(m12.5))
r <- resid(m12.5)
plot(fitted(m12.5), r)

performance::multicollinearity(m12.5) ##High correlation! Don't use
```

```{r}
m13 <- lmer(PC1 ~ excess_water.scaled + lfm.NAs.imputed + spp + site + year.month + (1 | individual), REML = F,  data = scaled.mem.data.alldates.noNAs)

performance::multicollinearity(m13) ##High correlation! Don't use
```

```{r}
m14 <- lmer(PC1 ~ excess_water.scaled + spp + site + year.month + (1 | individual), REML = F,  data = scaled.mem.data.alldates.noNAs)

performance::multicollinearity(m14)
```

```{r}
m15 <- lmer(PC1 ~ excess_water.scaled * spp + site + year.month + (1 | individual), REML = F,  data = scaled.mem.data.alldates.noNAs)

performance::multicollinearity(m15)
```

```{r}
tab_model(mpa_mod_fig.mpa, m11, m7, m10, m3, m6, m6.5, m12, m15,
          show.reflvl = TRUE, 
          digits = 3, 
          show.aic = TRUE, 
          show.ci = FALSE,
          show.icc = FALSE, 
          pred.labels = c('Intercept', 
                          'Water Potential (scaled)', 
                          'Species Effect (C. megacarpus)',
                          'Water Potential x Species Effect
                          (interaction)',
                          'LFM (%)',
                          'Water Weight (g)', 
                          'Dry Weight (g)',
                          'Site Effect (Site 1)',
                          'Excess Water (scaled)',
                          'Excess Water x Species Effect
                          (interaction)'),
         dv.labels = c("mpa_spp", 
                       "m11", 
                       "m7", 
                       "m7",
                       "m10",
                       "m3", 
                       "m6", 
                       "m6.5", 
                       "m12",
                       "m15"),
         string.pred = "Coeffcient", 
         title = "models with no collinearity",
  string.p = "P-Value", 
  p.style = "stars"
  #file = here('figures', 'MEM.figures','site.yearmonth.html')
  ) # saving .html into figures folder
```

## Site, species, year & month, sample.wt
```{r}
mpa_mod_fig.mpa <- lmer(PC1 ~ mpa.unscaled + spp + site + year + month + sample.wt + (1 | individual), REML = F,  data = scaled.mem.data.alldates.noNAs)

performance::multicollinearity(mpa_mod_fig.mpa)
```

```{r}
max_model <- lmer(PC1 ~ spp + dw.flam.sample + ww.flam.sample + mpa + lfm.NAs.imputed + excess_water.scaled + site + year + month + sample.wt + (1 | individual), REML = F,  data = scaled.mem.data.alldates.noNAs)

performance::multicollinearity(max_model)
```

```{r}
max_model_nolfm <- lmer(PC1 ~ spp + dw.flam.sample + ww.flam.sample + mpa + excess_water.scaled + site +  year + month + sample.wt + (1 | individual), REML = F,  data = scaled.mem.data.alldates.noNAs)

performance::multicollinearity(max_model_nolfm)
```

```{r}
max_model_nompa <- lmer(PC1 ~ spp + dw.flam.sample + ww.flam.sample + lfm.NAs.imputed + excess_water.scaled + site +  year + month + sample.wt + (1 | individual), REML = F,  data = scaled.mem.data.alldates.noNAs)

performance::multicollinearity(max_model_nompa)
```

```{r}
m3 <- lmer(PC1 ~ spp + ww.flam.sample + site + year + month + sample.wt + (1 | individual), REML = F,  data = scaled.mem.data.alldates.noNAs)

performance::multicollinearity(m3)
```

```{r}
m6 <- lmer(PC1 ~ dw.flam.sample + spp + site + year + month + sample.wt + (1 | individual), REML = F,  data = scaled.mem.data.alldates.noNAs)

performance::multicollinearity(m6)
```

```{r}
m6.5 <- lmer(PC1 ~ dw.flam.sample + ww.flam.sample + spp + site + year + month + sample.wt + (1 | individual), REML = F,  data = scaled.mem.data.alldates.noNAs)

performance::multicollinearity(m6.5)
```

```{r}
m7 <- lmer(PC1 ~ mpa + lfm.NAs.imputed + spp + site + year + month + sample.wt + (1 | individual), REML = F,  data = scaled.mem.data.alldates.noNAs)

performance::multicollinearity(m7)
```

```{r}
m10 <- lmer(PC1 ~ lfm.NAs.imputed + spp + site + year + month + sample.wt + (1 | individual), REML = F,  data = scaled.mem.data.alldates.noNAs)

performance::multicollinearity(m10)
```

```{r}
m11 <- lmer(PC1 ~ mpa*spp + site + year + month + sample.wt + (1 | individual), REML = F,  data = scaled.mem.data.alldates.noNAs)

performance::multicollinearity(m11)
```

```{r}
m12 <- lmer(PC1 ~ excess_water.scaled + spp + site + year + month + sample.wt + (1 | individual), REML = F,  data = scaled.mem.data.alldates.noNAs)

performance::multicollinearity(m12)
```

```{r}
m12.5 <- lmer(PC1 ~ excess_water.scaled + mpa + lfm.NAs.imputed + spp + site + year + month + sample.wt + (1 | individual), REML = F,  data = scaled.mem.data.alldates.noNAs)

performance::multicollinearity(m12.5) ##High correlation! Don't use
```

```{r}
m13 <- lmer(PC1 ~ excess_water.scaled + lfm.NAs.imputed + spp + site + year + month + sample.wt + (1 | individual), REML = F,  data = scaled.mem.data.alldates.noNAs)

performance::multicollinearity(m13) ##High correlation! Don't use
```

```{r}
m14 <- lmer(PC1 ~ excess_water.scaled + spp + site + year + month + sample.wt + (1 | individual), REML = F,  data = scaled.mem.data.alldates.noNAs)

performance::multicollinearity(m14)
```

```{r}
m15 <- lmer(PC1 ~ excess_water.scaled * spp + site + year + month + sample.wt + (1 | individual), REML = F,  data = scaled.mem.data.alldates.noNAs)

performance::multicollinearity(m15)
```

```{r}
tab_model(mpa_mod_fig.mpa, m11, m7, m10, m3, m6, m6.5, m12, m15,
          show.reflvl = TRUE, 
          digits = 3, 
          show.aic = TRUE, 
          show.ci = FALSE,
          show.icc = FALSE, 
          pred.labels = c('Intercept', 
                          'Water Potential (scaled)', 
                          'Species Effect (C. megacarpus)',
                          'Water Potential x Species Effect
                          (interaction)',
                          'LFM (%)',
                          'Water Weight (g)', 
                          'Dry Weight (g)',
                          'Site Effect (Site 1)',
                          'Excess Water (scaled)',
                          'Excess Water x Species Effect
                          (interaction)'),
         dv.labels = c("mpa_spp", 
                       "m11", 
                       "m7", 
                       "m7",
                       "m10",
                       "m3", 
                       "m6", 
                       "m6.5", 
                       "m12",
                       "m15"),
         string.pred = "Coeffcient", 
         title = "models with no collinearity",
  string.p = "P-Value", 
  p.style = "stars",
  file = here('figures', 'MEM.figures','site.yearANDmonth.samplewt.html')) # saving .html into figures folder
```

## Site, species, year & month
```{r}
mpa_mod_fig.mpa <- lmer(PC1 ~ mpa.unscaled + spp + site + year + month + (1 | individual), REML = F,  data = scaled.mem.data.alldates.noNAs)

performance::multicollinearity(mpa_mod_fig.mpa)
```

```{r}
max_model <- lmer(PC1 ~ spp + dw.flam.sample + ww.flam.sample + mpa + lfm.NAs.imputed + excess_water.scaled + site + year + month + (1 | individual), REML = F,  data = scaled.mem.data.alldates.noNAs)

performance::multicollinearity(max_model)
```

```{r}
max_model_nolfm <- lmer(PC1 ~ spp + dw.flam.sample + ww.flam.sample + mpa + excess_water.scaled + site +  year + month + (1 | individual), REML = F,  data = scaled.mem.data.alldates.noNAs)

performance::multicollinearity(max_model_nolfm)
```

```{r}
max_model_nompa <- lmer(PC1 ~ spp + dw.flam.sample + ww.flam.sample + lfm.NAs.imputed + excess_water.scaled + site +  year + month + (1 | individual), REML = F,  data = scaled.mem.data.alldates.noNAs)

performance::multicollinearity(max_model_nompa)
```

```{r}
m3 <- lmer(PC1 ~ spp + ww.flam.sample + site + year + month + (1 | individual), REML = F,  data = scaled.mem.data.alldates.noNAs)

performance::multicollinearity(m3)
```

```{r}
m6 <- lmer(PC1 ~ dw.flam.sample + spp + site + year + month + (1 | individual), REML = F,  data = scaled.mem.data.alldates.noNAs)

performance::multicollinearity(m6)
```

```{r}
m6.5 <- lmer(PC1 ~ dw.flam.sample + ww.flam.sample + spp + site + year + month + (1 | individual), REML = F,  data = scaled.mem.data.alldates.noNAs)

performance::multicollinearity(m6.5)
```

```{r}
m7 <- lmer(PC1 ~ mpa + lfm.NAs.imputed + spp + site + year + month + (1 | individual), REML = F,  data = scaled.mem.data.alldates.noNAs)

performance::multicollinearity(m7)
```

```{r}
m10 <- lmer(PC1 ~ lfm.NAs.imputed + spp + site + year + month + (1 | individual), REML = F,  data = scaled.mem.data.alldates.noNAs)

performance::multicollinearity(m10)
```

```{r}
m11 <- lmer(PC1 ~ mpa*spp + site + year + month + (1 | individual), REML = F,  data = scaled.mem.data.alldates.noNAs)

performance::multicollinearity(m11)
```

```{r}
m12 <- lmer(PC1 ~ excess_water.scaled + spp + site + year + month + (1 | individual), REML = F,  data = scaled.mem.data.alldates.noNAs)

performance::multicollinearity(m12)
```

```{r}
m12.5 <- lmer(PC1 ~ excess_water.scaled + mpa + lfm.NAs.imputed + spp + site + year + month + (1 | individual), REML = F,  data = scaled.mem.data.alldates.noNAs)

performance::multicollinearity(m12.5) ##High correlation! Don't use
```

```{r}
m13 <- lmer(PC1 ~ excess_water.scaled + lfm.NAs.imputed + spp + site + year + month + (1 | individual), REML = F,  data = scaled.mem.data.alldates.noNAs)

performance::multicollinearity(m13) ##High correlation! Don't use
```

```{r}
m14 <- lmer(PC1 ~ excess_water.scaled + spp + site + year + month + (1 | individual), REML = F,  data = scaled.mem.data.alldates.noNAs)

performance::multicollinearity(m14)
```

```{r}
m15 <- lmer(PC1 ~ excess_water.scaled * spp + site + year + month + (1 | individual), REML = F,  data = scaled.mem.data.alldates.noNAs)

performance::multicollinearity(m15)
```

```{r}
tab_model(mpa_mod_fig.mpa, m11, m7, m10, m3, m6, m6.5, m12, m15,
          show.reflvl = TRUE, 
          digits = 3, 
          show.aic = TRUE, 
          show.ci = FALSE,
          show.icc = FALSE, 
          pred.labels = c('Intercept', 
                          'Water Potential (scaled)', 
                          'Species Effect (C. megacarpus)',
                          'Water Potential x Species Effect
                          (interaction)',
                          'LFM (%)',
                          'Water Weight (g)', 
                          'Dry Weight (g)',
                          'Site Effect (Site 1)',
                          'Excess Water (scaled)',
                          'Excess Water x Species Effect
                          (interaction)'),
         dv.labels = c("mpa_spp", 
                       "m11", 
                       "m7", 
                       "m7",
                       "m10",
                       "m3", 
                       "m6", 
                       "m6.5", 
                       "m12",
                       "m15"),
         string.pred = "Coeffcient", 
         title = "models with no collinearity",
  string.p = "P-Value", 
  p.style = "stars",
  file = here('figures', 'MEM.figures','site.yearANDmonth.html')) # saving .html into figures folder
```
# ---------------------------------------
# MAIN MODEL SELECTION

From our first round of model selection, it appears that site, species, and year.month are all important variables to include, and sample.wt should probably be included as well despite worsening the models slightly.

Below, I am going to perform another round of model selection with all possible permutations of the following metrics: mpa, lfm, dw.flam.sample, ww.flam.sample, excess_water.scaled with the caveat that some combinations of metrics absolutely do NOT work. Those combinations are as follows:
  - dw.flam.sample, ww.flam.sample and sample.wt
  - dw.flam.sample, lfm, ww.flam.sample
  - lfm, dw.flam.sample
  - lfm, ww.flam.sample
  - lfm, excess_water.scaled
  - ww.flam.sample, excess_water.scaled
  - dw.flam.sample, excess_water.scaled (with sample.wt)
  - dw.flam.sample, ww.flam.sample (with sample.wt)
These do not work, as live fuel moisture was used in the calculations of these metrics, therefore they return very high VIF values that signal to us that they cannot be used in the mixed effects models together

```{r}
m1 <- lmer(PC1 ~ lfm + spp + site + year.month + sample.wt + (1 | individual), REML = F,  data = scaled.mem.data.alldates.noNAs)

performance::multicollinearity(m1)
```

```{r}
m2 <- lmer(PC1 ~ mpa + spp + site + year.month + sample.wt + (1 | individual), REML = F,  data = scaled.mem.data.alldates.noNAs)

performance::multicollinearity(m2)
```

```{r}
m3 <- lmer(PC1 ~ dw.flam.sample + spp + site + year.month + sample.wt + (1 | individual), REML = F,  data = scaled.mem.data.alldates.noNAs)

performance::multicollinearity(m3)
```

```{r}
m4 <- lmer(PC1 ~ ww.flam.sample + spp + site + year.month + sample.wt + (1 | individual), REML = F,  data = scaled.mem.data.alldates.noNAs)

performance::multicollinearity(m4)
```

```{r}
m5 <- lmer(PC1 ~ excess_water.scaled + spp + site + year.month + sample.wt + (1 | individual), REML = F,  data = scaled.mem.data.alldates.noNAs)

performance::multicollinearity(m5)
```

```{r}
m6 <- lmer(PC1 ~ lfm + mpa + spp + site + year.month + sample.wt + (1 | individual), REML = F,  data = scaled.mem.data.alldates.noNAs)

performance::multicollinearity(m6)
```

```{r}
m7 <- lmer(PC1 ~ mpa + dw.flam.sample + spp + site + year.month + sample.wt + (1 | individual), REML = F,  data = scaled.mem.data.alldates.noNAs)

performance::multicollinearity(m7)
```

```{r}
m8 <- lmer(PC1 ~ mpa + ww.flam.sample + spp + site + year.month + sample.wt + (1 | individual), REML = F,  data = scaled.mem.data.alldates.noNAs)

performance::multicollinearity(m8)
```

```{r}
m9 <- lmer(PC1 ~ mpa + excess_water.scaled + spp + site + year.month + sample.wt + (1 | individual), REML = F,  data = scaled.mem.data.alldates.noNAs)

performance::multicollinearity(m9)
```

```{r}
m10 <- lmer(PC1 ~ mpa + dw.flam.sample + ww.flam.sample + spp + site + year.month + (1 | individual), REML = F,  data = scaled.mem.data.alldates.noNAs)

performance::multicollinearity(m10)
```

```{r}
m11 <- lmer(PC1 ~ dw.flam.sample + excess_water.scaled + spp + site + year.month + (1 | individual), REML = F,  data = scaled.mem.data.alldates.noNAs)

performance::multicollinearity(m11)
```

```{r}
m12 <- lmer(PC1 ~ dw.flam.sample + ww.flam.sample + spp + site + year.month + (1 | individual), REML = F,  data = scaled.mem.data.alldates.noNAs)

performance::multicollinearity(m12)
```

```{r}
tab_model(m1, m2, m3, m4, m5, m6, m7, m8, m9, m10, m11, m12,
          show.reflvl = TRUE, 
          digits = 3, 
          show.aic = TRUE, 
          show.ci = FALSE,
          show.icc = FALSE, 
          pred.labels = c('Intercept', 'LFM (%)',
                          'Species Effect (C. megacarpus)',
                          'Site Effect (Site 1)',
                          'Sampling Date Effect (Jan 2018)',
                          'Sampling Date Effect (Dec 2019)',
                          'Sampling Date Effect (Jan 2020)',
                          'Sampling Date Effect (Sept 2020)',
                          'Sample Weight (g)',
                          'Water Potential (scaled)',
                          'Dry Weight (g)',
                          'Water Weight (g)',
                          'Excess Water (scaled)'),
         dv.labels = c("Model 1", "Model 2", "Model 3", "Model 4",
                       "Model 5", "Model 6", "Model 7", "Model 8", 
                       "Model 9","Model 10", 'Model 11', 'Model 12'),
         string.pred = "Coeffcient", 
         title = "models with no collinearity",
  string.p = "P-Value", 
  p.style = "stars",
  file = here('figures', 'MEM.figures','modelselection.html')) # saving .html into figures folder
```

## P Values
The p values in the MEM tables is incorrect, and the Kenward-Roger approximation should be used via 'lmerTest' as follows

```{r}
m1 <- lmer(PC1 ~ lfm + spp + site + year.month + sample.wt + (1 | individual), data = scaled.mem.data.alldates.noNAs, REML = T)
summary(m1, ddf = 'Kenward-Roger')

m2 <- lmer(PC1 ~ mpa + spp + site + year.month + sample.wt + (1 | individual), data = scaled.mem.data.alldates.noNAs, REML = T)
summary(m2, ddf = 'Kenward-Roger')

m3 <- lmer(PC1 ~ dw.flam.sample + spp + site + year.month + sample.wt + (1 | individual), data = scaled.mem.data.alldates.noNAs, REML = T)
summary(m3, ddf = 'Kenward-Roger')

m4 <- lmer(PC1 ~ ww.flam.sample + spp + site + year.month + sample.wt + (1 | individual), data = scaled.mem.data.alldates.noNAs, REML = T)
summary(m4, ddf = 'Kenward-Roger')

m5 <- lmer(PC1 ~ excess_water.scaled + spp + site + year.month + sample.wt + (1 | individual), data = scaled.mem.data.alldates.noNAs, REML = T)
summary(m5, ddf = 'Kenward-Roger')

m6 <- lmer(PC1 ~ lfm + mpa + spp + site + year.month + sample.wt + (1 | individual), data = scaled.mem.data.alldates.noNAs, REML = T)
summary(m6, ddf = 'Kenward-Roger')

m7 <- lmer(PC1 ~ mpa + dw.flam.sample + spp + site + year.month + sample.wt + (1 | individual), data = scaled.mem.data.alldates.noNAs, REML = T)
summary(m7, ddf = 'Kenward-Roger')

m8 <- lmer(PC1 ~ mpa + ww.flam.sample + spp + site + year.month + sample.wt + (1 | individual), data = scaled.mem.data.alldates.noNAs, REML = T)
summary(m8, ddf = 'Kenward-Roger')

m9 <- lmer(PC1 ~ mpa + excess_water.scaled + spp + site + year.month + sample.wt + (1 | individual), data = scaled.mem.data.alldates.noNAs, REML = T)
summary(m9, ddf = 'Kenward-Roger')

m10 <- lmer(PC1 ~ mpa + dw.flam.sample + ww.flam.sample + spp + site + year.month + (1 | individual), data = scaled.mem.data.alldates.noNAs, REML = T)
summary(m10, ddf = 'Kenward-Roger')

m11 <- lmer(PC1 ~ dw.flam.sample + excess_water.scaled + spp + site + year.month + (1 | individual), data = scaled.mem.data.alldates.noNAs, REML = T)
summary(m11, ddf = 'Kenward-Roger')

m12 <- lmer(PC1 ~ dw.flam.sample + ww.flam.sample + spp + site + year.month + (1 | individual), data = scaled.mem.data.alldates.noNAs, REML = T)
summary(m12, ddf = 'Kenward-Roger')
```

# KableExtra Table
Setting up dataframe:
```{r}
dep.var <- c("Intercept", "Species Effect (C. megacarpus)", "Site (St. Mary's)", "Date (Jan. 2018)", "Date (Dec. 2019)","Date (Jan. 2020)", "Date (Sept. 2020)", "Sample Weight (g)", "LFM (%)", "Water Potential", "Dry Weight (g)", "Water Weight (g)", "Excess Water", "Random Effects", paste("\U03C3", "2"), paste("\U03C4", "00"), " ", "Marginal R2/Conditional R2", "AIC")
m1.est <- c("0.575", "-0.925**", "0.467", "-0.558", "-0.458", "0.077", "-0.975**", "-0.121", "-0.296***", " ", " ", " ", " ", " ", "2.26", "0.20ind", " ", "0.192/0.258", "1687.590")
m2.est <- c("-0.106", "-0.824**", "0.487", "-0.567", "-0.755", "-0.272", "-0.964**", "-0.103"," ", "-0.142***", " ", " ", " ", " ", "2.15", "0.18ind", " ", "0.229/0.289", "1668.106")
m3.est <- c("0.571", "-0.972**", "0.449", "-0.564", "-0.358", "0.188", "-0.958**", "-0.276***"," ", " ", "0.399***", " ", " ", " ", "2.23", "0.20ind", " ", "0.201/0.268", "1681.577")
m4.est <- c("0.569", "-0.966**", "0.451", "-0.564", "-0.360", "0.171", "-0.959**", "-0.001"," ", " ", " ", "-0.396***", " ", " ", "2.23", "0.20ind", " ", "0.201/0.267", "1682.009")
m5.est <- c("0.570", "-0.971**", "0.450", "-0.563", "-0.356", "0.182", "-0.958**", "-0.138", " ", " ", " ", " ", "-0.373***", " ", "2.23", "0.20ind", " ", "0.201/0.268", "1681.903")
m6.est <- c("-0.077", "-0.833**", "0.485", "-0.580", "-0.742", "-0.253", "-0.973**", "-0.103", "-0.026", "-0.138***", " ", " ", " ", " ", "2.16", "0.19ind", " ", "0.228/0.289", "1672.881")
m7.est <- c("-0.060", "-0.840**", "0.482", "-0.580", "-0.724", "-0.232", "-0.971**", "-0.121", " ", "-0.134***", "0.044", " ", " ", " ", "2.16", "0.19ind", " ", "0.228/0.290", "1672.325")
m8.est <- c("-0.067", "-0.838**", "0.483", "-0.578", "-0.728", "-0.240", "-0.970**", "-0.092", " ", "-0.135***", " ", "-0.038", " ", " ", "2.16", "0.19ind", " ",  "0.228/0.289", "1672.357")
m9.est <- c("-0.063", "-0.839**", "0.483", "-0.579", "-0.725", "-0.236", "-0.971**", "-0.105", " ", "-0.135***", " ", " ", "-0.039", " ", "2.16", "0.19ind", " ", "0.228/0.290", "1672.468")
m10.est <- c("-0.065", "-0.837**", "0.484", "-0.580", "-0.729", "-0.242", "-0.972**", " ", " ", "-0.135***", "-0.126", "-0.167", " ", " ", "2.16", "0.19ind", " ", "0.228/0.290", "1671.735")
m11.est <- c("0.569", "-0.969***", "0.451", "-0.562", "-0.356", "0.179", "-0.960**", " ", " ", " ", "-0.383", " ", "-0.733***", " ", "2.23", "0.20ind", " ", "0.201/0.268", "1680.009")
m12.est <- c("0.570", "-0.967**", "0.451", "-0.565", "-0.362", "0.174", "-0.961**", " ", " ", " ", "0.007", "-0.391***", " ", " ", "2.23", "0.20ind", " ", "0.201/0.268", "1681.279")
main.mem.table.df <- data.frame(dep.var, m1.est, m2.est, m3.est, m4.est, m5.est, m6.est, m7.est, m8.est, m9.est, m10.est, m11.est, m12.est)
```

Kable table:
```{r}
main.mem.table.df %>%
  kable(format = 'html', escape = F, col.names = c(' ', 'Model 1', 'Model 2','Model 3', 'Model 4', 'Model 5', 'Model 6', 'Model 7', 'Model 8', 'Model 9', 'Model 10', 'Model 11', 'Model 12')) %>% 
  kable_styling(bootstrap_options = c('hover', 'bordered', 'condensed'), fixed_thead = T, stripe_color = "black") %>% 
  row_spec(c(14), background = "#D3D3D3", bold = T) %>%
  row_spec(c(17), background = "black") %>% 
  row_spec(c(18:19), bold = T) %>%
  row_spec(c(1:19), color = "black") %>% 
  column_spec(c(1), bold = T, border_right = T) %>% 
  save_kable(here('figures', 'supp-figures', 'TabS1_Full_PC1_MEM_Selection_kable.html'))
```

# -------------------------------
# Splitting by Species (Supplemental Table)
## PC1
```{r}
c.p1 <- lmer(PC1 ~ mpa.scaled + site + year.month + (1 | individual), data = mem.CEME.subset)

c.p2 <- lmer(PC1 ~ lfm + site + year.month + (1 | individual), data = mem.CEME.subset)

c.p3 <- lmer(PC1 ~ ww.flam.sample + site + year.month + (1 | individual), data = mem.CEME.subset)

c.p4 <- lmer(PC1 ~ dw.flam.sample + site + year.month + (1 | individual), data = mem.CEME.subset)

c.p5 <- lmer(PC1 ~ mpa.scaled + dw.flam.sample + sample.wt  + site + year.month + (1 | individual), data = mem.CEME.subset)

# Adding in MPa + WW and MPa + LFM despite their moderate collinearity to see how they affect model selection
c.p6 <- lmer(PC1 ~ mpa.scaled + ww.flam.sample + site + year.month + (1 | individual), data = mem.CEME.subset)

c.p7 <- lmer(PC1 ~ mpa.scaled + lfm + sample.wt + site + year.month + (1 | individual), data = mem.CEME.subset)
```

```{r}
a.p1 <- lmer(PC1 ~ mpa.scaled + site + year.month + (1 | individual), data = mem.adfa.subset)

a.p2 <- lmer(PC1 ~ lfm + site + year.month + (1 | individual), data = mem.adfa.subset)

a.p3 <- lmer(PC1 ~ ww.flam.sample + site + year.month + (1 | individual), data = mem.adfa.subset)

a.p4 <- lmer(PC1 ~ dw.flam.sample + site + year.month + (1 | individual), data = mem.adfa.subset)

a.p5 <- lmer(PC1 ~ mpa.scaled + dw.flam.sample + sample.wt + site + year.month + (1 | individual), data = mem.adfa.subset)

# Adding in MPa + WW and MPa + LFM despite their moderate collinearity to see how they affect model selection
a.p6 <- lmer(PC1 ~ mpa.scaled + ww.flam.sample + site + year.month + (1 | individual), data = mem.adfa.subset)

a.p7 <- lmer(PC1 ~ mpa.scaled + lfm + sample.wt + site + year.month + (1 | individual), data = mem.adfa.subset)
```

### Table
```{r}
tab_model(a.p1, a.p5, a.p3, a.p2, a.p4, c.p1, c.p3, c.p4, c.p5, c.p2, a.p6, a.p7, c.p6, c.p7,
          show.reflvl = FALSE, 
          digits = 3, 
          show.aic = TRUE,
          show.ci = FALSE,
          show.icc = FALSE, 
          pred.labels = c(),
         dv.labels = c("Model 1 (ADFA)", "Model 5 (ADFA)", "Model 3 (ADFA)", "Model 2 (ADFA)",  "Model 4 (ADFA)", "Model 1 (CEME)", "Model 3 (CEME)",  "Model 4 (CEME)", "Model 5 (CEME)",  "Model 2 (CEME)", "Model 6 (ADFA)", "Model 7 (ADFA)", "Model 6 (CEME)", "Model 7 (CEME)"),
         string.pred = "Coeffcient", 
         title = "models with no collinearity",
  string.p = "P-Value", 
  p.style = "stars",
  file = here('figures', 'MEM.figures','PC1.by.species.html')) # saving .html into figures folder
```

## Time to Ignition
```{r}
c.t1 <- lmer(tti ~ mpa.scaled + site + year.month + (1 | individual), data = mem.CEME.subset)

c.t2 <- lmer(tti ~ lfm + site + year.month + (1 | individual), data = mem.CEME.subset)

c.t3 <- lmer(tti ~ ww.flam.sample + site + year.month + (1 | individual), data = mem.CEME.subset)

c.t4 <- lmer(tti ~ dw.flam.sample + site + year.month + (1 | individual), data = mem.CEME.subset)

c.t5 <- lmer(tti ~ mpa.scaled + dw.flam.sample + sample.wt   + site + year.month + (1 | individual), data = mem.CEME.subset)

# Adding in MPa + WW and MPa + LFM despite their moderate collinearity to see how they affect model selection
c.t6 <- lmer(tti ~ mpa.scaled + ww.flam.sample + site + year.month + (1 | individual), data = mem.CEME.subset)

c.t7 <- lmer(tti ~ mpa.scaled + lfm + sample.wt + site + year.month + (1 | individual), data = mem.CEME.subset)
```

```{r}
a.t1 <- lmer(tti ~ mpa.scaled + site + year.month + (1 | individual), data = mem.adfa.subset)

a.t2 <- lmer(tti ~ lfm + site + year.month + (1 | individual), data = mem.adfa.subset)

a.t3 <- lmer(tti ~ ww.flam.sample + site + year.month + (1 | individual), data = mem.adfa.subset)

a.t4 <- lmer(tti ~ dw.flam.sample + site + year.month + (1 | individual), data = mem.adfa.subset)

a.t5 <- lmer(tti ~ mpa.scaled + dw.flam.sample + sample.wt + site + year.month + (1 | individual), data = mem.adfa.subset)

# Adding in MPa + WW and MPa + LFM despite their moderate collinearity to see how they affect model selection
a.t6 <- lmer(tti ~ mpa.scaled + ww.flam.sample + site + year.month + (1 | individual), data = mem.adfa.subset)

a.t7 <- lmer(tti ~ mpa.scaled + lfm + sample.wt + site + year.month + (1 | individual), data = mem.adfa.subset)
```

### Table
```{r}
tab_model(a.t5, a.t1, a.t3, a.t2, a.t4, c.t5, c.t1, c.t3, c.t4, c.t2, a.t6, a.t7, c.t6, c.t7,
          show.reflvl = FALSE, 
          digits = 3, 
          show.aic = TRUE, 
          show.ci = FALSE,
          show.icc = FALSE, 
          pred.labels = c(),
         dv.labels = c("Model 5 (ADFA)", "Model 1 (ADFA)", "Model 3 (ADFA)", "Model 2 (ADFA)",   "Model 4 (ADFA)", "Model 5 (CEME)", "Model 1 (CEME)", "Model 3 (CEME)",  "Model 4 (CEME)", "Model 2 (CEME)", "Model 6 (ADFA)", "Model 7 (ADFA)", "Model 6 (CEME)", "Model 7 (CEME)"),
         string.pred = "Coeffcient", 
         title = "models with no collinearity",
  string.p = "P-Value", 
  p.style = "stars",
  file = here('figures', 'MEM.figures','TTI.by.species.html')) # saving .html into figures folder
```

## Flame Duration
```{r}
c.f1 <- lmer(fd ~ mpa.scaled + site + year.month + (1 | individual), data = mem.CEME.subset)

c.f2 <- lmer(fd ~ lfm + site + year.month + (1 | individual), data = mem.CEME.subset)

c.f3 <- lmer(fd ~ ww.flam.sample + site + year.month + (1 | individual), data = mem.CEME.subset)

c.f4 <- lmer(fd ~ dw.flam.sample + site + year.month + (1 | individual), data = mem.CEME.subset)

c.f5 <- lmer(fd ~ mpa.scaled + dw.flam.sample + sample.wt   + site + year.month + (1 | individual), data = mem.CEME.subset)

# Adding in MPa + WW and MPa + LFM despite their moderate collinearity to see how they affect model selection
c.f6 <- lmer(fd ~ mpa.scaled + ww.flam.sample + site + year.month + (1 | individual), data = mem.CEME.subset)

c.f7 <- lmer(fd ~ mpa.scaled + lfm + sample.wt + site + year.month + (1 | individual), data = mem.CEME.subset)
```

```{r}
a.f1 <- lmer(fd ~ mpa.scaled + site + year.month + (1 | individual), data = mem.adfa.subset)

a.f2 <- lmer(fd ~ lfm + site + year.month + (1 | individual), data = mem.adfa.subset)

a.f3 <- lmer(fd ~ ww.flam.sample + site + year.month + (1 | individual), data = mem.adfa.subset)

a.f4 <- lmer(fd ~ dw.flam.sample + site + year.month + (1 | individual), data = mem.adfa.subset)

a.f5 <- lmer(fd ~ mpa.scaled + dw.flam.sample + sample.wt + site + year.month + (1 | individual), data = mem.adfa.subset)

# Adding in MPa + WW and MPa + LFM despite their moderate collinearity to see how they affect model selection
a.f6 <- lmer(fd ~ mpa.scaled + ww.flam.sample + site + year.month + (1 | individual), data = mem.adfa.subset)

a.f7 <- lmer(fd ~ mpa.scaled + lfm + sample.wt + site + year.month + (1 | individual), data = mem.adfa.subset)
```

### Table
```{r}
tab_model(a.f1, a.f4, a.f2, a.f3, a.f5, c.f4, c.f2, c.f1, c.f3, c.f5, a.f6, a.f7, c.f6, c.f7,
          show.reflvl = FALSE, 
          digits = 3, 
          show.aic = TRUE, 
          show.ci = FALSE,
          show.icc = FALSE, 
          pred.labels = c(),
         dv.labels = c("Model 1 (ADFA)",  "Model 4 (ADFA)", "Model 2 (ADFA)", "Model 3 (ADFA)", "Model 5 (ADFA)",  "Model 4 (CEME)", "Model 2 (CEME)", "Model 1 (CEME)", "Model 3 (CEME)", "Model 5 (CEME)", "Model 6 (ADFA)", "Model 7 (ADFA)", "Model 6 (CEME)", "Model 7 (CEME)"),
         string.pred = "Coeffcient", 
         title = "models with no collinearity",
  string.p = "P-Value", 
  p.style = "stars",
  file = here('figures', 'MEM.figures','FD.by.species.html')) # saving .html into figures folder
```

## Glow Duration
```{r}
c.g1 <- lmer(gd ~ mpa.scaled + site + year.month + (1 | individual), data = mem.CEME.subset)

c.g2 <- lmer(gd ~ lfm + site + year.month + (1 | individual), data = mem.CEME.subset)

c.g3 <- lmer(gd ~ ww.flam.sample + site + year.month + (1 | individual), data = mem.CEME.subset)

c.g4 <- lmer(gd ~ dw.flam.sample + site + year.month + (1 | individual), data = mem.CEME.subset)

c.g5 <- lmer(gd ~ mpa.scaled + dw.flam.sample + sample.wt + site + year.month + (1 | individual), data = mem.CEME.subset)

# Adding in MPa + WW and MPa + LFM despite their moderate collinearity to see how they affect model selection
c.g6 <- lmer(gd ~ mpa.scaled + ww.flam.sample + site + year.month + (1 | individual), data = mem.CEME.subset)

c.g7 <- lmer(gd ~ mpa.scaled + lfm + sample.wt + site + year.month + (1 | individual), data = mem.CEME.subset)
```

```{r}
a.g1 <- lmer(gd ~ mpa.scaled + site + year.month + (1 | individual), data = mem.adfa.subset)

a.g2 <- lmer(gd ~ lfm + site + year.month + (1 | individual), data = mem.adfa.subset)

a.g3 <- lmer(gd ~ ww.flam.sample + site + year.month + (1 | individual), data = mem.adfa.subset)

a.g4 <- lmer(gd ~ dw.flam.sample + site + year.month + (1 | individual), data = mem.adfa.subset)

a.g5 <- lmer(gd ~ mpa.scaled + dw.flam.sample + sample.wt + site + year.month + (1 | individual), data = mem.adfa.subset)

# Adding in MPa + WW and MPa + LFM despite their moderate collinearity to see how they affect model selection
a.g6 <- lmer(gd ~ mpa.scaled + ww.flam.sample + site + year.month + (1 | individual), data = mem.adfa.subset)

a.g7 <- lmer(gd ~ mpa.scaled + lfm + sample.wt + site + year.month + (1 | individual), data = mem.adfa.subset)
```

### Table
```{r}
tab_model(a.g5, a.g4, a.g2, a.g3, a.g1, c.g5, c.g4, c.g1, c.g2, c.g3, a.g6, a.g7, c.g6, c.g7,
          show.reflvl = FALSE, 
          digits = 3, 
          show.aic = TRUE, 
          show.ci = FALSE,
          show.icc = FALSE, 
          pred.labels = c(),
         dv.labels = c("Model 5 (ADFA)", "Model 4 (ADFA)", "Model 2 (ADFA)",  "Model 3 (ADFA)", "Model 1 (ADFA)", "Model 5 (CEME)", "Model 4 (CEME)", "Model 1 (CEME)",  "Model 2 (CEME)", "Model 3 (CEME)", "Model 6 (ADFA)", "Model 7 (ADFA)", "Model 6 (CEME)", "Model 7 (CEME)"),
         string.pred = "Coeffcient", 
         title = "models with no collinearity",
  string.p = "P-Value", 
  p.style = "stars",
  file = here('figures', 'MEM.figures','GD.by.species.html')) # saving .html into figures folder
```

## P Values
```{r}
a.p1 <- lmer(PC1 ~ mpa.scaled + site + year.month + (1 | individual), data = mem.adfa.subset, REML = T)

a.p2 <- lmer(PC1 ~ lfm + site + year.month + (1 | individual), data = mem.adfa.subset, REML = T)

a.p3 <- lmer(PC1 ~ ww.flam.sample + site + year.month + (1 | individual), data = mem.adfa.subset, REML = T)

a.p4 <- lmer(PC1 ~ dw.flam.sample + site + year.month + (1 | individual), data = mem.adfa.subset, REML = T)

a.p5 <- lmer(PC1 ~ mpa.scaled + dw.flam.sample + sample.wt + site + year.month + (1 | individual), data = mem.adfa.subset, REML = T)

c.p1 <- lmer(PC1 ~ mpa.scaled + site + year.month + (1 | individual), data = mem.CEME.subset, REML = T)

c.p2 <- lmer(PC1 ~ lfm + site + year.month + (1 | individual), data = mem.CEME.subset, REML = T)

c.p3 <- lmer(PC1 ~ ww.flam.sample + site + year.month + (1 | individual), data = mem.CEME.subset, REML = T)

c.p4 <- lmer(PC1 ~ dw.flam.sample + site + year.month + (1 | individual), data = mem.CEME.subset, REML = T)

c.p5 <- lmer(PC1 ~ mpa.scaled + dw.flam.sample + sample.wt  + site + year.month + (1 | individual), data = mem.CEME.subset, REML = T)

a.t1 <- lmer(tti ~ mpa.scaled + site + year.month + (1 | individual), data = mem.adfa.subset, REML = T)

a.t2 <- lmer(tti ~ lfm + site + year.month + (1 | individual), data = mem.adfa.subset, REML = T)

a.t3 <- lmer(tti ~ ww.flam.sample + site + year.month + (1 | individual), data = mem.adfa.subset, REML = T)

a.t4 <- lmer(tti ~ dw.flam.sample + site + year.month + (1 | individual), data = mem.adfa.subset, REML = T)

a.t5 <- lmer(tti ~ mpa.scaled + dw.flam.sample + sample.wt + site + year.month + (1 | individual), data = mem.adfa.subset, REML = T)

c.t1 <- lmer(tti ~ mpa.scaled + site + year.month + (1 | individual), data = mem.CEME.subset, REML = T)

c.t2 <- lmer(tti ~ lfm + site + year.month + (1 | individual), data = mem.CEME.subset, REML = T)

c.t3 <- lmer(tti ~ ww.flam.sample + site + year.month + (1 | individual), data = mem.CEME.subset, REML = T)

c.t4 <- lmer(tti ~ dw.flam.sample + site + year.month + (1 | individual), data = mem.CEME.subset, REML = T)

c.t5 <- lmer(tti ~ mpa.scaled + dw.flam.sample + sample.wt   + site + year.month + (1 | individual), data = mem.CEME.subset, REML = T)

a.f1 <- lmer(fd ~ mpa.scaled + site + year.month + (1 | individual), data = mem.adfa.subset, REML = T)

a.f2 <- lmer(fd ~ lfm + site + year.month + (1 | individual), data = mem.adfa.subset, REML = T)

a.f3 <- lmer(fd ~ ww.flam.sample + site + year.month + (1 | individual), data = mem.adfa.subset, REML = T)

a.f4 <- lmer(fd ~ dw.flam.sample + site + year.month + (1 | individual), data = mem.adfa.subset, REML = T)

a.f5 <- lmer(fd ~ mpa.scaled + dw.flam.sample + sample.wt + site + year.month + (1 | individual), data = mem.adfa.subset, REML = T)

c.f1 <- lmer(fd ~ mpa.scaled + site + year.month + (1 | individual), data = mem.CEME.subset, REML = T)

c.f2 <- lmer(fd ~ lfm + site + year.month + (1 | individual), data = mem.CEME.subset, REML = T)

c.f3 <- lmer(fd ~ ww.flam.sample + site + year.month + (1 | individual), data = mem.CEME.subset, REML = T)

c.f4 <- lmer(fd ~ dw.flam.sample + site + year.month + (1 | individual), data = mem.CEME.subset, REML = T)

c.f5 <- lmer(fd ~ mpa.scaled + dw.flam.sample + sample.wt   + site + year.month + (1 | individual), data = mem.CEME.subset, REML = T)

a.g1 <- lmer(gd ~ mpa.scaled + site + year.month + (1 | individual), data = mem.adfa.subset, REML = T)

a.g2 <- lmer(gd ~ lfm + site + year.month + (1 | individual), data = mem.adfa.subset, REML = T)

a.g3 <- lmer(gd ~ ww.flam.sample + site + year.month + (1 | individual), data = mem.adfa.subset, REML = T)

a.g4 <- lmer(gd ~ dw.flam.sample + site + year.month + (1 | individual), data = mem.adfa.subset, REML = T)

a.g5 <- lmer(gd ~ mpa.scaled + dw.flam.sample + sample.wt + site + year.month + (1 | individual), data = mem.adfa.subset, REML = T)

c.g1 <- lmer(gd ~ mpa.scaled + site + year.month + (1 | individual), data = mem.CEME.subset, REML = T)

c.g2 <- lmer(gd ~ lfm + site + year.month + (1 | individual), data = mem.CEME.subset, REML = T)

c.g3 <- lmer(gd ~ ww.flam.sample + site + year.month + (1 | individual), data = mem.CEME.subset, REML = T)

c.g4 <- lmer(gd ~ dw.flam.sample + site + year.month + (1 | individual), data = mem.CEME.subset, REML = T)

c.g5 <- lmer(gd ~ mpa.scaled + dw.flam.sample + sample.wt + site + year.month + (1 | individual), data = mem.CEME.subset, REML = T)

MEM.split.spp.list <- list(a.p1, a.p5, a.p3, a.p2, a.p4, c.p1, c.p3, c.p4, c.p5, c.p2, a.t5, a.t1, a.t3, a.t2, a.t4, c.t5, c.t1, c.t3, c.t4, c.t2, a.f1, a.f4, a.f2, a.f3, a.f5, c.f4, c.f2, c.f1, c.f3, c.f5, a.g5, a.g4, a.g2, a.g3, a.g1, c.g5, c.g4, c.g1, c.g2, c.g3)

lapply(MEM.split.spp.list, function(x) summary(x, ddf = 'Kenward-Roger'))
```



## SUPP: Kable Table
```{r}
dep.var <- c("PC1", " ", " ", " ", " ",  " ", " ", " ", " ", " ", "Time to Ignition (s)", " ", " ", " ", " "," ", " ", " ", " ",  " ", "Flame Duration (s)", " ", " ", " ", " "," ", " ", " ", " ",  " ", "Glow Duration (s)", " ", " ", " "," ", " ", " ", " ",  " ", " ")
sp.lab <- c("Adenostoma fasciculatum", " ", " ", " ", " ", "Ceanothus megacarpus", " ", " ", " ", " ", "Adenostoma fasciculatum", " ", " ", " ", " ", "Ceanothus megacarpus", " ", " ", " ", " ", "Adenostoma fasciculatum", " ", " ", " ", " ", "Ceanothus megacarpus", " ", " ", " ", " ", "Adenostoma fasciculatum", " ", " ", " ", " ", "Ceanothus megacarpus", " ", " ", " ", " ")
fix.fx <- c("Water Potential", "Water Potential, Sample Weight & Dry Weight", "Wet Weight", "Live Fuel Moisture", "Dry Weight", "Water Potential", "Wet Weight", "Dry Weight", "Water Potential, Sample Weight & Dry Weight", "Live Fuel Moisture", "Water Potential, Sample Weight & Dry Weight", "Water Potential", "Wet Weight", "Live Fuel Moisture", "Dry Weight", "Water Potential, Sample Weight & Dry Weight", "Water Potential", "Wet Weight", "Dry Weight", "Live Fuel Moisture", "Water Potential", "Dry Weight", "Live Fuel Moisture", "Wet Weight", "Water Potential, Sample Weight & Dry Weight", "Dry Weight", "Live Fuel Moisture", "Water Potential", "Wet Weight", "Water Potential, Sample Weight & Dry Weight", "Water Potential, Sample Weight & Dry Weight", "Dry Weight", "Wet Weight", "Live Fuel Moisture", "Water Potential", "Water Potential, Sample Weight & Dry Weight", "Dry Weight", "Water Potential", "Live Fuel Moisture", "Wet Weight")
est <- c('-0.455***',	'-0.447***', '-0.388***', '-0.288**', '0.225', '-0.429***', '-0.428***', '0.371**', '-0.321*', '-0.357**', "4.353***",	'4.502***', '3.899***', '2.750***', '-2.243**', '5.113***',	'5.454***', '4.823***', '-4.031***', '3.874***', '0.688*', '0.657*', '-0.556', '-0.574', '-0.588', '0.548', '-0.473', '-0.420', '-0.210', '-0.353', '-0.410', '3.345', '-1.755', '-1.682', '-1.559', '-2.218', '3.954', '-2.584', '-2.072', '-0.393')
aic.values <- c(991.932, 998.721, 1001.523,	1005.452,	1009.031,	675.597,	679.522, 682.241, 682.551, 683.855, 1934.922, 1938.040, 1960.263, 1970.769,	1977.598,	1379.405,	1380.281,	1402.308,	1410.268,	1413.893, 1513.504,	1515.244,	1515.939,	1516.295,	1517.224,	1051.365,	1052.229,	1052.325,	1053.778,	1054.336, 2392.661,	2395.074,	2397.632, 2397.718,	2397.866,	1735.815,	1741.946,	1743.660,	1744.259,	1745.052)
r2.values <- c('0.161 / 0.269',	'0.166 / 0.263',	'0.133 / 0.242',	'0.122 / 0.235',	'0.109 / 0.217',	'0.151 / 0.176',	'0.132 / 0.154',	'0.119 / 0.153',	'0.155 / 0.178',	'0.111 / 0.141', '0.253 / 0.335',	'0.242 / 0.335',	'0.182 / 0.279',	'0.152 / 0.252',	'0.129 / 0.219',	'0.283 / 0.327',	'0.280 / 0.328',	'0.193 / 0.240',	'0.159 / 0.216',	'0.143 / 0.199', '0.067 / 0.136',	'0.061 / 0.126',	'0.059 / 0.127',	'0.057 / 0.128',	'0.069 / 0.136',	'0.037 / 0.134',	'0.033 / 0.124',	'0.033 / 0.131',	'0.025 / 0.119',	'0.042 / 0.151', '0.229 / 0.345',	'0.229 / 0.348', '0.218 / 0.354',	'0.219 / 0.352',	'0.218 / 0.355',	'0.125 / 0.125',	'0.113 / NA',	'0.106 / 0.111',	'0.101 / 0.103',	'0.097 / 0.103')
d.aic <- c(0, 6.789, 9.591, 13.52, 17.099, 0, 3.925, 6.644, 6.954, 8.258, 0, 3.118, 25.341, 35.847, 42.676, 0, 0.876, 22.903, 30.863, 34.488, 0, 1.74, 2.435, 2.791, 3.72, 0, 0.864, 0.96, 2.413, 2.971, 0, 2.413, 4.971, 5.057, 5.205, 0, 6.131, 7.845, 8.444, 9.237)
mem.spp.table.df <- data.frame(dep.var, sp.lab, fix.fx, est, d.aic, r2.values)
```

```{r}
mem.spp.table.df %>%
  kable(format = 'html', escape = F, col.names = c('Dependent Variable', 'Species', 'Model Fixed Effects', 'Water Estimate','\U0394 AIC', expression('Marginal R2/Conditional R2'))) %>% 
  kable_styling(bootstrap_options = c('hover', 'bordered', 'condensed'), fixed_thead = T) %>% 
  row_spec(c(1, 6, 11, 16, 21, 26, 31, 36), background = "#D3D3D3") %>%
  row_spec(c(1:40), color = 'black') %>% 
  row_spec(c(1, 6, 11, 16:17, 21:22, 26:28, 31, 36), bold = T) %>% 
  column_spec(2, italic = T) %>% 
  column_spec(1, width = "0.1in") %>% 
  save_kable(here('figures', 'supp-figures', 'TabS2_Spp_Split_MEM_kable.html'))
```

# -----------------------------------
# Predicting PC2
From our first round of model selection, it appears that site, species, and year.month are all important variables to include, and sample.wt should probably be included as well despite worsening the models slightly.

Below, I am going to perform another round of model selection with all possible permutations of the following metrics: mpa, lfm, dw.flam.sample, ww.flam.sample, excess_water.scaled with the caveat that some combinations of metrics absolutely do NOT work. Those combinations are as follows:
  - dw.flam.sample, ww.flam.sample and sample.wt
  - dw.flam.sample, lfm, ww.flam.sample
  - lfm, dw.flam.sample
  - lfm, ww.flam.sample
  - lfm, excess_water.scaled
  - ww.flam.sample, excess_water.scaled
  - dw.flam.sample, excess_water.scaled (with sample.wt)
  - dw.flam.sample, ww.flam.sample (with sample.wt)
These do not work, as live fuel moisture was used in the calculations of these metrics, therefore they return very high VIF values that signal to us that they cannot be used in the mixed effects models together

```{r}
m1 <- lmer(PC2 ~ lfm + spp + site + year.month + sample.wt + (1 | individual), REML = F,  data = scaled.mem.data.alldates.noNAs)

performance::multicollinearity(m1)
```

```{r}
m2 <- lmer(PC2 ~ mpa + spp + site + year.month + sample.wt + (1 | individual), REML = F,  data = scaled.mem.data.alldates.noNAs)

performance::multicollinearity(m2)
```

```{r}
m3 <- lmer(PC2 ~ dw.flam.sample + spp + site + year.month + sample.wt + (1 | individual), REML = F,  data = scaled.mem.data.alldates.noNAs)

performance::multicollinearity(m3)
```

```{r}
m4 <- lmer(PC2 ~ ww.flam.sample + spp + site + year.month + sample.wt + (1 | individual), REML = F,  data = scaled.mem.data.alldates.noNAs)

performance::multicollinearity(m4)
```

```{r}
m5 <- lmer(PC2 ~ excess_water.scaled + spp + site + year.month + sample.wt + (1 | individual), REML = F,  data = scaled.mem.data.alldates.noNAs)

performance::multicollinearity(m5)
```

```{r}
m6 <- lmer(PC2 ~ lfm + mpa + spp + site + year.month + sample.wt + (1 | individual), REML = F,  data = scaled.mem.data.alldates.noNAs)

performance::multicollinearity(m6)
```

```{r}
m7 <- lmer(PC2 ~ mpa + dw.flam.sample + spp + site + year.month + sample.wt + (1 | individual), REML = F,  data = scaled.mem.data.alldates.noNAs)

performance::multicollinearity(m7)
```

```{r}
m8 <- lmer(PC2 ~ mpa + ww.flam.sample + spp + site + year.month + sample.wt + (1 | individual), REML = F,  data = scaled.mem.data.alldates.noNAs)

performance::multicollinearity(m8)
```

```{r}
m9 <- lmer(PC2 ~ mpa + excess_water.scaled + spp + site + year.month + sample.wt + (1 | individual), REML = F,  data = scaled.mem.data.alldates.noNAs)

performance::multicollinearity(m9)
```

```{r}
m10 <- lmer(PC2 ~ mpa + dw.flam.sample + ww.flam.sample + spp + site + year.month + (1 | individual), REML = F,  data = scaled.mem.data.alldates.noNAs)

performance::multicollinearity(m10)
```

```{r}
m11 <- lmer(PC2 ~ dw.flam.sample + excess_water.scaled + spp + site + year.month + (1 | individual), REML = F,  data = scaled.mem.data.alldates.noNAs)

performance::multicollinearity(m11)
```

```{r}
m12 <- lmer(PC2 ~ dw.flam.sample + ww.flam.sample + spp + site + year.month + (1 | individual), REML = F,  data = scaled.mem.data.alldates.noNAs)

performance::multicollinearity(m12)
```

```{r}
tab_model(m1, m2, m3, m4, m5, m6, m7, m8, m9, m10, m11, m12,
          show.reflvl = TRUE, 
          digits = 3, 
          show.aic = TRUE, 
          show.ci = FALSE,
          show.icc = FALSE, 
          pred.labels = c('Intercept', 'LFM (%)',
                          'Species Effect (C. megacarpus)',
                          'Site Effect (Site 1)',
                          'Sampling Date Effect (Jan 2018)',
                          'Sampling Date Effect (Dec 2019)',
                          'Sampling Date Effect (Jan 2020)',
                          'Sampling Date Effect (Sept 2020)',
                          'Sample Weight (g)',
                          'Water Potential (scaled)',
                          'Dry Weight (g)',
                          'Water Weight (g)',
                          'Excess Water (scaled)'),
         dv.labels = c("Model 1", "Model 2", "Model 3", "Model 4",
                       "Model 5", "Model 6", "Model 7", "Model 8", 
                       "Model 9","Model 10", 'Model 11', 'Model 12'),
         string.pred = "Coeffcient", 
         title = "models with no collinearity",
  string.p = "P-Value", 
  p.style = "stars",
  file = here('figures', 'MEM.figures','PC2.html')) # saving .html into figures folder
```
# -----------------------------------
# Visualizations
## Dry Weight Effect
### Models
```{r}
m.p1 <- lmer(PC1 ~ mpa + spp + site + year.month + sample.wt + (1 | individual), data = scaled.mem.data.alldates.noNAs)
  
m.p2 <- lmer(PC2 ~ mpa + spp + site + year.month + sample.wt + (1 | individual), data = scaled.mem.data.alldates.noNAs)
```

```{r}
r_pc1 <- remef(m.p1, fix = 2, ran = list(individual = c("(Intercept)")))

r_pc2 <- remef(m.p2, fix = 2, ran = list(individual = c("(Intercept)")))

r_scaled.mem.data.alldates.noNAs <- scaled.mem.data.alldates.noNAs

r_scaled.mem.data.alldates.noNAs$r_pc1 <- r_pc1
r_scaled.mem.data.alldates.noNAs$r_pc2 <- r_pc2
```
### Plots
```{r}
p1 <- ggplot(r_scaled.mem.data.alldates.noNAs, aes(x = dw.flam.sample, y = r_pc1)) +
  geom_point(alpha = 0.6, size = 1) +
  labs(x = "Dry Weight (scaled, centered)", y = "PC1 (Effects Removed)") +
  geom_smooth(method = 'lm', color = 'black') +
  facet_wrap(~spp) + 
  th

p2 <- ggplot(r_scaled.mem.data.alldates.noNAs, aes(x = dw.flam.sample, y = PC1)) +
  geom_point(alpha = 0.6, size = 1) +
  labs(x = "Dry Weight (scaled, centered)") +
  geom_smooth(method = 'lm', color = 'black') +
  facet_wrap(~spp) + 
  th

p3 <- ggplot(r_scaled.mem.data.alldates.noNAs, aes(x = dw.flam.sample, y = r_pc2)) +
  geom_point(alpha = 0.6, size = 1) +
  labs(x = "Dry Weight (scaled, centered)", y = "PC2 (Effects Removed)") +
  geom_smooth(method = 'lm', color = 'black') +
  facet_wrap(~spp) + 
  th

p4 <- ggplot(r_scaled.mem.data.alldates.noNAs, aes(x = dw.flam.sample, y = PC2)) +
  geom_point(alpha = 0.6, size = 1) +
  labs(x = "Dry Weight (scaled, centered)") +
  geom_smooth(method = 'lm', color = 'black') +
  facet_wrap(~spp) + 
  th

ggarrange(p1, p3, p2, p4)

ggsave(here('figures', 'extra-figures', 'remef.dry.weight.jpg'), height = 8, width = 13)
```

## Composite Model
```{r}
comp.model.df <- scaled.mem.data.alldates.noNAs %>% 
  mutate(site.n = case_when(
    site == "St. Marys" ~ 1,
    site == "Painted Caves" ~ 2)) %>% 
  mutate(spp.n = case_when(
    spp == "ADFA" ~ 1,
    spp == "CEME" ~ 2))
  

model <- lmer(PC1 ~ mpa + dw.flam.sample + spp.n + site.n + year.month + sample.wt + (1 | individual), data = comp.model.df)

comp.model.df$r_pc1 <- remef(model, fix = c("mpa", "sample.wt", "spp.n", "site.n"))
```

### Plots
```{r}
p1 <- ggplot(comp.model.df, aes(x = mpa, y = r_pc1)) +
  geom_point(alpha = 0.6, size = 1) +
  labs(x = "Water Potential (scaled, centered)", y = "PC1 (Effects Removed)") +
  geom_smooth(method = 'lm', color = 'black') +
  th

p2 <- ggplot(comp.model.df, aes(x = dw.flam.sample, y = r_pc1)) +
  geom_point(alpha = 0.6, size = 1) +
  labs(x = "Dry Weight (scaled, centered)", y = "PC1 (Effects Removed)") +
  geom_smooth(method = 'lm', color = 'black') +
  theme(axis.title.y = element_blank()) +
  th

p3 <- ggplot(comp.model.df, aes(x = sample.wt, y = r_pc1)) +
  geom_point(alpha = 0.6, size = 1) +
  labs(x = "Sample Weight (scaled, centered)", y = "PC1 (Effects Removed)") +
  geom_smooth(method = 'lm', color = 'black') +
  theme(axis.title.y = element_blank()) +
  th

ggarrange(p1,p2,p3, nrow = 1)
```


# -----------------------------------
## 2.0.2 Tables of models

"The variance inflation factor is a measure to analyze the magnitude of multicollinearity of model terms. A VIF less than 5 indicates a low correlation of that predictor with other predictors. A value between 5 and 10 indicates a moderate correlation, while VIF values larger than 10 are a sign for high, not tolerable correlation of model predictors.

The Increased SE column in the output indicates how much larger the standard error is due to the correlation with other predictors." (https://easystats.github.io/blog/posts/performance_check_collinearity/)
https://strengejacke.github.io/sjPlot/articles/tab_model_estimates.html 
##### Tables
```{r}
tab_model(mpa_mod_fig.mpa, m11, m7, m10, m3, m6, m6.5, m12, m15,
          show.reflvl = TRUE, 
          digits = 3, 
          show.aic = TRUE, 
          show.ci = FALSE,
          show.icc = FALSE, 
          pred.labels = c('Intercept', 
                          'Water Potential (scaled)', 
                          'Species Effect (C. megacarpus)',
                          'Water Potential x Species Effect
                          (interaction)',
                          'LFM (%)',
                          'Water Weight (g)', 
                          'Dry Weight (g)',
                          'Site Effect (Site 1)',
                          'Excess Water (scaled)',
                          'Excess Water x Species Effect
                          (interaction)'),
         dv.labels = c("mpa_spp", 
                       "m11", 
                       "m7", 
                       "m7",
                       "m10",
                       "m3", 
                       "m6", 
                       "m6.5", 
                       "m12",
                       "m15"),
         string.pred = "Coeffcient", 
         title = "models with no collinearity",
  string.p = "P-Value", 
  p.style = "stars",
  file = here('figures','MEM.figures','site.yearmonth.samplewt')) # saving .html into figures folder
```

```{r}
#tab_model(top_model, m7, m10, m3, m6) 
tab_model(max_model, m7, mpa_mod_fig.mpa, max_model_nolfm, max_model_nompa, m12.5, m13,
          show.reflvl = TRUE, 
          digits = 3, 
          show.aic = TRUE, 
          show.ci = FALSE,
          show.icc = FALSE, 
          pred.labels = c('Intercept', 
                          'Species Effect (C. megacarpus)',
                          'Dry Weight (g)',
                          'Water Weight (g)',
                          'Water Potential (scaled)',
                          'LFM (%)',
                          'Excess Water (scaled)',
                          'Site Effect (Site 1)',
                          'Water Potential (MPa, unscaled)'),
         dv.labels = c("max_model", 
                       "m7**", 
                       "mpa_spp**", 
                       "max_model_nolfm", 
                      "max_model_nompa", 
                      "m12.5", 
                      "m13"),
         string.pred = "Coeffcient", 
         title = "models WITH collinearity (except **)",
  string.p = "P-Value", 
  p.style = "stars")
```

The intra-class correlation coefficient (ICC) is a related statistic that quantifies the proportion of variance explained by a grouping (random) factor in multilevel/hierarchical data.

## 2.0.3 Other figures
### Predicted vs. Observed PC1
```{r}
predicted_pc1_df <- data.frame(predicted_pc1 = predict(mpa_mod_fig.mpa), observed_pc1 = scaled.mem.data.alldates.noNAs$PC1)

ggplot(data = predicted_pc1_df, aes(x = predicted_pc1, y = observed_pc1)) +
  geom_point(color = "lightblue", size = .7) +
  geom_abline() +
  labs(title = "Predicted vs. Observed PC1", subtitle = "All dates, mpa model",
       x = "Predicted PC1", y = "Observed PC1") +
  theme(panel.background = element_rect(fill='white', colour='black'), # Make background white and border black
          panel.grid.major = element_blank(),  # Hide major gridlines
          panel.grid.minor = element_blank()) +
  ylim(-3, 3) +
  xlim(-3, 3) +
  coord_equal()
```

```{r}
effects_mpa_fig.mpa <- effects::effect(term= "mpa.unscaled", 
                                   mod= mpa_mod_fig.mpa, 
                                   se = TRUE, 
                                   confidence.level=.95
                                   )
# Save the effects values as a df:
x_mpa_fig.mpa <- as.data.frame(effects_mpa_fig.mpa)
```

### MPa vs. PC1 Visualization

**Figure X.** Model estimate of flammability (principle component 1) from water potential. 
##### Species separation
```{r, message=FALSE, results='hide',fig.height= 4, fig.length = 8}
# Separating by Species
p <- ggplot() + 
  geom_point(data= scaled.mem.data.alldates.noNAs, aes(mpa.unscaled, PC1, color = Species), size = 1) + 
  geom_line(data=x_mpa_fig.mpa, aes(x=mpa.unscaled, y=fit), color="#db8872") +
  geom_ribbon(data= x_mpa_fig.mpa, aes(x=mpa.unscaled, ymin=lower, ymax=upper), alpha= 0.3, fill="#db8872") +
  labs(x="Water Potential (-Mpa)", y="Principle Component 1") +
  scale_color_manual(values = c("#b08ba5", "#ffb178"))
  th 

p +
  annotate("segment",x = -11, xend = -11, y = -3, yend = 3,
           colour = "#db8872", size = 1, arrow = arrow(ends = "last", angle = 25, type = "closed", length = unit(.2,"cm"))) +
  annotate("text", x= -11, y= 4, label= "More", size = 3) +
  annotate("text", x= -11, y= 3.5, label= "Flammable", size = 3)+
annotate("text", x= -11, y= -3.5, label= "Less", size = 3) +
  annotate("text", x= -11, y= -4, label= "Flammable", size = 3)  +
  scale_x_continuous(limits = c(-11.5, 0), breaks= c(-10, -8, -6, -4, -2, 0)) + 
  theme(
    legend.position = c(.99, .99),
    legend.justification = c("right", "top"),
    legend.box.just = "right",
    legend.margin = margin(4,4,4,4), 
    panel.background = element_rect(fill='white', colour='black'), # Make background white and border black
          panel.grid.major = element_blank(),  # Hide major gridlines
          panel.grid.minor = element_blank()
    ) 

ggsave(here("figures", 'extra-figures', "PC1.vs.mpa.jpg"))
```

##### Water status separation
```{r, message=FALSE, results='hide',fig.height= 4, fig.length = 8}
# Separating by Species
p <- ggplot() + 
  geom_point(data= scaled.mem.data.alldates.noNAs, aes(mpa.unscaled, PC1, color = hydration), size = .5) + 
  geom_line(data=x_mpa_fig.mpa, aes(x=mpa.unscaled, y=fit), color="#db8872") +
  geom_ribbon(data= x_mpa_fig.mpa, aes(x=mpa.unscaled, ymin=lower, ymax=upper), alpha= 0.3, fill="#db8872") +
  labs(x="Water Potential (-Mpa)",
       y="Principle Component 1", 
       color = "Timing")+
  scale_color_manual(values=met.brewer("Morgenstern", 7)) +
  th 
p +
  annotate("segment", x = -11, xend = -11, y = -3, yend = 3,
           colour = "#db8872", size = .7, arrow = arrow(ends = "last", angle = 25, type = "closed", length = unit(.2,"cm"))) +
  annotate("text", x= -11, y= 4, label= "More", size = 3) +
  annotate("text", x= -11, y= 3.5, label= "Flammable", size = 3)+
annotate("text", x= -11, y= -3.5, label= "Less", size = 3) +
  annotate("text", x= -11, y= -4, label= "Flammable", size = 3)  +
  scale_x_continuous(limits = c(-11.5, 0), breaks= c(-10, -8, -6, -4, -2, 0)) + 
  theme(
    legend.position = c(.99, .99),
    legend.justification = c("right", "top"),
    legend.box.just = "right",
    legend.margin = margin(4,4,4,4), 
    panel.background = element_rect(fill='white', colour='black'), # Make background white and border black
          panel.grid.major = element_blank(),  # Hide major gridlines
          panel.grid.minor = element_blank()) 
```


#------------------------------------
# 2.1. Predicting raw metrics (part 2)
## Time to Ignition
Models
```{r}
m1 <- lmer(tti ~ mpa.scaled + spp + site + year.month + (1 | individual), REML = F,  data = scaled.mem.data.alldates.noNAs)

m2 <- lmer(tti ~ lfm + spp + site + year.month + (1 | individual), REML = F,  data = scaled.mem.data.alldates.noNAs)

m3 <- lmer(tti ~ ww.flam.sample + spp + site + year.month + (1 | individual), REML = F,  data = scaled.mem.data.alldates.noNAs)

m4 <- lmer(tti ~ dw.flam.sample + spp + site + year.month + (1 | individual), REML = F,  data = scaled.mem.data.alldates.noNAs)

m5 <- lmer(tti ~ mpa.scaled + dw.flam.sample + sample.wt + spp + site + year.month + (1 | individual), REML = F,  data = scaled.mem.data.alldates.noNAs)

# Adding in MPa + WW and MPa + LFM despite their moderate collinearity to see how they affect model selection
m6 <- lmer(tti ~ mpa.scaled + ww.flam.sample + spp + site + year.month + (1 | individual), REML = F,  data = scaled.mem.data.alldates.noNAs)

m7 <- lmer(tti ~ mpa.scaled + lfm + sample.wt + spp + site + year.month + (1 | individual), REML = F,  data = scaled.mem.data.alldates.noNAs)
```

Table
```{r}
tab_model(m5, m1, m3, m2, m4, m6, m7,
          show.reflvl = FALSE, 
          digits = 3, 
          show.aic = TRUE, 
          show.ci = FALSE,
          show.icc = FALSE, 
          pred.labels = c(),
         dv.labels = c("Model 5", "Model 1", "Model 3", "Model 2", "Model 4", "Model 6", "Model 7"),
         string.pred = "Coeffcient", 
         title = "models with no collinearity",
  string.p = "P-Value", 
  p.style = "stars",
  file = here('figures','MEM.figures','mem.TTI.NEW.html')) # saving .html into figures folder
```

## Flame Height
Models
```{r}
m1 <- lmer(fh ~ mpa.scaled + spp + site + year.month + (1 | individual), REML = F,  data = scaled.mem.data.alldates.noNAs)

m2 <- lmer(fh ~ lfm + spp + site + year.month + (1 | individual), REML = F,  data = scaled.mem.data.alldates.noNAs)

m3 <- lmer(fh ~ ww.flam.sample + spp + site + year.month + (1 | individual), REML = F,  data = scaled.mem.data.alldates.noNAs)

m4 <- lmer(fh ~ dw.flam.sample + spp + site + year.month + (1 | individual), REML = F,  data = scaled.mem.data.alldates.noNAs)

m5 <- lmer(fh ~ mpa.scaled + dw.flam.sample + sample.wt + spp + site + year.month + (1 | individual), REML = F,  data = scaled.mem.data.alldates.noNAs)

# Adding in MPa + WW and MPa + LFM despite their moderate collinearity to see how they affect model selection
m6 <- lmer(fh ~ mpa.scaled + ww.flam.sample + spp + site + year.month + (1 | individual), REML = F,  data = scaled.mem.data.alldates.noNAs)

m7 <- lmer(fh ~ mpa.scaled + lfm + sample.wt + spp + site + year.month + (1 | individual), REML = F,  data = scaled.mem.data.alldates.noNAs)
```

Table
```{r}
tab_model(m1, m4, m2, m5, m3, m6, m7,
          show.reflvl = FALSE, 
          digits = 3, 
          show.aic = TRUE, 
          show.ci = FALSE,
          show.icc = FALSE, 
          pred.labels = c(),
         dv.labels = c("Model 1", "Model 4", "Model 2", "Model 5", "Model 3", "Model 6", "Model 7"),
         string.pred = "Coeffcient", 
         title = "models with no collinearity",
  string.p = "P-Value", 
  p.style = "stars",
  file = here('figures','MEM.figures','mem.FH.NEW.html')) # saving .html into figures folder
```

## Glow Duration
Models
```{r}
m1 <- lmer(gd ~ mpa.scaled + spp + site + year.month + (1 | individual), REML = F,  data = scaled.mem.data.alldates.noNAs)

m2 <- lmer(gd ~ lfm + spp + site + year.month + (1 | individual), REML = F,  data = scaled.mem.data.alldates.noNAs)

m3 <- lmer(gd ~ ww.flam.sample + spp + site + year.month + (1 | individual), REML = F,  data = scaled.mem.data.alldates.noNAs)

m4 <- lmer(gd ~ dw.flam.sample + spp + site + year.month + (1 | individual), REML = F,  data = scaled.mem.data.alldates.noNAs)

m5 <- lmer(gd ~ mpa.scaled + dw.flam.sample + sample.wt + spp + site + year.month + (1 | individual), REML = F,  data = scaled.mem.data.alldates.noNAs)

# Adding in MPa + WW and MPa + LFM despite their moderate collinearity to see how they affect model selection
m6 <- lmer(gd ~ mpa.scaled + ww.flam.sample + spp + site + year.month + (1 | individual), REML = F,  data = scaled.mem.data.alldates.noNAs)

m7 <- lmer(gd ~ mpa.scaled + lfm + sample.wt + spp + site + year.month + (1 | individual), REML = F,  data = scaled.mem.data.alldates.noNAs)
```

Table
```{r}
tab_model(m5, m4, m2, m1, m3, m6, m7,
          show.reflvl = FALSE, 
          digits = 3, 
          show.aic = TRUE, 
          show.ci = FALSE,
          show.icc = FALSE, 
          pred.labels = c(),
         dv.labels = c("Model 5", "Model 4", "Model 2", "Model 1", "Model 3", "Model 6", "Model 7"),
         string.pred = "Coeffcient", 
         title = "models with no collinearity",
  string.p = "P-Value", 
  p.style = "stars",
  file = here('figures','MEM.figures','mem.GD.html')) # saving .html into figures folder
```

## Flame Duration
Models
```{r}
m1 <- lmer(fd ~ mpa.scaled + spp + site + year.month + (1 | individual), REML = F,  data = scaled.mem.data.alldates.noNAs)

m2 <- lmer(fd ~ lfm + spp + site + year.month + (1 | individual), REML = F,  data = scaled.mem.data.alldates.noNAs)

m3 <- lmer(fd ~ ww.flam.sample + spp + site + year.month + (1 | individual), REML = F,  data = scaled.mem.data.alldates.noNAs)

m4 <- lmer(fd ~ dw.flam.sample + spp + site + year.month + (1 | individual), REML = F,  data = scaled.mem.data.alldates.noNAs)

m5 <- lmer(fd ~ mpa.scaled + dw.flam.sample + sample.wt + spp + site + year.month + (1 | individual), REML = F,  data = scaled.mem.data.alldates.noNAs)

# Adding in MPa + WW and MPa + LFM despite their moderate collinearity to see how they affect model selection
m6 <- lmer(fd ~ mpa.scaled + ww.flam.sample + spp + site + year.month + (1 | individual), REML = F,  data = scaled.mem.data.alldates.noNAs)

m7 <- lmer(fd ~ mpa.scaled + lfm + sample.wt + spp + site + year.month + (1 | individual), REML = F,  data = scaled.mem.data.alldates.noNAs)
```

Table
```{r}
tab_model(m1, m4, m2, m5, m3, m6, m7,
          show.reflvl = FALSE, 
          digits = 3, 
          show.aic = TRUE, 
          show.ci = FALSE,
          show.icc = FALSE, 
          pred.labels = c(),
         dv.labels = c("Model 1", "Model 4", "Model 2", "Model 5", "Model 3", "Model 6", "Model 7"),
         string.pred = "Coeffcient", 
         title = "models with no collinearity",
  string.p = "P-Value", 
  p.style = "stars",
  file = here('figures', 'MEM.figures','mem.FD.html')) # saving .html into figures folder
```

## Max. Temp.
Models
```{r}
m1 <- lmer(temp.max ~ mpa.scaled + spp + site + year.month + (1 | individual), REML = F,  data = scaled.mem.data.alldates.noNAs)

m2 <- lmer(temp.max ~ lfm + spp + site + year.month + (1 | individual), REML = F,  data = scaled.mem.data.alldates.noNAs)

m3 <- lmer(temp.max ~ ww.flam.sample + spp + site + year.month + (1 | individual), REML = F,  data = scaled.mem.data.alldates.noNAs)

m4 <- lmer(temp.max ~ dw.flam.sample + spp + site + year.month + (1 | individual), REML = F,  data = scaled.mem.data.alldates.noNAs)

m5 <- lmer(temp.max ~ mpa.scaled + dw.flam.sample + sample.wt + spp + site + year.month + (1 | individual), REML = F,  data = scaled.mem.data.alldates.noNAs)

# Adding in MPa + WW and MPa + LFM despite their moderate collinearity to see how they affect model selection
m6 <- lmer(temp.max ~ mpa.scaled + ww.flam.sample + spp + site + year.month + (1 | individual), REML = F,  data = scaled.mem.data.alldates.noNAs)

m7 <- lmer(temp.max ~ mpa.scaled + lfm + sample.wt + spp + site + year.month + (1 | individual), REML = F,  data = scaled.mem.data.alldates.noNAs)
```

Table
```{r}
tab_model(m1, m2, m3, m4, m5, m6, m7,
          show.reflvl = FALSE, 
          digits = 3, 
          show.aic = TRUE, 
          show.ci = FALSE,
          show.icc = FALSE, 
          pred.labels = c(),
         dv.labels = c("Model 1", "Model 2", "Model 3", "Model 4", "Model 5", "Model 6", "Model 7"),
         string.pred = "Coeffcient", 
         title = "models with no collinearity",
  string.p = "P-Value", 
  p.style = "stars",
  file = here('figures', 'MEM.figures','mem.MAXTEMP.html')) # saving .html into figures folder
```

## Post-Flame Glow
Models
```{r}
m1 <- lmer(pfg ~ mpa.scaled + spp + site + year.month + (1 | individual), REML = F,  data = scaled.mem.data.alldates.noNAs)

m2 <- lmer(pfg ~ lfm + spp + site + year.month + (1 | individual), REML = F,  data = scaled.mem.data.alldates.noNAs)

m3 <- lmer(pfg ~ ww.flam.sample + spp + site + year.month + (1 | individual), REML = F,  data = scaled.mem.data.alldates.noNAs)

m4 <- lmer(pfg ~ dw.flam.sample + spp + site + year.month + (1 | individual), REML = F,  data = scaled.mem.data.alldates.noNAs)

m5 <- lmer(pfg ~ mpa.scaled + dw.flam.sample + sample.wt + spp + site + year.month + (1 | individual), REML = F,  data = scaled.mem.data.alldates.noNAs)

# Adding in MPa + WW and MPa + LFM despite their moderate collinearity to see how they affect model selection
m6 <- lmer(pfg ~ mpa.scaled + ww.flam.sample + spp + site + year.month + (1 | individual), REML = F,  data = scaled.mem.data.alldates.noNAs)

m7 <- lmer(pfg ~ mpa.scaled + lfm + sample.wt + spp + site + year.month + (1 | individual), REML = F,  data = scaled.mem.data.alldates.noNAs)
```

Table
```{r}
tab_model(m1, m2, m3, m4, m5, m6, m7,
          show.reflvl = FALSE, 
          digits = 3, 
          show.aic = TRUE, 
          show.ci = FALSE,
          show.icc = FALSE, 
          pred.labels = c(),
         dv.labels = c("Model 1", "Model 2", "Model 3", "Model 4", "Model 5", "Model 6", "Model 7"),
         string.pred = "Coeffcient", 
         title = "models with no collinearity",
  string.p = "P-Value", 
  p.style = "stars",
  file = here('figures', 'MEM.figures','mem.PFG.html')) # saving .html into figures folder
```

## Time to First Glow
Models
```{r}
m1 <- lmer(ttfg ~ mpa.scaled + spp + site + year.month + (1 | individual), REML = F,  data = scaled.mem.data.alldates.noNAs)

m2 <- lmer(ttfg ~ lfm + spp + site + year.month + (1 | individual), REML = F,  data = scaled.mem.data.alldates.noNAs)

m3 <- lmer(ttfg ~ ww.flam.sample + spp + site + year.month + (1 | individual), REML = F,  data = scaled.mem.data.alldates.noNAs)

m4 <- lmer(ttfg ~ dw.flam.sample + spp + site + year.month + (1 | individual), REML = F,  data = scaled.mem.data.alldates.noNAs)

m5 <- lmer(ttfg ~ mpa.scaled + dw.flam.sample + sample.wt + spp + site + year.month + (1 | individual), REML = F,  data = scaled.mem.data.alldates.noNAs)

# Adding in MPa + WW and MPa + LFM despite their moderate collinearity to see how they affect model selection
m6 <- lmer(ttfg ~ mpa.scaled + ww.flam.sample + spp + site + year.month + (1 | individual), REML = F,  data = scaled.mem.data.alldates.noNAs)

m7 <- lmer(ttfg ~ mpa.scaled + lfm + sample.wt + spp + site + year.month + (1 | individual), REML = F,  data = scaled.mem.data.alldates.noNAs)
```

Table
```{r}
tab_model(m1, m2, m3, m4, m5, m6, m7,
          show.reflvl = FALSE, 
          digits = 3, 
          show.aic = TRUE, 
          show.ci = FALSE,
          show.icc = FALSE, 
          pred.labels = c(),
         dv.labels = c("Model 1", "Model 2", "Model 3", "Model 4", "Model 5", "Model 6", "Model 7"),
         string.pred = "Coeffcient", 
         title = "models with no collinearity",
  string.p = "P-Value", 
  p.style = "stars",
  file = here('figures', 'MEM.figures','mem.TTFG.html')) # saving .html into figures folder
```

## Glow to Ignition
Models
```{r}
m1 <- lmer(gti ~ mpa.scaled + spp + site + year.month + (1 | individual), REML = F,  data = scaled.mem.data.alldates.noNAs)

m2 <- lmer(gti ~ lfm + spp + site + year.month + (1 | individual), REML = F,  data = scaled.mem.data.alldates.noNAs)

m3 <- lmer(gti ~ ww.flam.sample + spp + site + year.month + (1 | individual), REML = F,  data = scaled.mem.data.alldates.noNAs)

m4 <- lmer(gti ~ dw.flam.sample + spp + site + year.month + (1 | individual), REML = F,  data = scaled.mem.data.alldates.noNAs)

m5 <- lmer(gti ~ mpa.scaled + dw.flam.sample + sample.wt + spp + site + year.month + (1 | individual), REML = F,  data = scaled.mem.data.alldates.noNAs)

# Adding in MPa + WW and MPa + LFM despite their moderate collinearity to see how they affect model selection
m6 <- lmer(gti ~ mpa.scaled + ww.flam.sample + spp + site + year.month + (1 | individual), REML = F,  data = scaled.mem.data.alldates.noNAs)

m7 <- lmer(gti ~ mpa.scaled + lfm + sample.wt + spp + site + year.month + (1 | individual), REML = F,  data = scaled.mem.data.alldates.noNAs)
```

Table
```{r}
tab_model(m1, m2, m3, m4, m5, m6, m7,
          show.reflvl = FALSE, 
          digits = 3, 
          show.aic = TRUE,
          show.ci = FALSE,
          show.icc = FALSE, 
          pred.labels = c(),
         dv.labels = c("Model 1", "Model 2", "Model 3", "Model 4", "Model 5", "Model 6", "Model 7"),
         string.pred = "Coeffcient", 
         title = "models with no collinearity",
  string.p = "P-Value", 
  p.style = "stars",
  file = here('figures', 'MEM.figures','mem.GTI.html')) # saving .html into figures folder
```

## P Values
The p values in the MEM tables is incorrect, and the Kenward-Roger approximation should be used via 'lmerTest' as follows
```{r}
# TTI
m1 <- lmer(tti ~ mpa.scaled + spp + site + year.month + (1 | individual), data = scaled.mem.data.alldates.noNAs, REML = T)

m2 <- lmer(tti ~ lfm + spp + site + year.month + (1 | individual), data = scaled.mem.data.alldates.noNAs, REML = T)

m3 <- lmer(tti ~ ww.flam.sample + spp + site + year.month + (1 | individual), data = scaled.mem.data.alldates.noNAs, REML = T)

m4 <- lmer(tti ~ dw.flam.sample + spp + site + year.month + (1 | individual), data = scaled.mem.data.alldates.noNAs, REML = T)

m5 <- lmer(tti ~ mpa.scaled + dw.flam.sample + sample.wt + spp + site + year.month + (1 | individual), data = scaled.mem.data.alldates.noNAs, REML = T)

summary(m5, ddf = 'Kenward-Roger')
summary(m1, ddf = 'Kenward-Roger')
summary(m3, ddf = 'Kenward-Roger')
summary(m2, ddf = 'Kenward-Roger')
summary(m4, ddf = 'Kenward-Roger')

# FD
m1 <- lmer(fd ~ mpa.scaled + spp + site + year.month + (1 | individual), data = scaled.mem.data.alldates.noNAs, REML = T)

m2 <- lmer(fd ~ lfm + spp + site + year.month + (1 | individual), data = scaled.mem.data.alldates.noNAs, REML = T)

m3 <- lmer(fd ~ ww.flam.sample + spp + site + year.month + (1 | individual), data = scaled.mem.data.alldates.noNAs, REML = T)

m4 <- lmer(fd ~ dw.flam.sample + spp + site + year.month + (1 | individual), data = scaled.mem.data.alldates.noNAs, REML = T)

m5 <- lmer(fd ~ mpa.scaled + dw.flam.sample + sample.wt + spp + site + year.month + (1 | individual), data = scaled.mem.data.alldates.noNAs, REML = T)

summary(m1, ddf = 'Kenward-Roger')
summary(m4, ddf = 'Kenward-Roger')
summary(m2, ddf = 'Kenward-Roger')
summary(m5, ddf = 'Kenward-Roger')
summary(m3, ddf = 'Kenward-Roger')

#GD
m1 <- lmer(gd ~ mpa.scaled + spp + site + year.month + (1 | individual), data = scaled.mem.data.alldates.noNAs, REML = T)

m2 <- lmer(gd ~ lfm + spp + site + year.month + (1 | individual), data = scaled.mem.data.alldates.noNAs, REML = T)

m3 <- lmer(gd ~ ww.flam.sample + spp + site + year.month + (1 | individual), data = scaled.mem.data.alldates.noNAs, REML = T)

m4 <- lmer(gd ~ dw.flam.sample + spp + site + year.month + (1 | individual), data = scaled.mem.data.alldates.noNAs, REML = T)

m5 <- lmer(gd ~ mpa.scaled + dw.flam.sample + sample.wt + spp + site + year.month + (1 | individual), data = scaled.mem.data.alldates.noNAs, REML = T)

summary(m5, ddf = 'Kenward-Roger')
summary(m4, ddf = 'Kenward-Roger')
summary(m2, ddf = 'Kenward-Roger')
summary(m1, ddf = 'Kenward-Roger')
summary(m3, ddf = 'Kenward-Roger')
```

# MAIN TABLE

All data below is from above MEM tables (2.1; for TTI, FD, GD) and from MAIN MODEL SELECTION (for PC1)

## BIC, AIC included
```{r}
dep.var <- c(
  "PC1",
  " ",
  " ",
  " ",
  " ",
  " ",
  "Time to Ignition (s)",
  " ",
  " ",
  " ",
  " ",
  " ",
  "Flame Duration (s)",
  " ",
  " ",
  " ",
  " ",
  " ",
  "Glow Duration (s)",
  " ",
  " ",
  " ",
  " ",
  " "
)
fix.fx <- c(
  " ", #PC1
  "Water Potential",
  "Composite Model",
  "Wet Weight",
  "Dry Weight",
  "Live Fuel Moisture",
  " ", #TTI
  "Composite Model",
  "Water Potential",
  "Wet Weight",
  "Live Fuel Moisture",
  "Dry Weight",
  " ", #FD
  "Water Potential",
  "Dry Weight",
  "Live Fuel Moisture",
  "Composite Model",
  "Wet Weight",
  " ", #GD
  "Composite Model",
  "Dry Weight",
  "Live Fuel Moisture",
  "Water Potential",
  "Wet Weight"
)
est1 <- c(
  " ",
  '-0.142***',
  '-0.134***',
  '-0.038***',
  '0.399***',
  '-0.296***',
  " ",
  '4.709***',
  '4.889***',
  '4.260***',
  '3.126***',
  '-3.005***',
  " ",
  '-0.574**',
  '0.612**',
  '-0.530*',
  '-0.491',
  '-0.426',
  " ",
  '-1.280',
  '3.741**',
  '-1.834',
  '-2.029',
  '-1.122'
)
aic.values <- c(
  " ",
  '1668.106',
  '1672.325',
  '1672.357',
  '1681.577',
  '1687.590',
  " ",
  '3316.212',
  '3318.910',
  '3363.524',
  '3385.911',
  '3390.382',
  " ",
  '2562.408',
  '2563.221',
  '2564.659',
  '2566.012',
  '2567.067',
  " ",
  '4141.659',
  '4145.782',
  '4151.125',
  '4150.366',
  '4152.267'
)
r2.values <- c(
  " ",
  '0.229 / 0.289',
  '0.228 / 0.290',
  '0.228 / 0.289',
  '0.201 / 0.268',
  '0.192 / 0.258',
  " ",
  '0.343 / 0.404',
  '0.337 / 0.403',
  '0.271 / 0.342',
  '0.236 / 0.312',
  '0.229 / 0.302',
  " ",
  '0.059 / 0.136',
  '0.058 / 0.131',
  '0.055 / 0.128',
  '0.064 / 0.140',
  '0.050 / 0.123',
  " ",
  '0.229 / 0.288',
  '0.225 / 0.286',
  '0.215 / 0.286',
  '0.215 / 0.290',
  '0.212 / 0.287'
)
d.aic <- c(
  " ",
  '0',
  '4.219',
  '4.251',
  '13.471',
  '19.484',
  " ",
  '0',
  '2.698',
  '47.312',
  '69.699',
  '74.17',
  " ",
  '0',
  '0.813',
  '2.251',
  '3.604',
  '4.659',
  " ",
  '0',
  '4.123',
  '9.466',
  '8.707',
  '10.608'
)
deltaAIC <- c('\U0394 AIC')

mem.table.df <- data.frame(dep.var, fix.fx, est1, d.aic, aic.values, r2.values)
mem.table.df %>%
  mutate(dep.var = cell_spec(dep.var, 'html', color = 'black', bold = T), fix.fx = cell_spec(fix.fx, 'html', color = 'black'), est1 = cell_spec(est1, 'html', color = 'black'), d.aic = cell_spec(d.aic, 'html', color = 'black'), aic.values = cell_spec(aic.values, 'html', color = 'black'), r2.values = cell_spec(r2.values, 'html', color = 'black')) %>% 
  kable(format = 'html', escape = F, col.names = c('Dependent Variable', 'Model Fixed Effects', 'Water Estimate','\U0394 AIC', 'AIC', expression('Marginal R2/Conditional R2'))) %>% 
  kable_styling(bootstrap_options = c('hover', 'bordered', 'condensed'), fixed_thead = T) %>% 
  row_spec(c(1, 7, 13, 19), background = "#D3D3D3", bold = T, font_size = 16) %>% 
  row_spec(c(2, 8, 14, 15, 20), bold = T) %>% 
  save_kable(here('figures', 'extra-figures', 'expanded_MEM_kable.html'))
```

## Kable
```{r}
mem.table2.df <- data.frame(dep.var, fix.fx, est1, d.aic, r2.values)
mem.table2.df %>%
  mutate(dep.var = cell_spec(dep.var, 'html', color = 'black', bold = T), fix.fx = cell_spec(fix.fx, 'html', color = 'black'), est1 = cell_spec(est1, 'html', color = 'black'), d.aic = cell_spec(d.aic, 'html', color = 'black'), r2.values = cell_spec(r2.values, 'html', color = 'black')) %>% 
  kable(format = 'html', escape = F, col.names = c('Dependent Variable', 'Model Fixed Effects', 'Water Estimate','\U0394 AIC', expression('Marginal R2/Conditional R2'))) %>% 
  kable_styling(bootstrap_options = c('hover', 'bordered', 'condensed'), fixed_thead = T) %>% 
  row_spec(c(1, 7, 13, 19), background = "#D3D3D3", bold = T, font_size = 16) %>% 
  row_spec(c(2, 8, 14, 15, 20), bold = T) %>% 
  save_kable(here('figures', 'main-figures', 'Table2_MEM_kable.html'))
```

## Shortened Table (for presentation)
```{r}
dep.var.s <- c("Ignitability-Combustibility Proxy (PC1)", "Time to Ignition (s)", "Flame Duration (s)", " ", " ", "Glow Duration (s)", "Flame Height (cm)", " ")
top.models <- c("Water Potential", "Dry Weight, Sample Weight & Water Potential", "Water Potential", "Dry Weight", "LFM", "Dry Weight, Sample Weight & Water Potential", "Dry Weight, Sample Weight & Water Potential", "Water Potential")
significance <- c("p<0.001", "p<0.001", "p<0.01", "p<0.01", "p<0.05", "non-significant", "p<0.01", "p<0.001")
shortened.table.df <- data.frame(dep.var.s, top.models, significance)

shortened.table.df %>%
  mutate(dep.var.s = cell_spec(dep.var.s, 'html', color = 'black', bold = T), top.models = cell_spec(top.models, 'html', color = 'black'), significance = cell_spec(significance, 'html', color = 'black')) %>% 
  kable(format = 'html', escape = F, col.names = c('Dependent Variable', 'Top Model(s)', 'Significance')) %>% 
  kable_styling(bootstrap_options = c('hover', 'bordered', 'condensed'), fixed_thead = T) %>% 
  save_kable(here('figures', 'extra-figures', 'shortened_MEM_kable.html'))
```

# ----------------------------------
# Residuals of flam ~ sample.wt, flam ~ dry weight
Based on Lee's suggestion, doing a variety of models and checking their residuals
```{r}
m1 <- lm(PC1 ~ sample.wt, data = scaled.mem.data.alldates.noNAs)
res.m1 <- resid(m1)

plot(fitted(m1), res.m1)
abline(0,0)

qqnorm(res.m1)
```

```{r}
m2 <- lm(PC1 ~ dw.flam.sample, data = scaled.mem.data.alldates.noNAs)
res.m2 <- resid(m2)

plot(fitted(m2), res.m2)
abline(0,0)

qqnorm(res.m2)
```

```{r}
m3 <- lm(PC1 ~ mpa + sample.wt + dw.flam.sample, data = scaled.mem.data.alldates.noNAs)
res.m3 <- resid(m3)

plot(fitted(m3), res.m3)
abline(0,0)

qqnorm(res.m3)
```

Note the effects of site, species, sampling date on residuals
```{r}
m4 <- lm(PC1 ~ mpa + site + spp + year.month, data = scaled.mem.data.alldates.noNAs)
res.m4 <- resid(m4)

plot(fitted(m4), res.m4)
abline(0,0)

qqnorm(res.m4)
```



# -----------------------------------
# Excess dry weight metric
In an attempt to capture the dry weight (effectively, woodiness), wet weight and total sample weight of each datapoint, trying to transform dry weight metric so that it is not colinear with sample weight

```{r}
scaled.mem.data.alldates.noNAs <- scaled.mem.data.alldates.noNAs %>% 
  mutate(excess.dw.centered = dw.flam.sample.unscaled - mean(dw.flam.sample.unscaled)) %>% 
  mutate(excess.dw.percent = (dw.flam.sample.unscaled - mean(dw.flam.sample.unscaled))/mean(dw.flam.sample.unscaled))
```

```{r}
m1 <- lmer(PC1 ~ spp + excess.dw.centered + excess_water.unscaled + sample.wt + site + year.month + (1 | individual), REML = F,  data = scaled.mem.data.alldates.noNAs)

performance::multicollinearity(m1)

# High correlation
```

```{r}
m2 <- lmer(PC1 ~ spp + excess.dw.centered + excess_water.unscaled + site + year.month + (1 | individual), REML = F,  data = scaled.mem.data.alldates.noNAs)

performance::multicollinearity(m2)

# Moderate correlation
```

```{r}
m3 <- lmer(PC1 ~ spp + excess.dw.percent + excess_water.unscaled + sample.wt + site + year.month + (1 | individual), REML = F,  data = scaled.mem.data.alldates.noNAs)

performance::multicollinearity(m3)

# High correlation
```

```{r}
m4 <- lmer(PC1 ~ spp + excess.dw.percent + excess_water.unscaled + site + year.month + (1 | individual), REML = F,  data = scaled.mem.data.alldates.noNAs)

performance::multicollinearity(m4)

# Moderate correlation
```

```{r}
m5 <- lmer(PC1 ~ spp + excess.dw.centered + mpa + sample.wt + site + year.month + (1 | individual), REML = F,  data = scaled.mem.data.alldates.noNAs)

performance::multicollinearity(m5)

# LOW correlation
```

```{r}
m6 <- lmer(PC1 ~ spp + excess.dw.centered + lfm + sample.wt + site + year.month + (1 | individual), REML = F,  data = scaled.mem.data.alldates.noNAs)

performance::multicollinearity(m6)

# Moderate correlation
```

```{r}
m7 <- lmer(PC1 ~ spp + dw.flam.sample + excess_water.scaled + sample.wt + site + year.month + (1 | individual), REML = F,  data = scaled.mem.data.alldates.noNAs)

performance::multicollinearity(m7)

# VERY High correlation
```

```{r}
m8 <- lmer(PC1 ~ spp + dw.flam.sample + mpa + sample.wt + site + year.month + (1 | individual), REML = F,  data = scaled.mem.data.alldates.noNAs)

performance::multicollinearity(m8)

# LOW correlation
```


```{r}
tab_model(m1, m2, m3, m4, m5, m6, m7, m8,
          show.reflvl = TRUE, 
          digits = 3, 
          show.aic = TRUE, 
          show.ci = FALSE,
          show.icc = FALSE, 
          pred.labels = c('Intercept', 
                          'Species Effect (C. megacarpus)',
                          'Excess Dry Weight (centered)',
                          'Excess Dry Weight (unscaled)',
                          'Sample Wt. (g)',
                          'Site Effect (Site 1)',
                          'Sampling Date (Jan. 2018)',
                          'Sampling Date (Dec. 2019)',
                          'Sampling Date (Jan. 2020)',
                          'Sampling Date (Sept. 2020)',
                          'Excess Dry Weight (percentile)',
                          'Water Potential (MPa)',
                          'LFM (%)',
                          'Dry Weight (g)',
                          'Excess Water (scaled)'),
         dv.labels = c('Model 1', 'Model 2', 'Model 3', 'Model 4', 'Model 5',
                       'Model 6', 'Model 7', 'Model 8'),
         string.pred = "Coeffcient", 
         title = "models with no collinearity",
  string.p = "P-Value", 
  p.style = "stars",
  file = here('figures', 'MEM.figures','excess.dry.weight.html')) # saving .html into figures folder
```
