==> scripts/3.1_season_and_species_effects.Rmd <==
---
title: "Season and Species Effects"
author: "Indra Boving & Joe Celebrezze"
date: "12/17/2021"
output: html_document
---

This script contains code necessary to formulate figures describing differences observed between species and between time points with different antecedent precipitations. It shows both figures that are in the paper as well as supplemental or exploratory figures.

# Setup
```{r setup, include=FALSE}
knitr::opts_chunk$set(echo = TRUE)

#lots of extra here, but oh well... 
library(ggplot2)
library(kableExtra)
library(gapminder)
library(data.table)
library(purrr) #for visualizing all variables in one big plot
library(naniar) #for dealing with NAs nicely 
library(tidyverse)
library(devtools)
library(ggfortify)
library(ggpubr)
library(jtools)
library(cowplot)
library(lmerTest)
library(ggeffects)  # install the package first if you haven't already, then load it
library(GGally) #for corr plots
require(kimisc) # has the nlist function to create a named list
require(AICcmodavg) # has the aictab function
library(psych)
#devtools::install_github("strengejacke/strengejacke")
library(strengejacke)
library(sjPlot) # table functions
library(sjmisc) # sample data
library(lme4) # fitting models
library(here)
library(effects) #for plotting model effects
library(sjstats) #use for r2 functions
library(TMB)
library(glmmTMB)
library(lattice)
library(equatiomatic)
library("broom.mixed")
library(ggbiplot)
select = dplyr::select
here = here::here
library(MuMIn)
library(modelsummary)
#install_github("BlakeRMills/MetBrewer") 
#library("BlakeRMills/MetBrewer")
library(segmented)
filter = dplyr::filter
mutate = dplyr::mutate
library(nlme)
library(MetBrewer)
#install.packages("segmented")
#devtools::install_github("hohenstein/remef")
library(remef)
library(gridExtra)
```

Set global options for color palettes and themes: 

```{r}
th <- theme(panel.background = element_rect(fill='white', colour='black'), # Make background white and border black
          panel.grid.major = element_blank(),  # Hide major gridlines
          panel.grid.minor = element_blank())

th.spp.wrap <- theme(legend.position = "bottom", 
        strip.text = element_text(face = "italic", size = 12) +
  scale_color_manual(values=met.brewer("Morgenstern", 7)))

Morgenstern = list(c("#7c668c", "#b08ba5", "#dfbbc8", "#ffc680", "#ffb178", "#db8872", "#a56457"), c(7, 5, 4, 6, 3, 2, 1), colorblind=TRUE)
```

#------------------------------------
# 1. Data wrangling

Seasonal mins and maxs from 2016 - 2020:

https://www.sbbg.org/about/onsite-weather-station-live-fuel-moisture 

```{r}
adfa_min <- 55
adfa_max <- 110

ceme_min <- 55
ceme_max <- 140
```

### All dates datasets
```{r}
seg.data.subset.alldates <- read.csv(here('processed-data', 'mem.data.subset.alldates.csv'))

seg.data.subset.alldates <- seg.data.subset.alldates %>% 
  na.omit()
```

Numerical 1 or 2 for site, dry vs. wet season
```{r}
seg.data.subset.alldates <- seg.data.subset.alldates %>% 
  mutate(season = case_when(
    precip.2mo < 0.5 ~ "Dry", 
    precip.2mo > 0.5 ~ "Wet")
  ) %>% 
  mutate(site.test = case_when( 
    site == "St. Marys" ~ 1,
    site == "Painted Caves" ~ 2
  )) # Making new variable, site.test, that uses numbers to represent the site instead of character values; this allows the site to be included in mixed effects models as a fixed effect.
```

```{r, message=FALSE, warning=FALSE, include=FALSE}
seg.adfa.subset <- seg.data.subset.alldates  %>% 
  filter(spp == "ADFA") %>% 
  filter(ignition == "1")

seg.CEME.subset <- seg.data.subset.alldates  %>% 
  filter(spp == "CEME")%>% 
  filter(ignition == "1")

seg.adfa.subset.allignitions <- seg.data.subset.alldates  %>% 
  filter(spp == "ADFA")

seg.CEME.subset.allignitions <- seg.data.subset.alldates %>% 
  filter(spp == "CEME")

seg.adfa.subset.limited <- seg.adfa.subset %>% 
  filter(lfm.unscaled < adfa_max, lfm.unscaled > adfa_min)

seg.CEME.subset <- seg.data.subset.alldates %>% 
  filter(spp == "CEME")

seg.CEME.subset.limited <- seg.CEME.subset %>% 
  filter(lfm.unscaled < ceme_max, lfm.unscaled > ceme_min)
```

Scale LFM based on sample data: 
```{r}
seg.data.subset.alldates <- seg.data.subset.alldates %>%
  group_by(year.month) %>% 
  mutate(lfm.scaled.by.year = scale(lfm), 
         mpa.scaled.by.year = scale(mpa))
```

Labels for species names:
```{r}
spp.labs <- c("A. fasciculatum", "C. megacarpus")
names(spp.labs) <- c("ADFA", "CEME")
```


#------------------------------------
# 2.0 Season Effects Plot(s)
## 2.0.1 REMEF
Although this step actually occurred AFTER the model selection, in the script it is included prior to the model selection, so that I could include the REMEF values in the models so I could see what the slopes/intercepts are for the REMEF visualizations of the mixed effects models
### PC1
```{r}
precip_mod <- lmer(PC1 ~ mpa.unscaled + site + spp + season + (1 | individual), data = seg.data.subset.alldates)

summary(precip_mod)

r1_1 <- remef(precip_mod, fix = c('sppCEME', 'siteSt. Marys'),  ran = "all", keep.intercept = FALSE)
#summary(r1_1)
scaled_r1_1_df <- as.data.frame(r1_1)
#https://stats.stackexchange.com/questions/258285/how-to-do-partial-regression-plots-with-linear-mixed-effect-models/258292#258292?newreg=3b55c635f610441db980674675bc0051
```

### PC2
```{r}
precip_mod.2 <- lmer(PC2 ~ mpa.unscaled + spp + season + site + (1 | individual), data = seg.data.subset.alldates)

r1_2 <- remef(precip_mod.2, fix = c('sppCEME', 'siteSt. Marys'),  ran = "all", keep.intercept = FALSE)

scaled_r1_2_df <- as.data.frame(r1_2)
```

```{r}
precip_groups_ranis <- cbind(seg.data.subset.alldates, scaled_r1_1_df, scaled_r1_2_df)

precip_groups_ranis.adfa <- precip_groups_ranis %>% 
  filter(spp == 'ADFA')
precip_groups_ranis.ceme <- precip_groups_ranis %>% 
  filter(spp == 'CEME')
```

### LFM, DW
```{r}
mpalfm_mod <- lmer(lfm ~ mpa + spp + season + site + (1 | individual), data = seg.data.subset.alldates)

summary(mpalfm_mod)

r1_lfm <- remef(mpalfm_mod, fix = c('sppCEME', 'siteSt. Marys'),  ran = "all", keep.intercept = FALSE)
scaled_r1_lfm_df <- as.data.frame(r1_lfm)
###
mpadw_mod <- lmer(dw.flam.sample ~ mpa + spp + season + site + (1 | individual), data = seg.data.subset.alldates)

summary(mpadw_mod)

r1_dw <- remef(mpadw_mod , fix = c('sppCEME', 'siteSt. Marys'),  ran = "all", keep.intercept = FALSE)
scaled_r1_dw_df <- as.data.frame(r1_dw)
```

```{r}
seg.data.subset.alldates <- seg.data.subset.alldates %>%
  mutate(season = case_when(
    timing %in% c("Dec. 2019", "Dec. 2016", "Jan. 2018", "Jan. 2020") ~ "Wet", 
    timing == "Sept. 2020" ~ "Dry"
  ))

mpa_rel_df <- cbind(scaled_r1_dw_df, scaled_r1_lfm_df, seg.data.subset.alldates)

mpa_rel_df.adfa <- mpa_rel_df %>% 
  filter(spp == 'ADFA')
mpa_rel_df.ceme <- mpa_rel_df %>% 
  filter(spp == 'CEME')
```

## 2.0.2 Model Selection
### Effect of Season - All Flam Metrics
#### ADFA
```{r}
pc1.m <- lmer(PC1 ~ mpa.unscaled + site.test + season + (1 | individual), data = seg.adfa.subset, REML = T)

pc1.x <- lmer(PC1 ~ mpa.unscaled + site.test + (1 | individual), data = seg.adfa.subset, REML = T)

pc2.m <- lmer(PC2 ~ mpa.unscaled + site.test + season + (1 | individual), data = seg.adfa.subset, REML = T)

pc2.x <- lmer(PC2 ~ mpa.unscaled + site.test + (1 | individual), data = seg.adfa.subset, REML = T)

tti.m <- lmer(tti ~ mpa.unscaled + site.test + season + (1 | individual), data = seg.adfa.subset, REML = T)

tti.x <- lmer(tti ~ mpa.unscaled + site.test + (1 | individual), data = seg.adfa.subset, REML = T)

fd.m <- lmer(fd ~ mpa.unscaled + site.test + season + (1 | individual), data = seg.adfa.subset, REML = T)

fd.x <- lmer(fd ~ mpa.unscaled + site.test + (1 | individual), data = seg.adfa.subset, REML = T)

gd.m <- lmer(gd ~ mpa.unscaled + site.test + season + (1 | individual), data = seg.adfa.subset, REML = T)

gd.x <- lmer(gd ~ mpa.unscaled + site.test + (1 | individual), data = seg.adfa.subset, REML = T)

fh.m <- lmer(fh ~ mpa.unscaled + site.test + season + (1 | individual), data = seg.adfa.subset, REML = T)

fh.x <- lmer(fh ~ mpa.unscaled + site.test + (1 | individual), data = seg.adfa.subset, REML = T)

gti.m <- lmer(gti ~ mpa.unscaled + site.test + season + (1 | individual), data = seg.adfa.subset, REML = T)

gti.x <- lmer(gti ~ mpa.unscaled + site.test + (1 | individual), data = seg.adfa.subset, REML = T)

ttfg.m <- lmer(ttfg ~ mpa.unscaled + site.test + season + (1 | individual), data = seg.adfa.subset, REML = T)

ttfg.x <- lmer(ttfg ~ mpa.unscaled + site.test + (1 | individual), data = seg.adfa.subset, REML = T)

pfg.m <- lmer(pfg ~ mpa.unscaled + site.test + season + (1 | individual), data = seg.adfa.subset, REML = T)

pfg.x <- lmer(pfg ~ mpa.unscaled + site.test + (1 | individual), data = seg.adfa.subset, REML = T)

temp.m <- lmer(temp.max ~ mpa.unscaled + site.test + season + (1 | individual), data = seg.adfa.subset, REML = T)

temp.x <- lmer(temp.max ~ mpa.unscaled + site.test + (1 | individual), data = seg.adfa.subset, REML = T)

tab_model(pc1.m, pc1.x, pc2.m, pc2.x, tti.m, tti.x, fd.m, fd.x, gd.m, gd.x, fh.m, fh.x, pfg.m, pfg.x, ttfg.m, ttfg.x, temp.m, temp.x, gti.m, gti.x,
          show.reflvl = TRUE, 
          digits = 3, 
          show.aic = TRUE, 
          show.ci = FALSE,
          show.icc = FALSE,
          string.pred = "Coeffcient", 
         title = "Season Plot: Does Season Improve All Models? ADFA",
         pred.labels = c('Intercept', 
                          'Water Potential (-MPa)',
                          'Effect of Site',
                          'Wet Season'),
         dv.labels = c("PC1", "PC1: No Season", 
                       "PC2", "PC2: No Season", 
                       "Time to Ignition", "TTI: No Season",
                       "Flame Duration", "FD: No Season",
                       "Glow Duration", "GD: No Season",
                       "Flame Height", "FH: No Season",
                       "Post-Flame Glow", "PFG: No Season",
                       "Time to First Glow", "TTFG: No Season",
                       "Maximum Temp.", "Max. Temp.: No Season",
                       "Glow to Ignition", "GTI: No Season"),
  string.p = "P-Value", 
  p.style = "stars",
  file = here('figures','MEM.figures','season.effects.ADFA.html'))
```

##### P Values
```{r}
adfa.szn.mem.list <- list(pc1.m, pc1.x, pc2.m, pc2.x, tti.m, tti.x, fd.m, fd.x, gd.m, gd.x, fh.m, fh.x, pfg.m, pfg.x, ttfg.m, ttfg.x, temp.m, temp.x, gti.m, gti.x)

lapply(adfa.szn.mem.list, function(x) summary(x, ddf = 'Kenward-Roger'))
```

#### CEME
```{r}
pc1.m <- lmer(PC1 ~ mpa.unscaled + site.test + season + (1 | individual), data = seg.CEME.subset, REML = T)

pc1.x <- lmer(PC1 ~ mpa.unscaled + site.test + (1 | individual), data = seg.CEME.subset, REML = T)

pc2.m <- lmer(PC2 ~ mpa.unscaled + site.test + season + (1 | individual), data = seg.CEME.subset, REML = T)

pc2.x <- lmer(PC2 ~ mpa.unscaled + site.test + (1 | individual), data = seg.CEME.subset, REML = T)

tti.m <- lmer(tti ~ mpa.unscaled + site.test + season + (1 | individual), data = seg.CEME.subset, REML = T)

tti.x <- lmer(tti ~ mpa.unscaled + site.test + (1 | individual), data = seg.CEME.subset, REML = T)

fd.m <- lmer(fd ~ mpa.unscaled + site.test + season + (1 | individual), data = seg.CEME.subset, REML = T)

fd.x <- lmer(fd ~ mpa.unscaled + site.test + (1 | individual), data = seg.CEME.subset, REML = T)

gd.m <- lmer(gd ~ mpa.unscaled + site.test + season + (1 | individual), data = seg.CEME.subset, REML = T)

gd.x <- lmer(gd ~ mpa.unscaled + site.test + (1 | individual), data = seg.CEME.subset, REML = T)

fh.m <- lmer(fh ~ mpa.unscaled + site.test + season + (1 | individual), data = seg.CEME.subset, REML = T)

fh.x <- lmer(fh ~ mpa.unscaled + site.test + (1 | individual), data = seg.CEME.subset, REML = T)

gti.m <- lmer(gti ~ mpa.unscaled + site.test + season + (1 | individual), data = seg.CEME.subset, REML = T)

gti.x <- lmer(gti ~ mpa.unscaled + site.test + (1 | individual), data = seg.CEME.subset, REML = T)

ttfg.m <- lmer(ttfg ~ mpa.unscaled + site.test + season + (1 | individual), data = seg.CEME.subset, REML = T)

ttfg.x <- lmer(ttfg ~ mpa.unscaled + site.test + (1 | individual), data = seg.CEME.subset, REML = T)

pfg.m <- lmer(pfg ~ mpa.unscaled + site.test + season + (1 | individual), data = seg.CEME.subset, REML = T)

pfg.x <- lmer(pfg ~ mpa.unscaled + site.test + (1 | individual), data = seg.CEME.subset, REML = T)

temp.m <- lmer(temp.max ~ mpa.unscaled + site.test + season + (1 | individual), data = seg.CEME.subset, REML = T)

temp.x <- lmer(temp.max ~ mpa.unscaled + site.test + (1 | individual), data = seg.CEME.subset, REML = T)

tab_model(pc1.m, pc1.x, pc2.m, pc2.x, tti.m, tti.x, fd.m, fd.x, gd.m, gd.x, fh.m, fh.x, pfg.m, pfg.x, ttfg.m, ttfg.x, temp.m, temp.x, gti.m, gti.x,
          show.reflvl = TRUE, 
          digits = 3, 
          show.aic = TRUE, 
          show.ci = FALSE,
          show.icc = FALSE,
          string.pred = "Coeffcient", 
         title = "Season Plot: Does Season Improve All Models? CEME",
         pred.labels = c('Intercept', 
                          'Water Potential (-MPa)',
                          'Effect of Site',
                          'Wet Season'),
         dv.labels = c("PC1", "PC1: No Season", 
                       "PC2", "PC2: No Season", 
                       "Time to Ignition", "TTI: No Season",
                       "Flame Duration", "FD: No Season",
                       "Glow Duration", "GD: No Season",
                       "Flame Height", "FH: No Season",
                       "Post-Flame Glow", "PFG: No Season",
                       "Time to First Glow", "TTFG: No Season",
                       "Maximum Temp.", "Max. Temp.: No Season",
                       "Glow to Ignition", "GTI: No Season"),
  string.p = "P-Value", 
  p.style = "stars",
  file = here('figures','MEM.figures','season.effects.CEME.html'))
```

##### P Values
```{r}
ceme.szn.mem.list <- list(pc1.m, pc1.x, pc2.m, pc2.x, tti.m, tti.x, fd.m, fd.x, gd.m, gd.x, fh.m, fh.x, pfg.m, pfg.x, ttfg.m, ttfg.x, temp.m, temp.x, gti.m, gti.x)

lapply(ceme.szn.mem.list, function(x) summary(x, ddf = 'Kenward-Roger'))
```

### SUPP: Kable Table
```{r}
dep.var <- c('A. fasciculatum','PC1', 'PC1 (w/o Season)','PC2', 'PC2 (w/o Season)','Time to Ignition', 'Time to Ignition (w/o Season)','Flame Duration', 'Flame Duration (w/o Season)', 'Glow Duration', 'Glow Duration (w/o Season)','Flame Height', 'Flame Height (w/o Season)','Post-Flame Glow', 'Post-Flame Glow (w/o Season)','Time to First Glow', 'Time to First Glow (w/o Season)','Max. Temp.', 'Max. Temp. (w/o Season)','Glow to Ignition', 'Glow to Ignition (w/o Season)', 'C. megacarpus', 'PC1', 'PC1 (w/o Season)','PC2', 'PC2 (w/o Season)','Time to Ignition', 'Time to Ignition (w/o Season)','Flame Duration', 'Flame Duration (w/o Season)', 'Glow Duration', 'Glow Duration (w/o Season)','Flame Height', 'Flame Height (w/o Season)','Post-Flame Glow', 'Post-Flame Glow (w/o Season)','Time to First Glow', 'Time to First Glow (w/o Season)','Max. Temp.', 'Max. Temp. (w/o Season)','Glow to Ignition', 'Glow to Ignition (w/o Season)')
col1 <- c(' ', '-0.143***',	'-0.138***',	'0.113***',	'0.117***',	'1.451***',	'1.427***',	'-0.189*',	'-0.204*',	'-0.538',	'-0.628',	'-0.415***',	'-0.404***',	'-1.133**',	'-1.204**',	'0.669***',	'0.664***',	'-2.653***',	'-2.593***', '0.756***',	'0.720***', ' ', '-0.147***',	'-0.152***',	'0.119***',	'0.114***',	'1.851***',	'1.887***',	'-0.091',	'-0.100',	'-0.560',	'-0.431',	'-0.264**',	'-0.286**',	'-1.765**',	'-1.650*',	'0.649***',	'0.676***',	'-2.472**',	'-2.347**',	'1.148***',	'1.155***')
col2 <- c(' ', '0.768', ' ',	'0.691', ' ', '-3.105', ' ',	'-1.309', ' ', '-20.348**', ' ',	'1.075', ' ',	'-16.767**', ' ',	'-0.835', ' ', '4.592', ' ',	'-1.945', ' ', ' ', '0.460', ' ',		'0.756*', ' ', '-3.229', ' ', '0.910', ' ', '-10.354', ' ', '1.333', ' ', '-10.410', ' ', '-2.291', ' ', '-55.867***', ' ', '-0.486', ' ')
col3 <- c(' ', '0.159 / 0.278',	'0.126 / 0.274',	'0.152 / 0.345',	'0.083 / 0.319',	'0.228 / 0.335',	'0.219 / 0.332',	'0.044 / 0.114',	'0.025 / 0.105',	'0.195 / 0.339',	'0.076 / 0.312',	'0.094 / 0.173',	'0.086 / 0.170',	'0.177 / 0.350',	'0.061 / 0.322',	'0.072 / 0.285',	'0.073 / 0.278',	'0.047 / 0.112',	'0.047 / 0.106',	'0.115 / 0.132',	'0.109 / 0.125', ' ', '0.137 / 0.172',	'0.122 / 0.164',	'0.131 / 0.226',	'0.093 / 0.231',	'0.283 / 0.331',	'0.270 / 0.324',	'0.033 / 0.133',	'0.026 / 0.124',	'0.044 / 0.070',	'0.023 / 0.065',	'0.073 / NA',	'0.054 / NA',	'0.088 / 0.132',	'0.067 / 0.131',	'0.155 / 0.203',	'0.143 / 0.198',	'0.398 / NA',	'0.141 / 0.435',	'0.097 / 0.155',	'0.097 / 0.148')
col4<- c(' ', '946.510',	'948.350',	'801.237',	'802.928',	'1854.497',	'1857.760',	'1453.906',	'1455.688',	'2300.492',	'2312.529',	'1415.211',	'1416.327',	'2214.319',	'2224.996',	'1692.521',	'1693.979',	'2488.123',	'2492.287',	'1858.022',	'1860.264', ' ', '651.681',	'651.482',	'612.007',	'614.999',	'1320.394',	'1323.970',	'1006.975',	'1007.627',	'1673.775',	'1680.490',	'997.282',	'999.947',	'1648.657',	'1655.467',	'1224.871',	'1227.666',	'1730.145',	'1771.595',	'1342.248',	'1343.777')

szn.diff.kable.df <- data.frame(dep.var, col1, col2, col3, col4)
```

```{r}
szn.diff.kable.df %>%
  kable(format = 'html', escape = F, col.names = c('Dependent Variable', 'Water Potential (MPa)', 'Season Effect (Wet Season)', 'Marginal R2/Conditional R2', 'AIC')) %>% 
  kable_styling(bootstrap_options = c('hover', 'bordered', 'condensed'), fixed_thead = T, stripe_color = "black") %>% 
  row_spec(c(1:42), color = "black") %>%
  column_spec(c(5), background = "#D3D3D3") %>% 
  row_spec(c(2, 4, 6, 10, 14, 18, 20, 25, 27, 31, 33, 35, 37, 39), bold = T) %>% 
  row_spec(c(2:5, 8:9, 12:13, 16:17, 23:24, 29:30, 41:42), italic = T) %>% 
  row_spec(c(1, 22), background = "#D3D3D3", bold = T, italic = T) %>% 
  column_spec(c(1), bold = T, border_right = T) %>% 
  column_spec(c(3), border_right = T) %>% 
  save_kable(here('figures', 'supp-figures', 'TabS7_MEM_SeasonDiff_kable.html'))
```

### PC1
```{r}
m1 <- lmer(PC1 ~ mpa.unscaled + site.test + Species + season + (1 | individual), data = seg.data.subset.alldates)

m2 <- lmer(PC1 ~ mpa.unscaled + site.test + Species + (1 | individual), data = seg.data.subset.alldates)

m3 <- lmer(PC1 ~ mpa.unscaled*season + site.test + Species + (1 | individual), data = seg.data.subset.alldates)

m.adfa <- lmer(PC1 ~ mpa.unscaled + site + season + (1 | individual), data = seg.adfa.subset)

m.ceme <- lmer(PC1 ~ mpa.unscaled + site + season + (1 | individual), data = seg.CEME.subset)

m.adfa.remef <- lmer(r1_1 ~ mpa.unscaled + site + season + (1 | individual), data = precip_groups_ranis.adfa)

m.ceme.remef <- lmer(r1_1 ~ mpa.unscaled + site + season + (1 | individual), data = precip_groups_ranis.ceme)

m.adfa.linear <- lm(r1_1 ~ mpa.unscaled + season, data = precip_groups_ranis.adfa)

m.ceme.linear <- lm(r1_1 ~ mpa.unscaled + season, data = precip_groups_ranis.ceme)

tab_model(m1, m2, m3, m.adfa, m.ceme, m.adfa.remef, m.ceme.remef, m.adfa.linear, m.ceme.linear,
          show.reflvl = TRUE, 
          digits = 3, 
          show.aic = TRUE, 
          show.ci = FALSE,
          show.icc = FALSE,
          string.pred = "Coeffcient", 
         title = "season plot model selection",
         dv.labels = c('Model 1', 'Model 2', 'Model 3', 'Model 1 - ADFA', 'Model 1 - CEME', 'REMEF - ADFA', 'REMEF - CEME', 'REMEF, Linear - ADFA', 'REMEF, Linear - CEME'),
  string.p = "P-Value", 
  p.style = "stars")
# Best model is as expected -- PC1 ~ mpa + season + species + site + (1|ind); added models split by species into table
```

### PC2
```{r}
m1 <- lmer(PC2 ~ mpa.unscaled + site.test + Species + season + (1 | individual), data = seg.data.subset.alldates)

m2 <- lmer(PC2 ~ mpa.unscaled + site.test + Species + (1 | individual), data = seg.data.subset.alldates)

m3 <- lmer(PC2 ~ mpa.unscaled*season + site.test + Species + (1 | individual), data = seg.data.subset.alldates)

m.adfa <- lmer(PC2 ~ mpa.unscaled + site + season + (1 | individual), data = seg.adfa.subset)

m.ceme <- lmer(PC2 ~ mpa.unscaled + site + season + (1 | individual), data = seg.CEME.subset)

m.adfa.remef <- lmer(r1_2 ~ mpa.unscaled + site + season + (1 | individual), data = precip_groups_ranis.adfa)

m.ceme.remef <- lmer(r1_2 ~ mpa.unscaled + site + season + (1 | individual), data = precip_groups_ranis.ceme)

m.adfa.linear <- lm(r1_2 ~ mpa.unscaled + season, data = precip_groups_ranis.adfa)

m.ceme.linear <- lm(r1_2 ~ mpa.unscaled + season, data = precip_groups_ranis.ceme)

tab_model(m1, m2, m3, m.adfa, m.ceme, m.adfa.remef, m.ceme.remef, m.adfa.linear, m.ceme.linear,
          show.reflvl = TRUE, 
          digits = 3, 
          show.aic = TRUE, 
          show.ci = FALSE,
          show.icc = FALSE,
          string.pred = "Coeffcient", 
         title = "season plot model selection",
         dv.labels = c('Model 1', 'Model 2', 'Model 3', 'Model 1 - ADFA', 'Model 1 - CEME', 'REMEF - ADFA', 'REMEF - CEME', 'REMEF, linear - ADFA', 'REMEF, linear - CEME'),
  string.p = "P-Value", 
  p.style = "stars")
```

### MPa vs. LFM
```{r}
m1 <- lmer(lfm.outliers.out ~ mpa.unscaled + site.test + Species + season + (1 | individual), data = seg.data.subset.alldates)

m2 <- lmer(lfm.outliers.out ~ mpa.unscaled + site.test + Species + (1 | individual), data = seg.data.subset.alldates)

m3 <- lmer(lfm.outliers.out ~ mpa.unscaled*season + site.test + Species + (1 | individual), data = seg.data.subset.alldates)

m.adfa <- lmer(lfm.outliers.out ~ mpa.unscaled + site + season + (1 | individual), data = seg.adfa.subset)

m.ceme <- lmer(lfm.outliers.out ~ mpa.unscaled + site + season + (1 | individual), data = seg.CEME.subset)

m.adfa.remef <- lmer(r1_lfm ~ mpa.unscaled + site + season + (1 | individual), data = mpa_rel_df.adfa)

m.ceme.remef <- lmer(r1_lfm ~ mpa.unscaled + site + season + (1 | individual), data = mpa_rel_df.ceme)

m.adfa.linear <- lm(r1_lfm ~ mpa.unscaled + season, data = mpa_rel_df.adfa)

m.ceme.linear <- lm(r1_lfm ~ mpa.unscaled + season, data = mpa_rel_df.ceme)

tab_model(m1, m2, m3, m.adfa, m.ceme, m.adfa.remef, m.ceme.remef, m.adfa.linear, m.ceme.linear,
          show.reflvl = TRUE, 
          digits = 3, 
          show.aic = TRUE, 
          show.ci = FALSE,
          show.icc = FALSE,
          string.pred = "Coeffcient", 
         title = "season plot model selection",
         dv.labels = c('Model 1', 'Model 2', 'Model 3', 'Model 1 - ADFA', 'Model 1 - CEME', 'REMEF - ADFA', 'REMEF - CEME', 'REMEF, linear - ADFA', 'REMEF, linear - CEME'),
  string.p = "P-Value", 
  p.style = "stars")
```

## 2.0.3 Visualizations

### PC1
#### ADFA
```{r}
pc1.vs.mpa.adfa <- precip_groups_ranis.adfa  %>%
  ggplot(aes(y = r1_1, 
             x = mpa.unscaled,
             fill = season,
             color = season,
             group = season)) +
  geom_point(alpha = .65, aes(color = season), size = 2) +
  geom_abline(slope = -0.143, intercept = -0.052,
              color = '#c1121f', cex = 1, alpha = .9,
              show.legend = FALSE) +
  geom_abline(slope = -0.143, intercept = 0.645,
              color = '#669bbc', cex = 1, alpha = .9,
              show.legend = FALSE) +
  geom_segment(aes(x = -9.5, xend = -9.5, y = -5, yend = -3.4), arrow = arrow(length = unit(0.15, 'inches')), color = 'black') + # CHANGE TO x = 9 if text above!
  #annotate('text', x = -9, y = -2.25, label = 'More \n Combustible \n & Ignitable', fontface = 2) + # THIS IS FOR TEXT ABOVE ARRROW
  annotate('text', x = -7.3, y = -3.75, label = 'More \n Combustible \n & Ignitable', fontface = 2, size = 6) + # THIS IS FOR TEXT TO RIGHT OF ARROW
  labs(x = "Water Potential (-MPa)", 
       y = "Ignitability-Combustibility Proxy", 
       fill = "Season",
       color = "Season") +
  scale_y_continuous(limits=c(-5,6.25), breaks=c(-5,-2.5,0,2.5,5)) +
  th +
  facet_wrap(~spp,
              labeller = labeller(spp = spp.labs),
             scales = "free" ) +
  guides(color = guide_legend(override.aes = list(linetype = 0, shape = 16))) +
  theme(legend.position = "none", 
        legend.title = element_text(face = "bold"),
        legend.key = element_rect(fill = 'white'),
        strip.text = element_text(face = "italic", size = 18),
        axis.title.y = element_text(face = 'bold', size = 18),
        axis.title.x = element_blank(),
        axis.text.x = element_blank(),
        axis.text.y = element_text(size = 14)) +
  guides(color = guide_legend(override.aes = list(size = 3))) +
  annotate('text', x = -9.7, y = 6, label = 'a', size = 12, face = 'bold') +
  scale_color_manual(values=c("#c1121f","#669BBC"))
pc1.vs.mpa.adfa
```

##### Site Effect
Upon our rethink of 'remef', I don't think these 'site effect' plots are at-all necessary. Since 'remef' removes the partial effect of site from the dependent variable (in this case, PC1), it does not make sense to fold in the effect of site into a mixed effects model prior to plotting the lines. Instead, we should have been using a basic linear model. I made the changes accordingly for our main plots. For now, I will keep the site effect and avg site effect for ADFA x PC1 as a reference in case we change our mind, but will delete it from other plots. 

```{r}
pc1.vs.mpa.adfa.site <- precip_groups_ranis  %>%
  filter(spp == 'ADFA') %>% 
  ggplot(aes(y = r1_1, 
             x = mpa.unscaled,
             fill = season,
             color = season,
             group = season)) +
    geom_segment(aes(x = -9, xend = -9, y = -5, yend = -3.4), arrow = arrow(length = unit(0.15, 'inches')), color = 'black') +
  annotate('text', x = -9, y = -2.25, label = 'More \n Combustible \n & Ignitable', fontface = 2) +
  geom_point(alpha = .65, aes(color = season), size = 2) +
  geom_abline(slope = -0.137, intercept = -0.022,
              color = '#c1121f', cex = 1, alpha = .9,
              show.legend = FALSE) +
  geom_abline(slope = -0.137, intercept = 0.215,
              color = '#c1121f', cex = 1, alpha = .4,
              show.legend = FALSE) +
  geom_abline(slope = -0.137, intercept = 0.723,
              color = '#669bbc', cex = 1, alpha = .9,
              show.legend = FALSE) +
  geom_abline(slope = -0.137, intercept = 0.96,
              color = '#669bbc', cex = 1, alpha = .4,
              show.legend = FALSE) +
  labs(x = "Water Potential (-MPa)", 
       y = "Ignitability-Combustibility Proxy", 
       fill = "Season",
       color = "Season") +
  scale_y_continuous(limits=c(-5,6.25), breaks=c(-5,-2.5,0,2.5,5)) +
  th +
  facet_wrap(~spp,
              labeller = labeller(spp = spp.labs),
             scales = "free" ) +
  guides(color = guide_legend(override.aes = list(linetype = 0, shape = 16))) +
  theme(legend.position = "none", 
        legend.title = element_text(face = "bold"),
        legend.key = element_rect(fill = 'white'),
        strip.text = element_text(face = "italic", size = 18),
        axis.title.y = element_text(face = 'bold', size = 18),
        axis.title.x = element_blank()) +
  guides(color = guide_legend(override.aes = list(size = 3))) +
  annotate('text', x = -9.7, y = 6, label = 'a', size = 12, face = 'bold') +
  scale_color_manual(values=c("#c1121f","#669BBC"))
pc1.vs.mpa.adfa.site
```

##### Avg Site Effect
```{r}
pc1.vs.mpa.adfa.avgsite <- precip_groups_ranis  %>%
  filter(spp == 'ADFA') %>% 
  ggplot(aes(y = r1_1, 
             x = mpa.unscaled,
             fill = season,
             color = season,
             group = season)) +
    geom_segment(aes(x = -9.5, xend = -9.5, y = -5, yend = -3.4), arrow = arrow(length = unit(0.15, 'inches')), color = 'black') +
  annotate('text', x = -8, y = -3.75, label = 'More \n Combustible \n & Ignitable', fontface = 2, size = 6) +
  geom_point(alpha = .65, aes(color = season), size = 2) +
  geom_abline(slope = -0.137, intercept = 0.0965,
              color = '#c1121f', cex = 1, alpha = .9,
              show.legend = FALSE) +
  geom_abline(slope = -0.137, intercept = 0.8415,
              color = '#669bbc', cex = 1, alpha = .9,
              show.legend = FALSE) +
  labs(x = "Water Potential (-MPa)", 
       y = "Ignitability-Combustibility Proxy", 
       fill = "Season",
       color = "Season") +
  scale_y_continuous(limits=c(-5,6.25), breaks=c(-5,-2.5,0,2.5,5)) +
  th +
  facet_wrap(~spp,
              labeller = labeller(spp = spp.labs),
             scales = "free" ) +
  guides(color = guide_legend(override.aes = list(linetype = 0, shape = 16))) +
  theme(legend.position = "none", 
        legend.title = element_text(face = "bold"),
        legend.key = element_rect(fill = 'white'),
        strip.text = element_text(face = "italic", size = 18),
        axis.title.y = element_text(face = 'bold', size = 18),
        axis.title.x = element_blank(),
        axis.text.x = element_blank(), 
        axis.text.y = element_text(size = 14)) +
  guides(color = guide_legend(override.aes = list(size = 3))) +
  annotate('text', x = -9.7, y = 6, label = 'a', size = 12, face = 'bold') +
  scale_color_manual(values=c("#c1121f","#669BBC"))
pc1.vs.mpa.adfa.avgsite
```

#### CEME
```{r}
pc1.vs.mpa.ceme <- precip_groups_ranis.ceme  %>%
  ggplot(aes(y = r1_1, 
             x = mpa.unscaled,
             fill = season,
             color = season,
             group = season)) +
  geom_point(alpha = .65, aes(color = season), size = 2) +
  geom_abline(slope = -0.145, intercept = 0.044,
              color = '#c1121f', cex = 1, alpha = .9,
              show.legend = FALSE) +
  geom_abline(slope = -0.145, intercept = 0.472,
              color = '#669bbc', cex = 1, alpha = .9,
              show.legend = FALSE) +
  geom_segment(aes(x = -9.5, xend = -9.5, y = -5, yend = -3.4), arrow = arrow(length = unit(0.15, 'inches')), color = 'black') + # CHANGE TO x = 9 if text above!
  #annotate('text', x = -9, y = -2.25, label = 'More \n Combustible \n & Ignitable', fontface = 2) + # THIS IS FOR TEXT ABOVE ARROW
  annotate('text', x = -7.3, y = -3.75, label = 'More \n Combustible \n & Ignitable', fontface = 2, size = 6) + # THIS IS FOR TEXT TO RIGHT OF ARROW
  labs(x = "Water Potential (-MPa)", 
       y = "Ignitability-Combustibility Proxy", 
       fill = "Season",
       color = "Season") +
  scale_y_continuous(limits=c(-5,6.25), breaks=c(-5,-2.5,0,2.5,5)) +
  th +
  facet_wrap(~spp,
              labeller = labeller(spp = spp.labs),
             scales = "free" ) +
  guides(color = guide_legend(override.aes = list(linetype = 0, shape = 16))) +
  theme(legend.position = "none", 
        legend.title = element_text(face = "bold"),
        legend.key = element_rect(fill = 'white'),
        strip.text = element_text(face = "italic", size = 18),
        axis.title = element_blank(),
        axis.text = element_blank()) +
  guides(color = guide_legend(override.aes = list(size = 3))) +
  annotate('text', x = -9.7, y = 6, label = 'b', size = 12, face = 'bold') +
  scale_color_manual(values=c("#c1121f","#669BBC"))
pc1.vs.mpa.ceme
```

### PC2
Consumability proxy

#### ADFA
```{r}
pc2.vs.mpa.adfa <- precip_groups_ranis.adfa  %>%
  ggplot(aes(y = r1_2, 
             x = mpa.unscaled,
             fill = season,
             color = season,
             group = season)) +
  geom_point(alpha = .65, aes(color = season), size = 2) +
  geom_abline(slope = 0.113, intercept = -0.010,
              color = '#c1121f', cex = 1, alpha = 0.9,
              show.legend = FALSE) +
  geom_abline(slope = 0.113, intercept = 0.695,
              color = '#669bbc', cex = 1, alpha = 0.9,
              show.legend = FALSE) +
  geom_segment(aes(x = -9.5, xend = -9.5, y = -5, yend = -3.3), arrow = arrow(length = unit(0.15, 'inches')), color = 'black') +
  annotate('text', x = -7.3, y = -4.2, label = 'Less \n Consumable', fontface = 2, size = 6) +
  labs(x = "Water Potential (-MPa)", 
       y = "Consumability Proxy", 
       fill = "Season",
       color = "Season") +
  scale_y_continuous(limits=c(-5,3.25), breaks=c(-5,-2.5,0,2.5,5)) +
  th +
  guides(color = guide_legend(override.aes = list(linetype = 0, shape = 16))) +
  theme(legend.position = "none", 
        legend.title = element_text(face = "bold"),
        legend.key = element_rect(fill = 'white'),
        strip.text = element_text(face = "italic", size = 18),
        axis.title.y = element_text(face = 'bold', size = 18),
        axis.title.x = element_blank(),
        axis.text.x = element_blank(),
        axis.text.y = element_text(size = 14)) +
  guides(color = guide_legend(override.aes = list(size = 3))) +
  annotate('text', x = -9.7, y = 3, label = 'c', size = 12, face = 'bold') +
  scale_color_manual(values=c("#c1121f","#669BBC"))
pc2.vs.mpa.adfa
```

#### CEME
```{r}
pc2.vs.mpa.ceme <- precip_groups_ranis.ceme  %>%
  ggplot(aes(y = r1_2, 
             x = mpa.unscaled,
             fill = season,
             color = season,
             group = season)) +
  geom_point(alpha = .65, aes(color = season), size = 2) +
  geom_abline(slope = 0.120, intercept = 0.015,
              color = '#c1121f', cex = 1, alpha = 0.9,
              show.legend = FALSE) +
  geom_abline(slope = 0.120, intercept = 0.744,
              color = '#669bbc', cex = 1, alpha = 0.9,
              show.legend = FALSE) +
  geom_segment(aes(x = -9.5, xend = -9.5, y = -5, yend = -3.3), arrow = arrow(length = unit(0.15, 'inches')), color = 'black') +
  annotate('text', x = -7.3, y = -4.2, label = 'Less \n Consumable', fontface = 2, size = 6) +
  labs(x = "Water Potential (-MPa)", 
       y = "Consumability Proxy", 
       fill = "Season",
       color = "Season") +
  scale_y_continuous(limits=c(-5,3.25), breaks=c(-5,-2.5,0,2.5,5)) +
  th +
  guides(color = guide_legend(override.aes = list(linetype = 0, shape = 16))) +
  theme(legend.position = "none", 
        legend.title = element_text(face = "bold"),
        legend.key = element_rect(fill = 'white'),
        strip.text = element_text(face = "italic", size = 18),
        axis.title = element_blank(),
        axis.text = element_blank()) +
  guides(color = guide_legend(override.aes = list(size = 3))) +
  annotate('text', x = -9.7, y = 3, label = 'd', size = 12, face = 'bold') +
  scale_color_manual(values=c("#c1121f","#669BBC"))
pc2.vs.mpa.ceme
```

### MPa vs LFM
#### ADFA
```{r}
mpa.vs.lfm.adfa <- mpa_rel_df %>% 
  filter(spp == 'ADFA') %>% 
  ggplot(aes(y = r1_lfm, 
             x = mpa.unscaled,
             fill = season,
             color = season,
             group = season)) +
  geom_point(alpha = .65, aes(color = season), size = 2) +
  geom_abline(slope = 0.170, intercept = 0.200,
              color = '#c1121f', cex = 1, alpha = 0.9,
              show.legend = FALSE) +
  geom_abline(slope = 0.170, intercept = 0.297,
              color = '#669bbc', cex = 1, alpha = 0.9,
              show.legend = FALSE) +
  labs(x = "Water Potential (MPa)", 
       y = "LFM (Scaled, Effects Removed)", 
       fill = "Season",
       color = "Season") +
  scale_y_continuous(limits=c(-2.5,2.5), breaks=c(-2.5,  0,  2.5 )) +
  th +
  guides(color = guide_legend(override.aes = list(linetype = 0, shape = 16))) +
  theme(legend.position = "none", 
        legend.title = element_text(face = "bold"),
        legend.key = element_rect(fill = 'white'),
        strip.text = element_text(face = "italic", size = 18),
        axis.title.y = element_text(face = 'bold', size = 18),
        axis.title.x = element_blank(),
        axis.text = element_text(size = 14)) +
  guides(color = guide_legend(override.aes = list(size = 3))) +
  annotate('text', x = -9.7, y = 2.25, label = 'e', size = 12, face = 'bold') +
  scale_color_manual(values=c("#c1121f","#669BBC"))
mpa.vs.lfm.adfa
```

#### CEME
```{r}
mpa.vs.lfm.ceme <- mpa_rel_df %>% 
  filter(spp == 'CEME') %>% 
  ggplot(aes(y = r1_lfm, 
             x = mpa.unscaled,
             fill = season,
             color = season,
             group = season)) +
  geom_point(alpha = .65, aes(color = season), size = 2) +
  geom_abline(slope = 0.128, intercept = -0.171,
              color = '#c1121f', cex = 1, alpha = 0.9,
              show.legend = FALSE) +
  geom_abline(slope = 0.128, intercept = 0.398,
              color = '#669bbc', cex = 1, alpha = 0.9,
              show.legend = FALSE) +
  labs(x = "Water Potential (MPa)", 
       y = "Live Fuel Moisture (Scaled, Effects Removed)", 
       fill = "Season",
       color = "Season") +
  scale_y_continuous(limits=c(-2.5,2.5), breaks=c(-2.5,  0,  2.5 )) +
  th +
  guides(color = guide_legend(override.aes = list(linetype = 0, shape = 16))) +
  theme(legend.position = "none", 
        legend.title = element_text(face = "bold"),
        legend.key = element_rect(fill = 'white'),
        strip.text = element_text(face = "italic", size = 18),
        axis.title = element_blank(),
        axis.text.x = element_text(size = 14),
        axis.text.y = element_blank()) +
  guides(color = guide_legend(override.aes = list(size = 3))) +
  annotate('text', x = -9.7, y = 2.25, label = 'f', size = 12, face = 'bold') +
  scale_color_manual(values=c("#c1121f","#669BBC"))
mpa.vs.lfm.ceme
```

### MAIN FIGURE
```{r}
legend.sz <- get_legend(
  # create some space to the left of the legend
  ggplot() + 
  geom_point(data= seg.data.subset.alldates, aes(mpa.unscaled, PC1, color = season), 
             size = 5, 
             alpha = .8) +
    labs(color = "Season") +
    scale_color_manual(values = c('#c1121f', '#669bbc')) +
    theme(legend.box.margin = margin(0, 0, 0, 0),
          legend.position = "top",
          legend.text = element_text(face = "italic", size = 18),
          legend.title = element_text(face = 'bold', size = 22),
          legend.key = element_rect(fill = "white")))

x.grob <- textGrob("Water Potential (MPa)", 
                   gp=gpar(fontface="bold", 
                           col="black", 
                           fontsize=24))
?theme
```


```{r}
library(patchwork)

plot.sz.mem <- pc1.vs.mpa.adfa +
        pc1.vs.mpa.ceme + 
        pc2.vs.mpa.adfa + 
      #  NULL,
        pc2.vs.mpa.ceme + 
        mpa.vs.lfm.adfa + 
      #  NULL,
        mpa.vs.lfm.ceme + 
  plot_layout(ncol = 2, heights = c(1, 1, 1), widths = c(1, 1))
plot.sz.mem <- as_grob(plot.sz.mem)

# plot.sz.mem <- cowplot::plot_grid(
#         pc1.vs.mpa.adfa + theme(plot.margin = unit(c(.25, 0, 0, 0), "cm")), 
#        # NULL,
#         pc1.vs.mpa.ceme + theme(plot.margin = unit(c(.25, 0, 0, .25), "cm")),
#         pc2.vs.mpa.adfa + theme(plot.margin = unit(c(.25, 0, 0, 0), "cm")), 
#       #  NULL,
#         pc2.vs.mpa.ceme + theme(plot.margin = unit(c(.25, 0, 0, .25), "cm")),
#         mpa.vs.lfm.adfa + theme(plot.margin = unit(c(.25, 0, 0, 0), "cm")),
#       #  NULL,
#         mpa.vs.lfm.ceme + theme(plot.margin = unit(c(.25, 0, 0, .25), "cm")),
#         align='vh', 
#         axis = 'lr',
#         vjust=0,
#         rel_widths = c(1, 1),
#         rel_heights = c(1, 1, 1),
#         scale = 1,
#         ncol = 2)
# plot.sz.mem

#add to plot
scatterplot.sz.mem <- gridExtra::grid.arrange(arrangeGrob(plot.sz.mem, bottom = x.grob)) 
scatterplot.sz.mem

plot_legend.sz.mem <- cowplot::plot_grid(legend.sz, scatterplot.sz.mem, ncol = 1, rel_heights = c(.7,10))
plot_legend.sz.mem

ggsave(here("figures", "main-figures", "Fig3_Season.Plot.jpg"),
       plot = plot_legend.sz.mem,
       width = 10, height = 15, dpi = 300)
```

### MPa vs. DW
Omitted from final figure
```{r}
mpa.vs.dw.plot <- mpa_rel_df %>% 
  ggplot(aes(y = dw.flam.sample, 
             x = mpa.unscaled,
             fill = season,
             color = season,
             group = season)) +
  geom_point(alpha = .5, aes(color = season)) +
  geom_smooth(method = lm, 
              se = TRUE, 
              aes(fill = season),
              show.legend = FALSE) +
  ggpubr::stat_regline_equation(show.legend = FALSE) +
  labs(x = "Water Potential (MPa)", 
       y = "Dry Weight (Scaled)", 
       fill = "Season",
       color = "Season") +
  th +
  facet_wrap(~spp,
              labeller = labeller(spp = spp.labs),
             scales = "free" ) +
  scale_y_continuous(limits=c(-2,2.5), breaks=c(-2, -1, 0, 1, 2)) +
  guides(color = guide_legend(override.aes = list(linetype = 0, shape = 16))) +
  theme(legend.position = "none", 
        legend.title = element_text(face = "bold"),
        legend.key = element_rect(fill = 'white'),
        strip.text = element_text(face = "italic", size = 12),
        axis.title.y = element_text(face = 'bold'),
        axis.title.x = element_blank()) +
  guides(color = guide_legend(override.aes = list(size = 3))) +
  scale_color_manual(values=c("#c1121f","#669BBC"))

mpa.vs.dw.plot
```

## 2.0.4 Statistical Tests
### Significance Tests
```{r}
# Is species significant predictor of PC1?
precip_mod <- lmer(PC1 ~ mpa.unscaled + Species*season + (1 | individual), data = seg.data.subset.alldates)
summary(precip_mod)
# Yes

# For CEME?
seg.data.subset.alldates.ceme <- seg.data.subset.alldates %>%
  filter(spp == "CEME")
precip_mod.ceme <- lmer(PC1 ~ mpa.unscaled + season + (1 | individual), data = seg.data.subset.alldates.ceme)
summary(precip_mod.ceme)
# Yes

# For ADFA?
seg.data.subset.alldates.adfa <- seg.data.subset.alldates %>%
  filter(spp == "ADFA")
precip_mod.adfa <- lmer(PC1 ~ mpa.unscaled + season + (1 | individual), data = seg.data.subset.alldates.adfa)
summary(precip_mod.adfa)
# Yes

# Is species significant predictor of PC2?
precip_mod2 <- lmer(PC2 ~ mpa.unscaled + Species*season + (1 | individual), data = seg.data.subset.alldates)
summary(precip_mod2)
# Yes

# For CEME?
precip_mod.ceme2 <- lmer(PC2 ~ mpa.unscaled + season + (1 | individual), data = seg.data.subset.alldates.ceme)
summary(precip_mod.ceme2)
# Yes

# For ADFA?
precip_mod.adfa2 <- lmer(PC2 ~ mpa.unscaled + season + (1 | individual), data = seg.data.subset.alldates.adfa)
summary(precip_mod.adfa2)
# Yes
```

### ANOVA
```{r}
# Is model with season sig. better than w/o?
## PC1
### ADFA
pc1.w.szn.a <- lmer(PC1 ~ mpa.unscaled + season + site + (1 | individual), data = seg.data.subset.alldates.adfa)
pc1.wo.szn.a <- lmer(PC1 ~ mpa.unscaled + site + (1 | individual), data = seg.data.subset.alldates.adfa)
anova(pc1.w.szn.a, pc1.wo.szn.a)
### CEME
pc1.w.szn.c <- lmer(PC1 ~ mpa.unscaled + season + site + (1 | individual), data = seg.data.subset.alldates.ceme)
pc1.wo.szn.c <- lmer(PC1 ~ mpa.unscaled + site + (1 | individual), data = seg.data.subset.alldates.ceme)
anova(pc1.w.szn.c, pc1.wo.szn.c)

## PC2
### ADFA
pc2.w.szn.a <- lmer(PC2 ~ mpa.unscaled + season + site + (1 | individual), data = seg.data.subset.alldates.adfa)
pc2.wo.szn.a <- lmer(PC2 ~ mpa.unscaled + site + (1 | individual), data = seg.data.subset.alldates.adfa)
anova(pc2.w.szn.a, pc2.wo.szn.a)
### CEME
pc2.w.szn.c <- lmer(PC2 ~ mpa.unscaled + season + site + (1 | individual), data = seg.data.subset.alldates.ceme)
pc2.wo.szn.c <- lmer(PC2 ~ mpa.unscaled + site + (1 | individual), data = seg.data.subset.alldates.ceme)
anova(pc2.w.szn.c, pc2.wo.szn.c)

## MPa vs. LFM
### ADFA
lfm.w.szn.a <- lmer(lfm.outliers.out ~ mpa.unscaled + site + season + (1 | individual), data = seg.data.subset.alldates.adfa)
lfm.wo.szn.a <- lmer(lfm.outliers.out ~ mpa.unscaled + site + (1 | individual), data = seg.data.subset.alldates.adfa)
anova(lfm.w.szn.a, lfm.wo.szn.a)
### CEME
lfm.w.szn.c <- lmer(lfm.outliers.out ~ mpa.unscaled + site + season + (1 | individual), data = seg.data.subset.alldates.ceme)
lfm.wo.szn.c <- lmer(lfm.outliers.out ~ mpa.unscaled + site + (1 | individual), data = seg.data.subset.alldates.ceme)
anova(lfm.w.szn.c, lfm.wo.szn.c)
```

#--------------------------------
# 2.1 Supplementary Season Effects Plots

# 2.1.1 Season Plot: LFM
## REMEF
```{r}
#### PC1
szn.lfm.pc1.mod <- lmer(PC1 ~ lfm.unscaled + site + spp + season + (1 | individual), data = seg.data.subset.alldates)

r1_L <- remef(szn.lfm.pc1.mod, fix = c('sppCEME', 'siteSt. Marys'),  ran = "all", keep.intercept = FALSE)
r1_L_df <- as.data.frame(r1_L)

### PC2
szn.lfm.pc1.mod <- lmer(PC2 ~ lfm.unscaled + site + spp + season + (1 | individual), data = seg.data.subset.alldates)

r2_L <- remef(szn.lfm.pc1.mod, fix = c('sppCEME', 'siteSt. Marys'),  ran = "all", keep.intercept = FALSE)
r2_L_df <- as.data.frame(r2_L)

### Combining into df
szn_lfm_ranis <- cbind(seg.data.subset.alldates, r1_L_df, r2_L_df)

### Splitting by species
szn_lfm_ranis.adfa <- szn_lfm_ranis %>% 
  filter(spp == "ADFA")
szn_lfm_ranis.ceme <- szn_lfm_ranis %>% 
  filter(spp == "CEME")
```

## Model Selection
```{r}
m.adfa.remef <- lmer(r1_L ~ lfm.unscaled + site + season + (1 | individual), data = szn_lfm_ranis.adfa)

m.ceme.remef <- lmer(r1_L ~ lfm.unscaled + site + season + (1 | individual), data = szn_lfm_ranis.ceme)

m.adfa.linear <- lm(r1_L ~ lfm.unscaled + season, data = szn_lfm_ranis.adfa)

m.ceme.linear <- lm(r1_L ~ lfm.unscaled + season, data = szn_lfm_ranis.ceme)

m.adfa.remef2 <- lmer(r2_L ~ lfm.unscaled + site + season + (1 | individual), data = szn_lfm_ranis.adfa)

m.ceme.remef2 <- lmer(r2_L ~ lfm.unscaled + site + season + (1 | individual), data = szn_lfm_ranis.ceme)

m.adfa.linear2 <- lm(r2_L ~ lfm.unscaled + season, data = szn_lfm_ranis.adfa)

m.ceme.linear2 <- lm(r2_L ~ lfm.unscaled + season, data = szn_lfm_ranis.ceme)

tab_model(m.adfa.remef, m.ceme.remef, m.adfa.linear, m.ceme.linear, m.adfa.remef2, m.ceme.remef2, m.adfa.linear2, m.ceme.linear2,
          show.reflvl = TRUE, 
          digits = 3, 
          show.aic = TRUE, 
          show.ci = FALSE,
          show.icc = FALSE,
          string.pred = "Coeffcient", 
         title = "season plot model selection: LFM",
         dv.labels = c('ADFA - PC1', 'CEME - PC1', 'ADFA - PC1, linear', 'CEME - PC1, linear', 'ADFA - PC2', 'CEME - PC2', 'ADFA - PC2, linear', 'CEME - PC2, linear'),
  string.p = "P-Value", 
  p.style = "stars")
```

## PC1
```{r}
#### ADFA
pc1.vs.lfm.adfa <- szn_lfm_ranis  %>%
  filter(spp == 'ADFA') %>% 
  ggplot(aes(y = r1_L, 
             x = lfm.unscaled,
             fill = season,
             color = season,
             group = season)) +
  geom_point(alpha = .65, aes(color = season), size = 2) +
  geom_abline(slope = -0.014, intercept = 0.082,
              color = '#c1121f', cex = 1, alpha = .9,
              show.legend = FALSE) +
  geom_abline(slope = -0.014, intercept = 0.854,
              color = '#669bbc', cex = 1, alpha = .9,
              show.legend = FALSE) +
  geom_segment(aes(x = 15, xend = 15, y = -5, yend = -3.4), arrow = arrow(length = unit(0.15, 'inches')), color = 'black') +
  annotate('text', x = 15, y = -2.25, label = 'More \n Combustible \n & Ignitable', fontface = 2) +
  labs(x = "Live Fuel Moisture (%)", 
       y = "Ignitability-Combustibility Proxy", 
       fill = "Season",
       color = "Season") +
  scale_y_continuous(limits=c(-5,6.25), breaks=c(-5,-2.5,0,2.5,5)) +
  th +
  facet_wrap(~spp,
              labeller = labeller(spp = spp.labs),
             scales = "free" ) +
  guides(color = guide_legend(override.aes = list(linetype = 0, shape = 16))) +
  theme(legend.position = "none", 
        legend.title = element_text(face = "bold"),
        legend.key = element_rect(fill = 'white'),
        strip.text = element_text(face = "italic", size = 18),
        axis.title.y = element_text(face = 'bold', size = 18),
        axis.title.x = element_blank(),
        axis.text.y = element_text(size = 14),
        axis.text.x = element_blank()) +
  guides(color = guide_legend(override.aes = list(size = 3))) +
  annotate('text', x = 10, y = 6, label = 'a', size = 12, face = 'bold') +
  scale_color_manual(values=c("#c1121f","#669BBC"))
pc1.vs.lfm.adfa

#### CEME
pc1.vs.lfm.ceme <- szn_lfm_ranis  %>%
  filter(spp == 'CEME') %>% 
  ggplot(aes(y = r1_L, 
             x = lfm.unscaled,
             fill = season,
             color = season,
             group = season)) +
  geom_point(alpha = .65, aes(color = season), size = 2) +
  geom_abline(slope = -0.013, intercept = -0.029,
              color = '#c1121f', cex = 1, alpha = .9,
              show.legend = FALSE) +
  geom_abline(slope = -0.013, intercept = 0.800,
              color = '#669bbc', cex = 1, alpha = .9,
              show.legend = FALSE) +
  geom_segment(aes(x = 20, xend = 20, y = -5, yend = -3.4), arrow = arrow(length = unit(0.15, 'inches')), color = 'black') +
  annotate('text', x = 20, y = -2.25, label = 'More \n Combustible \n & Ignitable', fontface = 2) +
  labs(x = "Live Fuel Moisture (%)", 
       y = "Ignitability-Combustibility Proxy", 
       fill = "Season",
       color = "Season") +
  scale_y_continuous(limits=c(-5,6.25), breaks=c(-5,-2.5,0,2.5,5)) +
  th +
  facet_wrap(~spp,
              labeller = labeller(spp = spp.labs),
             scales = "free" ) +
  guides(color = guide_legend(override.aes = list(linetype = 0, shape = 16))) +
  theme(legend.position = "none", 
        legend.title = element_text(face = "bold"),
        legend.key = element_rect(fill = 'white'),
        strip.text = element_text(face = "italic", size = 18),
        axis.title = element_blank(),
        axis.text = element_blank()) +
  guides(color = guide_legend(override.aes = list(size = 3))) +
  annotate('text', x = 10, y = 6, label = 'b', size = 12) +
  scale_color_manual(values=c("#c1121f","#669BBC"))
pc1.vs.lfm.ceme
```

## PC2
```{r}
#### ADFA
pc2.vs.lfm.adfa <- szn_lfm_ranis  %>%
  filter(spp == "ADFA") %>% 
  ggplot(aes(y = r2_L, 
             x = lfm.unscaled,
             fill = season,
             color = season,
             group = season)) +
  geom_point(alpha = .65, aes(color = season), size = 2) +
  geom_abline(slope = 0.016, intercept = -0.570,
              color = '#c1121f', cex = 1, alpha = 0.9,
              show.legend = FALSE) +
  geom_abline(slope = 0.016, intercept = 0.338,
              color = '#669bbc', cex = 1, alpha = 0.9,
              show.legend = FALSE) +
  geom_segment(aes(x = 15, xend = 15, y = -5, yend = -3.3), arrow = arrow(length = unit(0.15, 'inches')), color = 'black') +
  annotate('text', x = 15, y = -2.6, label = 'Less \n Consumable', fontface = 2) +
  labs(x = "Live Fuel Moisture (%)", 
       y = "Consumability Proxy", 
       fill = "Season",
       color = "Season") +
  scale_y_continuous(limits=c(-5,5), breaks=c(-5,-2.5,0,2.5,5)) +
  th +
  #facet_wrap(~spp,labeller = labeller(spp = spp.labs),scales = "free" ) +
  guides(color = guide_legend(override.aes = list(linetype = 0, shape = 16))) +
  theme(legend.position = "none", 
        legend.title = element_text(face = "bold"),
        legend.key = element_rect(fill = 'white'),
        strip.text = element_text(face = "italic", size = 18),
        axis.title.y = element_text(face = 'bold', size = 18),
        axis.title.x = element_blank(),
        axis.text = element_text(size = 14)) +
  guides(color = guide_legend(override.aes = list(size = 3))) +
  annotate('text', x = 10, y = 4.9, label = 'c', size = 12) +
  scale_color_manual(values=c("#c1121f","#669BBC"))
pc2.vs.lfm.adfa

#### CEME
pc2.vs.lfm.ceme <- szn_lfm_ranis  %>%
  filter(spp == 'CEME') %>% 
  ggplot(aes(y = r2_L, 
             x = lfm.unscaled,
             fill = season,
             color = season,
             group = season)) +
  geom_point(alpha = .65, aes(color = season), size = 2) +
  geom_abline(slope = 0.007, intercept = 0.223,
              color = '#c1121f', cex = 1, alpha = 0.9,
              show.legend = FALSE) +
  geom_abline(slope = 0.007, intercept = 0.773,
              color = '#669bbc', cex = 1, alpha = 0.9,
              show.legend = FALSE) +
  geom_segment(aes(x = 20, xend = 20, y = -5, yend = -3.3), arrow = arrow(length = unit(0.15, 'inches')), color = 'black') +
  annotate('text', x = 20, y = -2.6, label = 'Less \n Consumable', fontface = 2) +
  labs(x = "Live Fuel Moisture (%)", 
       y = "Consumability Proxy", 
       fill = "Season",
       color = "Season") +
  scale_y_continuous(limits=c(-5,5), breaks=c(-5,-2.5,0,2.5,5)) +
  th +
  #facet_wrap(~spp, labeller = labeller(spp = spp.labs), scales = "free" ) +
  guides(color = guide_legend(override.aes = list(linetype = 0, shape = 16))) +
  theme(legend.position = "none", 
        legend.title = element_text(face = "bold"),
        legend.key = element_rect(fill = 'white'),
        strip.text = element_text(face = "italic", size = 18),
        axis.title = element_blank(),
        axis.text.x = element_text(size = 14),
        axis.text.y = element_blank()) +
  guides(color = guide_legend(override.aes = list(size = 3))) +
  annotate('text', x = 10, y = 4.85, label = 'd', size = 12) +
  scale_color_manual(values=c("#c1121f","#669BBC"))
pc2.vs.lfm.ceme
```

## SUPP Figure: Arranging Plots
```{r}
# plot.sz.lfm <- cowplot::plot_grid(
#         pc1.vs.lfm.adfa + theme(plot.margin = unit(c(0.1, 0, 0.1, 0.3), "cm")), 
#         pc1.vs.lfm.ceme + theme(plot.margin = unit(c(0.25, 0.25, 0.25, 0.25), "cm")),
#         pc2.vs.lfm.adfa + theme(plot.margin = unit(c(0.1, 0, 0.1, 0.3), "cm")), 
#         pc2.vs.lfm.ceme + theme(plot.margin = unit(c(0.1, 0.1, 0.1, 0), "cm")),
#         align='vh', vjust=.5, scale = 1,
#         ncol = 2
#           )

plot.sz.lfm <- pc1.vs.lfm.adfa +
        pc1.vs.lfm.ceme + 
        pc2.vs.lfm.ceme + 
        pc2.vs.lfm.adfa + 
  plot_layout(ncol = 2, heights = c(1, 1, 1), widths = c(1, 1))

plot.sz.mem <- as_grob(plot.sz.lfm)

lfm.grob <- textGrob("Live Fuel Moisture (%)", 
                   gp=gpar(fontface="bold", 
                           col="black", 
                           fontsize=24))

#add to plot
scatterplot.sz.lfm <- gridExtra::grid.arrange(arrangeGrob(plot.sz.lfm, bottom = lfm.grob)) 

plot_legend.sz.lfm <- cowplot::plot_grid(legend.sz, scatterplot.sz.lfm, ncol = 1, rel_heights = c(1,15))
plot_legend.sz.lfm

ggsave(here("figures", "supp-figures", "FigS3_LFM.season.plot.suppfigure.jpg"),
       plot = plot_legend.sz.lfm,
       width = 17, height = 10, dpi = 300)
```

# 2.1.2 Extra Plots
## MPa vs PC1

Timing (Year Month)
```{r}
mpa_mod_fig.mpa <- lmer(PC1 ~ mpa.unscaled + Species + year.month + (1 | individual), data = seg.data.subset.alldates)

effects_mpa_fig.mpa <- effects::effect(term= c("mpa.unscaled","year.month"),
                                   mod= mpa_mod_fig.mpa, 
                                   se = TRUE, 
                                   confidence.level=.95
                                   )
# Save the effects values as a df:
x_mpa_fig.mpa <- as.data.frame(effects_mpa_fig.mpa)
```

```{r}
precip_plot <- seg.data.subset.alldates %>%
  ggplot(aes(y = PC1, x = mpa.unscaled, 
             #color = as.factor(timing)
             color = as.factor(precip.2mo)
             )) +
  geom_point(size = 1, alpha = .5) +
  #geom_line(data=x_mpa_fig.mpa, aes(x=mpa.unscaled, y=fit, color = year.month)) +
  geom_smooth(method = lm, 
              size = 1, 
              alpha = 1, 
              se = FALSE) +
  # ggpubr::stat_regline_equation(aes(label =  paste(..eq.label.., ..adj.rr.label.., sep = "~~~~")),
  #                                   label.y = 4) +
  #stat_cor() +
  labs(x = "Water Potential (MPa)", 
       y = "PC1 ", 
       color = "Precip. (in)") +
 # scale_color_discrete(name = "Site")+
  th +
 # scale_color_brewer(palette = "Dark2") +
  facet_wrap(~spp,
              labeller = labeller(spp = spp.labs),
             scales = "free" ) +
  theme(legend.position = "bottom", 
        strip.text = element_text(face = "italic", size = 12)) +
  scale_color_manual(values=met.brewer("Morgenstern", 11)) +
  facet_wrap(~Species) +
  theme(legend.position = "bottom") +
  scale_x_continuous(breaks = c(-10, -8, -6, -4, -2, 0))
precip_plot

ggsave(here("figures", "extra-figures", "precip_plot.jpg"),
       plot = precip_plot, 
       width = 6, height = 4, dpi = 300)
```


## Dry Weight vs PC1
```{r}
mod_dw <- lmer(PC1 ~ dw.flam.sample + Species * season + (1 | individual), data = seg.data.subset.alldates)

summary(mod_dw)
```

```{r}
precip_plot.dw <- seg.data.subset.alldates %>%
  ggplot(aes(y = PC1, x = dw.flam.sample,
             color = as.factor(precip.2mo)
             )) +
  geom_point(size = 1, alpha = .5) +
  geom_smooth(method = lm, 
              size = 1, 
              alpha = 1, 
              se = FALSE) +
  labs(x = "Dry Weight (g)", 
       y = "PC1 ", 
       color = "Precip. (in)") +
  th +
  facet_wrap(~spp,
              labeller = labeller(spp = spp.labs),
             scales = "free" ) +
  theme(legend.position = "bottom", 
        strip.text = element_text(face = "italic", size = 12)) +
  scale_color_manual(values=met.brewer("Morgenstern", 11)) +
  facet_wrap(~Species) +
  theme(legend.position = "bottom") +
  scale_x_continuous(breaks = c(-10, -8, -6, -4, -2, 0))
precip_plot.dw
```

## Timing and precip plot
Precip. designated by size of point, timing by color

Model: 
```{r}
precip_mod <- lmer(lfm.unscaled ~ mpa.unscaled + Species + timing + (1 | individual), data = seg.data.subset.alldates)

#remove fixed effect of year.month
#     see what names are associated: 
terms <- term2coef(precip_mod, "timing") %>% 
  list()
terms

r1_1 <- remef(precip_mod ,  ran = "all")

#https://stats.stackexchange.com/questions/258285/how-to-do-partial-regression-plots-with-linear-mixed-effect-models/258292#258292?newreg=3b55c635f610441db980674675bc0051

scaled_r1_1_df <- as.data.frame(r1_1)
precip_groups_ranis <- cbind(seg.data.subset.alldates, scaled_r1_1_df)
```

Plot: 
```{r}
timing_plot <- precip_groups_ranis %>%
ggplot(aes(y = lfm.unscaled, 
             x = mpa.unscaled, 
             color = as.factor(timing))) +
  geom_jitter(aes(size = precip.2mo),
    alpha = .5) +
            scale_size_continuous(limits = c(0,6), guide='none')+
  #geom_line(data=x_mpa_fig.mpa, aes(x=mpa.unscaled, y=fit, color = year.month)) +
  geom_smooth(method = lm, 
              size = 1, 
              alpha = 1, 
              se = FALSE) +
 ggpubr::stat_regline_equation() +
  labs(x = "Water Potential (-MPa)", 
       y = "Live Fuel Moisture (%)", 
       color = "Timing", 
       size = "Precip.") +
 # scale_color_discrete(name = "Site")+
  th +
 # scale_color_brewer(palette = "Dark2") +
  facet_wrap(~spp,
              labeller = labeller(spp = spp.labs),
             scales = "free" ) +
  theme(legend.position = "bottom", 
        strip.text = element_text(face = "italic", size = 12)) +
  scale_color_manual(values=met.brewer("Morgenstern", 10)) +
  facet_wrap(~Species) +
  #theme(legend.position = "bottom") +
  scale_x_continuous(breaks = c(-10, -8, -6, -4, -2, 0))
timing_plot

ggsave(here("figures", "extra-figures", "timing_precip_plot.jpg"),
       plot = timing_plot, 
       width = 6, height = 4, dpi = 300)
```

```{r}
mpa_lfm_mod <- lmer(lfm ~ mpa.unscaled + Species + year.month + (1| individual), data = seg.data.subset.alldates)
mpa_lfm_mod

tab_model(mpa_lfm_mod)
```

#------------------------------------
# 3.0 Species Effects Plots

Plotting options
```{r}
color_panels <- scale_color_manual(values=c("#9D0208","#F48C06")) 

th_panels <- theme(
        axis.title.x=element_blank(),
        axis.title.y=element_text(face = 'bold', size = 18),
        legend.position = "none", 
        aspect.ratio = 1, 
        plot.margin = unit(c(0,0,0,0), "cm")
        )
```

# 3.0.1 Data wrangling
Since September 2020 is the only period of time that we tested both species at once, to compare species differences, we must use only these tests
```{r}
scaled.mem.data.sept2020 <- seg.data.subset.alldates %>% 
  filter(year.month == '2020_September')
```

# 3.0.2 Model Selection
## Effect of Species - All Flam Metrics
```{r}
pc1.m <- lmer(PC1 ~ mpa.unscaled + Species + site + (1 | individual), data = scaled.mem.data.sept2020, REML = T)

pc1.x <- lmer(PC1 ~ mpa.unscaled + site + (1 | individual), data = scaled.mem.data.sept2020, REML = T)

pc2.m <- lmer(PC2 ~ mpa.unscaled + Species + site + (1 | individual), data = scaled.mem.data.sept2020, REML = T)

pc2.x <- lmer(PC2 ~ mpa.unscaled + site + (1 | individual), data = scaled.mem.data.sept2020, REML = T)

tti.m <- lmer(tti ~ mpa.unscaled + Species + site + (1 | individual), data = scaled.mem.data.sept2020, REML = T)

tti.x <- lmer(tti ~ mpa.unscaled + site + (1 | individual), data = scaled.mem.data.sept2020, REML = T)

fd.m <- lmer(fd ~ mpa.unscaled + Species + site + (1 | individual), data = scaled.mem.data.sept2020, REML = T)

fd.x <- lmer(fd ~ mpa.unscaled + site + (1 | individual), data = scaled.mem.data.sept2020, REML = T)

fh.m <- lmer(fh ~ mpa.unscaled + Species + site + (1 | individual), data = scaled.mem.data.sept2020, REML = T)

fh.x <- lmer(fh ~ mpa.unscaled + site + (1 | individual), data = scaled.mem.data.sept2020, REML = T)

gd.m <- lmer(gd ~ mpa.unscaled + Species + site + (1 | individual), data = scaled.mem.data.sept2020, REML = T)

gd.x <- lmer(gd ~ mpa.unscaled + site + (1 | individual), data = scaled.mem.data.sept2020, REML = T)

pfg.m <- lmer(pfg ~ mpa.unscaled + Species + site + (1 | individual), data = scaled.mem.data.sept2020, REML = T)

pfg.x <- lmer(pfg ~ mpa.unscaled + site + (1 | individual), data = scaled.mem.data.sept2020, REML = T)

ttfg.m <- lmer(ttfg ~ mpa.unscaled + Species + site + (1 | individual), data = scaled.mem.data.sept2020, REML = T)

ttfg.x <- lmer(ttfg ~ mpa.unscaled + site + (1 | individual), data = scaled.mem.data.sept2020, REML = T)

temp.m <- lmer(temp.max ~ mpa.unscaled + Species + site + (1 | individual), data = scaled.mem.data.sept2020, REML = T)

temp.x <- lmer(temp.max ~ mpa.unscaled + site + (1 | individual), data = scaled.mem.data.sept2020, REML = T)

gti.m <- lmer(gti ~ mpa.unscaled + Species + site + (1 | individual), data = scaled.mem.data.sept2020, REML = T)

gti.x <- lmer(gti ~ mpa.unscaled + site + (1 | individual), data = scaled.mem.data.sept2020, REML = T)

tab_model(pc1.m, pc1.x, pc2.m, pc2.x, tti.m, tti.x, fd.m, fd.x, gd.m, gd.x, fh.m, fh.x, pfg.m, pfg.x, ttfg.m, ttfg.x, temp.m, temp.x, gti.m, gti.x,
          show.reflvl = TRUE, 
          digits = 3, 
          show.aic = TRUE, 
          show.ci = FALSE,
          show.icc = FALSE,
          string.pred = "Coeffcient", 
         title = "species plot: species improve all models?",
  string.p = "P-Value", 
  p.style = "stars",
  file = here('figures','MEM.figures','species.effects.html'))
```

### P Values
```{r}
spp.mem.list <- list(pc1.m, pc1.x, pc2.m, pc2.x, tti.m, tti.x, fd.m, fd.x, gd.m, gd.x, fh.m, fh.x, pfg.m, pfg.x, ttfg.m, ttfg.x, temp.m, temp.x, gti.m, gti.x)

lapply(spp.mem.list, function(x) summary(x, ddf = 'Kenward-Roger'))
```


## SUPP: Kable Table 
Setting up dataframe
```{r}
dep.var <- c('PC1', 'PC1 (w/o Species)','PC2', 'PC2 (w/o Species)','Time to Ignition', 'Time to Ignition (w/o Species)','Flame Duration', 'Flame Duration (w/o Species)', 'Glow Duration', 'Glow Duration (w/o Species)','Flame Height', 'Flame Height (w/o Species)','Post-Flame Glow', 'Post-Flame Glow (w/o Species)','Time to First Glow', 'Time to First Glow (w/o Species)','Max. Temp.', 'Max. Temp. (w/o Species)','Glow to Ignition', 'Glow to Ignition (w/o Species)')
col1 <- c('-0.119***',	'-0.120***',	'0.151***',	'0.151***',	'1.453***',	'1.462***',	'-0.138',	'-0.136',	'-1.097',	'-1.098',	'-0.270**',	'-0.270**',	'-1.981***',	'-1.984***',	'0.525**',	'0.533**',	'-4.810***',	'-4.768***',	'0.924***',	'0.921***')
col2 <- c('-0.808**', ' ', '-0.051', ' ',	'7.469**', ' ', '-2.151*', ' ', '6.370', ' ', '-0.382', ' ', '4.807', ' ', '4.110**', ' ', '15.486*', ' ', '3.333', ' ')
col3 <- c('0.119 / 0.218',	'0.067 / 0.217',	'0.113 / 0.225',	'0.114 / 0.218', '0.245 / 0.314',	'0.158 / 0.317',	'0.063 / 0.142',	'0.011 / 0.140',	'0.035 / 0.139',	'0.023 / 0.133',	'0.042 / NA',	'0.040 / NA',	'0.069 / 0.170',	'0.060 / 0.163',	'0.094 / 0.166',	'0.039 / 0.162',	'0.151 / 0.154',	'0.118 / 0.149',	'0.102 / 0.188',	'0.081 / 0.186')
col4 <- c('842.962',	'846.605',	'749.166',	'746.299',	'1653.945',	'1666.668',	'1282.351',	'1288.217',	'2065.979',	'2070.586',	'1248.304',	'1247.451',	'2007.342',	'2011.323',	'1535.178',	'1543.170',	'2220.840',	'2231.133',	'1657.816',	'1661.730')
spp.diff.kable.df.long <- data.frame(dep.var, col1, col2, col3, col4)
```

```{r}
spp.diff.kable.df.long %>%
  kable(format = 'html', escape = F, col.names = c('Dependent Variable', 'Water Potential (MPa)', 'Species Effect (C. Megacarpus)', 'Marginal R2/Conditional R2', 'AIC')) %>% 
  kable_styling(bootstrap_options = c('hover', 'bordered', 'condensed'), fixed_thead = T, stripe_color = "black") %>% 
  row_spec(c(1:20), color = "black") %>%
  column_spec(c(5), background = "#D3D3D3") %>% 
  row_spec(c(1,4,5,7,9,13,15,17,19), bold = T) %>% 
  row_spec(c(11:12), italic = T) %>% 
  column_spec(c(1), bold = T, border_right = T) %>% 
  column_spec(c(3), border_right = T) %>% 
  save_kable(here('figures', 'supp-figures', 'TabS8_MEM_SpeciesDiff_kable.html'))
```

# 3.0.3 Statistical Tests
## ANOVA
Does species significantly improve model?
- Only testing variables in our plot (for now)
```{r}
# PC1
anova(pc1.m, pc1.x)
# TTI
anova(tti.m, tti.x)
# FD
anova(fd.m, fd.x)
# GD
anova(gd.m, gd.x)
```

# 3.0.4 REMEF: Removing effects of random effects
Because we are interested in the units for the flam. metrics, I did keep.intercept = TRUE; however, I am not totally sure if we should do keep.intercept = FALSE for one set of analyses and not the other. For now, I think the season plots are fine, as is, because the units of the proxies do not matter at all

## Time to ignition
```{r}
mpa_mod_fig_tti <- lmer(tti ~ mpa.unscaled + spp + site + (1 | individual), data = scaled.mem.data.sept2020)

r1_1 <- remef(mpa_mod_fig_tti, fix = 'siteSt. Marys',  ran = "all", keep.intercept = TRUE)
#https://stats.stackexchange.com/questions/258285/how-to-do-partial-regression-plots-with-linear-mixed-effect-models/258292#258292?newreg=3b55c635f610441db980674675bc0051

scaled_r1_1_df <- as.data.frame(r1_1)
tti_groups_ranis <- cbind(scaled.mem.data.sept2020, scaled_r1_1_df)
```

## Flame duration
```{r}
mpa_mod_fig_fd <- lmer(fd ~ mpa.unscaled + spp + site + (1 | individual), data = scaled.mem.data.sept2020)

fd_remef_year <- remef(mpa_mod_fig_fd, fix = 'siteSt. Marys',  ran = "all", keep.intercept = TRUE)

scaled_r1_1_df <- as.data.frame(fd_remef_year)
fd_remef_year_df <- cbind(scaled.mem.data.sept2020, scaled_r1_1_df)
```

## Glow duration
```{r}
mpa_mod_fig_gd <- lmer(gd ~ mpa.unscaled + spp + site + (1 | individual), data = scaled.mem.data.sept2020)

gd_remef_year <- remef(mpa_mod_fig_gd , fix = 'siteSt. Marys',  ran = "all", keep.intercept = T)

scaled_r1_1_df <- as.data.frame(gd_remef_year)
gd_remef_year_df <- cbind(scaled.mem.data.sept2020, scaled_r1_1_df)
```

## PC1
```{r}
mpa_mod_fig_pc1 <- lmer(PC1 ~ mpa.unscaled + spp + site + (1 | individual), data = scaled.mem.data.sept2020)

pc1_remef_year <- remef(mpa_mod_fig_pc1 , fix = 'siteSt. Marys',  ran = "all", keep.intercept = T)

scaled_r1_1_df <- as.data.frame(pc1_remef_year)
pc1_remef_year_df <- cbind(scaled.mem.data.sept2020, scaled_r1_1_df)
```

# 3.0.5 Visualization
#### Models
```{r}
m.tti <- lmer(r1_1 ~ mpa.unscaled + Species + site + (1 | individual), data = tti_groups_ranis)

m.tti.nospp <- lmer(r1_1 ~ mpa.unscaled + site + (1 | individual), data = tti_groups_ranis)

m.tti.linear <- lm(r1_1 ~ mpa.unscaled + Species, data = tti_groups_ranis)

m.fd <- lmer(fd_remef_year ~ mpa.unscaled + Species + site + (1 | individual), data = fd_remef_year_df)

m.fd.nospp <- lmer(fd_remef_year ~ mpa.unscaled + site + (1 | individual), data = fd_remef_year_df)

m.fd.linear <- lm(fd_remef_year ~ mpa.unscaled + Species, data = fd_remef_year_df)

m.gd <- lmer(gd_remef_year ~ mpa.unscaled + Species + site + (1 | individual), data = gd_remef_year_df)

m.gd.nospp <- lmer(gd_remef_year ~ mpa.unscaled + site + (1 | individual), data = gd_remef_year_df)

m.gd.linear <- lm(gd_remef_year ~ mpa.unscaled + Species, data = gd_remef_year_df)

m.pc1 <- lmer(pc1_remef_year ~ mpa.unscaled + Species + site + (1 | individual), data = pc1_remef_year_df)

m.pc1.nospp <- lmer(pc1_remef_year ~ mpa.unscaled + site + (1 | individual), data = pc1_remef_year_df)

m.pc1.linear <- lm(pc1_remef_year ~ mpa.unscaled + Species, data = pc1_remef_year_df)

tab_model(m.tti, m.tti.nospp, m.tti.linear, m.fd, m.fd.nospp, m.fd.linear, m.gd, m.gd.nospp, m.gd.linear, m.pc1, m.pc1.nospp, m.pc1.linear,
          show.reflvl = TRUE, 
          digits = 3, 
          show.aic = TRUE, 
          show.ci = FALSE,
          show.icc = FALSE,
          string.pred = "Coeffcient", 
         title = "species plot model selection",
         dv.labels = c('TTI Model', 'no spp', 'TTI, linear', 'FD Model', 'no spp', 'FD, linear', 'GD Model', 'no spp', 'GD, linear', 'PC1 Model', 'no spp', 'PC1, linear'),
  string.p = "P-Value", 
  p.style = "stars")
```

#### a. PC1
```{r}
pc1.mpa.spp <- ggplot(data= pc1_remef_year_df, aes(mpa.unscaled, pc1_remef_year, color = Species,shape = Species)) + 
  geom_point(size = 2, alpha = .65) + 
  geom_abline(slope = -0.118, intercept = -0.939, color = "#9D0208", cex = 1, alpha = 0.9) +
  geom_abline(slope = -0.118, intercept = -1.743, color = "#F48C06", cex = 1, alpha = 0.9) +
  labs(x="Water Potential (-Mpa)", y="Ignitability-Combustibility Proxy") +
  annotate('point', x = -5.411, y = -0.300502, color = "#9D0208", size = 5, alpha = 0.9) +
  annotate('point', x = -7.444, y = -0.864608, color = "#F48C06", size = 5, alpha = 0.9) +
  annotate('text', x = -4, y = -5.35, label = 'Species Difference: *** p < 0.001', size = 4.5) +
  geom_segment(aes(x = -9, xend = -9, y = -5.56, yend = -4.66), arrow = arrow(length = unit(0.15, 'inches')), color = 'black') +
  annotate('text', x = -9, y = -3.66, label = 'More \n Combustible \n & Ignitable', fontface = 2) +
  annotate('text', x = -9.7, y = 3.56, label = 'a', size = 12) +
  th +
  th_panels +
  theme(axis.text.x = element_blank(),
        axis.text.y = element_text(size = 14)) +
  color_panels
pc1.mpa.spp
```

#### b. Time to ignition
```{r}
tti.mpa.spp <- ggplot(data= tti_groups_ranis, aes(mpa.unscaled, r1_1, color = Species, shape = Species)) + 
  geom_point(size = 2, alpha = .65) + 
  geom_abline(slope = 1.443, intercept = 47.659, color = "#9D0208", cex = 1, alpha = 0.9) +
  geom_abline(slope = 1.443, intercept = 55.086, color = "#F48C06", cex = 1, alpha = 0.9) +
  annotate('point', x = -5.411, y = 39.85093, color = "#9D0208", size = 5, alpha = 0.9) +
  annotate('point', x = -7.444, y = 44.34431, color = "#F48C06", size = 5, alpha = 0.9) +
  annotate('text', x = -6.5, y = 15.5, label = 'Species Difference: *** p < 0.001', size = 4.5) +
  annotate('text', x = -9.7, y = 88.2, label = 'b', size = 12) +
  labs(x="Water Potential (-Mpa)", y="Time to Ignition")  +
  th +
  th_panels +
  theme(axis.text.x = element_blank(),
        axis.text.y = element_text(size = 14)) +
  color_panels
tti.mpa.spp
```

#### c. Flame duration
```{r}
fd.mpa.spp <- ggplot(data= fd_remef_year_df, aes(mpa.unscaled, fd_remef_year, color = Species,shape = Species)) + 
  geom_point(size = 2, alpha = .65) + 
  geom_abline(slope = -0.137, intercept = 7.137, color = "#9D0208", cex = 1, alpha = 0.9) +
  geom_abline(slope = -0.137, intercept = 4.99, color = "#F48C06", cex = 1, alpha = 0.9) +
  annotate('point', x = -5.411, y = 8.058307, color = "#9D0208", size = 5, alpha = 0.9) +
  annotate('point', x = -7.444, y = 6.009828, color = "#F48C06", size = 5, alpha = 0.9) +
  labs(x="Water Potential (-Mpa)", y="Flame Duration")  +
  annotate('text', x = -6.5, y = 0.1, label = 'Species Difference: *** p < 0.001', size = 4.5) +
  annotate('text', x = -9.7, y = 22.6, label = 'c', size = 12) +
  th +
  th_panels +
  theme(axis.text = element_text(size = 14)) +
  color_panels
fd.mpa.spp
```

#### d. Glow duration
```{r}
gd.mpa.spp <- ggplot(data= gd_remef_year_df, aes(mpa.unscaled, gd_remef_year, color = Species, shape = Species)) + 
  geom_point(size = 2, alpha = .65) + 
  geom_abline(slope = -1.090, intercept = 126.542, color = "#9D0208", cex = 1, alpha = 0.9) +
  geom_abline(slope = -1.090, intercept = 132.955, color = "#F48C06", cex = 1, alpha = 0.9) +
  annotate('point', x = -5.411, y = 132.44, color = "#9D0208", size = 5, alpha = 0.9) +
  annotate('point', x = -7.444, y = 141.069, color = "#F48C06", size = 5, alpha = 0.9) +
  labs(x="Water Potential (-Mpa)", y="Glow Duration")  +
  annotate('text', x = -6.35, y = 55, label = 'Species Difference: non-significant', size = 4.5) +
  annotate('text', x = -9.7, y = 216.5, label = 'd', size = 12) +
  th +
  th_panels +
  theme(axis.text = element_text(size = 14)) +
  color_panels
gd.mpa.spp
```

#### MAIN FIGURE: Arranging Above Plots
```{r}
legend <- get_legend(
  # create some space to the left of the legend
  ggplot() + 
  geom_point(data= scaled.mem.data.sept2020, aes(mpa.unscaled, PC1, color = Species, shape = Species), size = 5, alpha = .8) +
    color_panels +
    theme(legend.box.margin = margin(0, 0, 0, 0),
          legend.position = "top",
          legend.text = element_text(face = "italic", size = 18),
          legend.title = element_text(face = 'bold', size = 22),
          legend.key = element_rect(fill = "white"))
)

plot.spp <- cowplot::plot_grid(
        pc1.mpa.spp + theme(plot.margin = unit(c(0.25, 0.25, 0.25, 0.25), "cm")),
        tti.mpa.spp + theme(plot.margin = unit(c(0.25, 0.25, 0.25, 0.25), "cm")), 
        fd.mpa.spp + theme(plot.margin = unit(c(0.25, 0.25, 0.25, 0.25), "cm")), 
        gd.mpa.spp + theme(plot.margin = unit(c(0.25, 0.25, 0.25, 0.25), "cm")),
        align='vh', vjust=.5, scale = 1,
        ncol = 2)

#add to plot
scatterplot.spp <- gridExtra::grid.arrange(arrangeGrob(plot.spp, bottom = x.grob)) 
scatterplot.spp

plot_legend.spp <- cowplot::plot_grid (legend, scatterplot.spp, ncol = 1, rel_heights = c(1,12))
plot_legend.spp

ggsave(here("figures", 'main-figures', "Fig4_Species.Plot.jpg"),
       plot = plot_legend.spp, 
       width = 10, height = 10, dpi = 300)
```

#--------
## 3.1.0 LFM Species Effects (supplemental figure)
### Time to ignition
```{r}
lfm_mod_fig_tti <- lmer(tti ~ lfm.unscaled + spp + site + (1 | individual), data = scaled.mem.data.sept2020)

r1_1 <- remef(lfm_mod_fig_tti, fix = 'siteSt. Marys',  ran = "all", keep.intercept = TRUE)

scaled_r1_1_df <- as.data.frame(r1_1)
tti_groups_ranis <- cbind(scaled.mem.data.sept2020, scaled_r1_1_df)
```

### Flame duration
```{r}
lfm_mod_fig_fd <- lmer(fd ~ lfm.unscaled + spp + site + (1 | individual), data = scaled.mem.data.sept2020)

fd_remef_year <- remef(lfm_mod_fig_fd, fix = 'siteSt. Marys',  ran = "all", keep.intercept = TRUE)

scaled_r1_1_df <- as.data.frame(fd_remef_year)
fd_remef_year_df <- cbind(scaled.mem.data.sept2020, scaled_r1_1_df)
```

### Glow duration
```{r}
lfm_mod_fig_gd <- lmer(gd ~ lfm.unscaled + spp + site + (1 | individual), data = scaled.mem.data.sept2020)

gd_remef_year <- remef(lfm_mod_fig_gd, fix = 'siteSt. Marys',  ran = "all", keep.intercept = TRUE)

scaled_r1_1_df <- as.data.frame(gd_remef_year)
gd_remef_year_df <- cbind(scaled.mem.data.sept2020, scaled_r1_1_df)
```

### PC1
```{r}
lfm_mod_fig_pc1 <- lmer(PC1 ~ lfm.unscaled + spp + site + (1 | individual), data = scaled.mem.data.sept2020)

pc1_remef_year <- remef(lfm_mod_fig_pc1, fix = 'siteSt. Marys',  ran = "all", keep.intercept = TRUE)

scaled_r1_1_df <- as.data.frame(pc1_remef_year)
pc1_remef_year_df <- cbind(scaled.mem.data.sept2020, scaled_r1_1_df)
```

## 3.1.1 Visualization
### Models
```{r}
m.tti <- lmer(r1_1 ~ lfm.unscaled + Species + site + (1 | individual), data = tti_groups_ranis)

m.tti.nospp <- lmer(r1_1 ~ lfm.unscaled + site + (1 | individual), data = tti_groups_ranis)

m.tti.linear <- lm(r1_1 ~ lfm.unscaled + Species, data = tti_groups_ranis)

m.fd <- lmer(fd_remef_year ~ lfm.unscaled + Species + site + (1 | individual), data = fd_remef_year_df)

m.fd.nospp <- lmer(fd_remef_year ~ lfm.unscaled + site + (1 | individual), data = fd_remef_year_df)

m.fd.linear <- lm(fd_remef_year ~ lfm.unscaled + Species, data = fd_remef_year_df)

m.gd <- lmer(gd_remef_year ~ lfm.unscaled + Species + site + (1 | individual), data = gd_remef_year_df)

m.gd.nospp <- lmer(gd_remef_year ~ lfm.unscaled + site + (1 | individual), data = gd_remef_year_df)

m.gd.linear <- lm(gd_remef_year ~ lfm.unscaled + Species, data = gd_remef_year_df)

m.pc1 <- lmer(pc1_remef_year ~ lfm.unscaled + Species + site + (1 | individual), data = pc1_remef_year_df)

m.pc1.nospp <- lmer(pc1_remef_year ~ lfm.unscaled + site + (1 | individual), data = pc1_remef_year_df)

m.pc1.linear <- lm(pc1_remef_year ~ lfm.unscaled + Species, data = pc1_remef_year_df)

tab_model(m.tti, m.tti.nospp, m.tti.linear, m.fd, m.fd.nospp, m.fd.linear, m.gd, m.gd.nospp, m.gd.linear, m.pc1, m.pc1.nospp, m.pc1.linear,
          show.reflvl = TRUE, 
          digits = 3, 
          show.aic = TRUE, 
          show.ci = FALSE,
          show.icc = FALSE,
          string.pred = "Coeffcient", 
         title = "species plot model selection",
         dv.labels = c('TTI Model', 'no spp', 'TTI, linear', 'FD Model', 'no spp','FD, linear', 'GD Model', 'no spp', 'GD, linear', 'PC1 Model', 'no spp', 'PC1, linear'),
  string.p = "P-Value", 
  p.style = "stars")
```

### a. PC1
```{r}
pc1.lfm.mem <- ggplot(data= pc1_remef_year_df, aes(lfm.unscaled, pc1_remef_year, color = Species,shape = Species)) + 
  geom_point(size = 1.2, alpha = .6) + 
  geom_abline(slope = -0.015, intercept = 0.680, color = "#9D0208", cex = 1) +
  geom_abline(slope = -0.015, intercept = 0.002, color = "#F48C06", cex = 1) +
  annotate('point', x = 65.67, y = -0.30505, color = "#9D0208", size = 5, alpha = 0.9) +
  annotate('point', x = 59.55, y = -0.89125, color = "#F48C06", size = 5, alpha = 0.9) +
  labs(x="Live Fuel Moisture (%)", y="Ignitability-Combustibility Proxy") +
  annotate('text', x = 110, y = -4.8, label = 'Species Difference: ** p < 0.01', size = 4.5) +
  geom_segment(aes(x = 45, xend = 45, y = -5, yend = -4), arrow = arrow(length = unit(0.15, 'inches')), color = 'black') +
  annotate('text', x = 45, y = -3, label = 'More \n Combustible \n & Ignitable', fontface = 2) +
  annotate('text', x = 37, y = 3.7, label = 'a', size = 12) +
  th +
  th_panels +
  theme(axis.text.x = element_blank(),
        axis.text.y = element_text(size = 14)) +
  color_panels
pc1.lfm.mem
```

### b. Time to ignition
```{r}
tti.lfm.mem <- ggplot(data= tti_groups_ranis, aes(lfm.unscaled, r1_1, color = Species, shape = Species), xlim = c(0, 170)) + 
  geom_point(size = 1.2, alpha = .6) + 
 geom_abline(slope = 0.188, intercept = 27.670, color = "#9D0208", cex = 1) +
  geom_abline(slope = 0.188, intercept = 33.528, color = "#F48C06", cex = 1) +
  annotate('point', x = 65.67, y = 40.01596, color = "#9D0208", size = 5, alpha = 0.9) +
  annotate('point', x = 59.55, y = 44.7234, color = "#F48C06", size = 5, alpha = 0.9) +
  annotate('text', x = 75, y = 14, label = 'Species Difference: *** p < 0.001', size = 4.5) +
  annotate('text', x = 37, y = 85, label = 'b', size = 12) +
  labs(x="Live Fuel Moisture (%)", y="Time to Ignition")  +
  th +
  th_panels +
  theme(axis.text.x = element_blank(),
        axis.text.y = element_text(size = 14)) +
  color_panels
tti.lfm.mem
```

### c. Flame duration
```{r}
fd.lfm.mem <- ggplot(data= fd_remef_year_df, aes(lfm.unscaled, fd_remef_year, color = Species,shape = Species)) + 
  geom_point(size = 1.2, alpha = .6) + 
  geom_abline(slope = -0.011, intercept = 8.57, color = "#9D0208", cex = 1) +
  geom_abline(slope = -0.011, intercept = 6.517, color = "#F48C06", cex = 1) +
  annotate('point', x = 65.67, y = 7.84763, color = "#9D0208", size = 5, alpha = 0.9) +
  annotate('point', x = 59.55, y = 5.86195, color = "#F48C06", size = 5, alpha = 0.9) +
  labs(x="Live Fuel Moisture (%)", y="Flame Duration")  +
  annotate('text', x = 75, y = 0.95, label = 'Species Difference: *** p < 0.001', size = 4.5) +
  annotate('text', x = 37, y = 21.7, label = 'c', size = 12) +
  th +
  th_panels +
  theme(axis.text = element_text(size = 14)) +
  color_panels
fd.lfm.mem
```

### d. Glow duration
```{r}
gd.lfm.mem <- ggplot(data= gd_remef_year_df, aes(lfm.unscaled, gd_remef_year, color = Species,shape = Species)) + 
  geom_point(size = 1.2, alpha = .6) + 
  geom_abline(slope = -0.135, intercept = 141.131, color = "#9D0208", cex = 1) +
  geom_abline(slope = -0.135, intercept = 148.668, color = "#F48C06", cex = 1) +
  annotate('point', x = 65.67, y = 132.2655, color = "#9D0208", size = 5, alpha = 0.9) +
  annotate('point', x = 59.55, y = 140.6287, color = "#F48C06", size = 5, alpha = 0.9) +
  labs(x="Live Fuel Moisture (%)", y="Glow Duration")  +
  scale_y_continuous(limits = c(50,235), breaks = c(50, 100, 150, 200)) +
  annotate('text', x = 70, y = 55, label = 'Species Difference: * p < 0.05', size = 4.5) +
  annotate('text', x = 37, y = 225, label = 'd', size = 12) +
   th +
  th_panels +
  theme(axis.text = element_text(size = 14)) +
  color_panels
gd.lfm.mem
```

### Main Figure: Arranging plots
```{r}
legend <- get_legend(
  # create some space to the left of the legend
  ggplot() + 
  geom_point(data= scaled.mem.data.sept2020, aes(lfm.unscaled, PC1, color = Species, shape = Species), size = 5, alpha = .8) +
    color_panels +
    theme(legend.box.margin = margin(0, 0, 0, 0),
          legend.position = "top",
          legend.text = element_text(face = "italic", size = 18),
          legend.title = element_text(face = 'bold', size = 22),
          legend.key = element_rect(fill = "white"))
)

lfm.grob <- textGrob("Live Fuel Moisture (%)", 
                   gp=gpar(fontface="bold", 
                           col="black", 
                           fontsize=20)
                   )

plot.mem <- cowplot::plot_grid(
       pc1.lfm.mem + theme(plot.margin = unit(c(0.25, 0.25, 0.25, 0.25), "cm")),
       tti.lfm.mem + theme(plot.margin = unit(c(0.25, 0.25, 0.25, 0.25), "cm")), 
        fd.lfm.mem + theme(plot.margin = unit(c(0.25, 0.25, 0.25, 0.25), "cm")), 
        gd.lfm.mem + theme(plot.margin = unit(c(0.25, 0.25, 0.25, 0.25), "cm")), 
        align='vh', vjust=.5, scale = 1,
        ncol = 2
          )

#add to plot
scatterplot.mem <- gridExtra::grid.arrange(arrangeGrob(plot.mem, bottom = lfm.grob)) 
scatterplot.mem

plot_legend.mem <- cowplot::plot_grid(legend, scatterplot.mem, ncol = 1, rel_heights = c(1,12))
plot_legend.mem


ggsave(here("figures", 'supp-figures', "FigS4_LFM.species.plot.MEM.jpg"),
       plot = plot_legend.mem, 
       width = 10, height = 10, dpi = 300)
```


# ------
# LFM vs. Water Potential (raw values)
```{r}
ggplot(seg.data.subset.alldates, aes(y = lfm.unscaled, x = mpa)) +
  geom_point(fill = "#c1121f", alpha = 0.6) +
  facet_wrap(~spp, labeller = labeller(spp = spp.labs),
             scales = "free" ) +
  labs(x = "Water Potential (MPa)", y = "Live Fuel Moisture (%)") +
  theme(strip.text = element_text(face = "italic")) +
  th

ggsave(here('figures', 'extra-figures',
            'lfm.vs.mpa.raw.values.jpg'), height = 10, width = 18)
```
