==> scripts/4.0_segmented_rand_reg.Rmd <==
---
title: "Segmented Regressions - Mixed Effects Models"
author: "Indra Boving & Joe Celebrezze"
date: "4/15/2021"
output: html_document
---

```{r}
knitr::opts_chunk$set(echo = TRUE)
#lots of extra here, but oh well... 
library(ggplot2)
library(gapminder)
library(tidyverse)
library(lme4)
library(here)
#remove.packages('segmented')
library(remotes)
install_version('segmented', '1.5.0') # Because of how finnicky the function we wrote for visualization and difficulties of including random effects in segmented regressions (segmented.lme is in early stages and is currently incompatible with slope()), this specific version should be used
library(segmented)
library(nlme)
install_version('nlme', '3.1.157') # same comment as above
citation('nlme')
library(lubridate)
library(plotrix)
library(kableExtra)
filter = dplyr::filter 
here = here::here
```

# Reading in Dataframe
```{r}
seg.data.subset <- read.csv(here('processed-data', 'seg.reg.data.local.csv'))

seg.data.subset <- seg.data.subset %>% 
  na.omit()
```

# Color by Season
```{r}
seg.data.subset$color <- 'black'

seg.data.subset$color[seg.data.subset$year.month=='2020_September']="#c1121f"
seg.data.subset$color[seg.data.subset$year.month=='2020_January']='#669BBC'
seg.data.subset$color[seg.data.subset$year.month=='2019_December']="#b08ba5"
seg.data.subset$color[seg.data.subset$year.month=='2018_January']="#FF7262"
seg.data.subset$color[seg.data.subset$year.month=='2016_December']="#A4A7DE"
```

# Splitting by Species
```{r}
seg.ceme.subset <- seg.data.subset %>% 
  filter(spp == "CEME")
```

```{r}
seg.adfa.subset <- seg.data.subset %>% 
  filter(spp == "ADFA")
```

# Splitting by Sampling Date
```{r}
seg.ceme.s20.subset <- seg.ceme.subset %>% 
  filter(year.month == "2020_September")

seg.ceme.d19.subset <- seg.ceme.subset %>% 
  filter(year.month == "2019_December")

seg.adfa.s20.subset <- seg.adfa.subset %>% 
  filter(year.month == "2020_September")

seg.adfa.d16.subset <- seg.adfa.subset %>% 
  filter(year.month == "2016_December")
```


# Maximums for LFM, MPa
```{r}
max.lfm <- max(seg.data.subset$lfm)
max.mpa <- max(seg.data.subset$mpa)

max.lfm.c <- max(seg.ceme.subset$lfm)
max.mpa.c <- max(seg.ceme.subset$mpa)

max.lfm.a <- max(seg.adfa.subset$lfm)
max.mpa.a <- max(seg.adfa.subset$mpa)
```

# PV Curve Data
```{r}
pv_summary_df_timing <- read_csv(here("processed-data", "pv_summary_df_timing.csv"))

pv_local <- pv_summary_df_timing %>% 
  filter(spp %in% c("ADFA", "CEME"))
#ADFA
##LFM
mean_lfm_adfa <- pv_local %>% 
  filter(spp == "ADFA") %>% 
  summarise(mean(mean_lfm)) %>% 
  pull()
upr_lfm_adfa <- pv_local %>% 
  filter(spp == "ADFA") %>% 
  summarise(mean(upr_lfm)) %>% 
  pull()
lwr_lfm_adfa <- pv_local %>% 
  filter(spp == "ADFA") %>% 
  summarise(mean(lwr_lfm)) %>% 
  pull()
##MPa
mean_mpa_adfa <- pv_local %>% 
  filter(spp == "ADFA") %>% 
  summarise(mean(mean_tlp)) %>% 
  pull()
upr_mpa_adfa <- pv_local %>% 
  filter(spp == "ADFA") %>% 
  summarise(mean(upr)) %>% 
  pull()
lwr_mpa_adfa <- pv_local %>% 
  filter(spp == "ADFA") %>% 
  summarise(mean(lwr)) %>% 
  pull()
# CEME
##LFM
mean_lfm_ceme <- pv_local %>% 
  filter(spp == "CEME") %>% 
  summarise(mean(mean_lfm)) %>% 
  pull()
upr_lfm_ceme <- pv_local %>% 
  filter(spp == "CEME") %>% 
  summarise(mean(upr_lfm)) %>% 
  pull()
lwr_lfm_ceme <- pv_local %>% 
  filter(spp == "CEME") %>% 
  summarise(mean(lwr_lfm)) %>% 
  pull()
##MPa
mean_mpa_ceme <- pv_local %>% 
  filter(spp == "CEME") %>% 
  summarise(mean(mean_tlp)) %>% 
  pull()
upr_mpa_ceme <- pv_local %>% 
  filter(spp == "CEME") %>% 
  summarise(mean(upr)) %>% 
  pull()
lwr_mpa_ceme <- pv_local %>% 
  filter(spp == "CEME") %>% 
  summarise(mean(lwr)) %>% 
  pull()
```

# Function to plot mixed effects model segmented regressions

This took a while to get running, but I think it works well now. I had to use base plot to get it to work. Note: x.variable and y.variable must be inputted as character strings.

I split the function into two so that we could customize the plots for LFM and MPa separately and also make the lines of code shorter and neater by involving more default values for arguments in the function

```{r}
rsegplot.lfm <- function(model, dataframe, species, x.variable = 'lfm', y.variable, max.x = max.lfm, x.lab = 'Live Fuel Moisture (%)', y.lab = NULL, tlp = NULL, tlp.upper = NULL, tlp.lower = NULL, color = 'color') {
  
p <- summary.segmented(model)$psi[2]
m <- slope(model, .coef = fixef(model))[[1]][[1]]
b <- intercept(model, .coef = fixef(model))[[1]][[1]]
m2 <- slope(model, .coef = fixef(model))[[1]][[2]]
b2 <- intercept(model, .coef = fixef(model))[[1]][[2]]
yvar <- dataframe[,y.variable]
xvar <- dataframe[,x.variable]
colr <- dataframe[,color]
confint <- confint.segmented(model, x.variable, .coef=fixef(model))
lower <- confint[2]
upper <- confint[3]
ymax <- max(yvar)
ymin <- min(yvar)

wacky <- plot(xvar, yvar, xlim = c(0, max.x), ylim = c(ymin-0.05*ymin, ymax+0.05*ymax), 
              xlab = x.lab, ylab = y.lab,
              pch = 16, col = alpha(colr, 0.8),
              cex.lab = 1.3)
polygon(x = c(lower, upper, upper, lower), y = c(ymax + 0.2*ymax, ymax +0.2*ymax, ymin - 30, ymin - 30), col = alpha('gray', 0.3), lty = 2)
abline(v = tlp, col = alpha("#64A400", 0.7), cex = 3)
polygon(x = c(tlp.lower, tlp.upper, tlp.upper, tlp.lower), y = c(ymax +0.2*ymax, ymax +0.2*ymax, ymin -30, ymin -30), col = alpha("#64A400", 0.3), border = alpha("#64A400", 0.6))
legend('topleft', legend=if(species == 'ADFA'){
  c("September 2020", "January 2018", "December 2016")
}
else if(species == 'CEME'){
  c("September 2020", "January 2020", "December 2019")
},
       col=if(species == "ADFA") {
         c("#c1121f","#FF7262","#A4A7DE")
         }
        else if(species == "CEME") {
        c("#c1121f",'#669BBC',"#b08ba5")
        }, pch = 16, cex=1)
for(z in 1:length(xvar)){
  if(z < p){
    segments(x0 = 0, x1 = p, y0 = b, y1 = m*((b2 - b)/(m - m2))+b, lwd = 3, col = "black")
  }else{
    segments(x0 = p, x1 = max.x, y0 = m2*((b - b2)/(m2 - m))+b2 , y1 = m2*max.x + b2, lwd = 3, col = "black")}}
return(wacky)
}
```

```{r}
rsegplot.mpa <- function(model, dataframe, species, x.variable = 'mpa', y.variable, max.x = max.mpa, x.lab = 'Water Potential (MPa)', y.lab = NULL, tlp = NULL, tlp.upper = NULL, tlp.lower = NULL, color = 'color') 
  {
p <- summary.segmented(model)$psi[2]
m <- slope(model, .coef = fixef(model))[[1]][[1]]
b <- intercept(model, .coef = fixef(model))[[1]][[1]]
m2 <- slope(model, .coef = fixef(model))[[1]][[2]]
b2 <- intercept(model, .coef = fixef(model))[[1]][[2]]
yvar <- dataframe[,y.variable]
xvar <- dataframe[,x.variable]
colr <- dataframe[,color]
confint <- confint.segmented(model, x.variable, .coef=fixef(model))
lower <- confint[2]
upper <- confint[3]
ymax <- max(yvar)
ymin <- min(yvar)

wacky <- plot(xvar, yvar, xlim = c(-10, 0), ylim = c(ymin-1, ymax+1), 
              xlab = x.lab, ylab = y.lab,
              pch = 16, col = colr, cex.lab = 1.1)
polygon(x = c(lower, upper, upper, lower), y = c(ymax +1.5, ymax +1.5, ymin -1.5, ymin -1.5), col = alpha('gray', 0.3), lty = 2)
abline(v = tlp, col = alpha("#64A400", 0.8), cex = 3)
polygon(x = c(tlp.lower, tlp.upper, tlp.upper, tlp.lower), y = c(ymax +1.5, ymax +1.5, ymin -1.5, ymin -1.5), col = alpha("#64A400", 0.3), border = alpha("#64A400", 0.7))
abline(a = b, b = m, lwd = 3, col = "black")
legend(-2, ymax+0.05*ymax, legend=if(species == 'ADFA'){
  c("September 2020", "January 2018", "December 2016")
}
else if(species == 'CEME'){
  c("September 2020", "January 2020", "December 2019")
},
       col=if(species == "ADFA") {
         c("#c1121f","#FF7262","#A4A7DE")
         }
        else if(species == "CEME") {
        c("#c1121f",'#669BBC',"#b08ba5")
        }, pch = 16, cex=0.8)
for(j in 1:length(xvar)){
  if(j < p){
    segments(x0 = -10, x1 = p, y0 = (-10*m) + b , y1 = m2*((b - b2)/(m2 - m))+b2, lwd = 3, col = "black")
  }
  else{
    segments(x0 = p, x1 = max.x, y0 = m2*((b - b2)/(m2 - m))+b2, y1 = b2, lwd = 3, col = "black")
  }
    }
return(wacky)
}
```

For splitting by sampling date: no legend!
```{r}
rsegplot.lfm.nolegend <- function(model, dataframe, species, x.variable = 'lfm', y.variable, max.x = max.lfm, x.lab = 'Live Fuel Moisture (%)', y.lab = NULL, tlp = NULL, tlp.upper = NULL, tlp.lower = NULL, color = 'color') {
  
p <- summary.segmented(model)$psi[2]
m <- slope(model, .coef = fixef(model))[[1]][[1]]
b <- intercept(model, .coef = fixef(model))[[1]][[1]]
m2 <- slope(model, .coef = fixef(model))[[1]][[2]]
b2 <- intercept(model, .coef = fixef(model))[[1]][[2]]
yvar <- dataframe[,y.variable]
xvar <- dataframe[,x.variable]
colr <- dataframe[,color]
confint <- confint.segmented(model, x.variable, .coef=fixef(model))
lower <- confint[2]
upper <- confint[3]
ymax <- max(yvar)
ymin <- min(yvar)

wacky <- plot(xvar, yvar, xlim = c(0, max.x), ylim = c(ymin-0.05*ymin, ymax+0.05*ymax), 
              xlab = x.lab, ylab = y.lab,
              pch = 16, col = alpha(colr, 0.8),
              cex.lab = 1.3)
polygon(x = c(lower, upper, upper, lower), y = c(ymax + 0.2*ymax, ymax +0.2*ymax, ymin - 30, ymin - 30), col = alpha('gray', 0.3), lty = 2)
abline(v = tlp, col = alpha("#64A400", 0.7), cex = 3)
polygon(x = c(tlp.lower, tlp.upper, tlp.upper, tlp.lower), y = c(ymax +0.2*ymax, ymax +0.2*ymax, ymin -30, ymin -30), col = alpha("#64A400", 0.3), border = alpha("#64A400", 0.6))
for(z in 1:length(xvar)){
  if(z < p){
    segments(x0 = 0, x1 = p, y0 = b, y1 = m*((b2 - b)/(m - m2))+b, lwd = 3, col = "black")
  }else{
    segments(x0 = p, x1 = max.x, y0 = m2*((b - b2)/(m2 - m))+b2 , y1 = m2*max.x + b2, lwd = 3, col = "black")}}
return(wacky)
}
```

#---------------------------------
# CEME
## LFM
Scaling LFM (to try it out)
```{r}
seg.ceme.subset <- seg.ceme.subset %>% 
  mutate(lfm.scaled = scale(lfm))
```

### PC1 (Scaled LFM)
```{r}
o.mod<-lme(PC1~lfm.scaled + year.month + site, random=~1|id, data=seg.ceme.subset)
os.mod<-segmented.default(o.mod, ~lfm.scaled, npsi=list(lfm.scaled=1))
summary.segmented(os.mod)
rsegplot.lfm(os.mod, seg.ceme.subset, 'CEME', x.variable = "lfm.scaled", y.variable = 'PC1',y.lab = "PC1", tlp = mean_lfm_ceme, tlp.lower = lwr_lfm_ceme, tlp.upper = upr_lfm_ceme, max.x = 3)
```


### PC1
```{r}
o.mod.a<-lme(PC1~lfm + year.month + site, random=~1|id, data=seg.ceme.subset)
os.mod.a<-segmented.default(o.mod.a, ~lfm, npsi=list(lfm=1))
#summarizing results (note the '.coef' argument)
#m <- slope(os.mod.a, .coef = fixef(os.mod.a))[[1]][[1]]
summary.segmented(os.mod.a)

rsegplot.lfm(os.mod.a, seg.ceme.subset, 'CEME', y.variable = 'PC1',y.lab = "PC1", tlp = mean_lfm_ceme, tlp.lower = lwr_lfm_ceme, tlp.upper = upr_lfm_ceme)
```

### PC2
```{r}
o.mod<-lme(PC2~lfm + year.month + site, random=~1|id, data=seg.ceme.subset)
os.mod<-segmented.default(o.mod, ~lfm, npsi=list(lfm=1))
#summarizing results (note the '.coef' argument)
summary.segmented(os.mod)

rsegplot.lfm(os.mod, seg.ceme.subset, 'CEME', y.variable = 'PC2',y.lab = "PC2", tlp = mean_lfm_ceme, tlp.lower = lwr_lfm_ceme, tlp.upper = upr_lfm_ceme)
```
 
Ignitability Metrics:
### Time to Ignition
```{r}
o.mod<-lme(tti~lfm + year.month + site, random=~1|id, data=seg.ceme.subset)
os.mod.b<-segmented.default(o.mod, ~lfm, npsi=list(lfm=1))
#summarizing results (note the '.coef' argument)
summary.segmented(os.mod.b)

rsegplot.lfm(os.mod.b, seg.ceme.subset, 'CEME', y.variable = 'tti',y.lab = "Time to Ignition (s)", tlp = mean_lfm_ceme, tlp.lower = lwr_lfm_ceme, tlp.upper = upr_lfm_ceme)
```

### Prop. Ignite (5%)
```{r}
o.mod<-lme(prop.ignite.bins5_nodate~lfm + year.month + site, random=~1|id, data=seg.ceme.subset)
os.mod<-segmented.default(o.mod, ~lfm, npsi=list(lfm=1))
#summarizing results (note the '.coef' argument)
summary.segmented(os.mod)

rsegplot.lfm(os.mod, seg.ceme.subset, 'CEME', y.variable = 'prop.ignite.bins5_nodate',y.lab = "Proportion Ignted (5% Bins of LFM)", tlp = mean_lfm_ceme, tlp.lower = lwr_lfm_ceme, tlp.upper = upr_lfm_ceme)
```

### Glow to Ignition
```{r}
o.mod<-lme(gti~lfm + year.month + site, random=~1|id, data=seg.ceme.subset)
os.mod<-segmented.default(o.mod, ~lfm, npsi=list(lfm=1))
#summarizing results (note the '.coef' argument)
summary.segmented(os.mod)

rsegplot.lfm(os.mod, seg.ceme.subset, 'CEME', y.variable = 'gti',y.lab = "Glow to Ignition (s)", tlp = mean_lfm_ceme, tlp.lower = lwr_lfm_ceme, tlp.upper = upr_lfm_ceme)
```

Combustibility Metrics:
### Flame Height
```{r}
o.mod<-lme(fh~lfm + year.month + site, random=~1|id, data=seg.ceme.subset)
os.mod<-segmented.default(o.mod, ~lfm, npsi=list(lfm=1))
#summarizing results (note the '.coef' argument)
summary.segmented(os.mod)

rsegplot.lfm(os.mod, seg.ceme.subset, 'CEME', y.variable = 'fh',y.lab = "Flame Height (cm)", tlp = mean_lfm_ceme, tlp.lower = lwr_lfm_ceme, tlp.upper = upr_lfm_ceme)
```

### Max. Temp.
```{r}
o.mod<-lme(temp.max~lfm + year.month + site, random=~1|id, data=seg.ceme.subset)
os.mod<-segmented.default(o.mod, ~lfm, npsi=list(lfm=1))
#summarizing results (note the '.coef' argument)
summary.segmented(os.mod)

rsegplot.lfm(os.mod, seg.ceme.subset, 'CEME', y.variable = 'temp.max',y.lab = "Maximum Temperature (C)", tlp = mean_lfm_ceme, tlp.lower = lwr_lfm_ceme, tlp.upper = upr_lfm_ceme)
```

Sustainability Metrics:
### Flame Duration
```{r}
o.mod<-lme(fd~lfm + year.month + site, random=~1|id, data=seg.ceme.subset)
os.mod.f<-segmented.default(o.mod, ~lfm, npsi=list(lfm=1))
#summarizing results (note the '.coef' argument)
summary.segmented(os.mod.f)

rsegplot.lfm(os.mod, seg.ceme.subset, 'CEME', y.variable = 'fd',y.lab = "Flame Duration (s)", tlp = mean_lfm_ceme, tlp.lower = lwr_lfm_ceme, tlp.upper = upr_lfm_ceme)
```

Consumability Metrics:
### Glow Duration
```{r}
o.mod<-lme(gd~lfm + year.month + site, random=~1|id, data=seg.ceme.subset)
os.mod.h<-segmented.default(o.mod, ~lfm, npsi=list(lfm=1))
#summarizing results (note the '.coef' argument)
summary.segmented(os.mod.h)

rsegplot.lfm(os.mod, seg.ceme.subset, 'CEME', y.variable = 'gd',y.lab = "Glow Duration (s)", tlp = mean_lfm_ceme, tlp.lower = lwr_lfm_ceme, tlp.upper = upr_lfm_ceme)
```

### Post-Flame Glow
```{r}
o.mod<-lme(pfg~lfm + year.month + site, random=~1|id, data=seg.ceme.subset)
os.mod<-segmented.default(o.mod, ~lfm, npsi=list(lfm=1))
#summarizing results (note the '.coef' argument)
summary.segmented(os.mod)

rsegplot.lfm(os.mod, seg.ceme.subset, 'CEME', y.variable = 'pfg',y.lab = "Post-Flame Glow (s)", tlp = mean_lfm_ceme, tlp.lower = lwr_lfm_ceme, tlp.upper = upr_lfm_ceme)
```

### Time to First Glow
```{r}
o.mod<-lme(ttfg~lfm + year.month + site, random=~1|id, data=seg.ceme.subset)
os.mod<-segmented.default(o.mod, ~lfm, npsi=list(lfm=1))
#summarizing results (note the '.coef' argument)
summary.segmented(os.mod)

confint.segmented(os.mod, 'lfm', .coef=fixef(os.mod))

rsegplot.lfm(os.mod, seg.ceme.subset, 'CEME', y.variable = 'ttfg',y.lab = "Time to First Glow (s)", tlp = mean_lfm_ceme, tlp.lower = lwr_lfm_ceme, tlp.upper = upr_lfm_ceme)
```
# -------
## Water Potential
```{r}
seg.ceme.subset <- seg.ceme.subset %>% 
  mutate(mpa.scaled = scale(mpa)) # Scaling mpa
```

### PC1 (scaled)
```{r}
o.mod<-lme(PC1~mpa.scaled + year.month + site, random=~1|id, data=seg.ceme.subset)
os.mod<-segmented.default(o.mod, ~mpa.scaled, npsi=list(mpa.scaled=1))
#summarizing results (note the '.coef' argument)
summary.segmented(os.mod)

rsegplot.mpa(os.mod, seg.ceme.subset, 'CEME',x.variable = 'mpa.scaled', y.variable = 'PC1',y.lab = "PC1", tlp = mean_mpa_ceme, tlp.lower = lwr_mpa_ceme, tlp.upper = upr_mpa_ceme, max.x = 4)
```

### PC1
```{r}
o.mod<-lme(PC1~mpa + year.month + site, random=~1|id, data=seg.ceme.subset)
os.mod.a1<-segmented.default(o.mod, ~mpa, npsi=list(mpa=1))
#summarizing results (note the '.coef' argument)
summary.segmented(os.mod.a1)

rsegplot.mpa(os.mod.a1, seg.ceme.subset, 'CEME', y.variable = 'PC1',y.lab = "PC1", tlp = mean_mpa_ceme, tlp.lower = lwr_mpa_ceme, tlp.upper = upr_mpa_ceme)
```

### PC2
```{r}
o.mod<-lme(PC2~mpa + year.month + site, random=~1|id, data=seg.ceme.subset)
os.mod<-segmented.default(o.mod, ~mpa, npsi=list(mpa=1))
#summarizing results (note the '.coef' argument)
summary.segmented(os.mod)

rsegplot.mpa(os.mod, seg.ceme.subset, 'CEME', y.variable = 'PC2',y.lab = "PC2", tlp = mean_mpa_ceme, tlp.lower = lwr_mpa_ceme, tlp.upper = upr_mpa_ceme)
```
 
Ignitability Metrics:
### Time to Ignition
```{r}
o.mod<-lme(tti~mpa + year.month + site, random=~1|id, data=seg.ceme.subset)
os.mod.b1<-segmented.default(o.mod, ~mpa, npsi=list(mpa=1))
#summarizing results (note the '.coef' argument)
summary.segmented(os.mod.b1)

rsegplot.mpa(os.mod, seg.ceme.subset, 'CEME', y.variable = 'tti',y.lab = "Time to Ignition (s)", tlp = mean_mpa_ceme, tlp.lower = lwr_mpa_ceme, tlp.upper = upr_mpa_ceme)
```

### Prop. Ignite (5%)
```{r}
o.mod<-lme(prop.ignite.bins5_nodate~mpa + year.month + site, random=~1|id, data=seg.ceme.subset)
os.mod<-segmented.default(o.mod, ~mpa, npsi=list(mpa=1))
#summarizing results (note the '.coef' argument)
summary.segmented(os.mod)

rsegplot.mpa(os.mod, seg.ceme.subset, 'CEME', y.variable = 'prop.ignite.bins5_nodate',y.lab = "Proportion Ignted (5% Bins of LFM)", tlp = mean_mpa_ceme, tlp.lower = lwr_mpa_ceme, tlp.upper = upr_mpa_ceme)
```

### Glow to Ignition
```{r}
o.mod<-lme(gti~mpa + year.month + site, random=~1|id, data=seg.ceme.subset)
os.mod<-segmented.default(o.mod, ~mpa, npsi=list(mpa=1))
#summarizing results (note the '.coef' argument)
summary.segmented(os.mod)

rsegplot.mpa(os.mod, seg.ceme.subset, 'CEME', y.variable = 'gti',y.lab = "Glow to Ignition (s)", tlp = mean_mpa_ceme, tlp.lower = lwr_mpa_ceme, tlp.upper = upr_mpa_ceme)
```

Combustibility Metrics:
### Flame Height
```{r}
o.mod<-lme(fh~mpa + year.month + site, random=~1|id, data=seg.ceme.subset)
os.mod<-segmented.default(o.mod, ~mpa, npsi=list(mpa=1))
#summarizing results (note the '.coef' argument)
summary.segmented(os.mod)

rsegplot.mpa(os.mod, seg.ceme.subset, 'CEME', y.variable = 'fh',y.lab = "Flame Height (cm)", tlp = mean_mpa_ceme, tlp.lower = lwr_mpa_ceme, tlp.upper = upr_mpa_ceme)
```

### Max. Temp.
```{r}
o.mod<-lme(temp.max~mpa + year.month + site, random=~1|id, data=seg.ceme.subset)
os.mod<-segmented.default(o.mod, ~mpa, npsi=list(mpa=1))
#summarizing results (note the '.coef' argument)
summary.segmented(os.mod)

rsegplot.mpa(os.mod, seg.ceme.subset, 'CEME', y.variable = 'temp.max',y.lab = "Maximum Temperature (C)", tlp = mean_mpa_ceme, tlp.lower = lwr_mpa_ceme, tlp.upper = upr_mpa_ceme)
```

Sustainability Metrics:
### Flame Duration
```{r}
o.mod<-lme(fd~mpa + year.month + site, random=~1|id, data=seg.ceme.subset)
os.mod.f1<-segmented.default(o.mod, ~mpa, npsi=list(mpa=1))
#summarizing results (note the '.coef' argument)
summary.segmented(os.mod.f1)

rsegplot.mpa(os.mod, seg.ceme.subset, 'CEME', y.variable = 'fd',y.lab = "Flame Duration (s)", tlp = mean_mpa_ceme, tlp.lower = lwr_mpa_ceme, tlp.upper = upr_mpa_ceme)
```

Consumability Metrics:
### Glow Duration
```{r}
o.mod<-lme(gd~mpa + year.month + site, random=~1|id, data=seg.ceme.subset)
os.mod.h1<-segmented.default(o.mod, ~mpa, npsi=list(mpa=1))
#summarizing results (note the '.coef' argument)
summary.segmented(os.mod.h1)

rsegplot.mpa(os.mod, seg.ceme.subset, 'CEME', y.variable = 'gd',y.lab = "Glow Duration (s)", tlp = mean_mpa_ceme, tlp.lower = lwr_mpa_ceme, tlp.upper = upr_mpa_ceme)
```

### Post-Flame Glow
```{r}
o.mod<-lme(pfg~mpa + year.month + site, random=~1|id, data=seg.ceme.subset)
os.mod<-segmented.default(o.mod, ~mpa, npsi=list(mpa=1))
#summarizing results (note the '.coef' argument)
summary.segmented(os.mod)

rsegplot.mpa(os.mod, seg.ceme.subset, 'CEME', y.variable = 'pfg',y.lab = "Post-Flame Glow (s)", tlp = mean_mpa_ceme, tlp.lower = lwr_mpa_ceme, tlp.upper = upr_mpa_ceme)
```

### Time to First Glow
SIGNIFICANT 0.179 (-5.5897)
```{r}
o.mod<-lme(ttfg~mpa + year.month + site, random=~1|id, data=seg.ceme.subset)
os.mod<-segmented.default(o.mod, ~mpa, npsi=list(mpa=1))
#summarizing results (note the '.coef' argument)
summary.segmented(os.mod)

rsegplot.mpa(os.mod, seg.ceme.subset, 'CEME', y.variable = 'ttfg',y.lab = "Time to First Glow (s)", tlp = mean_mpa_ceme, tlp.lower = lwr_mpa_ceme, tlp.upper = upr_mpa_ceme)
```

#---------------------------------
# ADFA
## LFM
### PC1
```{r}
o.mod<-lme(PC1~ lfm + year.month + site, random=~1|id, data=seg.adfa.subset)
os.mod.c<-segmented.default(o.mod, ~lfm, npsi=list(lfm=1))
#summarizing results (note the '.coef' argument)
summary.segmented(os.mod.c)

rsegplot.lfm(os.mod.c, seg.adfa.subset, 'ADFA', y.variable = 'PC1',y.lab = "PC1", tlp = mean_lfm_adfa, tlp.lower = lwr_lfm_adfa, tlp.upper = upr_lfm_adfa, max.x = max.lfm.a)
```

### PC2
```{r}
o.mod<-lme(PC2~lfm + year.month + site, random=~1|id, data=seg.adfa.subset)
os.mod<-segmented.default(o.mod, ~lfm, npsi=list(lfm=1))
#summarizing results (note the '.coef' argument)
summary.segmented(os.mod)

rsegplot.lfm(os.mod, seg.adfa.subset, 'ADFA', y.variable = 'PC2',y.lab = "PC2", tlp = mean_lfm_adfa, tlp.lower = lwr_lfm_adfa, tlp.upper = upr_lfm_adfa, max.x = max.lfm.a)
```
 
Ignitability Metrics:
### Time to Ignition
```{r}
o.mod<-lme(tti~lfm + year.month + site, random=~1|id, data=seg.adfa.subset)
os.mod.d<-segmented.default(o.mod, ~lfm, npsi=list(lfm=1))
#summarizing results (note the '.coef' argument)
summary.segmented(os.mod.d)

rsegplot.lfm(os.mod.d, seg.adfa.subset, 'ADFA', y.variable = 'tti',y.lab = "Time to Ignition (s)", tlp = mean_lfm_adfa, tlp.lower = lwr_lfm_adfa, tlp.upper = upr_lfm_adfa, max.x = max.lfm.a)
```

### Prop. Ignite (5%)
```{r}
o.mod<-lme(prop.ignite.bins5_nodate~lfm + year.month + site, random=~1|id, data=seg.adfa.subset)
os.mod<-segmented.default(o.mod, ~lfm, npsi=list(lfm=1))
#summarizing results (note the '.coef' argument)
summary.segmented(os.mod)

rsegplot.lfm(os.mod, seg.adfa.subset, 'ADFA', y.variable = 'prop.ignite.bins5_nodate',y.lab = "Proportion Ignted (5% Bins of LFM)", tlp = mean_lfm_adfa, tlp.lower = lwr_lfm_adfa, tlp.upper = upr_lfm_adfa, max.x = max.lfm.a)
```

### Glow to Ignition
```{r}
o.mod<-lme(gti~lfm + year.month + site, random=~1|id, data=seg.adfa.subset)
os.mod<-segmented.default(o.mod, ~lfm, npsi=list(lfm=1))
#summarizing results (note the '.coef' argument)
summary.segmented(os.mod)

rsegplot.lfm(os.mod, seg.adfa.subset, 'ADFA', y.variable = 'gti',y.lab = "Glow to Ignition (s)", tlp = mean_lfm_adfa, tlp.lower = lwr_lfm_adfa, tlp.upper = upr_lfm_adfa, max.x = max.lfm.a)
```

Combustibility Metrics:
### Flame Height
```{r}
o.mod<-lme(fh~lfm + year.month + site, random=~1|id, data=seg.adfa.subset)
os.mod<-segmented.default(o.mod, ~lfm, npsi=list(lfm=1))
#summarizing results (note the '.coef' argument)
summary.segmented(os.mod)

rsegplot.lfm(os.mod, seg.adfa.subset, 'ADFA', y.variable = 'fh',y.lab = "Flame Height (cm)", tlp = mean_lfm_adfa, tlp.lower = lwr_lfm_adfa, tlp.upper = upr_lfm_adfa, max.x = max.lfm.a)
```

### Max. Temp.
```{r}
o.mod<-lme(temp.max~lfm + year.month + site, random=~1|id, data=seg.adfa.subset)
os.mod<-segmented.default(o.mod, ~lfm, npsi=list(lfm=1))
#summarizing results (note the '.coef' argument)
summary.segmented(os.mod)

rsegplot.lfm(os.mod, seg.adfa.subset, 'ADFA', y.variable = 'temp.max',y.lab = "Maximum Temperature (C)", tlp = mean_lfm_adfa, tlp.lower = lwr_lfm_adfa, tlp.upper = upr_lfm_adfa)
```

Sustainability Metrics:
### Flame Duration
```{r}
o.mod<-lme(fd~lfm + year.month + site, random=~1|id, data=seg.adfa.subset)
os.mod.e<-segmented.default(o.mod, ~lfm, npsi=list(lfm=1))
#summarizing results (note the '.coef' argument)
summary.segmented(os.mod.e)

rsegplot.lfm(os.mod, seg.adfa.subset, 'ADFA', y.variable = 'fd',y.lab = "Flame Duration (s)", tlp = mean_lfm_adfa, tlp.lower = lwr_lfm_adfa, tlp.upper = upr_lfm_adfa)
```

Consumability Metrics:
### Glow Duration
```{r}
o.mod<-lme(gd~lfm + year.month + site, random=~1|id, data=seg.adfa.subset)
os.mod.g<-segmented.default(o.mod, ~lfm, npsi=list(lfm=1))
#summarizing results (note the '.coef' argument)
summary.segmented(os.mod.g)

rsegplot.lfm(os.mod, seg.adfa.subset, 'ADFA', y.variable = 'gd',y.lab = "Glow Duration (s)", tlp = mean_lfm_adfa, tlp.lower = lwr_lfm_adfa, tlp.upper = upr_lfm_adfa)
```

### Post-Flame Glow
```{r}
o.mod<-lme(pfg~lfm + year.month + site, random=~1|id, data=seg.adfa.subset)
os.mod<-segmented.default(o.mod, ~lfm, npsi=list(lfm=1))
#summarizing results (note the '.coef' argument)
summary.segmented(os.mod)

rsegplot.lfm(os.mod, seg.adfa.subset, 'ADFA', y.variable = 'pfg',y.lab = "Post-Flame Glow (s)", tlp = mean_lfm_adfa, tlp.lower = lwr_lfm_adfa, tlp.upper = upr_lfm_adfa)
```

### Time to First Glow
```{r}
o.mod<-lme(ttfg~lfm + year.month + site, random=~1|id, data=seg.adfa.subset)
os.mod<-segmented.default(o.mod, ~lfm, npsi=list(lfm=1))
#summarizing results (note the '.coef' argument)
summary.segmented(os.mod)

rsegplot.lfm(os.mod, seg.adfa.subset, 'ADFA', y.variable = 'ttfg',y.lab = "Time to First Glow (s)", tlp = mean_lfm_adfa, tlp.lower = lwr_lfm_adfa, tlp.upper = upr_lfm_adfa)
```

# ------
## Water Potential

### PC1
```{r}
o.mod<-lme(PC1~mpa + year.month + site, random=~1|id, data=seg.adfa.subset)
os.mod.c1<-segmented.default(o.mod, ~mpa, npsi=list(mpa=1))
#summarizing results (note the '.coef' argument)
summary.segmented(os.mod.c1)

rsegplot.mpa(os.mod.c1, seg.adfa.subset, 'ADFA', y.variable = 'PC1',y.lab = "PC1", tlp = mean_mpa_adfa, tlp.lower = lwr_mpa_adfa, tlp.upper = upr_mpa_adfa)
```

### PC2
```{r}
o.mod<-lme(PC2~mpa + year.month + site, random=~1|id, data=seg.adfa.subset)
os.mod<-segmented.default(o.mod, ~mpa, npsi=list(mpa=1))
#summarizing results (note the '.coef' argument)
summary.segmented(os.mod)

rsegplot.mpa(os.mod, seg.adfa.subset, 'ADFA', y.variable = 'PC2',y.lab = "PC2", tlp = mean_mpa_adfa, tlp.lower = lwr_mpa_adfa, tlp.upper = upr_mpa_adfa)
```
 
Ignitability Metrics:
### Time to Ignition
```{r}
o.mod<-lme(tti~mpa + year.month + site, random=~1|id, data=seg.adfa.subset)
os.mod.d1<-segmented.default(o.mod, ~mpa, npsi=list(mpa=1))
#summarizing results (note the '.coef' argument)
summary.segmented(os.mod.d1)

rsegplot.mpa(os.mod, seg.adfa.subset, 'ADFA', y.variable = 'tti',y.lab = "Time to Ignition (s)", tlp = mean_mpa_adfa, tlp.lower = lwr_mpa_adfa, tlp.upper = upr_mpa_adfa)
```

### Prop. Ignite (5%)
```{r}
o.mod<-lme(prop.ignite.bins5_nodate~mpa + year.month + site, random=~1|id, data=seg.adfa.subset)
os.mod<-segmented.default(o.mod, ~mpa, npsi=list(mpa=1))
#summarizing results (note the '.coef' argument)
summary.segmented(os.mod)

rsegplot.mpa(os.mod, seg.adfa.subset, 'ADFA', y.variable = 'prop.ignite.bins5_nodate',y.lab = "Proportion Ignted (5% Bins of LFM)", tlp = mean_mpa_adfa, tlp.lower = lwr_mpa_adfa, tlp.upper = upr_mpa_adfa)
```

### Glow to Ignition
```{r}
o.mod<-lme(gti~mpa + year.month + site, random=~1|id, data=seg.adfa.subset)
os.mod<-segmented.default(o.mod, ~mpa, npsi=list(mpa=1))
#summarizing results (note the '.coef' argument)
summary.segmented(os.mod)

plot.segmented(os.mod, "mpa", .coef=fixef(os.mod), conf.level=.95)

rsegplot.mpa(os.mod, seg.adfa.subset, 'ADFA', y.variable = 'gti',y.lab = "Glow to Ignition (s)", tlp = mean_mpa_adfa, tlp.lower = lwr_mpa_adfa, tlp.upper = upr_mpa_adfa)
```

Combustibility Metrics:
### Flame Height
No threshold able to be identified (segmented.default returns error)
```{r}
o.mod<-lme(fh~mpa + year.month + site, random=~1|id, data=seg.adfa.subset)
os.mod<-segmented.default(o.mod, ~mpa, npsi=list(mpa=1))
#summarizing results (note the '.coef' argument)
summary.segmented(os.mod)

plot.segmented(os.mod, "mpa", .coef=fixef(os.mod), conf.level=.95)

rsegplot.mpa(os.mod, seg.adfa.subset, 'ADFA', y.variable = 'fh', y.lab = "Flame Height", tlp = mean_mpa_adfa, tlp.lower = lwr_mpa_adfa, tlp.upper = upr_mpa_adfa)
```

### Max. Temp.
No threshold able to be identified (segmented.default returns error)
```{r}
#o.mod<-lme(temp.max~mpa + year.month + site, random=~1|id, data=seg.adfa.subset)
#os.mod<-segmented.default(o.mod, ~mpa, npsi=list(mpa=1))
#summarizing results (note the '.coef' argument)
#summary.segmented(os.mod)

#rsegplot.mpa(os.mod, seg.adfa.subset, 'ADFA', y.variable = 'temp.max',y.lab = "Maxiumum Temp.", tlp = mean_mpa_adfa, tlp.lower = lwr_mpa_adfa, tlp.upper = upr_mpa_adfa)
```

Sustainability Metrics:
### Flame Duration
```{r}
o.mod<-lme(fd~mpa + year.month + site, random=~1|id, data=seg.adfa.subset)
os.mod<-segmented.default(o.mod, ~mpa, npsi=list(mpa=1))
#summarizing results (note the '.coef' argument)
summary.segmented(os.mod)

rsegplot.mpa(os.mod, seg.adfa.subset, 'ADFA', y.variable = 'fd',y.lab = "Flame Duration", tlp = mean_mpa_adfa, tlp.lower = lwr_mpa_adfa, tlp.upper = upr_mpa_adfa)
```

Consumability Metrics:
### Glow Duration
```{r}
o.mod<-lme(gd~mpa + year.month + site, random=~1|id, data=seg.adfa.subset)
os.mod.g1<-segmented.default(o.mod, ~mpa, npsi=list(mpa=1))
#summarizing results (note the '.coef' argument)
summary.segmented(os.mod.g1)

rsegplot.mpa(os.mod, seg.adfa.subset, 'ADFA', y.variable = 'gd',y.lab = "Glow Duration (s)", tlp = mean_mpa_adfa, tlp.lower = lwr_mpa_adfa, tlp.upper = upr_mpa_adfa)
```

### Post-Flame Glow
```{r}
o.mod<-lme(pfg~mpa + year.month + site, random=~1|id, data=seg.adfa.subset)
os.mod<-segmented.default(o.mod, ~mpa, npsi=list(mpa=1))
#summarizing results (note the '.coef' argument)
summary.segmented(os.mod)

rsegplot.mpa(os.mod, seg.adfa.subset, 'ADFA', y.variable = 'pfg',y.lab = "Post-Flame Glow (s)", tlp = mean_mpa_adfa, tlp.lower = lwr_mpa_adfa, tlp.upper = upr_mpa_adfa)
```

### Time to First Glow
```{r}
o.mod<-lme(ttfg~mpa + year.month + site, random=~1|id, data=seg.adfa.subset)
os.mod<-segmented.default(o.mod, ~mpa, npsi=list(mpa=1))
#summarizing results (note the '.coef' argument)
summary.segmented(os.mod)

plot.segmented(os.mod, "mpa", .coef=fixef(os.mod), conf.level=.95)

rsegplot.mpa(os.mod, seg.adfa.subset, 'ADFA', y.variable = 'ttfg',y.lab = "Time to First Glow", tlp = mean_mpa_adfa, tlp.lower = lwr_mpa_adfa, tlp.upper = upr_mpa_adfa)
```

# ---------------------------------
# MAIN Figure: Arranging Plots

For threshold values from above segmented regressions and the 95% CI (+/- 1.96*SE); in the following order (repeated twice; once for CEME, then ADFA): PC2, Prop. Ignite, GTI, FH, PFG, FD, MT, GD, TTFG 
```{r}
psi <- c(60.30118, 91.87861, 119.9111, 57.1915, 68.05055, 49.68505, 55.57693, 63.74955, 59.89132, 62.65922, 43.14614, 98.27882, 79.50582, 65.84311, 46.55515, 98.27882, 65.82738, 78.53595)
se.psi <- c(6.658953, 5.592003, 22.37689, 4.920527, 9.347637, 7.83918, 6.153706, 9.380537, 7.690371, 4.290069, 2.192261, 10.12274, 5.414246, 6.058802, 6.604431, 8.924882, 7.186473, 6.963152)
lwr.psi <- psi - 1.96*se.psi
upr.psi <- psi + 1.96*se.psi
psi
lwr.psi
upr.psi
```


```{r}
jpeg(here('figures', 'main-figures', 'Fig2_segmented.regressionsNEW.jpg'), width = 1200, height = 900, res = 300)
par(mfrow = c(2,2))
par(mar = c(4,5,3,2))
par(font.lab = 2)

rsegplot.lfm(os.mod.a, seg.ceme.subset, 'CEME', y.variable = 'PC1',y.lab = "Ignitability-Combustibility Proxy", x.lab = " ", tlp = mean_lfm_ceme, tlp.lower = lwr_lfm_ceme, tlp.upper = upr_lfm_ceme)
draw.circle(60.30118, -3.84, 2.4, col = 'black') # PC2 Threshold
segments(x0 = 47.24963, x1 = 73.35273, y0 = -3.84, y1 = -3.84, col = 'black')
text(x = 60.30118, y = -3.44, labels = c("Consumability"), cex = 1)
arrows(x0 = 15, x1 = 15, y0 = -4, y1 = -2, length = 0.15, lwd = 2)
text(x = 15, y = -1, labels = c("More \n Combustible \n & Ignitable"), cex = 1)
mtext("a", font = 2, cex = 2, side = 3, line = 0.5, at = 0)
text(labels = "C. megacarpus", font = 4, cex = 1.75, x = 190, y = 4)

#-----------------------------------------------------------------------
confint <- confint.segmented(os.mod.b, 'lfm', .coef=fixef(os.mod.b))
lower.b <- confint[2]
upper.b <- confint[3]

plot(seg.ceme.subset[,'lfm'], seg.ceme.subset[, 'tti'], xlim = c(0, max.lfm), ylim = c(-20, 86+0.05*86),
              xlab = ' ', ylab = 'Time to Ignition',
              pch = 16, col = alpha(seg.ceme.subset[,'color'], 0.8),
              cex.lab = 1.3)
polygon(x = c(lower.b, upper.b, upper.b, lower.b), y = c(86 + 0.2*86, 86 +0.2*86, -30, -30), col = alpha('gray', 0.3), lty = 2)
abline(v = mean_lfm_ceme, col = alpha("#64A400", 0.7), cex = 3)
polygon(x = c(lwr_lfm_ceme, upr_lfm_ceme, upr_lfm_ceme, lwr_lfm_ceme), y = c(86 +0.2*86, 86 +0.2*86, 0-22, -22), col = alpha("#64A400", 0.3), border = alpha("#64A400", 0.6))
legend('topleft', legend = c("September 2020", "January 2020", "December 2019"),
       col = c("#c1121f",'#669BBC',"#b08ba5"), pch = 16, cex=1)
for(z in 1:length(seg.ceme.subset[,'lfm'])){
  if(z < summary.segmented(os.mod.b)$psi[2]){
    segments(x0 = 0, x1 = summary.segmented(os.mod.b)$psi[2], y0 = intercept(os.mod.b, .coef = fixef(os.mod.b))[[1]][[1]], y1 = slope(os.mod.b, .coef = fixef(os.mod.b))[[1]][[1]]*((intercept(os.mod.b, .coef = fixef(os.mod.b))[[1]][[2]] - intercept(os.mod.b, .coef = fixef(os.mod.b))[[1]][[1]])/(slope(os.mod.b, .coef = fixef(os.mod.b))[[1]][[1]] - slope(os.mod.b, .coef = fixef(os.mod.b))[[1]][[2]]))+intercept(os.mod.b, .coef = fixef(os.mod.b))[[1]][[1]], lwd = 3, col = "black")
  }else{
    segments(x0 = summary.segmented(os.mod.b)$psi[2], x1 = max.lfm, y0 = slope(os.mod.b, .coef = fixef(os.mod.b))[[1]][[2]]*((intercept(os.mod.b, .coef = fixef(os.mod.b))[[1]][[1]] - intercept(os.mod.b, .coef = fixef(os.mod.b))[[1]][[2]])/(slope(os.mod.b, .coef = fixef(os.mod.b))[[1]][[2]] - slope(os.mod.b, .coef = fixef(os.mod.b))[[1]][[1]]))+intercept(os.mod.b, .coef = fixef(os.mod.b))[[1]][[2]] , y1 = slope(os.mod.b, .coef = fixef(os.mod.b))[[1]][[2]]*max.lfm + intercept(os.mod.b, .coef = fixef(os.mod.b))[[1]][[2]], lwd = 3, col = "black")}}
draw.circle(91.87861, -7, 2.4, col = 'black') # Prop ignite threshold
segments(x0 = 80.91828, x1 = 102.83894, y0 = -7, y1 = -7, col = 'black')
text(x = 93.87861, y = -2, labels = c("Prop. Ig."), cex = 1)
draw.circle(119.91110, -20, 2.4, col = 'black') # Glow to ignition threshold
segments(x0 = 76.05240, x1 = 163.76980, y0 = -20, y1 = -20, col = 'black')
text(x = 119.91110, y = -15, labels = c("GTI"), cex = 1)
draw.circle(57.19150, -12, 2.4, col = 'black') # Flame height threshold
segments(x0 = 47.54727, x1 = 66.83573, y0 = -12, y1 = -12, col = 'black')
text(x = 57.19150, y = -7, labels = c("FH"), cex = 1)
draw.circle(68.05055, -16, 2.4, col = 'black') # Post-flame glow threshold
segments(x0 = 49.72918, x1 = 86.37192, y0 = -16, y1 = -16, col = 'black')
text(x = 75.05055, y = -11, labels = c("PFG"), cex = 1)
draw.circle(49.68505, -20, 2.4) # Flame duration threshold
segments(x0 = 34.32026, x1 = 65.04984, y0 = -20, y1 = -20, col = 'black')
text(x = 44.68505, y = -15.25, labels = c("FD"), cex = 1)
draw.circle(55.57693, -2, 2.4, col = 'black') # Maximum temp. threshold
segments(x0 = 43.51567, x1 = 67.63819, y0 = -2, y1 = -2, col = 'black')
text(x = 46.57693, y = 2.5, labels = c("Max. T."), cex = 1)
draw.circle(63.74955, 5.75, 2.4) # Glow duration threshold
segments(x0 = 45.36370, x1 = 82.13540, y0 = 5.75, y1 = 5.75, col = 'black')
text(x = 52.74955, y = 8.75, labels = c("GD"), cex = 1)
draw.circle(59.89132, 13.5, 2.4) # Time to first glow threshold
segments(x0 = 44.81819, x1 = 74.96445, y0 = 13.5, y1 = 13.5, col = 'black')
text(x = 49.89132, y = 17.5, labels = c("TTFG"), cex = 1)
mtext("b", font = 2, cex = 2, side = 3, line = 0.5, at = 0)
text(labels = "C. megacarpus", font = 4, cex = 1.75, x = 190, y = 81)

#-----------------------------------------------------------------------

rsegplot.lfm(os.mod.c, seg.adfa.subset, 'ADFA', y.variable = 'PC1',y.lab = "Ignitability-Combustibility Proxy", x.lab = " ", tlp = mean_lfm_adfa, tlp.lower = lwr_lfm_adfa, tlp.upper = upr_lfm_adfa, max.x = max.lfm.a)
draw.circle(62.65922, -4.6, 1.25, col = 'black') # PC2 Threshold
segments(x0 = 54.25068, x1 = 71.06776, y0 = -4.6, y1 = -4.6, col = 'black')
text(x = 62.65922, y = -4.2, labels = c("Consumability"), cex = 1)
arrows(x0 = 7, x1 = 7, y0 = -4.8, y1 = -2.8, length = 0.15, lwd = 2)
text(x = 7, y = -1.8, labels = c("More \n Combustible \n & Ignitable"), cex = 1)
mtext("c", font = 2, cex = 2, side = 3, line = 0.5, at = 0)
text(labels = "A. fasciculatum", font = 4, cex = 1.75, x = 92, y = 3.8)

#-----------------------------------------------------------------------

confint <- confint.segmented(os.mod.d, 'lfm', .coef=fixef(os.mod.d))
lower.d <- confint[2]
upper.d <- confint[3]

plot(seg.adfa.subset[,'lfm'], seg.adfa.subset[, 'tti'], xlim = c(0, max.lfm.a), ylim = c(-17, 75+0.05*75),
              xlab = ' ', ylab = 'Time to Ignition',
              pch = 16, col = alpha(seg.adfa.subset[,'color'], 0.8),
              cex.lab = 1.3)
polygon(x = c(lower.d, upper.d, upper.d, lower.d), y = c(75 + 0.2*75, 75 +0.2*75, -22, -22), col = alpha('gray', 0.3), lty = 2)
abline(v = mean_lfm_adfa, col = alpha("#64A400", 0.7), cex = 3)
polygon(x = c(lwr_lfm_adfa, upr_lfm_adfa, upr_lfm_adfa, lwr_lfm_adfa), y = c(75 +0.2*75, 75 +0.2*75, 0-22, -22), col = alpha("#64A400", 0.3), border = alpha("#64A400", 0.6))
legend('topleft', legend = c("September 2020", "January 2018", "December 2016"),
       col = c("#c1121f","#FF7262","#A4A7DE"), pch = 16, cex=1)
for(z in 1:length(seg.adfa.subset[,'lfm'])){
  if(z < summary.segmented(os.mod.d)$psi[2]){
    segments(x0 = 0, x1 = summary.segmented(os.mod.d)$psi[2], y0 = intercept(os.mod.d, .coef = fixef(os.mod.d))[[1]][[1]], y1 = slope(os.mod.d, .coef = fixef(os.mod.d))[[1]][[1]]*((intercept(os.mod.d, .coef = fixef(os.mod.d))[[1]][[2]] - intercept(os.mod.d, .coef = fixef(os.mod.d))[[1]][[1]])/(slope(os.mod.d, .coef = fixef(os.mod.d))[[1]][[1]] - slope(os.mod.d, .coef = fixef(os.mod.d))[[1]][[2]]))+intercept(os.mod.d, .coef = fixef(os.mod.d))[[1]][[1]], lwd = 3, col = "black")
  }else{
    segments(x0 = summary.segmented(os.mod.d)$psi[2], x1 = max.lfm, y0 = slope(os.mod.d, .coef = fixef(os.mod.d))[[1]][[2]]*((intercept(os.mod.d, .coef = fixef(os.mod.d))[[1]][[1]] - intercept(os.mod.d, .coef = fixef(os.mod.d))[[1]][[2]])/(slope(os.mod.d, .coef = fixef(os.mod.d))[[1]][[2]] - slope(os.mod.d, .coef = fixef(os.mod.d))[[1]][[1]]))+intercept(os.mod.d, .coef = fixef(os.mod.d))[[1]][[2]] , y1 = slope(os.mod.d, .coef = fixef(os.mod.d))[[1]][[2]]*max.lfm + intercept(os.mod.d, .coef = fixef(os.mod.d))[[1]][[2]], lwd = 3, col = "black")}}
draw.circle(43.14614, -17, 1.25, col = 'black') # Prop ignite threshold
segments(x0 = 38.84931, x1 = 47.44297, y0 = -17, y1 = -17, col = 'black')
text(x = 43.14614, y = -12.5, labels = c("Prop. Ig."), cex = 1)
draw.circle(98.27882, -15, 1.25, col = 'black') # Glow to ignition threshold
segments(x0 = 78.43825, x1 = 118.11939, y0 = -15, y1 = -15, col = 'black')
text(x = 98.27882, y = -10.5, labels = c("GTI"), cex = 1)
draw.circle(79.50582, 2, 1.25, col = 'black') # Flame height threshold
segments(x0 = 68.89390, x1 = 90.11774, y0 = 2, y1 = 2, col = 'black')
text(x = 79.50582, y = 6.5, labels = c("FH"), cex = 1)
draw.circle(65.84311, -18, 1.25, col = 'black') # Post-flame glow threshold
segments(x0 = 53.96786, x1 = 77.71836, y0 = -18, y1 = -18, col = 'black')
text(x = 62.84311, y = -13.5, labels = c("PFG"), cex = 1)
draw.circle(46.55515, -7.75, 1.25) # Flame duration threshold
segments(x0 = 33.61047, x1 = 59.49983, y0 = -7.75, y1 = -7.75, col = 'black')
text(x = 45.55515, y = -3.25, labels = c("FD"), cex = 1)
draw.circle(98.27882, -4, 1.25, col = 'black') # Maximum temp. threshold
segments(x0 = 80.78605, x1 = 115.77159, y0 = -4, y1 = -4, col = 'black')
text(x = 98.27882, y = 0.5, labels = c("Max. T."), cex = 1)
draw.circle(65.82738, -2.5, 1.25, col = 'black') # Glow duration threshold
segments(x0 = 51.74189, x1 = 79.91287, y0 = -2.5, y1 = -2.5, col = 'black')
text(x = 64.42738, y = 2, labels = c("GD"), cex = 1)
draw.circle(78.53595, -11, 1.25, col = 'black') # Time to first glow threshold
segments(x0 = 64.88817, x1 = 92.18373, y0 = -11, y1 = -11, col = 'black')
text(x = 78.53595, y = -6.5, labels = c("TTFG"), cex = 1)
mtext("d", font = 2, cex = 2, side = 3, line = 0.5, at = 0)
text(labels = "A. fasciculatum", font = 4, cex = 1.75, x = 92, y = 72)
mtext("Live Fuel Moisture (%)", font = 2, side = 1, cex = 1.5, line = 2.7, at = -27)

dev.off()
```

# ---------------------------------
# Splitting by Sampling Date (LFM)
## CEME
### PC1
For some odd reason, the plots split by sampling date for CEME weren't working perfectly, so I added the second portion of the segmented regression
```{r}
o.mod<-lme(PC1~lfm + site, random=~1|id, data=seg.ceme.s20.subset)
os.mod.a1<-segmented.default(o.mod, ~lfm, npsi=list(lfm=1))
#summarizing results (note the '.coef' argument)
summary.segmented(os.mod.a1)

m2.1 <- slope(os.mod.a1, .coef = fixef(os.mod.a1))[[1]][[2]]
b2.1 <- intercept(os.mod.a1, .coef = fixef(os.mod.a1))[[1]][[2]]
p.1 <- summary.segmented(os.mod.a1)$psi[2]
m.1 <- slope(os.mod.a1, .coef = fixef(os.mod.a1))[[1]][[1]]
b.1 <- intercept(os.mod.a1, .coef = fixef(os.mod.a1))[[1]][[1]]

rsegplot.lfm(os.mod.a1, seg.ceme.s20.subset, 'CEME', y.variable = 'PC1',y.lab = "PC1", tlp = mean_lfm_ceme, tlp.lower = lwr_lfm_ceme, tlp.upper = upr_lfm_ceme)
segments(x0 = p.1, x1 = max.lfm.c, y0 = m2.1*((b.1 - b2.1)/(m2.1 - m.1))+b2.1 , y1 = m2.1*max.lfm.c + b2.1, lwd = 4, col = "black")
```

```{r}
o.mod<-lme(PC1~lfm, random=~1|id, data=seg.ceme.d19.subset)
os.mod.a2<-segmented.default(o.mod, ~lfm, npsi=list(lfm=1))
#summarizing results (note the '.coef' argument)
summary.segmented(os.mod.a2)

m2.2 <- slope(os.mod.a2, .coef = fixef(os.mod.a2))[[1]][[2]]
b2.2 <- intercept(os.mod.a2, .coef = fixef(os.mod.a2))[[1]][[2]]
p.2 <- summary.segmented(os.mod.a2)$psi[2]
m.2 <- slope(os.mod.a2, .coef = fixef(os.mod.a2))[[1]][[1]]
b.2 <- intercept(os.mod.a2, .coef = fixef(os.mod.a2))[[1]][[1]]

rsegplot.lfm(os.mod.a2, seg.ceme.d19.subset, 'CEME', y.variable = 'PC1',y.lab = "PC1", tlp = mean_lfm_ceme, tlp.lower = lwr_lfm_ceme, tlp.upper = upr_lfm_ceme)
segments(x0 = p.2, x1 = max.lfm.c, y0 = m2.2*((b.2 - b2.2)/(m2.2 - m.2))+b2.2 , y1 = m2.2*max.lfm.c + b2.2, lwd = 4, col = "black")
```

### Time to Ignition
```{r}
o.mod<-lme(tti~lfm + site, random=~1|id, data=seg.ceme.s20.subset)
os.mod.b1<-segmented.default(o.mod, ~lfm, npsi=list(lfm=1))
#summarizing results (note the '.coef' argument)
summary.segmented(os.mod.b1)

m2.3 <- slope(os.mod.b1, .coef = fixef(os.mod.b1))[[1]][[2]]
b2.3 <- intercept(os.mod.b1, .coef = fixef(os.mod.b1))[[1]][[2]]
p.3 <- summary.segmented(os.mod.b1)$psi[2]
m.3 <- slope(os.mod.b1, .coef = fixef(os.mod.b1))[[1]][[1]]
b.3 <- intercept(os.mod.b1, .coef = fixef(os.mod.b1))[[1]][[1]]

rsegplot.lfm(os.mod.b1, seg.ceme.s20.subset, 'CEME', y.variable = 'tti',y.lab = "Time to Ignition (s)", tlp = mean_lfm_ceme, tlp.lower = lwr_lfm_ceme, tlp.upper = upr_lfm_ceme)
segments(x0 = p.3, x1 = max.lfm.c, y0 = m2.3*((b.3 - b2.3)/(m2.3 - m.3))+b2.3 , y1 = m2.3*max.lfm.c + b2.3, lwd = 4, col = "black")
```

```{r}
o.mod<-lme(tti~lfm, random=~1|id, data=seg.ceme.d19.subset)
os.mod.b2<-segmented.default(o.mod, ~lfm, npsi=list(lfm=1))
#summarizing results (note the '.coef' argument)
summary.segmented(os.mod.b2)

m2.4 <- slope(os.mod.b2, .coef = fixef(os.mod.b2))[[1]][[2]]
b2.4 <- intercept(os.mod.b2, .coef = fixef(os.mod.b2))[[1]][[2]]
p.4 <- summary.segmented(os.mod.b2)$psi[2]
m.4 <- slope(os.mod.b2, .coef = fixef(os.mod.b2))[[1]][[1]]
b.4 <- intercept(os.mod.b2, .coef = fixef(os.mod.b2))[[1]][[1]]

rsegplot.lfm(os.mod.b2, seg.ceme.d19.subset, 'CEME', y.variable = 'tti',y.lab = "Time to Ignition (s)", tlp = mean_lfm_ceme, tlp.lower = lwr_lfm_ceme, tlp.upper = upr_lfm_ceme)
segments(x0 = p.4, x1 = max.lfm.c, y0 = m2.4*((b.4 - b2.4)/(m2.4 - m.4))+b2.4 , y1 = m2.4*max.lfm.c + b2.4, lwd = 4, col = "black")
```

### Flame Height
```{r}
o.mod<-lme(fh~lfm + site, random=~1|id, data=seg.ceme.s20.subset)
os.mod<-segmented.default(o.mod, ~lfm, npsi=list(lfm=1))
#summarizing results (note the '.coef' argument)
summary.segmented(os.mod)

rsegplot.lfm(os.mod, seg.ceme.s20.subset, 'CEME', y.variable = 'fh',y.lab = "Flame Height (cm)", tlp = mean_lfm_ceme, tlp.lower = lwr_lfm_ceme, tlp.upper = upr_lfm_ceme)
```

```{r}
o.mod<-lme(fh~lfm, random=~1|id, data=seg.ceme.d19.subset)
os.mod<-segmented.default(o.mod, ~lfm, npsi=list(lfm=1))
#summarizing results (note the '.coef' argument)
summary.segmented(os.mod)

rsegplot.lfm(os.mod, seg.ceme.d19.subset, 'CEME', y.variable = 'fh',y.lab = "Flame Height (cm)", tlp = mean_lfm_ceme, tlp.lower = lwr_lfm_ceme, tlp.upper = upr_lfm_ceme)
```

## ADFA
### PC1
```{r}
o.mod<-lme(PC1~ lfm + site, random=~1|id, data=seg.adfa.s20.subset)
os.mod.c1<-segmented.default(o.mod, ~lfm, npsi=list(lfm=1))
#summarizing results (note the '.coef' argument)
summary.segmented(os.mod.c1)

rsegplot.lfm(os.mod.c1, seg.adfa.s20.subset, 'ADFA', y.variable = 'PC1',y.lab = "PC1", tlp = mean_lfm_adfa, tlp.lower = lwr_lfm_adfa, tlp.upper = upr_lfm_adfa, max.x = max.lfm.a)
```

```{r}
o.mod<-lme(PC1~ lfm, random=~1|id, data=seg.adfa.d16.subset)
os.mod.c2<-segmented.default(o.mod, ~lfm, npsi=list(lfm=1))
#summarizing results (note the '.coef' argument)
summary.segmented(os.mod.c2)

rsegplot.lfm(os.mod.c1, seg.adfa.s20.subset, 'ADFA', y.variable = 'PC1',y.lab = "PC1", tlp = mean_lfm_adfa, tlp.lower = lwr_lfm_adfa, tlp.upper = upr_lfm_adfa, max.x = max.lfm.a)
rsegplot.lfm(os.mod.c2, seg.adfa.d16.subset, 'ADFA', y.variable = 'PC1',y.lab = "PC1", tlp = mean_lfm_adfa, tlp.lower = lwr_lfm_adfa, tlp.upper = upr_lfm_adfa, max.x = max.lfm.a)
```

### Time to Ignition
```{r}
o.mod<-lme(tti~lfm + site, random=~1|id, data=seg.adfa.s20.subset)
os.mod.d1<-segmented.default(o.mod, ~lfm, npsi=list(lfm=1))
#summarizing results (note the '.coef' argument)
summary.segmented(os.mod.d1)

rsegplot.lfm(os.mod.d1, seg.adfa.s20.subset, 'ADFA', y.variable = 'tti',y.lab = "Time to Ignition (s)", tlp = mean_lfm_adfa, tlp.lower = lwr_lfm_adfa, tlp.upper = upr_lfm_adfa, max.x = max.lfm.a)
```

```{r}
o.mod<-lme(tti~lfm, random=~1|id, data=seg.adfa.d16.subset)
os.mod.d2<-segmented.default(o.mod, ~lfm, npsi=list(lfm=1))
#summarizing results (note the '.coef' argument)
summary.segmented(os.mod.d2)

rsegplot.lfm(os.mod.d2, seg.adfa.d16.subset, 'ADFA', y.variable = 'tti',y.lab = "Time to Ignition (s)", tlp = mean_lfm_adfa, tlp.lower = lwr_lfm_adfa, tlp.upper = upr_lfm_adfa, max.x = max.lfm.a)
```

### Flame Height
```{r}
o.mod<-lme(fh~lfm + site, random=~1|id, data=seg.adfa.s20.subset)
os.mod<-segmented.default(o.mod, ~lfm, npsi=list(lfm=1))
#summarizing results (note the '.coef' argument)
summary.segmented(os.mod)

rsegplot.lfm(os.mod, seg.adfa.s20.subset, 'ADFA', y.variable = 'fh',y.lab = "Flame Height (cm)", tlp = mean_lfm_adfa, tlp.lower = lwr_lfm_adfa, tlp.upper = upr_lfm_adfa, max.x = max.lfm.a)
```

```{r}
o.mod<-lme(fh~lfm, random=~1|id, data=seg.adfa.d16.subset)
os.mod<-segmented.default(o.mod, ~lfm, npsi=list(lfm=1))
#summarizing results (note the '.coef' argument)
summary.segmented(os.mod)

rsegplot.lfm(os.mod, seg.adfa.d16.subset, 'ADFA', y.variable = 'fh',y.lab = "Flame Height (cm)", tlp = mean_lfm_adfa, tlp.lower = lwr_lfm_adfa, tlp.upper = upr_lfm_adfa, max.x = max.lfm.a)
```

# Supp Figure
```{r}
jpeg(here('figures', 'supp-figures', 'FigS2_seg.reg.by.dates.jpg'), width = 1500, height = 1950, res = 150)
par(mfrow = c(4,3))
par(font.lab = 2)
par(mar = c(5,4.75,3.5,1))
#####################################################
rsegplot.lfm(os.mod.a, seg.ceme.subset, 'CEME', y.variable = 'PC1',y.lab = "Ignitability-Combustibility Proxy", x.lab = " ", tlp = mean_lfm_ceme, tlp.lower = lwr_lfm_ceme, tlp.upper = upr_lfm_ceme)
arrows(x0 = 20, x1 = 20, y0 = -4, y1 = -2.75, length = 0.1, lwd = 1.3)
text(x = 20, y = -1.85, labels = c("More \n Combustible \n & Ignitable"), cex = 0.8)
text(labels = c("a"), font = 2, cex = 3, x = 230, y = 4)
mtext("C. megacarpus", font = 4, cex = 1.8, side = 3, line = 1, at = 100)
#####################################################
rsegplot.lfm.nolegend(os.mod.a1, seg.ceme.s20.subset, 'CEME', y.variable = 'PC1',y.lab = "Ignitability-Combustibility Proxy", x.lab = " ", tlp = mean_lfm_ceme, tlp.lower = lwr_lfm_ceme, tlp.upper = upr_lfm_ceme)
segments(x0 = p.1, x1 = max.lfm.c, y0 = m2.1*((b.1 - b2.1)/(m2.1 - m.1))+b2.1 , y1 = m2.1*max.lfm.c + b2.1, lwd = 4, col = "black")
text(labels = c("b"), font = 2, cex = 3, x = 230, y = 4.05)
mtext("September 2020", font = 1, cex = 1, side = 3, line = 0.8, at = 115)
#####################################################
rsegplot.lfm.nolegend(os.mod.a2, seg.ceme.d19.subset, 'CEME', y.variable = 'PC1',y.lab = "Ignitability-Combustibility Proxy", x.lab = " ", tlp = mean_lfm_ceme, tlp.lower = lwr_lfm_ceme, tlp.upper = upr_lfm_ceme)
segments(x0 = p.2, x1 = max.lfm.c, y0 = m2.2*((b.2 - b2.2)/(m2.2 - m.2))+b2.2 , y1 = m2.2*max.lfm.c + b2.2, lwd = 4, col = "black")
text(labels = c("c"), font = 2, cex = 3, x = 230, y = 1.9)
mtext("December 2019", font = 1, cex = 1, side = 3, line = 0.8, at = 115)
#####################################################
rsegplot.lfm(os.mod.b, seg.ceme.subset, 'CEME', y.variable = 'tti',y.lab = "Time to Ignition (s)", x.lab = " ", tlp = mean_lfm_ceme, tlp.lower = lwr_lfm_ceme, tlp.upper = upr_lfm_ceme)
text(labels = c("d"), font = 2, cex = 3, x = 230, y = 85)
#####################################################
rsegplot.lfm.nolegend(os.mod.b1, seg.ceme.s20.subset, 'CEME', y.variable = 'tti',y.lab = "Time to Ignition (s)", x.lab = " ", tlp = mean_lfm_ceme, tlp.lower = lwr_lfm_ceme, tlp.upper = upr_lfm_ceme)
segments(x0 = p.3, x1 = max.lfm.c, y0 = m2.3*((b.3 - b2.3)/(m2.3 - m.3))+b2.3 , y1 = m2.3*max.lfm.c + b2.3, lwd = 4, col = "black")
text(labels = c("e"), font = 2, cex = 3, x = 230, y = 85)
#####################################################
rsegplot.lfm.nolegend(os.mod.b2, seg.ceme.d19.subset, 'CEME', y.variable = 'tti',y.lab = "Time to Ignition (s)", x.lab = " ", tlp = mean_lfm_ceme, tlp.lower = lwr_lfm_ceme, tlp.upper = upr_lfm_ceme)
segments(x0 = p.4, x1 = max.lfm.c, y0 = m2.4*((b.4 - b2.4)/(m2.4 - m.4))+b2.4 , y1 = m2.4*max.lfm.c + b2.4, lwd = 4, col = "black")
text(labels = c("f"), font = 2, cex = 3, x = 230, y = 70)
#####################################################
rsegplot.lfm(os.mod.c, seg.adfa.subset, 'ADFA', y.variable = 'PC1',y.lab = "Ignitability-Combustibility Proxy", x.lab = " ", tlp = mean_lfm_adfa, tlp.lower = lwr_lfm_adfa, tlp.upper = upr_lfm_adfa, max.x = max.lfm.a)
arrows(x0 = 10, x1 = 10, y0 = -4.8, y1 = -3.55, length = 0.1, lwd = 1.3)
text(x = 10, y = -2.65, labels = c("More \n Combustible \n & Ignitable"), cex = 0.8)
text(labels = c("g"), font = 2, cex = 3, x = 110, y = 4)
mtext("A. fasciculatum", font = 4, cex = 1.8, side = 3, line = 1, at = 50)
#####################################################
rsegplot.lfm.nolegend(os.mod.c1, seg.adfa.s20.subset, 'ADFA', y.variable = 'PC1',y.lab = "Ignitability-Combustibility Proxy", x.lab = " ", tlp = mean_lfm_adfa, tlp.lower = lwr_lfm_adfa, tlp.upper = upr_lfm_adfa, max.x = max.lfm.a)
mtext("September 2020", font = 1, cex = 1, side = 3, line = 0.8, at = 55)
text(labels = c("h"), font = 2, cex = 3, x = 110, y = 3.5)
#####################################################
rsegplot.lfm.nolegend(os.mod.c2, seg.adfa.d16.subset, 'ADFA', y.variable = 'PC1',y.lab = "Ignitability-Combustibility Proxy", x.lab = " ", tlp = mean_lfm_adfa, tlp.lower = lwr_lfm_adfa, tlp.upper = upr_lfm_adfa, max.x = max.lfm.a)
mtext("December 2016", font = 1, cex = 1, side = 3, line = 0.8, at = 55)
text(labels = c("i"), font = 2, cex = 3, x = 110, y = 4)
#####################################################
rsegplot.lfm(os.mod.d, seg.adfa.subset, 'ADFA', y.variable = 'tti',y.lab = "Time to Ignition (s)", x.lab = " ", tlp = mean_lfm_adfa, tlp.lower = lwr_lfm_adfa, tlp.upper = upr_lfm_adfa, max.x = max.lfm.a)
text(labels = c("j"), font = 2, cex = 3, x = 110, y = 70)
#####################################################
rsegplot.lfm.nolegend(os.mod.d1, seg.adfa.s20.subset, 'ADFA', y.variable = 'tti',y.lab = "Time to Ignition (s)", x.lab = " ", tlp = mean_lfm_adfa, tlp.lower = lwr_lfm_adfa, tlp.upper = upr_lfm_adfa, max.x = max.lfm.a)
text(labels = c("k"), font = 2, cex = 3, x = 110, y = 70)
mtext("Live Fuel Moisture (%)", font = 2, cex = 1.8, side = 1, line = 3, at = 60)
#####################################################
rsegplot.lfm.nolegend(os.mod.d2, seg.adfa.d16.subset, 'ADFA', y.variable = 'tti',y.lab = "Time to Ignition (s)", x.lab = " ", tlp = mean_lfm_adfa, tlp.lower = lwr_lfm_adfa, tlp.upper = upr_lfm_adfa, max.x = max.lfm.a)
text(labels = c("l"), font = 2, cex = 3, x = 110, y = 53.75)
#####################################################
dev.off()
```

# ------------------------------
# Supp. Analysis: -1/MPa

# CEME
```{r}
seg.ceme.subset <- seg.ceme.subset %>% 
  mutate(mpa.inverse = -1/mpa)
```

### PC1
n.s. 0.1001 (-9.99)
```{r}
o.mod<-lme(PC1~mpa.inverse + year.month + site, random=~1|id, data=seg.ceme.subset)
os.mod.a1<-segmented.default(o.mod, ~mpa.inverse, npsi=list(mpa.inverse=1))
#summarizing results (note the '.coef' argument)
summary.segmented(os.mod.a1)
```

### PC2
sig (p = 0.017856280) 0.1546915 (-6.46)
```{r}
o.mod<-lme(PC2~mpa.inverse + year.month + site, random=~1|id, data=seg.ceme.subset)
os.mod<-segmented.default(o.mod, ~mpa.inverse, npsi=list(mpa.inverse=1))
#summarizing results (note the '.coef' argument)
summary.segmented(os.mod)
```
 
Ignitability Metrics:
### Time to Ignition
SIGNIFICANT (p = 0.0015312), 0.3219623 (-3.105954)
```{r}
o.mod<-lme(tti~mpa.inverse + year.month + site, random=~1|id, data=seg.ceme.subset)
os.mod.b1<-segmented.default(o.mod, ~mpa.inverse, npsi=list(mpa.inverse=1))
#summarizing results (note the '.coef' argument)
summary.segmented(os.mod.b1)
```

### Prop. Ignite (5%)
n.s., 0.1195791 (-8.36)
```{r}
o.mod<-lme(prop.ignite.bins5_nodate~mpa.inverse + year.month + site, random=~1|id, data=seg.ceme.subset)
os.mod<-segmented.default(o.mod, ~mpa.inverse, npsi=list(mpa.inverse=1))
#summarizing results (note the '.coef' argument)
summary.segmented(os.mod)
```

### Glow to Ignition
n.s. 0.1001 (-9.99)
```{r}
o.mod<-lme(gti~mpa.inverse + year.month + site, random=~1|id, data=seg.ceme.subset)
os.mod<-segmented.default(o.mod, ~mpa.inverse, npsi=list(mpa.inverse=1))
#summarizing results (note the '.coef' argument)
summary.segmented(os.mod)
```

Combustibility Metrics:
### Flame Height
SIGNIFICANT (p = 0.003973) 0.1275757 (-7.84)
```{r}
o.mod<-lme(fh~mpa.inverse + year.month + site, random=~1|id, data=seg.ceme.subset)
os.mod<-segmented.default(o.mod, ~mpa.inverse, npsi=list(mpa.inverse=1))
#summarizing results (note the '.coef' argument)
summary.segmented(os.mod)
```

### Max. Temp.
n.s. 0.1235898 (-8.09)
```{r}
o.mod<-lme(temp.max~mpa.inverse + year.month + site, random=~1|id, data=seg.ceme.subset)
os.mod<-segmented.default(o.mod, ~mpa.inverse, npsi=list(mpa.inverse=1))
#summarizing results (note the '.coef' argument)
summary.segmented(os.mod)
```

Sustainability Metrics:
### Flame Duration
n.s. 0.10007 (-9.99)
```{r}
o.mod<-lme(fd~mpa.inverse + year.month + site, random=~1|id, data=seg.ceme.subset)
os.mod.f1<-segmented.default(o.mod, ~mpa.inverse, npsi=list(mpa.inverse=1))
#summarizing results (note the '.coef' argument)
summary.segmented(os.mod.f1)
```

Consumability Metrics:
### Glow Duration
n.s. 0.1482018 (-6.75)
```{r}
o.mod<-lme(gd~mpa.inverse + year.month + site, random=~1|id, data=seg.ceme.subset)
os.mod.h1<-segmented.default(o.mod, ~mpa.inverse, npsi=list(mpa.inverse=1))
#summarizing results (note the '.coef' argument)
summary.segmented(os.mod.h1)
```

### Post-Flame Glow
n.s. 0.1508012 (-6.63)
```{r}
o.mod<-lme(pfg~mpa.inverse + year.month + site, random=~1|id, data=seg.ceme.subset)
os.mod<-segmented.default(o.mod, ~mpa.inverse, npsi=list(mpa.inverse=1))
#summarizing results (note the '.coef' argument)
summary.segmented(os.mod)
```

### Time to First Glow
n.s. 0.1123685 (-8.90)
```{r}
o.mod<-lme(ttfg~mpa.inverse + year.month + site, random=~1|id, data=seg.ceme.subset)
os.mod<-segmented.default(o.mod, ~mpa.inverse, npsi=list(mpa.inverse=1))
#summarizing results (note the '.coef' argument)
summary.segmented(os.mod)
```

# -----

# ADFA
```{r}
seg.adfa.subset <- seg.adfa.subset %>% 
  mutate(mpa.inverse = -1/mpa)
```

### PC1
SIGNIFICANT (p = 0.001931591) 0.4268678 (-2.34)
```{r}
o.mod<-lme(PC1~mpa.inverse + year.month + site, random=~1|id, data=seg.adfa.subset)
os.mod.c1<-segmented.default(o.mod, ~mpa.inverse, npsi=list(mpa.inverse=1))
#summarizing results (note the '.coef' argument)
summary.segmented(os.mod.c1)
```

### PC2
SIGNIFICANT (p = 0.002949292) 0.1742132 (-5.74)
```{r}
o.mod<-lme(PC2~mpa.inverse + year.month + site, random=~1|id, data=seg.adfa.subset)
os.mod<-segmented.default(o.mod, ~mpa.inverse, npsi=list(mpa.inverse=1))
#summarizing results (note the '.coef' argument)
summary.segmented(os.mod)
```
 
Ignitability Metrics:
### Time to Ignition
SIGNIFICANT (p = 6.4785 x 10^-6) 0.3814873 (-2.62)
```{r}
o.mod<-lme(tti~mpa.inverse + year.month + site, random=~1|id, data=seg.adfa.subset)
os.mod.d1<-segmented.default(o.mod, ~mpa.inverse, npsi=list(mpa.inverse=1))
#summarizing results (note the '.coef' argument)
summary.segmented(os.mod.d1)
```

### Prop. Ignite (5%)
n.s. 0.1062578 (-9.41)
```{r}
o.mod<-lme(prop.ignite.bins5_nodate~mpa.inverse + year.month + site, random=~1|id, data=seg.adfa.subset)
os.mod<-segmented.default(o.mod, ~mpa.inverse, npsi=list(mpa.inverse=1))
#summarizing results (note the '.coef' argument)
summary.segmented(os.mod)
```

### Glow to Ignition
SIGNIFICANT (p = 0.0070038) 0.4280231 (-2.336)
```{r}
o.mod<-lme(gti~mpa.inverse + year.month + site, random=~1|id, data=seg.adfa.subset)
os.mod<-segmented.default(o.mod, ~mpa.inverse, npsi=list(mpa.inverse=1))
#summarizing results (note the '.coef' argument)
summary.segmented(os.mod)
```

Combustibility Metrics:
### Flame Height
SIGNIFICANT (p = 0.002695) 0.3761366 (-2.659)
```{r}
o.mod<-lme(fh~mpa.inverse + year.month + site, random=~1|id, data=seg.adfa.subset)
os.mod<-segmented.default(o.mod, ~mpa.inverse, npsi=list(mpa.inverse=1))
#summarizing results (note the '.coef' argument)
summary.segmented(os.mod)
```

### Max. Temp.
n.s. 0.1848124 (-5.411)
```{r}
o.mod<-lme(temp.max~mpa.inverse + year.month + site, random=~1|id, data=seg.adfa.subset)
os.mod<-segmented.default(o.mod, ~mpa.inverse, npsi=list(mpa.inverse=1))
#summarizing results (note the '.coef' argument)
summary.segmented(os.mod)
```

Sustainability Metrics:
### Flame Duration
n.s. 0.1175808 (-8.50)
```{r}
o.mod<-lme(fd~mpa.inverse + year.month + site, random=~1|id, data=seg.adfa.subset)
os.mod<-segmented.default(o.mod, ~mpa.inverse, npsi=list(mpa.inverse=1))
#summarizing results (note the '.coef' argument)
summary.segmented(os.mod)
```

Consumability Metrics:
### Glow Duration
n.s. 8.310105 (-0.1203l)
```{r}
o.mod<-lme(gd~mpa.inverse + year.month + site, random=~1|id, data=seg.adfa.subset)
os.mod.g1<-segmented.default(o.mod, ~mpa.inverse, npsi=list(mpa.inverse=1))
#summarizing results (note the '.coef' argument)
summary.segmented(os.mod.g1)
```

### Post-Flame Glow
n.s. 0.1611068 (-6.21)
```{r}
o.mod<-lme(pfg~mpa.inverse + year.month + site, random=~1|id, data=seg.adfa.subset)
os.mod<-segmented.default(o.mod, ~mpa.inverse, npsi=list(mpa.inverse=1))
#summarizing results (note the '.coef' argument)
summary.segmented(os.mod)
```

### Time to First Glow
n.s. 0.2488277 (-4.0188)
```{r}
o.mod<-lme(ttfg~mpa.inverse + year.month + site, random=~1|id, data=seg.adfa.subset)
os.mod<-segmented.default(o.mod, ~mpa.inverse, npsi=list(mpa.inverse=1))
#summarizing results (note the '.coef' argument)
summary.segmented(os.mod)
```

# ----
# Kable Table
Setup:
```{r}
dep.var <- c('PC1','PC2','TTI','Prop Ignite (5%)','GTI','FH','Max Temp.','FD','GD','PFG','TTFG', 'Summary','PC1','PC2','TTI','Prop Ignite (5%)','GTI','FH','Max Temp.','FD','GD','PFG','TTFG', 'Summary')
spp <- c(rep('ADFA', 11), ' ', rep('CEME', 11), ' ')
sig.p <- c('Yes (p = 0.0019)','Yes (p = 0.0029)','Yes (p < 0.0005)','No','Yes (p = 0.0070)','Yes (p = 0.0027)','No','No','No','No','Yes (p = 0.013)', '6/11 Significant', 'No', 'Yes (p = 0.018)','Yes (p = 0.0015)','No','No','Yes (p = 0.0040)','No','No','No','No','No', '3/11 Significant')
threshold <- c('0.4269','0.1742','0.3815','0.1063','0.4280','0.3761','0.1848','0.1176','8.3102','0.1611','0.2488', 'Avg: 0.3393','0.1001','0.1547','0.3220','0.1196','0.1001','0.1276','0.1236','0.1001','0.1482','0.1508','0.1124', 'Avg: 0.2014')
threshold.mpa <- c('-2.34','-5.74','-2.62','-9.41','-2.34','-2.66','-5.41','-8.50','-0.12','-6.21','-4.02', 'Avg: -2.95','-9.99','-6.46','-3.11','-8.36', '-9.99','-7.84','-8.09', '-9.99','-6.75','-6.63','-8.90', 'Avg: -4.96')
seg.reg.table.df <- data.frame(dep.var, spp, sig.p, threshold, threshold.mpa)
```

Kable:
```{r}
seg.reg.table.df %>% 
   mutate(dep.var = cell_spec(dep.var, 'html', color = 'black', bold = T), spp = cell_spec(spp, 'html', color = 'black'), sig.p = cell_spec(sig.p, 'html', color = 'black'), threshold = cell_spec(threshold, 'html', color = 'black'), threshold.mpa = cell_spec(threshold.mpa, 'html', color = 'black')) %>% 
  kable(format = 'html', escape = F, col.names = c('Dependent Variable', 'Species', 'Significant?','Threshold (-1/MPa)', 'Threshold (MPa)'), caption = '-1/MPa') %>% 
  kable_styling(bootstrap_options = c('hover', 'bordered', 'condensed'), fixed_thead = T) %>% 
  row_spec(c(1:3, 5:6, 11, 14:15, 18), bold = T) %>% 
  row_spec(c(4, 7:10, 13, 16:17, 19:23), bold = F) %>% 
  row_spec(c(12, 24), background = "#D3D3D3", bold = T) %>%
  save_kable(here('figures', 'supp-figures', 'TabS6_inverse.mpa.segreg.summary.html'))
```
