==> scripts/4.1_segmented_non_random_reg.Rmd <==
---
title: "Segmented Regressions - Linear Models"
author: "Indra Boving & Joe Celebrezze"
date: "6/7/2022"
output: html_document
---
```{r setup, include = FALSE}
knitr::opts_chunk$set(echo = TRUE)
#lots of extra here, but oh well... 
library(ggplot2)
library(gapminder)
library(tidyverse)
library(lme4)
library(here)
library(segmented)
library(nlme)
library(lubridate)
library(MetBrewer)
library(lattice)
library(gt)
library(gtsummary)
filter = dplyr::filter 
rename = dplyr::rename
here = here::here
```

# Reading in Dataframes
Using all dates instead of just Sept. 2020 data
```{r}
seg.data.subset <- read.csv(here('processed-data', 'seg.reg.data.local.csv'))

seg.data.subset.alldates <- read.csv(here('processed-data', 'mem.data.subset.alldates.csv'))

scaled.mem.data.alldates <- read.csv(here('processed-data', 'mem.data.local.epi.alldates.csv'))
```

# Data Wrangling
For the segmented regressions, we debated using seasonal LFM minimums and maximums to limit the dataset, but instead -- since we wanted to keep the datasets similar for each analysis -- we used the dataset that was used in the mixed effects model analysis

## Species Split
```{r}
seg.adfa.subset <- seg.data.subset %>% 
  filter(spp == "ADFA")

seg.ceme.subset <- seg.data.subset %>% 
  filter(spp == "CEME")
```

## PV Data, TLP
Getting PV data, TLP stuff all ready for segmented regression visualizations (comparing breakpoint in segmented regression to turgor loss point derived from pressure-volume curves)

```{r}
pv_summary_df_timing <- read_csv(here("processed-data", "pv_summary_df_timing.csv"))

pv_local <- pv_summary_df_timing %>% 
  filter(spp %in% c("ADFA", "CEME"))
```

```{r}
mean_lfm_adfa <- pv_local %>% 
  filter(spp == "ADFA") %>% 
  summarise(mean(mean_lfm)) %>% 
  pull()

upr_lfm_adfa <- pv_local %>% 
  filter(spp == "ADFA") %>% 
  summarise(mean(upr_lfm)) %>% 
  pull()

lwr_lfm_adfa <- pv_local %>% 
  filter(spp == "ADFA") %>% 
  summarise(mean(lwr_lfm)) %>% 
  pull()

mean_mpa_adfa <- pv_local %>% 
  filter(spp == "ADFA") %>% 
  summarise(mean(mean_tlp)) %>% 
  pull()

upr_mpa_adfa <- pv_local %>% 
  filter(spp == "ADFA") %>% 
  summarise(mean(upr)) %>% 
  pull()

lwr_mpa_adfa <- pv_local %>% 
  filter(spp == "ADFA") %>% 
  summarise(mean(lwr)) %>% 
  pull()
```

```{r}
mean_lfm_ceme <- pv_local %>% 
  filter(spp == "CEME") %>% 
  summarise(mean(mean_lfm)) %>% 
  pull()

upr_lfm_ceme <- pv_local %>% 
  filter(spp == "CEME") %>% 
  summarise(mean(upr_lfm)) %>% 
  pull()

lwr_lfm_ceme <- pv_local %>% 
  filter(spp == "CEME") %>% 
  summarise(mean(lwr_lfm)) %>% 
  pull()

mean_mpa_ceme <- pv_local %>% 
  filter(spp == "CEME") %>% 
  summarise(mean(mean_tlp)) %>% 
  pull()

upr_mpa_ceme <- pv_local %>% 
  filter(spp == "CEME") %>% 
  summarise(mean(upr)) %>% 
  pull()

lwr_mpa_ceme <- pv_local %>% 
  filter(spp == "CEME") %>% 
  summarise(mean(lwr)) %>% 
  pull()
```

```{r}
scaled.mem.data.alldates.noNAs <- scaled.mem.data.alldates %>% 
  drop_na(mpa, lfm, PC1)

tmp <- scaled.mem.data.alldates.noNAs %>% 
  mutate(tlp = case_when(
    spp == "ADFA" ~ mean_mpa_adfa, 
    spp == "CEME" ~ mean_mpa_ceme
  )) %>% 
  mutate(tlp_split = case_when(
    mpa > tlp ~ "tlp_above", 
    mpa < tlp ~ "tlp_below"
  )) 
tmp
```

# Global Options: Colors, Themes
```{r}
th <- theme(panel.background = element_rect(fill='white', colour='black'), # Make background white and border black
          panel.grid.major = element_blank(),  # Hide major gridlines
          panel.grid.minor = element_blank())

th.spp.wrap <- theme(legend.position = "bottom", 
        strip.text = element_text(face = "italic", size = 12) +
  scale_color_manual(values=met.brewer("Morgenstern", 7)))

Morgenstern = list(c("#7c668c", "#b08ba5", "#dfbbc8", "#ffc680", "#ffb178", "#db8872", "#a56457"), c(7, 5, 4, 6, 3, 2, 1), colorblind=TRUE)
```


# Function: Segmented Regression Visualization
```{r}
segplot <- function(data, x, y, z, tlp_value, lwr, upr, lwr_seg, upr_seg, color = NULL) {
  ggplot(data, aes({{x}}, {{y}}, color = {{color}})) +
  geom_point()+
  geom_line(aes({{x}}, {{z}}), #segmented line
            color = "blue2") +
  geom_vline(xintercept = tlp_value, #tlp value line
             color = "gold1") +
  geom_ribbon(aes(xmin=lwr, xmax=upr), #tlp CI ribbon
              alpha = .3,
              color = "gold1") +
  geom_vline(xintercept = lwr_seg, #segmented CI line lower
             color = "blue2")+
  geom_vline(xintercept = upr_seg, #segmented CI line upper
             color = "blue2") +
  geom_ribbon(aes(xmin=lwr_seg, xmax=upr_seg), #segmented ribbon
              alpha = .3, 
              color = "blue2") +
   th #theme!
}
```
# ----------------------------------
# Start of Segmented Regressions
# ----------------------------------

# - Live Fuel Moisture -
# PC1

#### ADFA
```{r}
seg.adfa.subset <- seg.adfa.subset %>% 
  filter(lfm < 150)

out.lm<-lm(PC1 ~ lfm + year.month + site, data = seg.adfa.subset)
davies.test(out.lm, ~lfm, k = 50)
# Result: insignificant (p = 0.0888)

segmented.mod <-segmented(out.lm, seg.Z = ~lfm) #1 breakpoint for lfm
segmented.mod
summary(segmented.mod)

confint <- confint(segmented.mod)
confint

lower <- confint[2]
lower

upper <- confint[3]
upper

fit <- numeric(length(seg.adfa.subset$lfm)) * NA

fit[complete.cases(rowSums(cbind(seg.adfa.subset$PC1, seg.adfa.subset$lfm)))] <- broken.line(segmented.mod)$fit

segplot(seg.adfa.subset, lfm, PC1, fit, mean_lfm_adfa, lwr_lfm_adfa, upr_lfm_adfa, lower, upper, color = year.month)
```

#### CEME

```{r}
seg.ceme.subset <- seg.ceme.subset %>% 
  filter(lfm < 300)

out.lm<-lm(PC1 ~ lfm + site + year.month, data = seg.ceme.subset)
davies.test(out.lm, ~lfm, k = 50)
# Result: marginally insignificant (p = 0.05644)

segmented.mod <-segmented(out.lm, seg.Z = ~lfm) #1 breakpoint for x
segmented.mod

confint <- confint(segmented.mod)
confint

lower <- confint[2]
lower

upper <- confint[3]
upper

fit <- numeric(length(seg.ceme.subset$lfm)) * NA

fit[complete.cases(rowSums(cbind(seg.ceme.subset$PC1, seg.ceme.subset$lfm)))] <- broken.line(segmented.mod)$fit

segplot(seg.ceme.subset, lfm, PC1, fit, mean_lfm_ceme, lwr_lfm_ceme, upr_lfm_ceme, lower, upper, color = year.month)
```

# PC2

#### ADFA
```{r}
out.lm<-lm(PC2 ~ lfm, data = seg.adfa.subset)
davies.test(out.lm, ~lfm, k = 50)
# Result: significant (p = 0.0142)

segmented.mod <-segmented(out.lm, seg.Z = ~lfm) #1 breakpoint for x
segmented.mod

confint <- confint(segmented.mod)
confint

lower <- confint[2]
lower

upper <- confint[3]
upper

fit <- numeric(length(seg.adfa.subset$lfm)) * NA

fit[complete.cases(rowSums(cbind(seg.adfa.subset$PC1, seg.adfa.subset$lfm)))] <- broken.line(segmented.mod)$fit

segplot(seg.adfa.subset, lfm, PC2, fit, mean_lfm_adfa, lwr_lfm_adfa, upr_lfm_adfa, lower, upper)
```

#### CEME

```{r}
out.lm<-lm(PC2 ~ lfm + site + year.month + sample.wt, data = seg.ceme.subset)
davies.test(out.lm, ~lfm, k = 50)
# Result: significant (p = 0.02631)

segmented.mod <-segmented(out.lm, seg.Z = ~lfm) #1 breakpoint for x
segmented.mod
summary(segmented.mod)


confint <- confint(segmented.mod)
confint

lower <- confint[2]
lower

upper <- confint[3]
upper

fit <- numeric(length(seg.ceme.subset$lfm)) * NA

fit[complete.cases(rowSums(cbind(seg.ceme.subset$PC1, seg.ceme.subset$lfm)))] <- broken.line(segmented.mod)$fit

segplot(seg.ceme.subset, lfm, PC2, fit, mean_lfm_ceme, lwr_lfm_ceme, upr_lfm_ceme, lower, upper)
```

# PC3

#### ADFA
```{r}
out.lm<-lm(PC3 ~ lfm, data = seg.adfa.subset)
davies.test(out.lm, ~lfm, k = 50)
# Result: significant (p = 0.0303)

segmented.mod <-segmented(out.lm) #1 breakpoint for x
segmented.mod

confint <- confint(segmented.mod)
confint

lower <- confint[2]
lower

upper <- confint[3]
upper

fit <- numeric(length(seg.adfa.subset$lfm)) * NA

fit[complete.cases(rowSums(cbind(seg.adfa.subset$PC1, seg.adfa.subset$lfm)))] <- broken.line(segmented.mod)$fit

segplot(seg.adfa.subset, lfm, PC3, fit, mean_lfm_adfa, lwr_lfm_adfa, upr_lfm_adfa, lower, upper)
```

#### CEME
```{r}
out.lm<-lm(PC3 ~ lfm, data = seg.ceme.subset)
davies.test(out.lm, ~lfm, k = 50)
# Result: significant (p = 0.01461)

segmented.mod <-segmented(out.lm) #1 breakpoint for x
segmented.mod
summary(segmented.mod)


confint <- confint(segmented.mod)
confint

lower <- confint[2]
lower

upper <- confint[3]
upper

fit <- numeric(length(seg.ceme.subset$lfm)) * NA

fit[complete.cases(rowSums(cbind(seg.ceme.subset$PC1, seg.ceme.subset$lfm)))] <- broken.line(segmented.mod)$fit

segplot(seg.ceme.subset, lfm, PC3, fit, mean_lfm_ceme, lwr_lfm_ceme, upr_lfm_ceme, lower, upper)
```

# Prop. Ignited (10% Bins)
## ADFA
```{r}
out.lm<-lm(prop.ignite.bins10_nodate ~ lfm, data = seg.adfa.subset) 
davies.test(out.lm, ~lfm, k = 50)
# Highly significant (p < 0.00005)

segmented.mod <-segmented(out.lm, seg.Z = ~lfm) #1 breakpoint for x
segmented.mod

confint(segmented.mod)

fit <- numeric(length(seg.adfa.subset$lfm)) * NA

fit[complete.cases(rowSums(cbind(seg.adfa.subset$prop.ignite.bins10_nodate, seg.adfa.subset$lfm)))] <- broken.line(segmented.mod)$fit

confint <- confint(segmented.mod)
confint

lower <- confint[2]
lower

upper <- confint[3]
upper

segplot(seg.adfa.subset, lfm, prop.ignite.bins10_nodate, fit, mean_lfm_ceme, lwr_lfm_ceme, upr_lfm_ceme, lower, upper)
```

## CEME
```{r}
out.lm<-lm(prop.ignite.bins10_nodate ~ lfm, data = seg.ceme.subset)
davies.test(out.lm, ~lfm, k = 50)
#Result: marginally significant (p = 0.04587)
 
segmented.mod <-segmented(out.lm) #1 breakpoint for x
segmented.mod
summary(segmented.mod)

confint <- confint(segmented.mod)
confint

lower <- confint[2]
lower

upper <- confint[3]
upper

fit <- numeric(length(seg.ceme.subset$lfm)) * NA

fit[complete.cases(rowSums(cbind(seg.ceme.subset$prop.ignite.bins10_nodate, seg.ceme.subset$lfm)))] <- broken.line(segmented.mod)$fit

segplot(seg.ceme.subset, lfm, prop.ignite.bins10_nodate, fit, mean_lfm_ceme, lwr_lfm_ceme, upr_lfm_ceme, lower, upper)
```

# *** Prop. Ignited (10% Bins)
***: one point per bin

### ADFA
```{r}
prop.ignite.df.adfa <- seg.adfa.subset %>% 
  filter(sample.wt == 0.5210 | sample.wt == 0.4984 & tti == 25| sample.wt == 0.4913 | sample.wt == 0.4799 & tti == 43| sample.wt == 0.5079 | sample.wt == 0.5518 | sample.wt == 0.4862 | sample.wt == 0.5344 | sample.wt == 0.5240 & tti == 20)
```

```{r}
out.lm<-lm(prop.ignite.bins10_nodate ~ lfm, data = prop.ignite.df.adfa) 
davies.test(out.lm, ~lfm, k = 50)
# Highly significant (p < 0.00005)

segmented.mod <-segmented(out.lm, seg.Z = ~lfm) #1 breakpoint for x
segmented.mod

confint(segmented.mod)

fit <- numeric(length(prop.ignite.df.adfa$lfm)) * NA

fit[complete.cases(rowSums(cbind(prop.ignite.df.adfa$prop.ignite.bins10_nodate, prop.ignite.df.adfa$lfm)))] <- broken.line(segmented.mod)$fit

confint <- confint(segmented.mod)
confint

lower <- confint[2]
lower

upper <- confint[3]
upper

segplot(prop.ignite.df.adfa, lfm, prop.ignite.bins10_nodate, fit, mean_lfm_ceme, lwr_lfm_ceme, upr_lfm_ceme, lower, upper)
```

### CEME
```{r}
prop.ignite.df.ceme <- seg.ceme.subset %>% 
  filter(sample.wt == 0.5793 | sample.wt == 0.4870 & tti == 32| sample.wt == 0.5347 | sample.wt == 0.4488 & tti == 49| sample.wt == 0.5266 | sample.wt == 0.5246 | sample.wt == 0.5478 | sample.wt == 0.4908 | sample.wt == 0.4924 & tti == 56 | sample.wt == 0.5940 & tti == 43 | sample.wt == 0.5138 & tti == 40 | sample.wt == 0.4620 & tti == 49 | sample.wt == 0.5320 & tti == 49 | sample.wt == 0.5340 & tti == 52 | sample.wt == 0.5739 & tti == 33 | sample.wt == 0.5520 & tti == 42 | sample.wt == 0.5110 & tti == 34 | sample.wt == 0.5130 & tti == 48 | sample.wt == 0.5240 & tti == 46)
```

```{r}
out.lm<-lm(prop.ignite.bins10_nodate ~ lfm, data = prop.ignite.df.ceme) 
davies.test(out.lm, ~lfm, k = 50)
# Highly significant (p < 0.00005)

segmented.mod <-segmented(out.lm, seg.Z = ~lfm) #1 breakpoint for x
segmented.mod

confint(segmented.mod)

fit <- numeric(length(prop.ignite.df.ceme$lfm)) * NA

fit[complete.cases(rowSums(cbind(prop.ignite.df.ceme$prop.ignite.bins10_nodate, prop.ignite.df.ceme$lfm)))] <- broken.line(segmented.mod)$fit

confint <- confint(segmented.mod)
confint

lower <- confint[2]
lower

upper <- confint[3]
upper

segplot(prop.ignite.df.ceme, lfm, prop.ignite.bins10_nodate, fit, mean_lfm_ceme, lwr_lfm_ceme, upr_lfm_ceme, lower, upper)
```

# TTI

### ADFA
```{r}
out.lm<-lm(tti ~ lfm + year.month + site, data = seg.adfa.subset)
davies.test(out.lm, ~lfm, k = 50)
#Result: significant

segmented.mod <-segmented(out.lm, seg.Z = ~lfm) #1 breakpoint for x
segmented.mod
summary(segmented.mod)

confint <- confint(segmented.mod)
confint

lower <- confint[2]
lower

upper <- confint[3]
upper

fit <- numeric(length(seg.adfa.subset$lfm)) * NA

fit[complete.cases(rowSums(cbind(seg.adfa.subset$tti, seg.adfa.subset$lfm)))] <- broken.line(segmented.mod)$fit

segplot(seg.adfa.subset, lfm, tti, fit, mean_lfm_adfa, lwr_lfm_adfa, upr_lfm_adfa, lower, upper, color = year.month)
```
### CEME

```{r}
out.lm<-lm(tti ~ lfm + site + year.month, data = seg.ceme.subset)
davies.test(out.lm, ~lfm, k = 50)
# Result: highly significant (p = 0.0009369)

segmented.mod <-segmented(out.lm, seg.Z = ~lfm) #1 breakpoint for x
segmented.mod
summary(segmented.mod)

confint <- confint(segmented.mod)
confint

lower <- confint[2]
lower

upper <- confint[3]
upper

fit <- numeric(length(seg.ceme.subset$lfm)) * NA

fit[complete.cases(rowSums(cbind(seg.ceme.subset$tti, seg.ceme.subset$lfm)))] <- broken.line(segmented.mod)$fit

segplot(seg.ceme.subset, lfm, tti, fit, mean_lfm_ceme, lwr_lfm_ceme, upr_lfm_ceme, lower, upper, color = year.month)
```


# FD

### ADFA

```{r}
out.lm<-lm(fd ~ lfm + site + year.month + sample.wt, data = seg.adfa.subset)
davies.test(out.lm, ~lfm, k=50)
#Result: insignificant (p = 0.6856)

segmented.mod <-segmented(out.lm, seg.Z = ~lfm) #1 breakpoint for x
segmented.mod

confint <- confint(segmented.mod)
confint

lower <- confint[2]
lower

upper <- confint[3]
upper

fit <- numeric(length(seg.adfa.subset$lfm)) * NA

fit[complete.cases(rowSums(cbind(seg.adfa.subset$fd, seg.adfa.subset$lfm)))] <- broken.line(segmented.mod)$fit

segplot(seg.adfa.subset, lfm, fd, fit, mean_lfm_adfa, lwr_lfm_adfa, upr_lfm_adfa, lower, upper)
```

### CEME
```{r}
out.lm<-lm(fd ~ lfm + site + year.month + sample.wt, data = seg.ceme.subset)
davies.test(out.lm, ~lfm, k = 50)
#Result: insignificant (p = 0.4814)

segmented.mod <-segmented(out.lm, seg.Z = ~lfm) #1 breakpoint for x
segmented.mod

confint <- confint(segmented.mod)
confint

lower <- confint[2]
lower

upper <- confint[3]
upper

fit <- numeric(length(seg.ceme.subset$lfm)) * NA

fit[complete.cases(rowSums(cbind(seg.ceme.subset$fd, seg.ceme.subset$lfm)))] <- broken.line(segmented.mod)$fit

segplot(seg.ceme.subset, lfm, fd, fit, mean_lfm_ceme, lwr_lfm_ceme, upr_lfm_ceme, lower, upper)
```

# FH

### ADFA

```{r}
out.lm<-lm(fh ~ lfm + site + year.month + sample.wt, data = seg.adfa.subset)
davies.test(out.lm, ~lfm, k = 50)
# result: significant (p = 0.001313)

segmented.mod <-segmented(out.lm, seg.Z = ~lfm) #1 breakpoint for x
segmented.mod

confint <- confint(segmented.mod)
confint

lower <- confint[2]
lower

upper <- confint[3]
upper

fit <- numeric(length(seg.adfa.subset$lfm)) * NA

fit[complete.cases(rowSums(cbind(seg.adfa.subset$fh, seg.adfa.subset$lfm)))] <- broken.line(segmented.mod)$fit

segplot(seg.adfa.subset, lfm, fh, fit, mean_lfm_adfa, lwr_lfm_adfa, upr_lfm_adfa, lower, upper)

```

### CEME
```{r}
out.lm<-lm(fh ~ lfm + site + year.month + sample.wt, data = seg.ceme.subset)
davies.test(out.lm, ~lfm, k = 50)
#Result: significant (p = 0.03032)

segmented.mod <-segmented(out.lm, seg.Z = ~lfm) #1 breakpoint for x
segmented.mod

confint <- confint(segmented.mod)
confint

lower <- confint[2]
lower

upper <- confint[3]
upper

fit <- numeric(length(seg.ceme.subset$lfm)) * NA

fit[complete.cases(rowSums(cbind(seg.ceme.subset$fh, seg.ceme.subset$lfm)))] <- broken.line(segmented.mod)$fit

segplot(seg.ceme.subset, lfm, fh, fit, mean_lfm_ceme, lwr_lfm_ceme, upr_lfm_ceme, lower, upper)
```

# GD

### ADFA
```{r}
out.lm<-lm(gd ~ lfm + site + year.month + sample.wt, data = seg.adfa.subset)
davies.test(out.lm, ~lfm, k = 50)
# result: insignificant

segmented.mod <-segmented(out.lm, seg.Z = ~lfm) #1 breakpoint for x
segmented.mod

confint <- confint(segmented.mod)
confint

lower <- confint[2]
lower

upper <- confint[3]
upper

fit <- numeric(length(seg.adfa.subset$lfm)) * NA

fit[complete.cases(rowSums(cbind(seg.adfa.subset$gd, seg.adfa.subset$lfm)))] <- broken.line(segmented.mod)$fit

segplot(seg.adfa.subset, lfm, gd, fit, mean_lfm_adfa, lwr_lfm_adfa, upr_lfm_adfa, lower, upper)

```

### CEME
```{r}
seg.ceme.subset.gd <- seg.ceme.subset %>% 
  filter(gd < 360) #Removing unrealistic glow durations
```


```{r}
out.lm <- lm(gd ~ lfm + site + year.month + sample.wt, data = seg.ceme.subset.gd)
davies.test(out.lm, ~lfm, k = 50)
#Result: insignificant

segmented.mod <-segmented(out.lm, seg.Z = ~lfm) #1 breakpoint for x
segmented.mod

confint <- confint(segmented.mod)
confint

lower <- confint[2]
lower

upper <- confint[3]
upper

fit <- numeric(length(seg.ceme.subset.gd$lfm)) * NA

fit[complete.cases(rowSums(cbind(seg.ceme.subset.gd$gd, seg.ceme.subset.gd$lfm)))] <- broken.line(segmented.mod)$fit

segplot(seg.ceme.subset.gd, lfm, gd, fit, mean_lfm_ceme, lwr_lfm_ceme, upr_lfm_ceme, lower, upper)
```


# GTI
No significant thresholds identified
### ADFA
```{r}
out.lm<-lm(gti ~ lfm + site + year.month + sample.wt, data = seg.adfa.subset)
davies.test(out.lm, ~lfm, k = 50)
# result: insignificant

segmented.mod <-segmented(out.lm, seg.Z = ~lfm) #1 breakpoint for x
segmented.mod

confint <- confint(segmented.mod)
confint

lower <- confint[2]
lower

upper <- confint[3]
upper

fit <- numeric(length(seg.adfa.subset$lfm)) * NA

fit[complete.cases(rowSums(cbind(seg.adfa.subset$gti, seg.adfa.subset$lfm)))] <- broken.line(segmented.mod)$fit

segplot(seg.adfa.subset, lfm, gti, fit, mean_lfm_adfa, lwr_lfm_adfa, upr_lfm_adfa, lower, upper)

```

### CEME
```{r}
out.lm <- lm(gti ~ lfm + site + year.month + sample.wt, data = seg.ceme.subset)
davies.test(out.lm, ~lfm, k = 50)
#Result: insignificant

segmented.mod <-segmented(out.lm, seg.Z = ~lfm) #1 breakpoint for x
segmented.mod

confint <- confint(segmented.mod)
confint

lower <- confint[2]
lower

upper <- confint[3]
upper

fit <- numeric(length(seg.ceme.subset$lfm)) * NA

fit[complete.cases(rowSums(cbind(seg.ceme.subset$gti, seg.ceme.subset$lfm)))] <- broken.line(segmented.mod)$fit

segplot(seg.ceme.subset, lfm, gti, fit, mean_lfm_ceme, lwr_lfm_ceme, upr_lfm_ceme, lower, upper)
```

# PFG
No significant thresholds identified
### ADFA
```{r}
out.lm<-lm(pfg ~ lfm + site + year.month + sample.wt, data = seg.adfa.subset)
davies.test(out.lm, ~lfm, k = 50)
# result: insignificant

segmented.mod <-segmented(out.lm, seg.Z = ~lfm) #1 breakpoint for x
segmented.mod

confint <- confint(segmented.mod)
confint

lower <- confint[2]
lower

upper <- confint[3]
upper

fit <- numeric(length(seg.adfa.subset$lfm)) * NA

fit[complete.cases(rowSums(cbind(seg.adfa.subset$pfg, seg.adfa.subset$lfm)))] <- broken.line(segmented.mod)$fit

segplot(seg.adfa.subset, lfm, pfg, fit, mean_lfm_adfa, lwr_lfm_adfa, upr_lfm_adfa, lower, upper)

```

### CEME
```{r}
out.lm <- lm(pfg ~ lfm + site + year.month + sample.wt, data = seg.ceme.subset)
davies.test(out.lm, ~lfm, k = 50)
#Result: marginally insignificant (p = 0.0667)

segmented.mod <-segmented(out.lm, seg.Z = ~lfm) #1 breakpoint for x
segmented.mod

confint <- confint(segmented.mod)
confint

lower <- confint[2]
lower

upper <- confint[3]
upper

fit <- numeric(length(seg.ceme.subset$lfm)) * NA

fit[complete.cases(rowSums(cbind(seg.ceme.subset$pfg, seg.ceme.subset$lfm)))] <- broken.line(segmented.mod)$fit

segplot(seg.ceme.subset, lfm, pfg, fit, mean_lfm_ceme, lwr_lfm_ceme, upr_lfm_ceme, lower, upper)
```



# Max. Temp.
### ADFA
```{r}
out.lm<-lm(temp.max ~ lfm + site + year.month + sample.wt, data = seg.adfa.subset)
davies.test(out.lm, ~lfm, k = 50)
# result: insignificant

segmented.mod <-segmented(out.lm, seg.Z = ~lfm) #1 breakpoint for x
segmented.mod

confint <- confint(segmented.mod)
confint

lower <- confint[2]
lower

upper <- confint[3]
upper

fit <- numeric(length(seg.adfa.subset$lfm)) * NA

fit[complete.cases(rowSums(cbind(seg.adfa.subset$temp.max, seg.adfa.subset$lfm)))] <- broken.line(segmented.mod)$fit

segplot(seg.adfa.subset, lfm, temp.max, fit, mean_lfm_adfa, lwr_lfm_adfa, upr_lfm_adfa, lower, upper)
```

### CEME
```{r}
out.lm <- lm(temp.max ~ lfm + site + year.month + sample.wt, data = seg.ceme.subset)
davies.test(out.lm, ~lfm, k = 50)
#Result: significant (p = 0.000395)

segmented.mod <-segmented(out.lm, seg.Z = ~lfm) #1 breakpoint for x
segmented.mod

confint <- confint(segmented.mod)
confint

lower <- confint[2]
lower

upper <- confint[3]
upper

fit <- numeric(length(seg.ceme.subset$lfm)) * NA

fit[complete.cases(rowSums(cbind(seg.ceme.subset$temp.max, seg.ceme.subset$lfm)))] <- broken.line(segmented.mod)$fit

segplot(seg.ceme.subset, lfm, temp.max, fit, mean_lfm_ceme, lwr_lfm_ceme, upr_lfm_ceme, lower, upper)
```

# TTFG

### ADFA
For time to first glow, there is one strange value remaining where it probably didn't glow or didn't ignite (ttfg = 490)
```{r}
seg.adfa.subset.ttfg <- seg.adfa.subset %>% 
  filter(ttfg < 300)
```

```{r}
out.lm<-lm(ttfg ~ lfm + site + year.month + sample.wt, data = seg.adfa.subset.ttfg)
davies.test(out.lm, ~lfm, k = 50)
# result: insignificant

segmented.mod <-segmented(out.lm, seg.Z = ~lfm) #1 breakpoint for x
segmented.mod

confint <- confint(segmented.mod)
confint

lower <- confint[2]
lower

upper <- confint[3]
upper

fit <- numeric(length(seg.adfa.subset$lfm)) * NA

fit[complete.cases(rowSums(cbind(seg.adfa.subset$ttfg, seg.adfa.subset$lfm)))] <- broken.line(segmented.mod)$fit

segplot(seg.adfa.subset, lfm, ttfg, fit, mean_lfm_adfa, lwr_lfm_adfa, upr_lfm_adfa, lower, upper)
```

### CEME
```{r}
out.lm <- lm(ttfg ~ lfm + site + year.month + sample.wt, data = seg.ceme.subset)
davies.test(out.lm, ~lfm, k = 50)
#Result: significant (p = 0.005595)

segmented.mod <-segmented(out.lm, seg.Z = ~lfm) #1 breakpoint for x
segmented.mod

confint <- confint(segmented.mod)
confint

lower <- confint[2]
lower

upper <- confint[3]
upper

fit <- numeric(length(seg.ceme.subset$lfm)) * NA

fit[complete.cases(rowSums(cbind(seg.ceme.subset$ttfg, seg.ceme.subset$lfm)))] <- broken.line(segmented.mod)$fit

segplot(seg.ceme.subset, lfm, ttfg, fit, mean_lfm_ceme, lwr_lfm_ceme, upr_lfm_ceme, lower, upper)
```
# -----------------------------------
# - Water Potential - 
Repeating the same analysis above, but using water potential as primary predictor instead of LFM

# PC1
#### ADFA 
no breakpoint
```{r}
out.lm<-lm(PC1 ~ mpa + site + year.month + sample.wt, data = seg.adfa.subset)
davies.test(out.lm, ~mpa, k = 50)
#Result: insignifcant (p = 0.6182)

segmented.mod <-segmented(out.lm, seg.Z = ~mpa) #1 breakpoint for x
segmented.mod

confint <- confint(segmented.mod)
confint

lower <- confint[2]
lower

upper <- confint[3]
upper

fit <- numeric(length(seg.adfa.subset$mpa)) * NA

#fit[complete.cases(rowSums(cbind(seg.adfa.subset$PC1, seg.adfa.subset$mpa)))] <- broken.line(segmented.mod)$fit

segplot(seg.adfa.subset, mpa, PC1, fit, mean_mpa_adfa, lwr_mpa_adfa, upr_mpa_adfa, lower, upper)
```
#### CEME
```{r}
out.lm<-lm(PC1 ~ mpa + site + year.month + sample.wt, data = seg.ceme.subset)
davies.test(out.lm, ~mpa, k = 50)
#Result: insignificant (p = 0.6775)

segmented.mod <-segmented(out.lm, seg.Z = ~mpa) #1 breakpoint for x
#segmented.mod
#summary(segmented.mod)


confint <- confint(segmented.mod)
confint

lower <- confint[2]
lower

upper <- confint[3]
upper

fit <- numeric(length(seg.ceme.subset$mpa)) * NA

fit[complete.cases(rowSums(cbind(seg.ceme.subset$PC1, seg.ceme.subset$mpa)))] <- broken.line(segmented.mod)$fit

segplot(seg.ceme.subset, mpa, PC1, fit, mean_mpa_ceme, lwr_mpa_ceme, upr_mpa_ceme, lower, upper)
```

# PC2

#### ADFA
```{r}
out.lm<-lm(PC2 ~ mpa + site + year.month + sample.wt, data = seg.adfa.subset)
davies.test(out.lm, ~mpa, k = 50)
# Result: insignificant

segmented.mod <-segmented(out.lm, seg.Z = ~mpa) #1 breakpoint for x
segmented.mod

confint <- confint(segmented.mod)
confint

lower <- confint[2]
lower

upper <- confint[3]
upper

fit <- numeric(length(seg.adfa.subset$mpa)) * NA

fit[complete.cases(rowSums(cbind(seg.adfa.subset$PC1, seg.adfa.subset$mpa)))] <- broken.line(segmented.mod)$fit

segplot(seg.adfa.subset, mpa, PC2, fit, mean_mpa_adfa, lwr_mpa_adfa, upr_mpa_adfa, lower, upper)
```

#### CEME

```{r}
out.lm<-lm(PC2 ~ mpa + site + year.month + sample.wt, data = seg.ceme.subset)
davies.test(out.lm, ~mpa, k = 50)
# Result: insignificant (p = 0.3722)

segmented.mod <-segmented(out.lm, seg.Z = ~mpa) #1 breakpoint for x
segmented.mod
summary(segmented.mod)


confint <- confint(segmented.mod)
confint

lower <- confint[2]
lower

upper <- confint[3]
upper

fit <- numeric(length(seg.ceme.subset$mpa)) * NA

fit[complete.cases(rowSums(cbind(seg.ceme.subset$PC1, seg.ceme.subset$mpa)))] <- broken.line(segmented.mod)$fit

segplot(seg.ceme.subset, mpa, PC2, fit, mean_mpa_ceme, lwr_mpa_ceme, upr_mpa_ceme, lower, upper)
```

# PC3

#### ADFA
```{r}
out.lm<-lm(PC3 ~ mpa + site + year.month + sample.wt, data = seg.adfa.subset)
davies.test(out.lm, ~mpa, k = 50)
# Result: insignificant (p = 1)

segmented.mod <-segmented(out.lm, seg.Z = ~mpa) #1 breakpoint for x
segmented.mod

confint <- confint(segmented.mod)
confint

lower <- confint[2]
lower

upper <- confint[3]
upper

fit <- numeric(length(seg.adfa.subset$mpa)) * NA

fit[complete.cases(rowSums(cbind(seg.adfa.subset$PC1, seg.adfa.subset$mpa)))] <- broken.line(segmented.mod)$fit

segplot(seg.adfa.subset, mpa, PC3, fit, mean_mpa_adfa, lwr_mpa_adfa, upr_mpa_adfa, lower, upper)
```

#### CEME
```{r}
out.lm<-lm(PC3 ~ mpa + site + year.month + sample.wt, data = seg.ceme.subset)
davies.test(out.lm, ~mpa, k = 50)
# Result: significant (p = 0.0311)

segmented.mod <-segmented(out.lm, seg.Z = ~mpa) #1 breakpoint for x
segmented.mod
summary(segmented.mod)


confint <- confint(segmented.mod)
confint

lower <- confint[2]
lower

upper <- confint[3]
upper

fit <- numeric(length(seg.ceme.subset$mpa)) * NA

fit[complete.cases(rowSums(cbind(seg.ceme.subset$PC1, seg.ceme.subset$mpa)))] <- broken.line(segmented.mod)$fit

segplot(seg.ceme.subset, mpa, PC3, fit, mean_mpa_ceme, lwr_mpa_ceme, upr_mpa_ceme, lower, upper)
```

# Prop. Ignited (10% Bins)
## ADFA
```{r}
out.lm<-lm(prop.ignite.bins10_nodate ~ mpa + site + year.month + sample.wt, data = seg.adfa.subset) 
davies.test(out.lm, ~mpa, k = 50)
# insignificant

segmented.mod <-segmented(out.lm, seg.Z = ~mpa) #1 breakpoint for x
segmented.mod

confint(segmented.mod)

fit <- numeric(length(seg.adfa.subset$mpa)) * NA

fit[complete.cases(rowSums(cbind(seg.adfa.subset$prop.ignite.bins5_nodate, seg.adfa.subset$mpa)))] <- broken.line(segmented.mod)$fit

confint <- confint(segmented.mod)
confint

lower <- confint[2]
lower

upper <- confint[3]
upper

segplot(seg.adfa.subset, mpa, prop.ignite.bins5_nodate, fit, mean_mpa_ceme, lwr_mpa_ceme, upr_mpa_ceme, lower, upper)
```

## CEME
```{r}
out.lm<-lm(prop.ignite.bins10_nodate ~ mpa + site + year.month + sample.wt, data = seg.ceme.subset)
davies.test(out.lm, ~mpa, k = 50)
#Result: insignificant (p = 1)

segmented.mod <-segmented(out.lm, seg.Z = ~mpa) #1 breakpoint for x
segmented.mod
summary(segmented.mod)

confint <- confint(segmented.mod)
confint

lower <- confint[2]
lower

upper <- confint[3]
upper

fit <- numeric(length(seg.ceme.subset$mpa)) * NA

fit[complete.cases(rowSums(cbind(seg.ceme.subset$prop.ignite.bins10_nodate, seg.ceme.subset$mpa)))] <- broken.line(segmented.mod)$fit

segplot(seg.ceme.subset, mpa, prop.ignite.bins10_nodate, fit, mean_mpa_ceme, lwr_mpa_ceme, upr_mpa_ceme, lower, upper)
```

# TTI
#### ADFA
```{r}
out.lm<-lm(tti ~ mpa + site + year.month + sample.wt, data = seg.adfa.subset)
davies.test(out.lm, ~mpa, k=50)
# Result: insignificant (p = 0.6741)

segmented.mod <-segmented(out.lm, seg.Z = ~mpa) #1 breakpoint for x
segmented.mod

confint <- confint(segmented.mod)
confint

lower <- confint[2]
lower

upper <- confint[3]
upper

fit <- numeric(length(seg.adfa.subset$mpa)) * NA

fit[complete.cases(rowSums(cbind(seg.adfa.subset$tti, seg.adfa.subset$mpa)))] <- broken.line(segmented.mod)$fit

segplot(seg.adfa.subset, mpa, tti, fit, mean_mpa_adfa, lwr_mpa_adfa, upr_mpa_adfa, lower, upper)
```

#### CEME
```{r}
out.lm<-lm(tti ~ mpa + site + year.month + sample.wt, data = seg.ceme.subset)
davies.test(out.lm, ~mpa, k = 50)
#Result: significant (p = 0.01178)

segmented.mod <-segmented(out.lm, seg.Z = ~mpa) #1 breakpoint for x
segmented.mod

confint <- confint(segmented.mod)
confint

lower <- confint[2]
lower

upper <- confint[3]
upper

fit <- numeric(length(seg.ceme.subset$mpa)) * NA

fit[complete.cases(rowSums(cbind(seg.ceme.subset$tti, seg.ceme.subset$mpa)))] <- broken.line(segmented.mod)$fit

segplot(seg.ceme.subset, mpa, tti, fit, mean_mpa_ceme, lwr_mpa_ceme, upr_mpa_ceme, lower, upper)
```

# FD
### ADFA
```{r}
out.lm<-lm(fd ~ mpa + site + year.month + sample.wt, data = seg.adfa.subset)
davies.test(out.lm, ~mpa, k=50)
# Result: insignificant (p = 0.7258)

segmented.mod <-segmented(out.lm, seg.Z = ~mpa) #1 breakpoint for x
segmented.mod

confint <- confint(segmented.mod)
confint

lower <- confint[2]
lower

upper <- confint[3]
upper

fit <- numeric(length(seg.adfa.subset$mpa)) * NA

fit[complete.cases(rowSums(cbind(seg.adfa.subset$fd, seg.adfa.subset$mpa)))] <- broken.line(segmented.mod)$fit

segplot(seg.adfa.subset, mpa, fd, fit, mean_mpa_adfa, lwr_mpa_adfa, upr_mpa_adfa, lower, upper)
```

### CEME
```{r}
out.lm<-lm(fd ~ mpa + site + year.month + sample.wt, data = seg.ceme.subset)
davies.test(out.lm, ~mpa, k = 50)
#Result: insignificant (p = 0.7735)

segmented.mod <-segmented(out.lm, seg.Z = ~mpa) #1 breakpoint for x
segmented.mod

confint <- confint(segmented.mod)
confint

lower <- confint[2]
lower

upper <- confint[3]
upper

fit <- numeric(length(seg.ceme.subset$mpa)) * NA

fit[complete.cases(rowSums(cbind(seg.ceme.subset$fd, seg.ceme.subset$mpa)))] <- broken.line(segmented.mod)$fit

segplot(seg.ceme.subset, mpa, fd, fit, mean_mpa_ceme, lwr_mpa_ceme, upr_mpa_ceme, lower, upper)
```

# FH
#### ADFA
No breakpoint estimated

#### CEME
```{r}
out.lm<-lm(fh ~ mpa + site + year.month + sample.wt, data = seg.ceme.subset)
davies.test(out.lm, ~mpa, k = 50)
# result: significant (p = 0.001903)

segmented.mod <-segmented(out.lm, seg.Z = ~mpa) #1 breakpoint for x
segmented.mod

confint <- confint(segmented.mod)
confint

lower <- confint[2]
lower

upper <- confint[3]
upper

fit <- numeric(length(seg.ceme.subset$mpa)) * NA

fit[complete.cases(rowSums(cbind(seg.ceme.subset$fh, seg.ceme.subset$mpa)))] <- broken.line(segmented.mod)$fit

segplot(seg.ceme.subset, mpa, fh, fit, mean_mpa_ceme, lwr_mpa_ceme, upr_mpa_ceme, lower, upper)
```

# GD
#### ADFA
```{r}
out.lm<-lm(gd ~ mpa + site + year.month + sample.wt, data = seg.adfa.subset)
davies.test(out.lm, ~mpa, k = 50)
# Result: insignificant

segmented.mod <-segmented(out.lm, seg.Z = ~mpa) #1 breakpoint for x
segmented.mod

confint <- confint(segmented.mod)
confint

lower <- confint[2]
lower

upper <- confint[3]
upper

fit <- numeric(length(seg.adfa.subset$mpa)) * NA

fit[complete.cases(rowSums(cbind(seg.adfa.subset$gd, seg.adfa.subset$mpa)))] <- broken.line(segmented.mod)$fit

segplot(seg.adfa.subset, mpa, gd, fit, mean_mpa_adfa, lwr_mpa_adfa, upr_mpa_adfa, lower, upper)
```

#### CEME
```{r}
out.lm<-lm(gd ~ mpa + site + year.month + sample.wt, data = seg.ceme.subset)
davies.test(out.lm, ~mpa, k = 50)
# result: insignificant

segmented.mod <-segmented(out.lm, seg.Z = ~mpa) #1 breakpoint for x
segmented.mod

confint <- confint(segmented.mod)
confint

lower <- confint[2]
lower

upper <- confint[3]
upper

fit <- numeric(length(seg.ceme.subset$mpa)) * NA

fit[complete.cases(rowSums(cbind(seg.ceme.subset$gd, seg.ceme.subset$mpa)))] <- broken.line(segmented.mod)$fit

segplot(seg.ceme.subset, mpa, gd, fit, mean_mpa_ceme, lwr_mpa_ceme, upr_mpa_ceme, lower, upper)
```

# GTI
### ADFA
```{r}
out.lm<-lm(gti ~ mpa + site + year.month + sample.wt, data = seg.adfa.subset)
davies.test(out.lm, ~mpa, k = 50)
# result: insignificant

segmented.mod <-segmented(out.lm, seg.Z = ~mpa) #1 breakpoint for x
segmented.mod

confint <- confint(segmented.mod)
confint

lower <- confint[2]
lower

upper <- confint[3]
upper

fit <- numeric(length(seg.adfa.subset$mpa)) * NA

fit[complete.cases(rowSums(cbind(seg.adfa.subset$gti, seg.adfa.subset$mpa)))] <- broken.line(segmented.mod)$fit

segplot(seg.adfa.subset, mpa, gti, fit, mean_mpa_adfa, lwr_mpa_adfa, upr_mpa_adfa, lower, upper)

```

### CEME
```{r}
out.lm <- lm(gti ~ mpa + site + year.month + sample.wt, data = seg.ceme.subset)
davies.test(out.lm, ~mpa, k = 50)
#Result: significant (0.01812)

segmented.mod <-segmented(out.lm, seg.Z = ~mpa) #1 breakpoint for x
segmented.mod

confint <- confint(segmented.mod)
confint

lower <- confint[2]
lower

upper <- confint[3]
upper

fit <- numeric(length(seg.ceme.subset$mpa)) * NA

fit[complete.cases(rowSums(cbind(seg.ceme.subset$gti, seg.ceme.subset$mpa)))] <- broken.line(segmented.mod)$fit

segplot(seg.ceme.subset, mpa, gti, fit, mean_mpa_ceme, lwr_mpa_ceme, upr_mpa_ceme, lower, upper)
```
# PFG
### ADFA
```{r}
out.lm<-lm(pfg ~ mpa + site + year.month + sample.wt, data = seg.adfa.subset)
davies.test(out.lm, ~mpa, k = 50)
# result: insignificant

segmented.mod <-segmented(out.lm, seg.Z = ~mpa) #1 breakpoint for x
segmented.mod

confint <- confint(segmented.mod)
confint

lower <- confint[2]
lower

upper <- confint[3]
upper

fit <- numeric(length(seg.adfa.subset$mpa)) * NA

fit[complete.cases(rowSums(cbind(seg.adfa.subset$pfg, seg.adfa.subset$mpa)))] <- broken.line(segmented.mod)$fit

segplot(seg.adfa.subset, mpa, pfg, fit, mean_mpa_adfa, lwr_mpa_adfa, upr_mpa_adfa, lower, upper)

```

### CEME
```{r}
out.lm <- lm(pfg ~ mpa + site + year.month + sample.wt, data = seg.ceme.subset)
davies.test(out.lm, ~mpa, k = 50)
#Result: insignificant

segmented.mod <-segmented(out.lm, seg.Z = ~mpa) #1 breakpoint for x
segmented.mod

confint <- confint(segmented.mod)
confint

lower <- confint[2]
lower

upper <- confint[3]
upper

fit <- numeric(length(seg.ceme.subset$mpa)) * NA

fit[complete.cases(rowSums(cbind(seg.ceme.subset$pfg, seg.ceme.subset$mpa)))] <- broken.line(segmented.mod)$fit

segplot(seg.ceme.subset, mpa, pfg, fit, mean_mpa_ceme, lwr_mpa_ceme, upr_mpa_ceme, lower, upper)
```



# Max. Temp.
### ADFA
```{r}
out.lm<-lm(temp.max ~ mpa + site + year.month + sample.wt, data = seg.adfa.subset)
davies.test(out.lm, ~mpa, k = 50)
# result: insignificant

segmented.mod <-segmented(out.lm, seg.Z = ~mpa) #1 breakpoint for x
segmented.mod

confint <- confint(segmented.mod)
confint

lower <- confint[2]
lower

upper <- confint[3]
upper

fit <- numeric(length(seg.adfa.subset$mpa)) * NA

fit[complete.cases(rowSums(cbind(seg.adfa.subset$temp.max, seg.adfa.subset$mpa)))] <- broken.line(segmented.mod)$fit

segplot(seg.adfa.subset, mpa, temp.max, fit, mean_mpa_adfa, lwr_mpa_adfa, upr_mpa_adfa, lower, upper)
```

### CEME
```{r}
out.lm <- lm(temp.max ~ mpa + site + year.month + sample.wt, data = seg.ceme.subset)
davies.test(out.lm, ~mpa, k = 50)
#Result: insignificant

segmented.mod <-segmented(out.lm, seg.Z = ~mpa) #1 breakpoint for x
segmented.mod

confint <- confint(segmented.mod)
confint

lower <- confint[2]
lower

upper <- confint[3]
upper

fit <- numeric(length(seg.ceme.subset$mpa)) * NA

fit[complete.cases(rowSums(cbind(seg.ceme.subset$temp.max, seg.ceme.subset$mpa)))] <- broken.line(segmented.mod)$fit

segplot(seg.ceme.subset, mpa, temp.max, fit, mean_mpa_ceme, lwr_mpa_ceme, upr_mpa_ceme, lower, upper)
```

# TTFG

### ADFA
```{r}
out.lm<-lm(ttfg ~ mpa + site + year.month + sample.wt, data = seg.adfa.subset.ttfg)
davies.test(out.lm, ~mpa, k = 50)
# result: insignificant

segmented.mod <-segmented(out.lm, seg.Z = ~mpa) #1 breakpoint for x
segmented.mod

confint <- confint(segmented.mod)
confint

lower <- confint[2]
lower

upper <- confint[3]
upper

fit <- numeric(length(seg.adfa.subset$mpa)) * NA

fit[complete.cases(rowSums(cbind(seg.adfa.subset$ttfg, seg.adfa.subset$mpa)))] <- broken.line(segmented.mod)$fit

segplot(seg.adfa.subset, mpa, ttfg, fit, mean_mpa_adfa, lwr_mpa_adfa, upr_mpa_adfa, lower, upper)
```

### CEME
```{r}
out.lm <- lm(ttfg ~ mpa + site + year.month + sample.wt, data = seg.ceme.subset)
davies.test(out.lm, ~mpa, k = 50)
#Result: insignificant

segmented.mod <-segmented(out.lm, seg.Z = ~mpa) #1 breakpoint for x
segmented.mod

confint <- confint(segmented.mod)
confint

lower <- confint[2]
lower

upper <- confint[3]
upper

fit <- numeric(length(seg.ceme.subset$mpa)) * NA

fit[complete.cases(rowSums(cbind(seg.ceme.subset$ttfg, seg.ceme.subset$mpa)))] <- broken.line(segmented.mod)$fit

segplot(seg.ceme.subset, mpa, ttfg, fit, mean_mpa_ceme, lwr_mpa_ceme, upr_mpa_ceme, lower, upper)
```
