==> scripts/4.2_bootstrapped_seg_reg.Rmd <==
---
title: "Segmented Regressions: Bootstrapping Prior to Analysis"
author: "Indra Boving & Joe Celebrezze"
date: "7/21/2022"
output: html_document
---

```{r setup, include = FALSE}
knitr::opts_chunk$set(echo = TRUE)
#lots of extra here, but oh well... 
library(ggplot2)
library(gapminder)
library(tidyverse)
library(lme4)
library(here)
library(segmented)
library(nlme)
library(lubridate)
library(MetBrewer)
library(lattice)
library(gt)
library(gtsummary)
library(kableExtra)
filter = dplyr::filter 
rename = dplyr::rename
here = here::here
group_by = dplyr::group_by
```

# Reading in Dataframes
Using all dates instead of just Sept. 2020 data
```{r}
seg.data.subset <- read.csv(here('processed-data', 'seg.reg.data.local.csv'))
```

# Data Wrangling
For the segmented regressions, we debated using seasonal LFM minimums and maximums to limit the dataset, but instead -- since we wanted to keep the datasets similar for each analysis -- we used the dataset that was used in the mixed effects model analysis

## Species Split
```{r}
seg.adfa.subset <- seg.data.subset %>% 
  filter(spp == "ADFA") %>% 
  group_by(lfm.bin = cut(lfm, breaks= seq(0, 120, by = 6))) %>%  #Fixing the bins
  drop_na(PC1, lfm, year.month, site, id) %>% 
  select(PC1, PC2, tti, fh, temp.max, gd, gti, pfg, ttfg, fd, prop.ignite.bins5_nodate, lfm, lfm.bin, mpa, year.month, site, id)

seg.ceme.subset <- seg.data.subset %>% 
  filter(spp == "CEME") %>% 
  filter(lfm < 220) %>% 
  group_by(lfm.bin = cut(lfm, breaks= seq(0, 216, by = 12))) %>%  #Fixing the bins
  drop_na(PC1, lfm, year.month, site, id) %>% 
  select(PC1, PC2, tti, fh, temp.max, gd, gti, pfg, ttfg, fd, prop.ignite.bins5_nodate, lfm, lfm.bin, mpa, year.month, site, id)
```

# CEME

# 1. Pick the median bin density
```{r}
ceme.bin.n <- seg.ceme.subset %>% 
  group_by(lfm.bin) %>% 
  count()

median(ceme.bin.n$n) # Median: 8
```

# 2. Resampling points from each bin
```{r}
ceme.binned.df <- seg.ceme.subset %>% 
  group_by(lfm.bin) %>% 
  sample_n(8, replace = T)
```

# 3. Segmented Regression
```{r}
out.lm<-lm(PC2 ~ lfm + site + year.month, data = ceme.binned.df)
davies.test(out.lm, ~lfm, k = 50)

segmented.mod <- segmented(out.lm, seg.Z = ~lfm) #1 breakpoint for x
summary(segmented.mod)
```

# 4. Iterating Above Process 100 Times
Holding vectors
```{r}
psi.PC1 <- c(rep(NaN, 100))
psi.PC2 <- c(rep(NaN, 100))
psi.tti <- c(rep(NaN, 100))
psi.fh <- c(rep(NaN, 100))
psi.fd <- c(rep(NaN, 100))
psi.gd <- c(rep(NaN, 100))
psi.gti <- c(rep(NaN, 100))
psi.pfg <- c(rep(NaN, 100))
psi.ttfg <- c(rep(NaN, 100))
psi.temp.max <- c(rep(NaN, 100))
psi.prop.ignite <- c(rep(NaN, 100))
ceme.thresholds <- data.frame(psi.PC1, psi.PC2, psi.tti, psi.fh, psi.temp.max, psi.gd, psi.gti, psi.pfg, psi.ttfg, psi.prop.ignite, psi.fd)
```

For loop
It is a bit clunky, but seems to do the trick. Could not figure out how to iterate dependent variables in the linear models.

Note: I did not include flame duration, glow-to-ignition or flame height below, as those variables were interrupting the for loop when they did not return a threshold

Note: I set a seed so that the code would consistently run and so that our analyses are repeatable
```{r}
for(i in 1:10){
   for(j in 1:nrow(ceme.thresholds)){
set.seed(seed = j, kind = 'default', normal.kind = 'default', sample.kind = 'default')
     
  ceme.binned.df <- seg.ceme.subset %>% 
  group_by(lfm.bin) %>% 
  sample_n(8, replace = T)
  if(i == 1){
    out.lm<-lm(PC1 ~ lfm + site + year.month, data = ceme.binned.df)
  segmented.mod <- segmented(out.lm, seg.Z = ~lfm)
  ceme.thresholds[j,i] <- summary(segmented.mod)$psi[2]
  }
  else if(i == 2) {
  out.lm<-lm(PC2 ~ lfm + site + year.month, data = ceme.binned.df)
  segmented.mod <- segmented(out.lm, seg.Z = ~lfm)
  ceme.thresholds[j,i] <- summary(segmented.mod)$psi[2]
  }
  else if(i == 3) {
    out.lm<-lm(tti ~ lfm + site + year.month, data = ceme.binned.df)
  segmented.mod <- segmented(out.lm, seg.Z = ~lfm)
  ceme.thresholds[j,i] <- summary(segmented.mod)$psi[2]
  }
  else if(i == 5) {
    out.lm<-lm(temp.max ~ lfm + site + year.month, data = ceme.binned.df)
  segmented.mod <- segmented(out.lm, seg.Z = ~lfm)
  ceme.thresholds[j,i] <- summary(segmented.mod)$psi[2]
  }
  else if(i == 6) {
    out.lm<-lm(gd ~ lfm + site + year.month, data = ceme.binned.df)
  segmented.mod <- segmented(out.lm, seg.Z = ~lfm)
  ceme.thresholds[j,i] <- summary(segmented.mod)$psi[2]
  }
  else if(i == 8) {
    out.lm<-lm(pfg ~ lfm + site + year.month, data = ceme.binned.df)
  segmented.mod <- segmented(out.lm, seg.Z = ~lfm)
  ceme.thresholds[j,i] <- summary(segmented.mod)$psi[2]
  }
  else if(i == 9) {
    out.lm<-lm(ttfg ~ lfm + site + year.month, data = ceme.binned.df)
  segmented.mod <- segmented(out.lm, seg.Z = ~lfm)
  ceme.thresholds[j,i] <- summary(segmented.mod)$psi[2]
  }
  else if(i == 10) {
    out.lm<-lm(prop.ignite.bins5_nodate ~ lfm + site + year.month, data = ceme.binned.df)
  segmented.mod <- segmented(out.lm, seg.Z = ~lfm)
  ceme.thresholds[j,i] <- summary(segmented.mod)$psi[2]
  }
   }}
```

If code is getting interrupted by this above chunk, just rerun the chunk -- usually by second time, it'll successfully run. Random seeds lead to no breakpoint being estimated and the for loop terrminating (see note below for work-around for some of the metrics that this happened most often with)

Since the above for loop kept getting interrupted when thresholds were not discovered, I set up a new for loop to figure out which seeds are causing the interruptions and manually removing those seeds. For CEME, I did this for flame height, glow to ignition and flame duration


Flame Height:
```{r}
for(j in 1:nrow(ceme.thresholds)){
    set.seed(seed = j, kind = 'default', normal.kind = 'default', sample.kind = 'default')
  
  ceme.binned.df <- seg.ceme.subset %>% 
  group_by(lfm.bin) %>% 
  sample_n(8, replace = T)

  out.lm<-lm(fh ~ lfm + site + year.month, data = ceme.binned.df)
  segmented.mod <- segmented(out.lm, seg.Z = ~lfm)
  ceme.thresholds$psi.fh[j] <- summary(segmented.mod)$psi[2]}
```
Huh.. Luckily, seeds 1:100 did not cause the for loop to be interrupted!

Glow to Ignition:
```{r}
for(j in 1:17){
  set.seed(seed = j, kind = 'default', normal.kind = 'default', sample.kind = 'default')
  
  ceme.binned.df <- seg.ceme.subset %>% 
  group_by(lfm.bin) %>% 
  sample_n(8, replace = T)
  
 out.lm<-lm(gti ~ lfm + site + year.month, data = ceme.binned.df)
  segmented.mod <- segmented(out.lm, seg.Z = ~lfm)
  ceme.thresholds$psi.gti[j] <- summary(segmented.mod)$psi[2]}

for(j in 18:48){
  set.seed(seed = j+1, kind = 'default', normal.kind = 'default', sample.kind = 'default')
  
  ceme.binned.df <- seg.ceme.subset %>% 
  group_by(lfm.bin) %>% 
  sample_n(8, replace = T)
  
 out.lm<-lm(gti ~ lfm + site + year.month, data = ceme.binned.df)
  segmented.mod <- segmented(out.lm, seg.Z = ~lfm)
  ceme.thresholds$psi.gti[j] <- summary(segmented.mod)$psi[2]}

for(j in 49:69){
  set.seed(seed = j+2, kind = 'default', normal.kind = 'default', sample.kind = 'default')
  
  ceme.binned.df <- seg.ceme.subset %>% 
  group_by(lfm.bin) %>% 
  sample_n(8, replace = T)
  
 out.lm<-lm(gti ~ lfm + site + year.month, data = ceme.binned.df)
  segmented.mod <- segmented(out.lm, seg.Z = ~lfm)
  ceme.thresholds$psi.gti[j] <- summary(segmented.mod)$psi[2]}

for(j in 70:80){
  set.seed(seed = j+4, kind = 'default', normal.kind = 'default', sample.kind = 'default')
  
  ceme.binned.df <- seg.ceme.subset %>% 
  group_by(lfm.bin) %>% 
  sample_n(8, replace = T)
  
 out.lm<-lm(gti ~ lfm + site + year.month, data = ceme.binned.df)
  segmented.mod <- segmented(out.lm, seg.Z = ~lfm)
  ceme.thresholds$psi.gti[j] <- summary(segmented.mod)$psi[2]}

for(j in 81:91){
  set.seed(seed = j+6, kind = 'default', normal.kind = 'default', sample.kind = 'default')
  
  ceme.binned.df <- seg.ceme.subset %>% 
  group_by(lfm.bin) %>% 
  sample_n(8, replace = T)
  
 out.lm<-lm(gti ~ lfm + site + year.month, data = ceme.binned.df)
  segmented.mod <- segmented(out.lm, seg.Z = ~lfm)
  ceme.thresholds$psi.gti[j] <- summary(segmented.mod)$psi[2]}

for(j in 91:100){
  set.seed(seed = j+17, kind = 'default', normal.kind = 'default', sample.kind = 'default')
  
  ceme.binned.df <- seg.ceme.subset %>% 
  group_by(lfm.bin) %>% 
  sample_n(8, replace = T)
  
 out.lm<-lm(gti ~ lfm + site + year.month, data = ceme.binned.df)
  segmented.mod <- segmented(out.lm, seg.Z = ~lfm)
  ceme.thresholds$psi.gti[j] <- summary(segmented.mod)$psi[2]}
```

Flame Duration:
```{r}
for(j in 1:68){
    set.seed(seed = j, kind = 'default', normal.kind = 'default', sample.kind = 'default')
  
  ceme.binned.df <- seg.ceme.subset %>% 
  group_by(lfm.bin) %>% 
  sample_n(8, replace = T)

  out.lm<-lm(fd ~ lfm + site + year.month, data = ceme.binned.df)
  segmented.mod <- segmented(out.lm, seg.Z = ~lfm)
  ceme.thresholds$psi.fd[j] <- summary(segmented.mod)$psi[2]}

for(j in 69:100){
    set.seed(seed = j+1, kind = 'default', normal.kind = 'default', sample.kind = 'default')
  
  ceme.binned.df <- seg.ceme.subset %>% 
  group_by(lfm.bin) %>% 
  sample_n(8, replace = T)

  out.lm<-lm(fd ~ lfm + site + year.month, data = ceme.binned.df)
  segmented.mod <- segmented(out.lm, seg.Z = ~lfm)
  ceme.thresholds$psi.fd[j] <- summary(segmented.mod)$psi[2]}
```


# 5. Mean Threshold Values
```{r}
summary(ceme.thresholds)
```

#-------------------------------

# ADFA

# 1. Pick the median bin density
```{r}
adfa.bin.n <- seg.adfa.subset %>% 
  group_by(lfm.bin) %>% 
  count()

median(adfa.bin.n$n) # Median: 18.5; round to 19
```

# 2. Resampling points from each bin
```{r}
adfa.binned.df <- seg.adfa.subset %>% 
  group_by(lfm.bin) %>% 
  sample_n(19, replace = T)
```

# 3. Segmented Regression
```{r}
out.lm<-lm(PC1 ~ lfm + site + year.month, data = adfa.binned.df)
davies.test(out.lm, ~lfm, k = 50)

segmented.mod <- segmented(out.lm, seg.Z = ~lfm) #1 breakpoint for x
summary(segmented.mod)
```

# 4. Iterating Above Process 1000 Times
Holding vectors
```{r}
psi.PC1 <- c(rep(NaN, 100))
psi.PC2 <- c(rep(NaN, 100))
psi.tti <- c(rep(NaN, 100))
psi.fh <- c(rep(NaN, 100))
psi.fd <- c(rep(NaN, 100))
psi.gd <- c(rep(NaN, 100))
psi.gti <- c(rep(NaN, 100))
psi.pfg <- c(rep(NaN, 100))
psi.ttfg <- c(rep(NaN, 100))
psi.fd <- c(rep(NaN, 100))
psi.prop.ignite <- c(rep(NaN, 100))
psi.temp.max <- c(rep(NaN, 100))
adfa.thresholds <- data.frame(psi.PC1, psi.PC2, psi.tti, psi.fh, psi.fd, psi.gd, psi.gti, psi.pfg, psi.ttfg, psi.prop.ignite, psi.temp.max)
```

For loop
It is a bit clunky, but seems to do the trick. Could not figure out how to iterate dependent variables in the linear models.

Note: I set a seed so that the code would consistently run and so that our analyses are repeatable
```{r}
for(i in 1:10){
   for(j in 1:nrow(adfa.thresholds)){
  set.seed(seed = j+150, kind = 'default', normal.kind = 'default', sample.kind = 'default')
     
  adfa.binned.df <- seg.adfa.subset %>% 
  group_by(lfm.bin) %>% 
  sample_n(19, replace = T)
  if(i == 1){
    out.lm<-lm(PC1 ~ lfm + site + year.month, data = adfa.binned.df)
  segmented.mod <- segmented(out.lm, seg.Z = ~lfm)
  adfa.thresholds[j,i] <- summary(segmented.mod)$psi[2]
  }
  else if(i == 2) {
  out.lm<-lm(PC2 ~ lfm + site + year.month, data = adfa.binned.df)
  segmented.mod <- segmented(out.lm, seg.Z = ~lfm)
  adfa.thresholds[j,i] <- summary(segmented.mod)$psi[2]
  }
  else if(i == 3) {
    out.lm<-lm(tti ~ lfm + site + year.month, data = adfa.binned.df)
  segmented.mod <- segmented(out.lm, seg.Z = ~lfm)
  adfa.thresholds[j,i] <- summary(segmented.mod)$psi[2]
  }
  else if(i == 4) {
    out.lm<-lm(fh ~ lfm + site + year.month, data = adfa.binned.df)
  segmented.mod <- segmented(out.lm, seg.Z = ~lfm)
  adfa.thresholds[j,i] <- summary(segmented.mod)$psi[2]
  }
  else if(i == 5) {
    out.lm<-lm(fd ~ lfm + site + year.month, data = adfa.binned.df)
  segmented.mod <- segmented(out.lm, seg.Z = ~lfm)
  adfa.thresholds[j,i] <- summary(segmented.mod)$psi[2]
  }
  else if(i == 8) {
    out.lm<-lm(pfg ~ lfm + site + year.month, data = adfa.binned.df)
  segmented.mod <- segmented(out.lm, seg.Z = ~lfm)
  adfa.thresholds[j,i] <- summary(segmented.mod)$psi[2]
  }
  else if(i == 9) {
    out.lm<-lm(ttfg ~ lfm + site + year.month, data = adfa.binned.df)
  segmented.mod <- segmented(out.lm, seg.Z = ~lfm)
  adfa.thresholds[j,i] <- summary(segmented.mod)$psi[2]
  }
   }}
```

Since the above for loop kept getting interrupted when thresholds were not discovered, I set up a new for loop to figure out which seeds are causing the interruptions and manually removing those seeds. For ADFA, I did this for glow duration, glow to ignition, proportion ignited, and max. temp.

Glow Duration:
```{r}
for(j in 1:3){
  set.seed(seed = j, kind = 'default', normal.kind = 'default', sample.kind = 'default')
  
  adfa.binned.df <- seg.adfa.subset %>% 
  group_by(lfm.bin) %>% 
  sample_n(19, replace = T)

  out.lm<-lm(gd ~ lfm + site + year.month, data = adfa.binned.df)
  segmented.mod <- segmented(out.lm, seg.Z = ~lfm)
  adfa.thresholds$psi.gd[j] <- summary(segmented.mod)$psi[2]}

for(j in 4:10){
  set.seed(seed = j+6, kind = 'default', normal.kind = 'default', sample.kind = 'default')
  
  adfa.binned.df <- seg.adfa.subset %>% 
  group_by(lfm.bin) %>% 
  sample_n(19, replace = T)

  out.lm<-lm(gd ~ lfm + site + year.month, data = adfa.binned.df)
  segmented.mod <- segmented(out.lm, seg.Z = ~lfm)
  adfa.thresholds$psi.gd[j] <- summary(segmented.mod)$psi[2]}

for(j in 11:29){
  set.seed(seed = j+11, kind = 'default', normal.kind = 'default', sample.kind = 'default')
  
  adfa.binned.df <- seg.adfa.subset %>% 
  group_by(lfm.bin) %>% 
  sample_n(19, replace = T)

  out.lm<-lm(gd ~ lfm + site + year.month, data = adfa.binned.df)
  segmented.mod <- segmented(out.lm, seg.Z = ~lfm)
  adfa.thresholds$psi.gd[j] <- summary(segmented.mod)$psi[2]}

for(j in 30:33){
  set.seed(seed = j+15, kind = 'default', normal.kind = 'default', sample.kind = 'default')
  
  adfa.binned.df <- seg.adfa.subset %>% 
  group_by(lfm.bin) %>% 
  sample_n(19, replace = T)

  out.lm<-lm(gd ~ lfm + site + year.month, data = adfa.binned.df)
  segmented.mod <- segmented(out.lm, seg.Z = ~lfm)
  adfa.thresholds$psi.gd[j] <- summary(segmented.mod)$psi[2]}

for(j in 34:38){
  set.seed(seed = j+21, kind = 'default', normal.kind = 'default', sample.kind = 'default')
  
  adfa.binned.df <- seg.adfa.subset %>% 
  group_by(lfm.bin) %>% 
  sample_n(19, replace = T)

  out.lm<-lm(gd ~ lfm + site + year.month, data = adfa.binned.df)
  segmented.mod <- segmented(out.lm, seg.Z = ~lfm)
  adfa.thresholds$psi.gd[j] <- summary(segmented.mod)$psi[2]}

for(j in 38:41){
  set.seed(seed = j+24, kind = 'default', normal.kind = 'default', sample.kind = 'default')
  
  adfa.binned.df <- seg.adfa.subset %>% 
  group_by(lfm.bin) %>% 
  sample_n(19, replace = T)

  out.lm<-lm(gd ~ lfm + site + year.month, data = adfa.binned.df)
  segmented.mod <- segmented(out.lm, seg.Z = ~lfm)
  adfa.thresholds$psi.gd[j] <- summary(segmented.mod)$psi[2]}

for(j in 42:64){
  set.seed(seed = j+100, kind = 'default', normal.kind = 'default', sample.kind = 'default')
  
  adfa.binned.df <- seg.adfa.subset %>% 
  group_by(lfm.bin) %>% 
  sample_n(19, replace = T)

  out.lm<-lm(gd ~ lfm + site + year.month, data = adfa.binned.df)
  segmented.mod <- segmented(out.lm, seg.Z = ~lfm)
  adfa.thresholds$psi.gd[j] <- summary(segmented.mod)$psi[2]}

for(j in 65:71){
  set.seed(seed = j+126, kind = 'default', normal.kind = 'default', sample.kind = 'default')
  
  adfa.binned.df <- seg.adfa.subset %>% 
  group_by(lfm.bin) %>% 
  sample_n(19, replace = T)

  out.lm<-lm(gd ~ lfm + site + year.month, data = adfa.binned.df)
  segmented.mod <- segmented(out.lm, seg.Z = ~lfm)
  adfa.thresholds$psi.gd[j] <- summary(segmented.mod)$psi[2]}

for(j in 72:85){
  set.seed(seed = j+200, kind = 'default', normal.kind = 'default', sample.kind = 'default')
  
  adfa.binned.df <- seg.adfa.subset %>% 
  group_by(lfm.bin) %>% 
  sample_n(19, replace = T)

  out.lm<-lm(gd ~ lfm + site + year.month, data = adfa.binned.df)
  segmented.mod <- segmented(out.lm, seg.Z = ~lfm)
  adfa.thresholds$psi.gd[j] <- summary(segmented.mod)$psi[2]}

for(j in 86:91){
  set.seed(seed = j+320, kind = 'default', normal.kind = 'default', sample.kind = 'default')
  
  adfa.binned.df <- seg.adfa.subset %>% 
  group_by(lfm.bin) %>% 
  sample_n(19, replace = T)

  out.lm<-lm(gd ~ lfm + site + year.month, data = adfa.binned.df)
  segmented.mod <- segmented(out.lm, seg.Z = ~lfm)
  adfa.thresholds$psi.gd[j] <- summary(segmented.mod)$psi[2]}

for(j in 92:96){
  set.seed(seed = j+400, kind = 'default', normal.kind = 'default', sample.kind = 'default')
  
  adfa.binned.df <- seg.adfa.subset %>% 
  group_by(lfm.bin) %>% 
  sample_n(19, replace = T)

  out.lm<-lm(gd ~ lfm + site + year.month, data = adfa.binned.df)
  segmented.mod <- segmented(out.lm, seg.Z = ~lfm)
  adfa.thresholds$psi.gd[j] <- summary(segmented.mod)$psi[2]}

for(j in 97:100){
  set.seed(seed = j+411, kind = 'default', normal.kind = 'default', sample.kind = 'default')
  
  adfa.binned.df <- seg.adfa.subset %>% 
  group_by(lfm.bin) %>% 
  sample_n(19, replace = T)

  out.lm<-lm(gd ~ lfm + site + year.month, data = adfa.binned.df)
  segmented.mod <- segmented(out.lm, seg.Z = ~lfm)
  adfa.thresholds$psi.gd[j] <- summary(segmented.mod)$psi[2]}
```
Oh boy that took a while.

Glow to Ignition:
```{r}
for(j in 1:24){
  set.seed(seed = j+4, kind = 'default', normal.kind = 'default', sample.kind = 'default')
  
  adfa.binned.df <- seg.adfa.subset %>% 
  group_by(lfm.bin) %>% 
  sample_n(19, replace = T)

  out.lm<-lm(gti ~ lfm + site + year.month, data = adfa.binned.df)
  segmented.mod <- segmented(out.lm, seg.Z = ~lfm)
  adfa.thresholds$psi.gti[j] <- summary(segmented.mod)$psi[2]}

for(j in 25:75){
  set.seed(seed = j+10, kind = 'default', normal.kind = 'default', sample.kind = 'default')
  
  adfa.binned.df <- seg.adfa.subset %>% 
  group_by(lfm.bin) %>% 
  sample_n(19, replace = T)

  out.lm<-lm(gti ~ lfm + site + year.month, data = adfa.binned.df)
  segmented.mod <- segmented(out.lm, seg.Z = ~lfm)
  adfa.thresholds$psi.gti[j] <- summary(segmented.mod)$psi[2]}

for(j in 76:100){
  set.seed(seed = j+50, kind = 'default', normal.kind = 'default', sample.kind = 'default')
  
  adfa.binned.df <- seg.adfa.subset %>% 
  group_by(lfm.bin) %>% 
  sample_n(19, replace = T)

  out.lm<-lm(gti ~ lfm + site + year.month, data = adfa.binned.df)
  segmented.mod <- segmented(out.lm, seg.Z = ~lfm)
  adfa.thresholds$psi.gti[j] <- summary(segmented.mod)$psi[2]}
```

Proportion Ignition:
No breakpoints estimated!

Maximum Temp.:
```{r}
for(j in 1:6){
  set.seed(seed = j, kind = 'default', normal.kind = 'default', sample.kind = 'default')
  
  adfa.binned.df <- seg.adfa.subset %>% 
  group_by(lfm.bin) %>% 
  sample_n(19, replace = T)

  out.lm<-lm(temp.max ~ lfm + site + year.month, data = adfa.binned.df)
  segmented.mod <- segmented(out.lm, seg.Z = ~lfm)
  adfa.thresholds$psi.temp.max[j] <- summary(segmented.mod)$psi[2]}

for(j in 7:15){
  set.seed(seed = j+9, kind = 'default', normal.kind = 'default', sample.kind = 'default')
  
  adfa.binned.df <- seg.adfa.subset %>% 
  group_by(lfm.bin) %>% 
  sample_n(19, replace = T)

  out.lm<-lm(temp.max ~ lfm + site + year.month, data = adfa.binned.df)
  segmented.mod <- segmented(out.lm, seg.Z = ~lfm)
  adfa.thresholds$psi.temp.max[j] <- summary(segmented.mod)$psi[2]}

for(j in 16:30){
  set.seed(seed = j+20, kind = 'default', normal.kind = 'default', sample.kind = 'default')
  
  adfa.binned.df <- seg.adfa.subset %>% 
  group_by(lfm.bin) %>% 
  sample_n(19, replace = T)

  out.lm<-lm(temp.max ~ lfm + site + year.month, data = adfa.binned.df)
  segmented.mod <- segmented(out.lm, seg.Z = ~lfm)
  adfa.thresholds$psi.temp.max[j] <- summary(segmented.mod)$psi[2]}

for(j in 31:44){
  set.seed(seed = j+21, kind = 'default', normal.kind = 'default', sample.kind = 'default')
  
  adfa.binned.df <- seg.adfa.subset %>% 
  group_by(lfm.bin) %>% 
  sample_n(19, replace = T)

  out.lm<-lm(temp.max ~ lfm + site + year.month, data = adfa.binned.df)
  segmented.mod <- segmented(out.lm, seg.Z = ~lfm)
  adfa.thresholds$psi.temp.max[j] <- summary(segmented.mod)$psi[2]}

for(j in 45:67){
  set.seed(seed = j+30, kind = 'default', normal.kind = 'default', sample.kind = 'default')
  
  adfa.binned.df <- seg.adfa.subset %>% 
  group_by(lfm.bin) %>% 
  sample_n(19, replace = T)

  out.lm<-lm(temp.max ~ lfm + site + year.month, data = adfa.binned.df)
  segmented.mod <- segmented(out.lm, seg.Z = ~lfm)
  adfa.thresholds$psi.temp.max[j] <- summary(segmented.mod)$psi[2]}

for(j in 68:77){
  set.seed(seed = j+50, kind = 'default', normal.kind = 'default', sample.kind = 'default')
  
  adfa.binned.df <- seg.adfa.subset %>% 
  group_by(lfm.bin) %>% 
  sample_n(19, replace = T)

  out.lm<-lm(temp.max ~ lfm + site + year.month, data = adfa.binned.df)
  segmented.mod <- segmented(out.lm, seg.Z = ~lfm)
  adfa.thresholds$psi.temp.max[j] <- summary(segmented.mod)$psi[2]}

for(j in 78:91){
  set.seed(seed = j+63, kind = 'default', normal.kind = 'default', sample.kind = 'default')
  
  adfa.binned.df <- seg.adfa.subset %>% 
  group_by(lfm.bin) %>% 
  sample_n(19, replace = T)

  out.lm<-lm(temp.max ~ lfm + site + year.month, data = adfa.binned.df)
  segmented.mod <- segmented(out.lm, seg.Z = ~lfm)
  adfa.thresholds$psi.temp.max[j] <- summary(segmented.mod)$psi[2]}

for(j in 91:100){
  set.seed(seed = j+78, kind = 'default', normal.kind = 'default', sample.kind = 'default')
  
  adfa.binned.df <- seg.adfa.subset %>% 
  group_by(lfm.bin) %>% 
  sample_n(19, replace = T)

  out.lm<-lm(temp.max ~ lfm + site + year.month, data = adfa.binned.df)
  segmented.mod <- segmented(out.lm, seg.Z = ~lfm)
  adfa.thresholds$psi.temp.max[j] <- summary(segmented.mod)$psi[2]}
```

# 5. Mean Threshold Values
```{r}
summary(adfa.thresholds)
```

# -------------------------------
# Summary Table: LFM
```{r}
dep.var <- c('PC1','PC2','TTI','Prop Ignite (5%)','GTI','FH','Max Temp.','FD','GD','PFG','TTFG', 'Summary','PC1','PC2','TTI','Prop Ignite (5%)','GTI','FH','Max Temp.','FD','GD','PFG','TTFG', 'Summary')
spp <- c(rep('ADFA', 11), 'TLP = 74.33 +/- 7.37', rep('CEME', 11), 'TLP = 88.96 +/- 2.53')
sig.p <- c('Yes (p < 0.0005)','Yes (p < 0.0005)','Yes (p < 0.0005)','Yes (p < 0.0005)','Yes (p = 0.025)','Yes (p < 0.0005)','Yes (p < 0.0005)','No','Yes (p = 0.032)','Yes (p = 0.0063)','Yes (p < 0.0005)', '10/11 Significant','Yes (p < 0.0005)','Yes (p = 0.0206)','Yes (p < 0.0005)','Yes (p < 0.0005)','Yes (p < 0.0005)','Yes (p = 0.0058)','Yes (p = 0.044)','No','No','Yes (p = 0.037)','No', '8/11 Significant') 
thresholds <- c('98.279','62.659','98.059','43.146','98.279','79.506','98.279',
                '46.555','65.827', '65.843', '78.536', 'Avg: 78.84',
                '132.532','60.301','120.141','91.879','119.911',
                '57.192','55.577','49.685','63.750','68.051','59.891',
                'Avg: 88.198')
bootstrap.thresholds <- c('101.00', '80.38', '87.86', 'NA', '105.27', '87.32', '99.57',
                          '88.62', '80.93', '64.91', '70.25','Avg: 86.61','111.10', '112.50',
                          '109.50', '92.28', '133.35', '78.71', '66.29', '135.80', '153.2',
                          '136.86','85.82','Avg: 110.49')
seg.reg.table.df <- data.frame(dep.var, spp, sig.p, thresholds, bootstrap.thresholds)

seg.reg.table.df %>%
  mutate(dep.var = cell_spec(dep.var, 'html', color = 'black', bold = T), spp = cell_spec(spp, 'html', color = 'black'), sig.p = cell_spec(sig.p, 'html', color = 'black'), thresholds = cell_spec(thresholds, 'html', color = 'black'), bootstrap.thresholds = cell_spec(bootstrap.thresholds, 'html', color = 'black')) %>% 
  kable(format = 'html', escape = F, col.names = c('Dependent Variable', 'Species', 'Significant?','Threshold', 'Bootstrapped Threshold'), caption = 'Live Fuel Moisture') %>% 
  kable_styling(bootstrap_options = c('hover', 'bordered', 'condensed'), fixed_thead = T) %>% 
  row_spec(c(1:7, 9:11, 13:19, 22), bold = T) %>% 
  row_spec(c(8, 20:21, 23), bold = F) %>% 
  row_spec(c(12, 24), background = "#D3D3D3", bold = T) %>%
  save_kable(here('figures', 'supp-figures', 'TabS4_LFM.segreg.summary.html'))
```

Mean, St. Dev., Median
```{r}
adfa.sig.lfm.psi <- c(98.279,62.659,98.059,43.146,98.279,79.506,98.279,65.827,65.843,78.536)
mean(adfa.sig.lfm.psi)
sd(adfa.sig.lfm.psi)
median(adfa.sig.lfm.psi)
ceme.sig.lfm.psi <- c(132.532,60.301,120.141,91.879,119.911,57.192,55.577,68.051)
mean(ceme.sig.lfm.psi)
sd(ceme.sig.lfm.psi)
median(ceme.sig.lfm.psi)
both.sig.lfm.psi <- c(98.279,62.659,98.059,43.146,98.279,79.506,98.279,65.827,65.843,78.536,132.532,60.301,120.141,91.879,119.911,57.192,55.577,68.051)
mean(both.sig.lfm.psi)
sd(both.sig.lfm.psi)
median(both.sig.lfm.psi)
```


# Summary Table: Water Potential
```{r}
dep.var <- c('PC1','PC2','TTI','Prop Ignite (5%)','GTI','FH','Max Temp.','FD','GD','PFG','TTFG', 'Summary','PC1','PC2','TTI','Prop Ignite (5%)','GTI','FH','Max Temp.','FD','GD','PFG','TTFG', 'Summary')
spp <- c(rep('ADFA', 11), ' ', rep('CEME', 11), ' ')
sig.p <- c('Yes (p < 0.0005)','Yes (p < 0.0005)','Yes (p < 0.0005)','No','Yes (p = 0.0042)','No','No','Yes (p = 0.0086)','No','Yes (p = 0.0015)','No', '6/11 Significant', 'No', 'No','No','No','No','Yes (p = 0.0072)','No','No','No','No','Yes (p < 0.0005)', '2/11 Significant')
thresholds <- c('-0.4','-0.4','-0.4','-0.4','-0.504','-9.88','NA','-0.4','-0.331','-3.45','-9.96', 'Avg:  -0.925','-9.99','-9.99','-10','-10','-10','-8.05','-10','-9.99','-10','-10','-0.76', 'Avg: -4.41')
seg.reg.table2.df <- data.frame(dep.var, spp, sig.p, thresholds)

seg.reg.table2.df %>% 
   mutate(dep.var = cell_spec(dep.var, 'html', color = 'black', bold = T), spp = cell_spec(spp, 'html', color = 'black'), sig.p = cell_spec(sig.p, 'html', color = 'black'), thresholds = cell_spec(thresholds, 'html', color = 'black')) %>% 
  kable(format = 'html', escape = F, col.names = c('Dependent Variable', 'Species', 'Significant?','Threshold'), caption = 'Water Potential') %>% 
  kable_styling(bootstrap_options = c('hover', 'bordered', 'condensed'), fixed_thead = T) %>% 
  row_spec(c(1:3, 5, 8, 10, 18, 23), bold = T) %>% 
  row_spec(c(4, 6:7, 9, 11, 13:17, 19:22), bold = F) %>% 
  row_spec(c(12, 24), background = "#D3D3D3", bold = T) %>%
  save_kable(here('figures', 'supp-figures', 'TabS5_mpa.segreg.summary.html'))
```
