==> scripts/6.0_LDMC_and_RWC.Rmd <==
---
title: "RWC and LDMC"
author: "Indra Boving & Joe Celebrezze"
date: "4/20/2021"
output: html_document
---

This script contains data and analysis for calculating RWC and LDMC from rehydrated leaves, and determining if LDMC and/or RWC predict 1) TLP, 2) Flammability in the manner similar to or different from water potential or LFM. 

# Setup
```{r setup, include=FALSE}
knitr::opts_chunk$set(echo = TRUE)
# install the package first if you haven't already, then load it
library(tidyverse)
library(here)
library(ggplot2)
library(MetBrewer)
library(lme4)
library(sjPlot)
library(AICcmodavg)
library(remef)
library(ggpubr)
library(janitor)
library(lubridate)                             # Install & load scales
library("scales")
library(kableExtra)
```

Set global options for color palettes and themes: 

```{r}
th <- theme(panel.background = element_rect(fill='white', colour='black'), # Make background white and border black
          panel.grid.major = element_blank(),  # Hide major gridlines
          panel.grid.minor = element_blank())

th.spp.wrap <- theme(legend.position = "bottom", 
        strip.text = element_text(face = "italic", size = 12) +
  scale_color_manual(values=met.brewer("Morgenstern", 7)))

Morgenstern = list(c("#7c668c", "#b08ba5", "#dfbbc8", "#ffc680", "#ffb178", "#db8872", "#a56457"), c(7, 5, 4, 6, 3, 2, 1), colorblind=TRUE)
```


Plotting options
```{r}
color_panels <- scale_color_manual(values=c("#9D0208",'#669bbc')) 

th_panels <- theme(
        axis.title.x=element_blank(),
        axis.title.y=element_text(face = 'bold', size = 18),
        legend.position = "none", 
        aspect.ratio = 1, 
        plot.margin = unit(c(0,0,0,0), "cm")
        )
```
#------------------------------------
# 1. Data wrangling

Reading in data and calculating Saturated Water Content (SWC), Relative Water Content (RWC), and Leaf Dry Matter Content (LDMC)

To start, try calculated SWC based on the samples with the highest LFM or the most hydrated mpa.scaled(whichever is higher):

```{r}
rwc_raw <- read.csv(here('processed-data', 'mem.data.subset.alldates.csv')) #%>% 
  # select(individual, spp, mpa, lfm.unscaled, dw.flam.sample.unscaled, ww.flam.sample.unscaled, flam.sample.wt.unscaled, PC1, tti, fd, gd) 

#Find the maximum hydrated samples based on LFM and calculate LDMC, SWC:

rwc_max_lfm_df <- rwc_raw %>% 
  group_by(individual) %>% 
  slice_max(lfm.unscaled, with_ties = T) %>% 
  ungroup() %>% 
  mutate(ldmc = (dw.flam.sample.unscaled/flam.sample.wt.unscaled), 
         swc = (ww.flam.sample.unscaled/dw.flam.sample.unscaled), 
         value = lfm.unscaled) %>% 
  #filter(mpa.scaled> -.5) %>% 
    select(individual, ldmc, swc, value) 

#Same but with MPa: 
rwc_max_mpa_df <- rwc_raw %>% 
  group_by(individual) %>% 
  slice_max(mpa, with_ties = T) %>% 
  ungroup() %>% 
  mutate(ldmc = (dw.flam.sample.unscaled/flam.sample.wt.unscaled), 
         swc = (ww.flam.sample.unscaled/dw.flam.sample.unscaled), 
         value = mpa) %>% 
  #filter(mpa.scaled> -.5) %>% 
    select(individual, ldmc, swc, value) 

#Check out those datasets: 
swc_max_highest <- bind_rows(rwc_max_lfm_df, rwc_max_mpa_df) %>% 
  group_by(individual) %>% 
  slice_max(swc, with_ties = F) %>% 
  select(individual, swc)

ldmc_min_lowest <- bind_rows(rwc_max_lfm_df, rwc_max_mpa_df) %>% #lower LDMC = less water
  group_by(individual) %>% 
  slice_min(ldmc, with_ties = F) %>% 
  select(individual, ldmc)

#Combine into one df with individual, the most hydrated SWC and the most hydrated LDMC: 
rwc_max_highest <- merge(swc_max_highest, ldmc_min_lowest, by = c("individual"))

rwc_df <- merge(rwc_raw, rwc_max_highest, by = c("individual"), all = T) %>% 
  mutate(water.content = (ww.flam.sample.unscaled/dw.flam.sample.unscaled), 
         rwc.unscaled = (water.content/swc),
         ldmc.unscaled = ldmc) 
```

The highest SWC is always the most hydrated LFM, which makes more sense to use as using mpa.scalednearly always gives us RWC > 100%. 

This works, but is not exactly how a PV curve would calculate SWC. Try with the intercept of the relationship between water.content and mpa:

```{r}
rwc_df %>% 
  ggplot(aes(y = water.content, x = mpa.unscaled, color = individual)) +
  geom_point() +
  theme(legend.position = "none") +
  geom_smooth(method = "lm", se = F)

rwc_int_subset <- rwc_df %>% 
  filter(mpa.unscaled > -2)

swc_mod <- lmer(water.content ~ mpa.unscaled + (1|individual), data = rwc_int_subset)

ind_ints <- as.data.frame(coef(swc_mod)$individual) %>% 
  rownames_to_column(var = "individual") %>% 
  rename(slope = mpa.unscaled, 
         intercept = '(Intercept)') %>% 
  select(individual, 
          intercept, slope)

ind_ints

#combine with previous attempt to get a swc:

rwc_swc_df <- merge(rwc_df, ind_ints, by = c("individual")) %>% 
 # select(individual, mpa, swc, intercept, water.content, rwc.unscaled, ldmc) %>% 
  mutate(rwc.int.unscaled = (water.content/intercept))

rwc_swc_df %>% 
  ggplot(aes(y = rwc.unscaled, x = rwc.int.unscaled))+
  geom_point()
```


When we extrapolate, we get a lot of RWCs that are greater than 100...which is a bit strange, conceptually. Force those to be 100. 

```{r}
rwc_swc_df2 <- rwc_swc_df %>% 
  mutate(rwc.int.unscaled = case_when(
    #rwc.int.unscaled > 1.001 ~ NA_real_, 
    rwc.int.unscaled > 1.001 ~ 1, 
    TRUE ~ rwc.int.unscaled
  )) 

rwc_swc_df2 %>% 
  ggplot(aes(y = rwc.unscaled, x = rwc.int.unscaled))+
  geom_point()
```
This is better! Let's go with this version for SWC, as it's the most similar to PV curves. For LDMC, we'll stick with the previous way (dry matter content of the most hydrated leaf in terms of LFM), as LDMF does not have as many other measurements hinging on it. (Could be convinced otherwise here though - could subset rows where RWC = 100 and write ldmc = (dw.flam.sample/(dw.flam.sample + sat.water.content)))

Combine rwc_df with TLP:

```{r}
rwc_pv_raw <- read.csv(here("processed-data", "pv_summary_df_timing.csv")) %>% 
  filter(timing == "fall") %>% 
  select(-timing)

rwc_pv_df <- merge(rwc_pv_raw, rwc_swc_df2, by = c("spp")) %>% 
  mutate(swc.int = intercept) %>% 
  group_by(spp) %>% 
  mutate(rwc.noint = scale(rwc.unscaled), 
         rwc = scale(rwc.int.unscaled),
         ldmc = scale(ldmc)) %>% 
  mutate(season = case_when(
    precip.2mo < 0.5 ~ "Dry", 
    precip.2mo > 0.5 ~ "Wet")
  ) %>% 
  mutate(site.test = case_when( 
    site == "St. Marys" ~ 1,
    site == "Painted Caves" ~ 2
  )) %>%
  mutate(season = case_when(
    timing %in% c("Dec. 2019", "Dec. 2016", "Jan. 2018", "Jan. 2020") ~ "Wet", 
    timing == "Sept. 2020" ~ "Dry"
  ))
```

# ---------------------------------------

#Comment 1: RWC?

"One suggestion I have for the authors is for them to also included discussion (and likely data) of relative water content. Relative water content and live fuel moisture should be related
(considering that both rely on fresh and dry weight, though RWC also requires saturated water
content) and the authors should have all of the raw data to be able to also determine RWC. 
This would be a further link between fire scientists and eco-physiologists. I see they report the
pressure-volume curve TLP, but I believe they could also determine a RWC for each LFM data
point they have. If so, this might best fit in the Discussion section “4.2 | Both plant hydration
and dry matter content affect flammability.”

      - Reviewer 3
      

# MAIN MODEL SELECTION - RWC and LDMC

#2.0 Compare to prev. identified top models

From previous analysis: 

"From our first round of model selection, it appears that site, species, and year.month are all important variables to include, and sample.wt should probably be included as well despite worsening the models slightly.

Below, I am going to perform another round of model selection with all possible permutations of the following metrics: mpa, lfm, dw.flam.sample, ww.flam.sample, excess_water.scaledwith the caveat that some combinations of metrics absolutely do NOT work. Those combinations are as follows:
  - dw.flam.sample, ww.flam.sample and sample.wt
  - dw.flam.sample, lfm, ww.flam.sample
  - lfm, dw.flam.sample
  - lfm, ww.flam.sample
  - lfm, excess_water.scaled
  - ww.flam.sample, excess_water.scaled
  - dw.flam.sample, excess_water.scaled (with sample.wt)
  - dw.flam.sample, ww.flam.sample (with sample.wt)
  
These do not work, as live fuel moisture was used in the calculations of these metrics, therefore they return very high VIF values that signal to us that they cannot be used in the mixed effects models together."

Additionally, in this second analysis, the following combinations will not work: 

  - rwc and ww.flam.sample
  - rwc and dw.flam.sample
  - rwc and excess.water.scaled

##PC1

```{r}
m01 <- lmer(PC1 ~ mpa.scaled + spp + site + year.month + sample.wt + (1 | individual), REML = F, data = rwc_pv_df) #mpa.scaled

AIC(m01)
AICc(m01)

m02 <- lmer(PC1 ~ mpa.scaled + dw.flam.sample + spp + site + year.month + sample.wt + (1 | individual), REML = F, data = rwc_pv_df) #composite

m03 <- lmer(PC1 ~ ww.flam.sample + spp + site + year.month + sample.wt + (1 | individual), REML = F, data = rwc_pv_df) #WW

m04 <- lmer(PC1 ~ dw.flam.sample + spp + site + year.month + sample.wt + (1 | individual), REML = F, data = rwc_pv_df) #DW

m05 <- lmer(PC1 ~ lfm + spp + site + year.month + sample.wt + (1 | individual), REML = F, data = rwc_pv_df) #LFM
AICc(m05)

m06 <- lmer(PC1 ~ rwc + spp + site + year.month + sample.wt + (1 | individual), REML = F, data = rwc_pv_df)

AICc(m06)
AIC(m06)

m07 <- lmer(PC1 ~ rwc + spp + site + year.month + sample.wt +  mpa.scaled + (1 | individual), REML = F, data = rwc_pv_df)

m08 <- lmer(PC1 ~ mpa.scaled + ldmc + spp + site + year.month + sample.wt + (1 | individual), REML = F, data = rwc_pv_df)

AICc(m08)
AIC(m08)

performance::multicollinearity(m08)

m09 <- lmer(PC1 ~ ldmc + spp + site + year.month + sample.wt + (1 | individual), REML = F, data = rwc_pv_df)

m010 <- lmer(PC1 ~ ldmc + rwc + spp + site + year.month + sample.wt + (1 | individual), REML = F, data = rwc_pv_df)

model_list <- c(m01, m02, m03, m04, m05, m06, m07, m08, m09, m010)

performance::multicollinearity(m06)

aictab(model_list)
m06

tab_model(m08, m01, m06, m07, m02, m05,
          show.reflvl = FALSE, 
          digits = 3, 
          show.aic = TRUE, 
          show.ci = FALSE,
          show.icc = FALSE, 
          pred.labels = c(),
         # dv.labels = c("Model 5", "Model 1", "Model 3", "Model 2", "Model 4", "Model 6", "Model 7"),
         string.pred = "Coeffcient", 
         title = "models with no collinearity",
  string.p = "P-Value", 
  p.style = "stars",
  file = here('figures','MEM.figures','mem2.PC1.norml.html')) # saving .html into figures folder
```

```{r}
m08_pc1 <- lmer(PC1 ~ mpa.scaled + ldmc + spp + site + year.month + sample.wt + (1 | individual), REML = T, data = rwc_pv_df)
performance::multicollinearity(m08_pc1)

m01_pc1 <- lmer(PC1 ~ mpa.scaled + spp + site + year.month + sample.wt + (1 | individual), REML = T, data = rwc_pv_df) #mpa.scaled
performance::multicollinearity(m01_pc1)

m06_pc1 <- lmer(PC1 ~ rwc + spp + site + year.month + sample.wt + (1 | individual), REML = T, data = rwc_pv_df)
performance::multicollinearity(m06_pc1)


m05_pc1 <- lmer(PC1 ~ lfm + spp + site + year.month + sample.wt + (1 | individual), REML = T, data = rwc_pv_df) #LFM
performance::multicollinearity(m05_pc1)

```

Table
```{r}
tab_model(m08_pc1, m01_pc1, m06_pc1, m05_pc1,
          show.reflvl = FALSE, 
          digits = 3, 
          show.aic = TRUE, 
          show.ci = FALSE,
          show.icc = FALSE, 
          pred.labels = c(),
          dv.labels = c("Model 5", "Model 1", "Model 3", "Model 2", "Model 4"),
         string.pred = "Coeffcient", 
         title = "PC1",
  string.p = "P-Value", 
  p.style = "stars",
  file = here('figures','MEM.figures','mem2.PC1.html')) # saving .html into figures folder
```

##TTI

```{r}
m01 <- lmer(tti ~ mpa.scaled + spp + site + year.month + sample.wt + (1 | individual), REML = F, data = rwc_pv_df) #Mpa

m02 <- lmer(tti ~ mpa.scaled + dw.flam.sample + spp + site + year.month + sample.wt + (1 | individual), REML = F, data = rwc_pv_df) #composite

m03 <- lmer(tti ~ ww.flam.sample + spp + site + year.month + sample.wt + (1 | individual), REML = F, data = rwc_pv_df) #WW

m04 <- lmer(tti ~ dw.flam.sample + spp + site + year.month + sample.wt + (1 | individual), REML = F, data = rwc_pv_df) #DW

m05 <- lmer(tti ~ lfm + spp + site + year.month + sample.wt + (1 | individual), REML = F, data = rwc_pv_df) #LFM

m06 <- lmer(tti ~ rwc + spp + site + year.month + sample.wt + (1 | individual), REML = F, data = rwc_pv_df)

m08 <- lmer(tti ~ mpa.scaled + ldmc + spp + site + year.month + sample.wt + (1 | individual), REML = F, data = rwc_pv_df)

m09 <- lmer(tti ~ ldmc + spp + site + year.month + sample.wt + (1 | individual), REML = F, data = rwc_pv_df)

m010 <- lmer(tti ~ ldmc + rwc + spp + site + year.month + sample.wt + (1 | individual), REML = F, data = rwc_pv_df)

m011 <- lmer(tti ~ ldmc + lfm + spp + site + year.month + sample.wt + (1 | individual), REML = F, data = rwc_pv_df)

summary(m011)

model_list <- c(m01, m02, m03, m04, m05, m06, m08, m09, m010, m011)

aictab(model_list)
```
```{r}
tab_model(m08, m01, m06, m05,
          show.reflvl = FALSE, 
          digits = 3, 
          show.aic = TRUE, 
          show.ci = FALSE,
          show.icc = FALSE, 
          pred.labels = c(),
         # dv.labels = c("Model 5", "Model 1", "Model 3", "Model 2", "Model 4", "Model 6", "Model 7"),
         string.pred = "Coeffcient", 
         title = "models with no collinearity",
  string.p = "P-Value", 
  p.style = "stars",
  file = here('figures','MEM.figures','mem2.TTI.noreml.html')) # saving .html into figures folder
```


```{r}
m08_tti <- lmer(tti ~ mpa.scaled + ldmc + spp + site + year.month + sample.wt + (1 | individual), REML = T, data = rwc_pv_df)

m01_tti <- lmer(tti ~ mpa.scaled + spp + site + year.month + sample.wt + (1 | individual), REML = T, data = rwc_pv_df) #Mpa

m06_tti <- lmer(tti ~ rwc + spp + site + year.month + sample.wt + (1 | individual), REML = T, data = rwc_pv_df)

m05_tti <- lmer(tti ~ lfm + spp + site + year.month + sample.wt + (1 | individual), REML = T, data = rwc_pv_df) #LFM
```

Table
```{r}
tab_model(m08_tti, m01_tti, m06_tti, m05_tti,
          show.reflvl = FALSE, 
          digits = 3, 
          show.aic = TRUE, 
          show.ci = FALSE,
          show.icc = FALSE, 
          pred.labels = c(),
         # dv.labels = c("Model 5", "Model 1", "Model 3", "Model 2", "Model 4", "Model 6", "Model 7"),
         string.pred = "Coeffcient", 
         title = "Time to Ignition",
  string.p = "P-Value", 
  p.style = "stars",
  file = here('figures','MEM.figures','mem2.TTI.html')) # saving .html into figures folder
```

##FD

```{r}
m01 <- lmer(fd ~ mpa.scaled + spp + site + year.month + sample.wt + (1 | individual), REML = F, data = rwc_pv_df) #mpa.scaled

m02 <- lmer(fd ~ mpa.scaled + dw.flam.sample + spp + site + year.month + sample.wt + (1 | individual), REML = F, data = rwc_pv_df) #composite, multicollinear?

m03 <- lmer(fd ~ ww.flam.sample + spp + site + year.month + sample.wt + (1 | individual), REML = F, data = rwc_pv_df) #WW

m04 <- lmer(fd ~ dw.flam.sample + spp + site + year.month + sample.wt + (1 | individual), REML = F, data = rwc_pv_df) #DW

m05 <- lmer(fd ~ lfm + spp + site + year.month + sample.wt + (1 | individual), REML = F, data = rwc_pv_df) #LFM

m06 <- lmer(fd ~ rwc + spp + site + year.month + sample.wt + (1 | individual), REML = F, data = rwc_pv_df)

m07 <- lmer(fd ~ mpa.scaled + rwc + spp + site + year.month + sample.wt + (1 | individual), REML = F, data = rwc_pv_df)

m08 <- lmer(fd ~ mpa.scaled + ldmc + spp + site + year.month + sample.wt + (1 | individual), REML = F, data = rwc_pv_df)

m09 <- lmer(fd ~ ldmc + spp + site + year.month + sample.wt + (1 | individual), REML = F, data = rwc_pv_df)

m010 <- lmer(fd ~ ldmc + rwc + spp + site + year.month + sample.wt + (1 | individual), REML = F, data = rwc_pv_df)

m011 <- lmer(fd ~ ldmc + lfm + spp + site + year.month + sample.wt + (1 | individual), REML = F, data = rwc_pv_df)

summary(m011)

model_list <- c(m01, m02, m03, m04, m05, m06, m07, m08, m09, m010, m011)

aictab(model_list)
```

```{r}
tab_model(#m01, m02, m08, m04, m07, m03, m05, m06, m011, m010,
          m08, m01, m06, m05, m04,m02, 
          show.reflvl = FALSE, 
          digits = 3, 
          show.aic = TRUE, 
          show.ci = FALSE,
          show.icc = FALSE, 
          pred.labels = c(),
         # dv.labels = c("Model 5", "Model 1", "Model 3", "Model 2", "Model 4", "Model 6", "Model 7"),
         string.pred = "Coeffcient", 
         title = "models with no collinearity",
  string.p = "P-Value", 
  p.style = "stars",
  file = here('figures','MEM.figures','mem2.FD.remlF.html'))
```

```{r}
m01_fd <- lmer(fd ~ mpa.scaled + spp + site + year.month + sample.wt + (1 | individual), REML = T, data = rwc_pv_df) #mpa.scaled

m02_fd <- lmer(fd ~ mpa.scaled + dw.flam.sample + spp + site + year.month + sample.wt + (1 | individual), REML = T, data = rwc_pv_df) #composite

m08_fd <- lmer(fd ~ mpa.scaled + ldmc + spp + site + year.month + sample.wt + (1 | individual), REML = T, data = rwc_pv_df)

m04_fd <- lmer(fd ~ dw.flam.sample + spp + site + year.month + sample.wt + (1 | individual), REML = T, data = rwc_pv_df) #DW

m05_fd <- lmer(fd ~ lfm + spp + site + year.month + sample.wt + (1 | individual), REML = T, data = rwc_pv_df) #LFM

m06_fd <- lmer(fd ~ rwc + spp + site + year.month + sample.wt + (1 | individual), REML = T, data = rwc_pv_df) #LFM
```

Table
```{r}
tab_model(m01_fd, m02_fd, m08_fd, m04_fd, m05_fd, m06_fd,
          show.reflvl = FALSE, 
          digits = 3, 
          show.aic = TRUE, 
          show.ci = FALSE,
          show.icc = FALSE, 
          pred.labels = c(),
         # dv.labels = c("Model 5", "Model 1", "Model 3", "Model 2", "Model 4", "Model 6", "Model 7"),
         string.pred = "Coeffcient", 
         title = "models with no collinearity",
  string.p = "P-Value", 
  p.style = "stars",
  file = here('figures','MEM.figures','mem2.FD.html')) # saving .html into figures folder
```
##GD

```{r}
m01 <- lmer(gd ~ mpa.scaled + spp + site + year.month + sample.wt + (1 | individual), REML = F, data = rwc_pv_df) #Mpa

m02 <- lmer(gd ~ mpa.scaled + dw.flam.sample + spp + site + year.month + sample.wt + (1 | individual), REML = F, data = rwc_pv_df) #composite

m03 <- lmer(gd ~ ww.flam.sample + spp + site + year.month + sample.wt + (1 | individual), REML = F, data = rwc_pv_df) #WW

m04 <- lmer(gd ~ dw.flam.sample + spp + site + year.month + sample.wt + (1 | individual), REML = F, data = rwc_pv_df) #DW

m05 <- lmer(gd ~ lfm + spp + site + year.month + sample.wt + (1 | individual), REML = F, data = rwc_pv_df) #LFM

m06 <- lmer(gd ~ rwc + spp + site + year.month + sample.wt + (1 | individual), REML = F, data = rwc_pv_df)

m07 <- lmer(gd ~ mpa.scaled + rwc + spp + site + year.month + sample.wt + (1 | individual), REML = F, data = rwc_pv_df)

m08 <- lmer(gd ~ mpa.scaled + ldmc + spp + site + year.month + sample.wt + (1 | individual), REML = F, data = rwc_pv_df)

m010 <- lmer(gd ~ ldmc + rwc + spp + site + year.month + sample.wt + (1 | individual), REML = F, data = rwc_pv_df)

m011 <- lmer(gd ~ ldmc + lfm + spp + site + year.month + sample.wt + (1 | individual), REML = F, data = rwc_pv_df)

summary(m011)

model_list <- c(m01, m02, m03, m04, m05, m06, m07, m08, m010)

aictab(model_list)
```

Table
```{r}
tab_model(m08, m04, m010, m03, m01, m05, m06,
          show.reflvl = FALSE, 
          digits = 3, 
          show.aic = TRUE, 
          show.ci = FALSE,
          show.icc = FALSE, 
          pred.labels = c(),
         # dv.labels = c("Model 5", "Model 1", "Model 3", "Model 2", "Model 4", "Model 6", "Model 7"),
         string.pred = "Coeffcient", 
         title = "Flame Duration",
  string.p = "P-Value", 
  p.style = "stars",
  file = here('figures','MEM.figures','mem2.GD.noreml.html')) # saving .html into figures folder
```

```{r}
m01_gd <- lmer(gd ~ mpa.scaled + spp + site + year.month + sample.wt + (1 | individual), REML = T, data = rwc_pv_df) #Mpa

m05_gd  <- lmer(gd ~ lfm + spp + site + year.month + sample.wt + (1 | individual), REML = T, data = rwc_pv_df) #LFM

m011_gd  <- lmer(gd ~ ldmc + rwc + spp + site + year.month + sample.wt + (1 | individual), REML = T, data = rwc_pv_df)

m08_gd  <- lmer(gd ~ mpa.scaled  + ldmc + spp + site + year.month + sample.wt + (1 | individual), REML = T, data = rwc_pv_df)

m010_gd  <- lmer(gd ~ ldmc + rwc + spp + site + year.month + sample.wt + (1 | individual), REML = T, data = rwc_pv_df)

m06_gd <- lmer(gd ~ rwc + spp + site + year.month + sample.wt + (1 | individual), REML = F, data = rwc_pv_df)

```

```{r}
tab_model(m010_gd , m011_gd , m08_gd , m05_gd , m01_gd , m06_gd,
          show.reflvl = FALSE, 
          digits = 3, 
          show.aic = TRUE, 
          show.ci = FALSE,
          show.icc = FALSE, 
          pred.labels = c(),
         # dv.labels = c("Model 5", "Model 1", "Model 3", "Model 2", "Model 4", "Model 6", "Model 7"),
         string.pred = "Coeffcient", 
         title = "Glow Duration",
  string.p = "P-Value", 
  p.style = "stars",
  file = here('figures','MEM.figures','mem2.GD.html')) # saving .html into figures folder
```

#SI Table: 
```{r}
tab_model(m08_pc1, m01_pc1, m06_pc1, m05_pc1,
          m08_tti, m01_tti, m06_tti, m05_tti,
          m08_fd, m01_fd, m06_fd, m05_fd, m04_fd, m02_fd, 
          m08_gd, m01_gd, m06_gd, m05_gd, m010_gd,
          show.reflvl = FALSE, 
          digits = 3, 
          show.aic = FALSE, 
         # show.aicc = TRUE,
          show.ci = FALSE,
          show.icc = FALSE, 
          pred.labels = c(),
         # dv.labels = c("Model 5", "Model 1", "Model 3", "Model 2", "Model 4", "Model 6", "Model 7"),
         string.pred = "Coeffcient", 
         title = "All Top Models",
  string.p = "P-Value", 
  p.style = "stars",
  file = here('figures','supp-figures','FEBovingST10.html'))
```


# MAIN TABLE

All data below is from above MEM tables:

```{r}
dep.var <- c(
  "PC1",
  " ",
  " ",
  " ",
  " ",
  "Time to Ignition (s)",
  " ",
  " ",
  " ",
  " ",
 # " ",
  "Flame Duration (s)",
 " ",
  " ",
  " ",
  " ",
 " ",
  " ",
   "Glow Duration (s)",
  " ",
  " ",
  " ",
  " ",
  " "
)
fix.fx <- c(
  " ", #PC1---------------
  "Water Potential + LDMC",
  "Water Potential",
  "RWC",
  "LFM",
   #DONE
  " ", #TTI ---------------
  "Water Potential + LDMC",
  "Water Potential",
  "RWC",
  "LFM",
 # "RWC + LDMC", #DONE
  " ", #FD---------------
  "Water Potential",
  "Water Potential + Dry Weight",
  "Water Potential + LDMC",
  "Dry Weight",
  "LFM",
  "RWC",
  " ", #GD---------------
  # "Water Potential + LDMC",
  # "Dry Weight",
  # "RWC + LDMC",
  # "Wet Weight",
  # 'Water Potential',
  # "LFM",
  # "RWC",
 # "RWC",
  "Water Potential + LDMC",
  "RWC + LDMC",
  'Water Potential',
  "LFM",
  "RWC"

 
)

est1 <- c(
  " ",#---------------
  '-0.459***',  #MPa
  '-0.456***', #Mpa
  '-0.404***', #RWC
  '-0.334***', #MP
  " ", #TTI---------------
  '4.940***',
  '4.927***',
  "4.220***",
  '3.259***', #DONE
  " ",#---------------
  '-0.677**',
  "-0.494",
  '-0.674***',
  '',
  '-0.658**', #DONE
  '-0.574 **',
  " ",#---------------
  '-2.020',
  '-1.866', #here
  '-2.063',
  '-1.950',
  '-1.846'
)

est2 <- c(
  " ", #---------------
  '4.721',  #ldmc
  ' ', #Mpa
  ' ', #MPa
  ' ', #RWC
 " ", #tti #---------------
  '2.988**', #ldmc
  ' ',
  ' ',
 " ",
#  '2.439*', #DONE #ldmc
  " ", #---------------
  ' ',
  "0.331",
  '0.272',
  '0.762**',
  ' ',
  ' ',
 # '0.272', #DONE
   " ", #---------------
  '4.721', #ldmc, and all below
  '4.942',
   ' ',
    ' ',
  ' ' #dw
)
# aic.values <- c(
#   " ", #---------------
#   '1530.660',
#   '1531.568', #.908
#  # '1535.015', #4.335
#  # '1538.457', #7.797
#   " ",#---------------
#   '3023.799',
#  # '3030.336', #6.557
#  # '3031.106', #7.327
#  # '3055.093', #DONE 31.314
#   " ",#---------------
#   '2323.369',
#   '2324.773', #.264
#   '2325.127', #.488
#  # '2325.764', #.637
# #  '2327.018', #DONE 1.891
#   " ",#---------------
#   '3778.262', 
#   '3778.930', #.668
#   '3780.012' #1.75
#  # '3782.924', #2.445
#  # '3784.533' #3.385
# )
r2.values <- c(
  " ",#---------------
  '0.259 / 0.305',
  '0.243 / 0.305',
  '0.227 / 0.282',
  '0.204 / 0.276', #DONE
  " ",#---------------
  '0.382 / 0.412',
  '0.362 / 0.416',
  '0.320 / 0.359',
  '0.263 / 0.327', #DONE
  #'0.263 / 0.327',
  " ",#---------------
  '0.076 / 0.131',
  "0.078 / 0.131",
  '0.078 / 0.137', #DONE
  '0.072 / 0.121',
  '0.071 / 0.121',
  '0.068 / 0.122',
   " ",#---------------
  '0.253 / 0.312',
  '0.252 / 0.311', 
  '0.248 / 0.293',
  '0.242 / 0.303',
  '0.248 / 0.293'
)
# d.aic <- c(
#   " ",#---------------
#   '0',
#   '0.908',
#   '10.216',
#   '21.879',
#  # '7.797', #DONE
#   " ",#---------------
#   '0',
#  "6.537",
#    '7.307',
#    '68.577',
# #  '31.314',
#   " ",#---------------
#   '0',
#   "1.369",
#   '1.404',
#   '1.758',
#   '2.395',
#   '3.507',
#   " ",#---------------
#   '0',
#   '0.668',
#   '0.175',
#   '4.662',
#   '5.279',
#   '6.271'
# )

d.aic <- c( #With REML = T
  " ",#---------------
  '0',
  '0.626',
  '9.937',
  '21.605',
 # '7.797', #DONE
  " ",#---------------
  '0',
   "6.973",
   '36.120',
   '70.144',
  " ",#---------------
  '0', #m1 MPa
  "1.369", #m2 MPA + DW
  '1.404', #m8 MPA + LDMC
  '1.758', #m4  DW
  '2.395', #m5 LFM
  '3.507', #m6 RWC
  " ",#---------------
  '0',
  '0.844',
  '4.694',
  '5.317',
  '5.384'
)


deltaAIC <- c('\U0394 AIC')

mem.table.df <- data.frame(dep.var, fix.fx, est1, est2, d.aic, #aic.values,
                           r2.values)

mem.table.df %>%
  mutate(
    dep.var = cell_spec(dep.var, 'html', color = 'black', bold = T),
    fix.fx = cell_spec(fix.fx, 'html', color = 'black'),
    est1 = cell_spec(est1, 'html', color = 'black'),
    est2 = cell_spec(est2, 'html', color = 'black'),
    d.aic = cell_spec(d.aic, 'html', color = 'black'),
   # aic.values = cell_spec(aic.values, 'html', color = 'black'),
    r2.values = cell_spec(r2.values, 'html', color = 'black'
    )) %>% 
  kable(format = 'html', escape = F, col.names = c(
    'Dependent Variable',
    'Model Fixed Effects',
    'Hydration',
    "Dry Matter",
    '\U0394 AIC',
   # 'AIC',
    expression('Marg. R2/Cond. R2')
  )) %>% 
  kable_styling(bootstrap_options = c('hover', 'bordered', 'condensed'), fixed_thead = T) %>% 
  row_spec(c(1, 6, 11, 18), background = "#e5e5e5", bold = T, font_size = 16) %>% 
  row_spec(c(0), background = "lightgrey", bold = T, font_size = 16) %>% 
  column_spec (1:6, border_left = T, border_right = T) %>%
  row_spec(c(1,2, 3, 6,7,11, 12, 13, 14, 15, 18, 19, 20), bold = T) %>% 
  kableExtra::save_kable(here('figures', 'extra-figures', 'expanded_MEM_kable2.html'))
```


```{r}
rwc_pv_df %>% 
  ggplot(aes(y = tti, x = rwc.int.unscaled, color = spp)) +
  geom_point()

rwc_pv_df %>% 
  ggplot(aes(y = PC1, x = rwc.int.unscaled, color = spp)) +
  geom_point()

rwc_pv_df %>% 
  ggplot(aes(y = PC2, x = rwc.int.unscaled, color = spp)) +
  geom_point()

rwc_pv_df %>% 
  ggplot(aes(y = gd, x = rwc.int.unscaled, color = spp)) +
  geom_point()

rwc_pv_df %>% 
  ggplot(aes(y = fh, x = rwc.int.unscaled, color = spp)) +
  geom_point()

rwc_pv_df %>% 
  ggplot(aes(y = fd, x = rwc.int.unscaled, color = spp)) +
  geom_point()
```

Methods: 

To determine relative water content, we performed an extrapolated linear regression on points more hydrated than the TLP (~-3.0 MPa) as is done in PC curve analysis. Unlike PV Curves, our dry down utilized different leaves of varying dry masses, leading to occasional estimates of RWC that were greater than 100% (e.g., when samples measured at water potentials closest to zero had lower overall dry matter than samples measured at less hydrated water potentials); in these cases RWC was assumed to be 100% and was notated accordingly. Results were similar when we used the highest observed LFM or mpa.scaledas the saturated water content.


#3.0 RWC dynamics: 

#####3.1 Remef

Try scaling MPa from 0-1 and not changing RWC (since it's already on that 0-1 scale). 
```{r}
rwc_pv_df <- rwc_pv_df %>% 
  group_by(spp) %>% 
  mutate(mpa.0.1.scaled = rescale(mpa))
  
mparwc_mod <- lmer(rwc.int.unscaled ~ mpa.0.1.scaled + spp + season + site + (1 | individual), data = rwc_pv_df)

summary(mparwc_mod)

r1_rwc <- remef(mparwc_mod, fix = c('sppCEME', 'siteSt. Marys'),  ran = "all", keep.intercept = FALSE)

scaled_r1_rwc_df <- as.data.frame(r1_rwc)
###

rwc_rel_df <- cbind(scaled_r1_rwc_df, rwc_pv_df)

rwc_rel_df.adfa <- rwc_rel_df %>% 
  filter(spp == 'ADFA')
rwc_rel_df.ceme <- rwc_rel_df %>% 
  filter(spp == 'CEME')
```

```{r}
rwc_pv_df %>% 
  ggplot(aes(y = rwc.int.unscaled, x = year.month, fill = spp))+
  geom_boxplot()

rwc_pv_df %>% 
  ggplot(aes(y = rwc, x = year.month, fill = spp))+
  geom_boxplot()
```


####3.1 Figures

with remef: 

```{r}
mean_tlp_all <- rwc_pv_df %>% 
  mutate(mean_tlp_all = mean(mean_tlp)) %>% 
  select(mean_tlp_all)
  
lfm_rwc_plot_adfa<- rwc_rel_df  %>% 
  filter(spp == "ADFA") %>% 
 # mutate(mean_tlp_all = mean(mean_tlp)) %>% 
  ggplot(aes(
    y = 100 * r1_rwc,
    x = lfm.unscaled,
    color = season,
    shape = season
  )) +
  geom_point(alpha = .5,size = 1)+
 # geom_smooth(method = 'lm', se = F)+
 # ylim(0, 100) +
 # geom_abline() +
 # geom_vline(xintercept = mean_tlp_all)+
 # facet_wrap(~Species) +
  theme(legend.position = "none",
        #axis.text.x = element_blank(),
        axis.text = element_text(size = 14), 
        axis.title = element_text(size = 16)) +
  labs(x="Live Fuel Moisture (%)", y="Relative Water Content (%)") +
  annotate('text', x = 0, y = 95, label = 'a', size = 10) +
  th +
  scale_color_manual(values= c("#9D0208",'#669bbc'))


lfm_rwc_plot_ceme<- rwc_rel_df  %>% 
  filter(spp == "CEME") %>% 
 # mutate(mean_tlp_all = mean(mean_tlp)) %>% 
  ggplot(aes(
    y = 100 * r1_rwc,
    x = lfm.unscaled,
    color = season,
    shape = season
  )) +
  geom_point(alpha = .5,size = 1)+
 # geom_smooth(method = 'lm', se = F)+
 # ylim(0, 100) +
 # geom_abline() +
 # geom_vline(xintercept = mean_tlp_all)+
 # facet_wrap(~Species) +
 theme(legend.position = "none",
        #axis.text.x = element_blank(),
        axis.text = element_text(size = 14), 
        axis.title = element_text(size = 16)) +
  labs(x="Live Fuel Moisture (%)", y="Relative Water Content (%)") +
  annotate('text', x = 0, y = 95, label = 'b', size = 10) +
  th +
  scale_color_manual(values= c("#9D0208",'#669bbc'))
lfm_rwc_plot


lfm_rwc_plot <- cowplot::plot_grid(lfm_rwc_plot_adfa, lfm_rwc_plot_ceme, nrow = 1)



mpa_rwc_plot_adfa <- rwc_rel_df  %>% 
  filter(spp == "ADFA") %>% 
  ggplot(aes(
    y = 100 * r1_rwc,
    x = mpa.unscaled,
    color = season,
    shape = season,
  )) +
  geom_point(alpha = .5,size = 1)+
 # geom_smooth(method = 'lm', se = F) +
 # ylim(0, 100) +
 # facet_wrap(~Species) +
 theme(legend.position = "none",
        #axis.text.x = element_blank(),
        axis.text = element_text(size = 14), 
        axis.title = element_text(size = 16)) +
  labs(x="Water Potential (-Mpa)", y="Relative Water Content (%)") +
  annotate('text', x = -10, y = 95, label = 'c', size = 10) +
  th +
  scale_color_manual(values= c("#9D0208",'#669bbc'))

mpa_rwc_plot_ceme <- rwc_rel_df  %>% 
  filter(spp == "CEME") %>% 
  ggplot(aes(
    y = 100 * r1_rwc,
    x = mpa.unscaled,
    color = season,
    shape = season,
  )) +
  geom_point(alpha = .5,size = 1)+
 # geom_smooth(method = 'lm', se = F, size = 1) +
 # ylim(0, 100) +
 # facet_wrap(~Species) +
  theme(legend.position = "none",
        #axis.text.x = element_blank(),
        axis.text = element_text(size = 14), 
        axis.title = element_text(size = 16)) +
  labs(x="Water Potential (-Mpa)", y="Relative Water Content (%)") +
  annotate('text', x = -10, y = 95, label = 'd', size = 10) +
  th +
  scale_color_manual(values= c("#9D0208",'#669bbc'))

mpa_rwc_plot <- cowplot::plot_grid(mpa_rwc_plot_ceme, mpa_rwc_plot_adfa, nrow = 1)

rwc_plot1 <- cowplot::plot_grid(lfm_rwc_plot, mpa_rwc_plot, nrow = 2)

rwc_plot1

rwc_rel_df <-rwc_rel_df %>% 
  mutate(Season = season)

legend <- get_legend(
  # create some space to the left of the legend
  ggplot() + 
    geom_point(data= rwc_rel_df, aes(mpa.unscaled, PC1, color = Season, shape = Season), size = 3, alpha = .8) +
    color_panels +
    theme(legend.box.margin = margin(0, 0, 0, 0),
          legend.position = "top",
          legend.text = element_text(face = "italic", size = 14),
          legend.title = element_text(face = 'bold', size = 18),
          legend.key = element_rect(fill = "white"))
) # NOTE: CAUSING ERROR!! SCALED.MEM.DATA.SEPT2020 NOT FOUND; COLOR_PANELS NOT FOUND

rwc_plot2 <- cowplot::plot_grid (legend, rwc_plot1, ncol = 1, rel_heights = c(1,12), rel_widths = c(1,.5))
rwc_plot2

ggsave(here("figures", 'reviewer-figures', "Fig1_RWC_LFM_MPa.jpg"),
       plot = rwc_plot2, 
       width = 5, height = 3, dpi = 300)
```


without remef: 

```{r}
mean_tlp_all <- rwc_pv_df %>% 
  mutate(mean_tlp_all = mean(mean_tlp)) %>% 
  select(mean_tlp_all)
  
lfm_rwc_plot_adfa<- rwc_rel_df  %>% 
  filter(spp == "ADFA") %>% 
 # mutate(mean_tlp_all = mean(mean_tlp)) %>% 
  ggplot(aes(
    y = 100 * rwc.int.unscaled,
    x = lfm.unscaled,
    color = season,
    shape = season
  )) +
  geom_point(alpha = .5,size = 1)+
  ylim(20, 100)+
 # geom_smooth(method = 'lm', se = F)+
 # ylim(0, 100) +
 # geom_abline() +
 # geom_vline(xintercept = mean_tlp_all)+
 # facet_wrap(~Species) +
  theme(legend.position = "none",
        #axis.text.x = element_blank(),
        axis.text = element_text(size = 14), 
        axis.title = element_text(size = 16)) +
  labs(x="Live Fuel Moisture (%)", y="Relative Water Content (%)") +
  annotate('text', x = 0, y = 95, label = 'a', size = 10) +
  th +
  scale_color_manual(values= c("#9D0208",'#669bbc'))


lfm_rwc_plot_ceme<- rwc_rel_df  %>% 
  filter(spp == "CEME") %>% 
 # mutate(mean_tlp_all = mean(mean_tlp)) %>% 
  ggplot(aes(
    y = 100 * rwc.int.unscaled,
    x = lfm.unscaled,
    color = season,
    shape = season
  )) +
  geom_point(alpha = .5,size = 1)+
   ylim(20, 100)+
 theme(legend.position = "none",
        #axis.text.x = element_blank(),
        axis.text = element_text(size = 14), 
        axis.title = element_text(size = 16)) +
  labs(x="Live Fuel Moisture (%)", y="Relative Water Content (%)") +
  annotate('text', x = 0, y = 95, label = 'b', size = 10) +
  th +
  scale_color_manual(values= c("#9D0208",'#669bbc'))
lfm_rwc_plot


lfm_rwc_plot <- cowplot::plot_grid(lfm_rwc_plot_adfa, lfm_rwc_plot_ceme, nrow = 1)



mpa_rwc_plot_adfa <- rwc_rel_df  %>% 
  filter(spp == "ADFA") %>% 
  ggplot(aes(
    y = 100 * rwc.int.unscaled,
    x = mpa.unscaled,
    color = season,
    shape = season,
  )) +
  geom_point(alpha = .5,size = 1)+
   ylim(20, 100)+
 theme(legend.position = "none",
        axis.text = element_text(size = 14), 
        axis.title = element_text(size = 16)) +
  labs(x="Water Potential (-Mpa)", y="Relative Water Content (%)") +
  annotate('text', x = -10, y = 95, label = 'c', size = 10) +
  th +
  scale_color_manual(values= c("#9D0208",'#669bbc'))

mpa_rwc_plot_ceme <- rwc_rel_df  %>% 
  filter(spp == "CEME") %>% 
  ggplot(aes(
    y = 100 * rwc.int.unscaled,
    x = mpa.unscaled,
    color = season,
    shape = season,
  )) +
  geom_point(alpha = .5,size = 1)+
 ylim(20, 100)+
  theme(legend.position = "none",
        axis.text = element_text(size = 14), 
        axis.title = element_text(size = 16)) +
  labs(x="Water Potential (-Mpa)", y="Relative Water Content (%)") +
  annotate('text', x = -10, y = 95, label = 'd', size = 10) +
  th +
  scale_color_manual(values= c("#9D0208",'#669bbc'))

mpa_rwc_plot <- cowplot::plot_grid(mpa_rwc_plot_ceme, mpa_rwc_plot_adfa, nrow = 1)

rwc_plot1 <- cowplot::plot_grid(lfm_rwc_plot, mpa_rwc_plot, nrow = 2)

rwc_plot1

rwc_rel_df <-rwc_rel_df %>% 
  mutate(Season = season)

legend <- get_legend(
  # create some space to the left of the legend
  ggplot() + 
    geom_point(data= rwc_rel_df, aes(mpa.unscaled, PC1, color = Season, shape = Season), size = 3, alpha = .8) +
    color_panels +
    theme(legend.box.margin = margin(0, 0, 0, 0),
          legend.position = "top",
          legend.text = element_text(face = "italic", size = 14),
          legend.title = element_text(face = 'bold', size = 18),
          legend.key = element_rect(fill = "white"))
) # NOTE: CAUSING ERROR!! SCALED.MEM.DATA.SEPT2020 NOT FOUND; COLOR_PANELS NOT FOUND

rwc_plot2 <- cowplot::plot_grid (legend, rwc_plot1, ncol = 1, rel_heights = c(1,12), rel_widths = c(1,.5))
rwc_plot2

ggsave(here("figures", 'reviewer-figures', "Fig1_RWC_LFM_MPa_NOREMEF.jpg"),
       plot = rwc_plot2, 
       width = 5, height = 3, dpi = 300)
```

Without remef: 

```{r}
mean_tlp_all <- rwc_pv_df %>% 
  mutate(mean_tlp_all = mean(mean_tlp)) %>% 
  select(mean_tlp_all)
  
lfm_rwc_plot<- rwc_pv_df  %>% 
 # mutate(mean_tlp_all = mean(mean_tlp)) %>% 
  ggplot(aes(
    y = 100 * rwc.int.unscaled,
    x = lfm.unscaled,
    color = season,
    shape = season
  )) +
  geom_point(alpha = .5,size = 2, )+
  geom_smooth(method = 'lm', se = F)+
  ylim(0, 100) +
  facet_wrap(~Species)+
 # geom_abline() +
 # geom_vline(xintercept = mean_tlp_all)+
  #facet_wrap(~year.month) +
  theme(legend.position = "none",
        #axis.text.x = element_blank(),
        axis.text = element_text(size = 14)) +
  labs(x="Live Fuel Moisture (%)", y="Relative Water Content (%)") +
  annotate('text', x = 0, y = 97, label = 'a', size = 12) +
  th +
  scale_color_manual(values= c("#9D0208",'#669bbc'))
lfm_rwc_plot
```


```{r}
mpa_rwc_plot <- rwc_pv_df %>% 
  ggplot(aes(
    y = 100 * rwc.int.unscaled,
    x = mpa.unscaled,
    color = season,
    shape = season
  )) +
  geom_point(alpha = .5,size = 2)+
  geom_smooth(method = 'lm', se = F) +
  ylim(0, 100) +
  facet_wrap(~Species) +
  theme(legend.position = "none",
        #axis.text.x = element_blank(),
        axis.text = element_text(size = 14)) +
  labs(x="Water Potential (-Mpa)", y="") +
  annotate('text', x = -10, y = 97, label = 'b', size = 12) +
  th +
  scale_color_manual(values= c("#9D0208",'#669bbc'))

rwc_plot1 <- cowplot::plot_grid(lfm_rwc_plot, mpa_rwc_plot, nrow = 1)

rwc_plot1

legend <- get_legend(
  # create some space to the left of the legend
  ggplot() + 
    geom_point(data=rwc_pv_df, aes(mpa.unscaled, PC1, color = season, shape = season), size = 5, alpha = .8) +
    color_panels +
    theme(legend.box.margin = margin(0, 0, 0, 0),
          legend.position = "top",
          legend.text = element_text(face = "italic", size = 18),
          legend.title = element_text(face = 'bold', size = 22),
          legend.key = element_rect(fill = "white"))
)

rwc_plot2 <- cowplot::plot_grid (legend, rwc_plot1, ncol = 1, rel_heights = c(1,12), rel_widths = c(1,.5))
rwc_plot2

ggsave(here("figures", 'reviewer-figures', "Fig1_NOREMEF_RWC_LFM_MPa.jpg"),
       plot = rwc_plot2, 
       width = 6, height = 3, dpi = 300)
```

#CONCLUSIONS:

Results: RWC showed a similar relationship with MPa LFM, but with much less distinguishment between species (maybe because of remef??). Need to think on this more.. 

#4.0 Segmented regression: 

Note: does not seem to make a difference if the 100% (i.e. some that had LFMs that were very high) are excluded - detected threshold is the same.

# Splitting by Species
```{r}
seg.ceme.subset <- rwc_pv_df %>% 
  filter(spp == "CEME") %>% 
  filter(rwc.int.unscaled < 99)
```

```{r}
seg.adfa.subset <- rwc_pv_df %>% 
  filter(spp == "ADFA")
```

#CEME

### PC1

n.s., 0.5039579
```{r}
o.mod.a<-lme(PC1~rwc.int.unscaled + year.month , random=~1|id, data=seg.ceme.subset)

os.mod.a<-segmented.default(o.mod.a, ~rwc.int.unscaled, npsi=list(rwc.int.unscaled=1))
#summarizing results (note the '.coef' argument)
#m <- slope(os.mod.a, .coef = fixef(os.mod.a))[[1]][[1]]
summary.segmented(os.mod.a)
```

###PC2
0.01812 , Est =0.591 
```{r}
o.mod.a<-lm(PC2~rwc.int.unscaled + site + year.month, random=~1|id, 
             data=seg.ceme.subset)

os.mod.a<-segmented(o.mod.a, ~rwc.int.unscaled, npsi=list(rwc.int.unscaled=1))
#summarizing results (note the '.coef' argument)
#m <- slope(os.mod.a, .coef = fixef(os.mod.a))[[1]][[1]]
summary.segmented(os.mod.a)
```

###TTI
p = 3.692689e-07, Est = 0.9855 (98.5% RWC)
```{r}
o.mod.a<-lme(tti~rwc.int.unscaled + year.month + site, random=~1|id, data=seg.ceme.subset)

os.mod.a<-segmented.default(o.mod.a, ~rwc.int.unscaled, npsi=list(rwc.int.unscaled=1))
#summarizing results (note the '.coef' argument)
#m <- slope(os.mod.a, .coef = fixef(os.mod.a))[[1]][[1]]
summary.segmented(os.mod.a)
```

###FH
p = 5.459356e-04, Est = 0.5494623 (55%)
```{r}
o.mod.a<-lme(fh~rwc.int.unscaled + year.month + site, random=~1|id, data=seg.ceme.subset)

os.mod.a<-segmented.default(o.mod.a, ~rwc.int.unscaled, npsi=list(rwc.int.unscaled=1))
#summarizing results (note the '.coef' argument)
#m <- slope(os.mod.a, .coef = fixef(os.mod.a))[[1]][[1]]
summary.segmented(os.mod.a)
```
###GD
n.s.0.3738128
```{r}
o.mod.a<-lme(gd~rwc.int.unscaled + year.month + site, random=~1|id, data=seg.ceme.subset)

os.mod.a<-segmented.default(o.mod.a, ~rwc.int.unscaled, npsi=list(rwc.int.unscaled=1))
#summarizing results (note the '.coef' argument)
#m <- slope(os.mod.a, .coef = fixef(os.mod.a))[[1]][[1]]
summary.segmented(os.mod.a)
```

###Temp Max. 
2.407598e-01, Est = 0.4200656  (42%)

```{r}
o.mod.a<-lme(temp.max~rwc.int.unscaled + year.month + site, random=~1|id, data=seg.ceme.subset)

os.mod.a<-segmented.default(o.mod.a, ~rwc.int.unscaled, npsi=list(rwc.int.unscaled=1))
#summarizing results (note the '.coef' argument)
#m <- slope(os.mod.a, .coef = fixef(os.mod.a))[[1]][[1]]
summary.segmented(os.mod.a)
```

###Flame duration
n.s.0.5458679 
```{r}
o.mod.a<-lme(fd~rwc.int.unscaled + year.month + site, random=~1|id, data=seg.ceme.subset)

os.mod.a<-segmented.default(o.mod.a, ~rwc.int.unscaled, npsi=list(rwc.int.unscaled=1))
#summarizing results (note the '.coef' argument)
#m <- slope(os.mod.a, .coef = fixef(os.mod.a))[[1]][[1]]
summary.segmented(os.mod.a)
```
###PFG

n.s. 0.3738128
```{r}
o.mod.a<-lme(pfg~rwc.int.unscaled + year.month + site, random=~1|id, data=seg.ceme.subset)

os.mod.a<-segmented.default(o.mod.a, ~rwc.int.unscaled, npsi=list(rwc.int.unscaled=1))
#summarizing results (note the '.coef' argument)
#m <- slope(os.mod.a, .coef = fixef(os.mod.a))[[1]][[1]]
summary.segmented(os.mod.a)
```
###Prop ignite

3.269609e-07, Est = 89.85%

```{r}
o.mod.a<-lme(prop.ignite~rwc.int.unscaled + year.month + site, random=~1|id, data=seg.ceme.subset)

os.mod.a<-segmented.default(o.mod.a, ~rwc.int.unscaled, npsi=list(rwc.int.unscaled=1))
#summarizing results (note the '.coef' argument)
#m <- slope(os.mod.a, .coef = fixef(os.mod.a))[[1]][[1]]
summary.segmented(os.mod.a)
```
##GTI
n.s. 0.3738128
```{r}
o.mod.a<-lme(gti~rwc.int.unscaled + year.month + site, random=~1|id, data=seg.ceme.subset)

os.mod.a<-segmented.default(o.mod.a, ~rwc.int.unscaled, npsi=list(rwc.int.unscaled=1))
#summarizing results (note the '.coef' argument)
#m <- slope(os.mod.a, .coef = fixef(os.mod.a))[[1]][[1]]
summary.segmented(os.mod.a)
```
##TTFG

0.009771142, Est = 0.5009272

```{r}
o.mod.a<-lme(ttfg~rwc.int.unscaled + year.month + site, random=~1|id, data=seg.ceme.subset)

os.mod.a<-segmented.default(o.mod.a, ~rwc.int.unscaled, npsi=list(rwc.int.unscaled=1))
#summarizing results (note the '.coef' argument)
#m <- slope(os.mod.a, .coef = fixef(os.mod.a))[[1]][[1]]
summary.segmented(os.mod.a)
````



#ADFA: 

### PC1

n.s. 0.527 

```{r}
o.mod.a<-lm(PC1~rwc.int.unscaled + site + year.month, #random=~1|id, 
             data=seg.adfa.subset)

os.mod.a<-segmented(o.mod.a, ~rwc.int.unscaled, npsi=list(rwc.int.unscaled=1))
#summarizing results (note the '.coef' argument)
#m <- slope(os.mod.a, .coef = fixef(os.mod.a))[[1]][[1]]
summary.segmented(os.mod.a)
```
###PC2
7.68e-05, Est = 0.698 (69.8%)
```{r}
o.mod.a<-lm(PC2~rwc.int.unscaled + site + year.month, random=~1|id, 
             data=seg.adfa.subset)

os.mod.a<-segmented(o.mod.a, ~rwc.int.unscaled, npsi=list(rwc.int.unscaled=1))
#summarizing results (note the '.coef' argument)
#m <- slope(os.mod.a, .coef = fixef(os.mod.a))[[1]][[1]]
summary.segmented(os.mod.a)
```

###TTI
n.s. 0.5128414
```{r}
o.mod.a<-lme(tti~rwc.int.unscaled + year.month + site, random=~1|id, data=seg.adfa.subset)

os.mod.a<-segmented.default(o.mod.a, ~rwc.int.unscaled, npsi=list(rwc.int.unscaled=1))
#summarizing results (note the '.coef' argument)
#m <- slope(os.mod.a, .coef = fixef(os.mod.a))[[1]][[1]]
summary.segmented(os.mod.a)
```
###FD
0.4844877, n.s. 
```{r}
o.mod.a<-lme(fd~rwc.int.unscaled + year.month + site, random=~1|id, data=seg.adfa.subset)

os.mod.a<-segmented.default(o.mod.a, ~rwc.int.unscaled, npsi=list(rwc.int.unscaled=1))
#summarizing results (note the '.coef' argument)
#m <- slope(os.mod.a, .coef = fixef(os.mod.a))[[1]][[1]]
summary.segmented(os.mod.a)
```

###FH
p = 2.239756e-07, Est. = 93.14%
```{r}
o.mod.a<-lme(fh~rwc.int.unscaled + year.month + site, random=~1|id, data=seg.adfa.subset)

os.mod.a<-segmented.default(o.mod.a, ~rwc.int.unscaled, npsi=list(rwc.int.unscaled=1))
#summarizing results (note the '.coef' argument)
#m <- slope(os.mod.a, .coef = fixef(os.mod.a))[[1]][[1]]
summary.segmented(os.mod.a)
```
###GD
p = 1.507566e-01, Est = 0.9752026 (97.5%)
```{r}
o.mod.a<-lme(gd~rwc.int.unscaled + year.month + site, random=~1|id, data=seg.adfa.subset)

os.mod.a<-segmented.default(o.mod.a, ~rwc.int.unscaled, npsi=list(rwc.int.unscaled=1))
#summarizing results (note the '.coef' argument)
#m <- slope(os.mod.a, .coef = fixef(os.mod.a))[[1]][[1]]
summary.segmented(os.mod.a)
```
###Temp Max. 
p = 1.848159e-02, Est = 0.6638564  (63%)
```{r}
o.mod.a<-lme(temp.max~rwc.int.unscaled + year.month + site, random=~1|id, data=seg.adfa.subset)

os.mod.a<-segmented.default(o.mod.a, ~rwc.int.unscaled, npsi=list(rwc.int.unscaled=1))
#summarizing results (note the '.coef' argument)
#m <- slope(os.mod.a, .coef = fixef(os.mod.a))[[1]][[1]]
summary.segmented(os.mod.a)
```

###Flame duration
n.s. 0.4844877
```{r}
o.mod.a<-lme(fd~rwc.int.unscaled + year.month + site, random=~1|id, data=seg.adfa.subset)

os.mod.a<-segmented.default(o.mod.a, ~rwc.int.unscaled, npsi=list(rwc.int.unscaled=1))
#summarizing results (note the '.coef' argument)
#m <- slope(os.mod.a, .coef = fixef(os.mod.a))[[1]][[1]]
summary.segmented(os.mod.a)
```
###PFG
6.712821e-04, Est = 0.9479989 
```{r}
o.mod.a<-lme(pfg~rwc.int.unscaled + year.month + site, random=~1|id, data=seg.adfa.subset)

os.mod.a<-segmented.default(o.mod.a, ~rwc.int.unscaled, npsi=list(rwc.int.unscaled=1))
#summarizing results (note the '.coef' argument)
#m <- slope(os.mod.a, .coef = fixef(os.mod.a))[[1]][[1]]
summary.segmented(os.mod.a)
```

###Prop ignite

5.328317e-01, Est = 46%

```{r}
o.mod.a<-lme(prop.ignite~rwc.int.unscaled + year.month + site, random=~1|id, data=seg.adfa.subset)

os.mod.a<-segmented.default(o.mod.a, ~rwc.int.unscaled, npsi=list(rwc.int.unscaled=1))
#summarizing results (note the '.coef' argument)
#m <- slope(os.mod.a, .coef = fixef(os.mod.a))[[1]][[1]]
summary.segmented(os.mod.a)
```
##TTFG

3.544659e-05, Est = 0.9676977

```{r}
o.mod.a<-lme(ttfg~rwc.int.unscaled + year.month + site, random=~1|id, data=seg.adfa.subset)

os.mod.a<-segmented.default(o.mod.a, ~rwc.int.unscaled, npsi=list(rwc.int.unscaled=1))
#summarizing results (note the '.coef' argument)
#m <- slope(os.mod.a, .coef = fixef(os.mod.a))[[1]][[1]]
summary.segmented(os.mod.a)
```

##GTI
n.s. 0.527


```{r}
o.mod.a<-lm(gti~rwc.int.unscaled + year.month + site, data=seg.adfa.subset)

os.mod.a<-segmented(o.mod.a, ~rwc.int.unscaled, npsi=list(rwc.int.unscaled=1))
#summarizing results (note the '.coef' argument)
#m <- slope(os.mod.a, .coef = fixef(os.mod.a))[[1]][[1]]
summary.segmented(os.mod.a)
```

#CONCLUSIONS: 



# Kable Table
Setup:
```{r}
dep.var <- c(
  'PC1',
  'PC2',
  'TTI',
  'Prop Ignite (5%)',
  'GTI',
  'FH',
  'Max Temp.',
  'FD',
  'GD',
  'PFG',
  'TTFG',
  'Summary',
  'PC1',
  'PC2',
  'TTI',
  'Prop Ignite (5%)',
  'GTI',
  'FH',
  'Max Temp.',
  'FD',
  'GD',
  'PFG',
  'TTFG',
  'Summary'
)
spp <- c(rep('ADFA', 11), 'TLP: 80.33 +/- 8.48 ', rep('CEME', 11), 'TLP: 84.82 +/- 2.13')
sig.p <- c(
  'No', #PC1
  'Yes (p < 0.0005)', #PC2
  'No ', #TTI
  'No', #prop.ignite
  'No', #GTI
  'Yes (p < 0.0001)', #FH
  'Yes (p = 0.0185)', #Max temp
  'No', #FD
  'Yes (p = 0.0151)', #GD
  'Yes (p < 0.0001)', #PFG
  'Yes (p < 0.0001)', #TTFG
  '6/11 Significant',
  'No', #PC1
  'Yes (p = 0.0181)', #PC2
  'Yes (p < 0.0001)', #TTI
  'Yes (p < 0.0001)', #prop.ignite
  'No', #GTI
  'Yes (p < 0.0001)', #FH
  'Yes (p = 0.0241)', #Max temp
  'No', #FD
  'No', #GD
  'No', #PFG
  'Yes (p < 0.0098)', #TTFG
  '6/11 Significant'
)
threshold <- c(
  '52.70', #PC1
  '69.80', #PC2
  '51.28 ', #TTI
  '53.28', #prop.ignite
  '52.70', #GTI
  '93.14', #FH
  '66.38', #Max temp
  '48.45', #FD
  '97.52', #GD
  '94.80', #PFG
  '96.77', #TTFG
  'Avg: 86.57 +/- 14.45',
  '50.93', #PC1
  '59.10', #PC2
  '98.55 ', #TTI
  '89.85', #prop.ignite
  '37.38', #GTI
  '54.94', #FH
  '42.00', #Max temp
  '54.59', #FD
  'NA', #GD
  'NA', #PFG
  '50.00', #TTFG
  'Avg: 65.74 +/- 22.93'
)


seg.reg.table.df <- data.frame(dep.var, spp, sig.p, threshold)
```

```{r}
mean.adfa <- mean(c(#52.70,
                    69.80,
                   # 51.28,
                   # 53.28,
                    #52.70,
                    93.14,
                    66.38,
                   # 48.45,
                    97.52,
                    94.80,
                    97.77))

sd.adfa <- sd(c(#52.70,
                    69.80,
                   # 51.28,
                   # 53.28,
                    #52.70,
                    93.14,
                    66.38,
                   # 48.45,
                    97.52,
                    94.80,
                    97.77))

mean.adfa
sd.adfa

mean.ceme <- mean(c(#50.92, 
                    59.1, 98.55, 89.85, #37.38,
                    54.94, 42, 
                    #54.59, 
                    50.00))

sd.ceme <- sd(c(#50.92, 
                    59.1, 98.55, 89.85, #37.38,
                    54.94, 42, 
                    #54.59, 
                    50.00))

mean.ceme
sd.ceme

```

Kable:
```{r}
seg.reg.table.df %>% 
   mutate(dep.var = cell_spec(dep.var, 'html', color = 'black', bold = T), spp = cell_spec(spp, 'html', color = 'black'), sig.p = cell_spec(sig.p, 'html', color = 'black'), threshold = cell_spec(threshold, 'html', color = 'black')) %>% 
  kable(format = 'html', escape = F, col.names = c('Dependent Variable', 'Species', 'Significant?','Threshold (RWC)'), caption = 'RWC') %>% 
  kable_styling(bootstrap_options = c('hover', 'bordered', 'condensed'), fixed_thead = T) %>% 
  row_spec(c(2, 6, 7, 9:11, 12, 14:16, 18:19, 23:24), bold = T) %>% 
  row_spec(c(1, 3, 4,5,8,13,17,20:22), bold = F) %>% 
  row_spec(c(12, 24), background = "#D3D3D3", bold = T) %>%
  save_kable(here('figures', 'supp-figures', 'TabS5_rwc.segreg.summary.html'))
```

# ---------------------------------------

#Comment 2: Does LDMC correlate with TLP?

Reviewer 1: 

Stears et al. 2022 (Ecology 103 (11) Water availability dictates how plant traits predict demographic rates) found that TLP and LDMC were strongly correlated for sixteen grass and forb species (pearson correlation of 0.831).  It would be neat to check whether TLP and LDMC were also strongly correlated for the study species you encountered.  

####Data wrangling:

```{r}
local_pv_curves <- read_csv(here::here("raw-data", "allcurves.csv"), show_col_types = FALSE) %>% 
  clean_names() %>% 
  filter(#spp %in% c("CEME", "ADFA"), 
         tlp < -1.5 ) %>% 
  mutate(fresh_weight_saturated = as.numeric(fresh_weight_saturated), 
         ldmc = dry_weight/fresh_weight_saturated, 
         date = mdy(date), 
         year = year(date)) 

summary_pv_curves_notclean <- local_pv_curves %>% 
  group_by(spp, timing) %>% 
  summarise(mean_tlp = mean(tlp), 
            sd = sd(tlp), 
            mean_rwc = mean(rwc_at_tlp), 
            sd_rwc = sd(rwc_at_tlp),
            mean_lfm = mean(lfm_at_tlp, na.rm = T), 
            sd_lfm = sd(lfm_at_tlp, na.rm = T),
            mean_po = mean(po, na.rm = T), 
            sd_po = sd(po, na.rm = T))

local_pv_curves %>% 
  #filter(year != 2016)%>% 
  ggplot(aes(y = ldmc, x = tlp, color = date)) +
  geom_point() +
  facet_wrap(~spp)+
  geom_smooth(method = "lm")

local_pv_curves %>% 
  #filter(year != 2016)%>% 
  ggplot(aes(y = ldmc, x = tlp, color = as.factor(date), shape = spp)) +
  geom_point() +
 # facet_wrap(~spp)+
  geom_smooth(method = "lm")
```
The code below is maybe what we actually used in the paper? Mean values are slightly different when I use the above. Starting to think I might have taken the PV parameters straight from the excel sheet (which is fine, but then need to get all that data in proper order too).

```{r, include = F, eval=FALSE}
# Section 1: Pressure Volume Curves
pv_data_all <- read_csv(here("raw-data", "allcurves_clean.csv"), show_col_types = FALSE) %>% 
  mutate(sample = as.factor(sample),
         date = lubridate::dmy(date),
         month = month(date),
         dataset = "pv") %>% 
  janitor::clean_names()

#by timing:
pv_summary_df_timing_mpa <- pv_data_all %>% 
  group_by(spp, timing) %>% 
  dplyr::summarise(ci = list(mean_cl_normal(tlp) %>% 
                        dplyr::rename(mean_tlp=y, lwr=ymin, upr=ymax))) %>% 
  unnest

pv_summary_df_timing_lfm <- pv_data_all %>% 
  drop_na(lfm_at_tlp) %>% 
  group_by(spp, timing) %>% 
  dplyr::summarise(ci = list(mean_cl_normal(lfm_at_tlp) %>% 
                        dplyr::rename(mean_lfm=y, lwr_lfm=ymin, upr_lfm=ymax))) %>% 
  unnest

pv_summary_df_timing <- merge(pv_summary_df_timing_mpa, pv_summary_df_timing_lfm, by = c("spp", "timing"), all = T)

write_csv(pv_summary_df_timing, here("processed-data", "pv_summary_df_timing1.csv"))

#by species:

pv_summary_df_spp_mpa <- pv_data_all %>% 
  group_by(spp) %>% 
  dplyr::summarise(ci = list(mean_cl_normal(tlp) %>% 
                        dplyr::rename(mean_tlp=y, lwr=ymin, upr=ymax)), 
            ci = list(mean_cl_normal(lfm_at_tlp))) %>% 
  unnest

pv_summary_df_spp_lfm <- pv_data_all %>% 
  drop_na(lfm_at_tlp) %>% 
  group_by(spp) %>% 
  dplyr::summarise(ci = list(mean_cl_normal(lfm_at_tlp) %>% 
                        dplyr::rename(mean_lfm=y, lwr_lfm=ymin, upr_lfm=ymax))) %>% 
  unnest

pv_summary_spp_timing <- merge(pv_summary_df_spp_mpa, pv_summary_df_spp_lfm, by = c("spp"), all = T)

write_csv(pv_summary_spp_timing, here("processed-data", "pv_summary_df_spp1.csv"))
```

#CONCLUSIONS: 

This is an interesting idea for further application of our dataset, but with only two species we are unable to really dig into this idea. There is large variability in both measured TLP and LDMC (across years, mostly). IF LDMC changes across seasons and TLP doesn't, then they wouldn't be strongly related. 


# ---------------------------------------

#Comment 3: LDMC and predicting flammability

Reviewer 1:

Since water potential and TLP were very predictive of flammability and thresholds, it would be great to understand whether the more easily measured and commonly available LDMC is equally as predictive and shows a similar threshold like behavior.  One potential complication, and fundamental way that LDMC measures differ from LFM is that LDMC is measured at “full hydration”, meaning leaves were fully hydrated in the lab.  We don’t know whether the authors took these types of measurements.  We see in the methods that branchlets were fully hydrated and LFM then measured periodically, so this might be a potential source for an LDMC comparison."

Reviewer 3: 

"We also think there is a potential opportunity to broaden the utility of this approach by looking at a more commonly (and easily) measured plant functional trait that has also been linked to drought response: Leaf dry matter content (LDMC). 

# MAIN LINEAR MODEL SELECTION - LDMC

#1.0 LDMC dynamics

```{r}
rwc_pv_df %>% 
  ggplot(aes(x = year.month, y = ldmc, fill = season)) +
  geom_boxplot() +
  facet_wrap(~spp)

rwc_pv_df %>% 
  ggplot(aes(x = year.month, y = ldmc.unscaled, fill = season)) +
  geom_boxplot() +
  facet_wrap(~spp)
```
Are they statistically different?

```{r}
ldmc_mod <- lm(ldmc ~ spp*season, data = rwc_pv_df)

aov(ldmc_mod)

summary(aov(ldmc_mod))


TukeyHSD(aov(ldmc_mod))
```


#2.0 PC1

From our first round of model selection, it appears that site, species, and year.month are all important variables to include, and sample.wt should probably be included as well despite worsening the models slightly.

Below, I am going to perform another round of model selection with all possible permutations of the following metrics: mpa, lfm, dw.flam.sample, ww.flam.sample, excess_water.scaledwith the caveat that some combinations of metrics absolutely do NOT work. Those combinations are as follows:
  - dw.flam.sample, ww.flam.sample and sample.wt
  - dw.flam.sample, lfm, ww.flam.sample
  - lfm, dw.flam.sample
  - lfm, ww.flam.sample
  - lfm, excess_water.scaled
  - ww.flam.sample, excess_water.scaled
  - dw.flam.sample, excess_water.scaled (with sample.wt)
  - dw.flam.sample, ww.flam.sample (with sample.wt)
  
These do not work, as live fuel moisture was used in the calculations of these metrics, therefore they return very high VIF values that signal to us that they cannot be used in the mixed effects models together

```{r}
m1 <- lmer(PC1 ~ lfm + spp + site + year.month + sample.wt + (1 | individual), data = rwc_pv_df)

performance::multicollinearity(m1)
```


```{r}
m2 <- lmer(PC1 ~ mpa.scaled+ spp + site + year.month + sample.wt + (1 | individual), data = rwc_pv_df)

performance::multicollinearity(m2)
```

```{r}
m3 <- lmer(PC1 ~ dw.flam.sample + spp + site + year.month + sample.wt + (1 | individual), data = rwc_pv_df)

performance::multicollinearity(m3)
```

```{r}
m4 <- lmer(PC1 ~ ww.flam.sample + spp + site + year.month + sample.wt + (1 | individual), data = rwc_pv_df)

performance::multicollinearity(m4)
```

```{r}
m5 <- lmer(PC1 ~ excess_water.scaled + spp + site + year.month + sample.wt + (1 | individual), data = rwc_pv_df)

performance::multicollinearity(m5)
```

```{r}
m6 <- lmer(PC1 ~ lfm + mpa.scaled+ spp + site + year.month + sample.wt + (1 | individual), data = rwc_pv_df)

performance::multicollinearity(m6)
```

```{r}
m7 <- lmer(PC1 ~ mpa.scaled+ dw.flam.sample + spp + site + year.month + sample.wt + (1 | individual), data = rwc_pv_df)

performance::multicollinearity(m7)
```

```{r}
m8 <- lmer(PC1 ~ mpa.scaled+ ww.flam.sample + spp + site + year.month + sample.wt + (1 | individual), data = rwc_pv_df)

performance::multicollinearity(m8)
```

```{r}
m9 <- lmer(PC1 ~ mpa.scaled+ excess_water.scaled + spp + site + year.month + sample.wt + (1 | individual), data = rwc_pv_df)

performance::multicollinearity(m9)
```

```{r}
m10 <- lmer(PC1 ~ mpa.scaled+ dw.flam.sample + ww.flam.sample + spp + site + year.month + (1 | individual), data = rwc_pv_df)

performance::multicollinearity(m10)
```

```{r}
m11 <- lmer(PC1 ~ dw.flam.sample + excess_water.scaled + spp + site + year.month + (1 | individual), data = rwc_pv_df)

performance::multicollinearity(m11)
```

```{r}
m12 <- lmer(PC1 ~ dw.flam.sample + ww.flam.sample + spp + site + year.month + (1 | individual), data = rwc_pv_df)

performance::multicollinearity(m12)
```

```{r}
tab_model(m1, m2, m3, m4, m5, m6, m7, m8, m9, m10, m11, m12,
          show.reflvl = TRUE, 
          digits = 3, 
          show.aic = TRUE, 
          show.ci = FALSE,
          show.icc = FALSE, 
          pred.labels = c('Intercept', 'LFM (%)',
                          'Species Effect (C. megacarpus)',
                          'Site Effect (Site 1)',
                          'Sampling Date Effect (Jan 2018)',
                          'Sampling Date Effect (Dec 2019)',
                          'Sampling Date Effect (Jan 2020)',
                          'Sampling Date Effect (Sept 2020)',
                          'Sample Weight (g)',
                          'Water Potential (scaled)',
                          'Dry Weight (g)',
                          'Water Weight (g)',
                          'Excess Water (scaled)'),
         dv.labels = c("Model 1", "Model 2", "Model 3", "Model 4",
                       "Model 5", "Model 6", "Model 7", "Model 8", 
                       "Model 9","Model 10", 'Model 11', 'Model 12'),
         string.pred = "Coeffcient", 
         title = "models with no collinearity",
  string.p = "P-Value", 
  p.style = "stars",
  file = here('figures', 'MEM.figures','modelselection.html')) # saving .html into figures folder
```


Rerunning this analysis with LDMC included in the model selection:

```{r}
m2.5 <- lmer(PC1 ~ mpa.scaled + spp + site + year.month + sample.wt + ldmc + (1 | individual), data = rwc_pv_df)


m3.5 <- lmer(PC1 ~ spp + site + year.month + sample.wt + ldmc + (1 | individual), data = rwc_pv_df)

m4.5 <- lmer(PC1 ~ spp + site + year.month + ldmc + (1 | individual), data = rwc_pv_df)

m5.5 <- lmer(PC1 ~ spp + mpa.scaled + site + year.month + ldmc + (1 | individual), data = rwc_pv_df)

performance::multicollinearity(m5.5)
```

Cant include the following combination in the same model (VERY collinear): 
- dw.flam.sample and LDMC
- ww.flam.sample and LDMC
- excess.water.sclaed and LDMC
- LFM and LDMC


```{r}
tab_model(m1, m2, m3, m4, m5, m6, m7, m8, m9, m10, m11, m12, m2.5, m3.5, m4.5, m5.5,
          show.reflvl = TRUE, 
          digits = 3, 
          show.aic = TRUE, 
          show.ci = FALSE,
          show.icc = FALSE, 
         #  pred.labels = c('Intercept', 'LFM (%)',
         #                  'Species Effect (C. megacarpus)',
         #                  'Site Effect (Site 1)',
         #                  'Sampling Date Effect (Jan 2018)',
         #                  'Sampling Date Effect (Dec 2019)',
         #                  'Sampling Date Effect (Jan 2020)',
         #                  'Sampling Date Effect (Sept 2020)',
         #                  'Sample Weight (g)',
         #                  'Water Potential (scaled)',
         #                  'Dry Weight (g)',
         #                  'Water Weight (g)',
         #                  'Excess Water (scaled)'),
         dv.labels = c("Model 1", "Model 2", "Model 3", "Model 4",
                       "Model 5", "Model 6", "Model 7", "Model 8",
                       "Model 9","Model 10", 'Model 11', 'Model 12', 
                       "Model 2.5", "Model 3.5"),
         string.pred = "Coeffcient", 
         title = "models with no collinearity",
  string.p = "P-Value", 
  p.style = "stars",
  file = here('figures', 'MEM.figures','modelselection.LDMC.html')) # saving .html into figures folder
```


Looks like adding in LDMC does change things a bit, but not a lot: the 'best' model are still the following:

Top by far: 

PC1 ~ mpa.scaled+ spp + site + year.month + sample.wt + (1 | individual)

lmer(PC1 ~ spp + mpa.scaled+ site + year.month + ldmc + (1 | individual), data = rwc_pv_df)




```{r}
model_list <- c(m1, m2, m3, m4, m5, m6, m7, m8, m9, m10, m11, m12, m2.5, m3.5, m4.5, m5.5)

aictab(model_list)
```
```{r}
m5.5
```



# 2.1. Predicting raw metrics (part 2)

## Time to Ignition
Models
```{r}
m1 <- lmer(tti ~ mpa.scaled + spp + site + year.month + (1 | individual), data = rwc_pv_df)

m2 <- lmer(tti ~ lfm + spp + site + year.month + (1 | individual), data = rwc_pv_df)

m3 <- lmer(tti ~ ww.flam.sample + spp + site + year.month + (1 | individual), data = rwc_pv_df)

m4 <- lmer(tti ~ dw.flam.sample + spp + site + year.month + (1 | individual), data = rwc_pv_df)

m5 <- lmer(tti ~ mpa.scaled + dw.flam.sample + sample.wt + spp + site + year.month + (1 | individual), data = rwc_pv_df)

# Adding in mpa.scaled+ WW and mpa.scaled+ LFM despite their moderate collinearity to see how they affect model selection
m6 <- lmer(tti ~ mpa.scaled + ww.flam.sample + spp + site + year.month + (1 | individual), data = rwc_pv_df)

m7 <- lmer(tti ~ mpa.scaled + lfm + sample.wt + spp + site + year.month + (1 | individual), data = rwc_pv_df)

#LDMC models: 
m2.5 <- lmer(tti ~ mpa.scaled+ spp + site + year.month + sample.wt + ldmc + (1 | individual), data = rwc_pv_df)

m3.5 <- lmer(tti ~ spp + site + year.month + sample.wt + ldmc + (1 | individual), data = rwc_pv_df)

performance::multicollinearity(m2.5)

m4.5 <- lmer(tti ~ spp + site + year.month + ldmc + (1 | individual), data = rwc_pv_df)

m5.5 <- lmer(tti ~ spp + mpa.scaled+ site + year.month + ldmc + (1 | individual), data = rwc_pv_df)
```

```{r}
model_list <- c(m1, m2, m3, m4, m5, m6, m7, m2.5, m3.5, m4.5, m5.5)

aictab(model_list)
```

conclusion: For TTI no models within delta 2.

## Flame Height
Models
```{r}
m1 <- lmer(fh ~ mpa.scaled + spp + site + year.month + (1 | individual), data = rwc_pv_df)

m2 <- lmer(fh ~ lfm + spp + site + year.month + (1 | individual), data = rwc_pv_df)

m3 <- lmer(fh ~ ww.flam.sample + spp + site + year.month + (1 | individual), data = rwc_pv_df)

m4 <- lmer(fh ~ dw.flam.sample + spp + site + year.month + (1 | individual), data = rwc_pv_df)

m5 <- lmer(fh ~ mpa.scaled + dw.flam.sample + sample.wt + spp + site + year.month + (1 | individual), data = rwc_pv_df)

# Adding in mpa.scaled+ WW and mpa.scaled+ LFM despite their moderate collinearity to see how they affect model selection
m6 <- lmer(fh ~ mpa.scaled + ww.flam.sample + spp + site + year.month + (1 | individual), data = rwc_pv_df)

m7 <- lmer(fh ~ mpa.scaled + lfm + sample.wt + spp + site + year.month + (1 | individual), data = rwc_pv_df)

m2.5 <- lmer(fh ~ mpa.scaled+ spp + site + year.month + sample.wt + ldmc + (1 | individual), data = rwc_pv_df)

performance::multicollinearity(m2.5)



m3.5 <- lmer(fh ~ spp + site + year.month + sample.wt + ldmc + (1 | individual), data = rwc_pv_df)

performance::multicollinearity(m3.5)
m4.5 <- lmer(fh ~ spp + site + year.month + ldmc + (1 | individual), data = rwc_pv_df)

m5.5 <- lmer(fh ~ spp + mpa.scaled+ site + year.month + ldmc + (1 | individual), data = rwc_pv_df)

```


```{r}
model_list <- c(m1, m2, m3, m4, m5, m6, m7, m2.5, m3.5, m4.5, m5.5)

aictab(model_list)
```
Conclusion: For FH, no models with LDMC were in the top mod.

## Glow Duration
Models
```{r}
m1 <- lmer(gd ~ mpa.scaled + spp + site + year.month + (1 | individual), data = rwc_pv_df)

m2 <- lmer(gd ~ lfm + spp + site + year.month + (1 | individual), data = rwc_pv_df)

m3 <- lmer(gd ~ ww.flam.sample + spp + site + year.month + (1 | individual), data = rwc_pv_df)

m4 <- lmer(gd ~ dw.flam.sample + spp + site + year.month + (1 | individual), data = rwc_pv_df)

m5 <- lmer(gd ~ mpa.scaled + dw.flam.sample + sample.wt + spp + site + year.month + (1 | individual), data = rwc_pv_df)

# Adding in mpa.scaled+ WW and mpa.scaled+ LFM despite their moderate collinearity to see how they affect model selection
m6 <- lmer(gd ~ mpa.scaled + ww.flam.sample + spp + site + year.month + (1 | individual), data = rwc_pv_df)

m7 <- lmer(gd ~ mpa.scaled + lfm + sample.wt + spp + site + year.month + (1 | individual), data = rwc_pv_df)

m2.5 <- lmer(gd ~ mpa.scaled+ spp + site + year.month + sample.wt + ldmc + (1 | individual), data = rwc_pv_df)

performance::multicollinearity(m2.5)



m3.5 <- lmer(gd ~ spp + site + year.month + sample.wt + ldmc + (1 | individual), data = rwc_pv_df)

performance::multicollinearity(m3.5)
m4.5 <- lmer(gd ~ spp + site + year.month + ldmc + (1 | individual), data = rwc_pv_df)

m5.5 <- lmer(gd ~ spp + mpa.scaled+ site + year.month + ldmc + (1 | individual), data = rwc_pv_df)

```


```{r}
model_list <- c(m1, m2, m3, m4, m5, m6, m7, m2.5, m3.5, m4.5, m5.5)

aictab(model_list)
```
Conclusion: For glow duration, adding in LDMC made it into the top models (close for one, not close for another)

tti ~ mpa.scaled+ spp + site + year.month + sample.wt + ldmc + (1 |  individual) (Delta AIC = 2.53)

tti ~ spp + site + year.month + sample.wt + ldmc + (1 |  individual) (Delta AIC = 1.68)


## Flame Duration


Models
```{r}
m1 <- lmer(fd ~ mpa.scaled + spp + site + year.month + (1 | individual), data = rwc_pv_df)

m2 <- lmer(fd ~ lfm + spp + site + year.month + (1 | individual), data = rwc_pv_df)

m3 <- lmer(fd ~ ww.flam.sample + spp + site + year.month + (1 | individual), data = rwc_pv_df)

m4 <- lmer(fd ~ dw.flam.sample + spp + site + year.month + (1 | individual), data = rwc_pv_df)

m5 <- lmer(fd ~ mpa.scaled + dw.flam.sample + sample.wt + spp + site + year.month + (1 | individual), data = rwc_pv_df)

# Adding in mpa.scaled+ WW and mpa.scaled+ LFM despite their moderate collinearity to see how they affect model selection
m6 <- lmer(fd ~ mpa.scaled + ww.flam.sample + spp + site + year.month + (1 | individual), data = rwc_pv_df)

m7 <- lmer(fd ~ mpa.scaled + lfm + sample.wt + spp + site + year.month + (1 | individual), data = rwc_pv_df)

m2.5 <- lmer(fd ~ mpa.scaled+ spp + site + year.month + sample.wt + ldmc + (1 | individual), data = rwc_pv_df)

m3.5 <- lmer(fd ~ spp + site + year.month + sample.wt + ldmc + (1 | individual), data = rwc_pv_df)

#performance::multicollinearity(m3.5)
m4.5 <- lmer(fd ~ spp + site + year.month + ldmc + (1 | individual), data = rwc_pv_df)

m5.5 <- lmer(fd ~ spp + mpa.scaled+ site + year.month + ldmc + (1 | individual), data = rwc_pv_df)

```


```{r}
model_list <- c(m1, m2, m3, m4, m5, m6, m7, m2.5, m3.5, m4.5, m5.5)

aictab(model_list)
```
Conclusion: For flame duration, the following model was within 2 of top model: 

m4.5 <- lmer(fd ~ spp + site + year.month + ldmc + (1 | individual), data = rwc_pv_df)

## Max. Temp.

Models
```{r}
m1 <- lmer(temp.max ~ mpa.scaled + spp + site + year.month + (1 | individual), data = rwc_pv_df)

m2 <- lmer(temp.max ~ lfm + spp + site + year.month + (1 | individual), data = rwc_pv_df)

m3 <- lmer(temp.max ~ ww.flam.sample + spp + site + year.month + (1 | individual), data = rwc_pv_df)

m4 <- lmer(temp.max ~ dw.flam.sample + spp + site + year.month + (1 | individual), data = rwc_pv_df)

m5 <- lmer(temp.max ~ mpa.scaled + dw.flam.sample + sample.wt + spp + site + year.month + (1 | individual), data = rwc_pv_df)

# Adding in mpa.scaled+ WW and mpa.scaled+ LFM despite their moderate collinearity to see how they affect model selection
m6 <- lmer(temp.max ~ mpa.scaled + ww.flam.sample + spp + site + year.month + (1 | individual), data = rwc_pv_df)

m7 <- lmer(temp.max ~ mpa.scaled + lfm + sample.wt + spp + site + year.month + (1 | individual), data = rwc_pv_df)

m2.5 <- lmer(temp.max ~ mpa.scaled+ spp + site + year.month + sample.wt + ldmc + (1 | individual), data = rwc_pv_df)

#performance::multicollinearity(m2.5)



m3.5 <- lmer(temp.max ~ spp + site + year.month + sample.wt + ldmc + (1 | individual), data = rwc_pv_df)

#performance::multicollinearity(m3.5)

m4.5 <- lmer(temp.max ~ spp + site + year.month + ldmc + (1 | individual), data = rwc_pv_df)

m5.5 <- lmer(temp.max ~ spp + mpa.scaled+ site + year.month + ldmc + (1 | individual), data = rwc_pv_df)

```


```{r}
model_list <- c(m1, m2, m3, m4, m5, m6, m7, m2.5, m3.5, m4.5, m5.5)

aictab(model_list)
```
Conclusion: For temp max, LDMC not under delta 2


## Post-Flame Glow
Models
```{r}
m1 <- lmer(pfg ~ mpa.scaled + spp + site + year.month + (1 | individual), data = rwc_pv_df)

m2 <- lmer(pfg ~ lfm + spp + site + year.month + (1 | individual), data = rwc_pv_df)

m3 <- lmer(pfg ~ ww.flam.sample + spp + site + year.month + (1 | individual), data = rwc_pv_df)

m4 <- lmer(pfg ~ dw.flam.sample + spp + site + year.month + (1 | individual), data = rwc_pv_df)

m5 <- lmer(pfg ~ mpa.scaled + dw.flam.sample + sample.wt + spp + site + year.month + (1 | individual), data = rwc_pv_df)

# Adding in mpa.scaled+ WW and mpa.scaled+ LFM despite their moderate collinearity to see how they affect model selection
m6 <- lmer(pfg ~ mpa.scaled + ww.flam.sample + spp + site + year.month + (1 | individual), data = rwc_pv_df)

m7 <- lmer(pfg ~ mpa.scaled + lfm + sample.wt + spp + site + year.month + (1 | individual), data = rwc_pv_df)

m2.5 <- lmer(pfg ~ mpa.scaled+ spp + site + year.month + sample.wt + ldmc + (1 | individual), data = rwc_pv_df)

performance::multicollinearity(m2.5)



m3.5 <- lmer(pfg ~ spp + site + year.month + sample.wt + ldmc + (1 | individual), data = rwc_pv_df)

performance::multicollinearity(m3.5)

m4.5 <- lmer(pfg ~ spp + site + year.month + ldmc + (1 | individual), data = rwc_pv_df)

m5.5 <- lmer(pfg ~ spp + mpa.scaled+ site + year.month + ldmc + (1 | individual), data = rwc_pv_df)

```


```{r}
model_list <- c(m1, m2, m3, m4, m5, m6, m7, m2.5, m3.5, m4.5, m5.5)

aictab(model_list)
```
conclusion: 
```{r}
m5
```


## Time to First Glow
Models
```{r}
m1 <- lmer(ttfg ~ mpa.scaled + spp + site + year.month + (1 | individual), data = rwc_pv_df)

m2 <- lmer(ttfg ~ lfm + spp + site + year.month + (1 | individual), data = rwc_pv_df)

m3 <- lmer(ttfg ~ ww.flam.sample + spp + site + year.month + (1 | individual), data = rwc_pv_df)

m4 <- lmer(ttfg ~ dw.flam.sample + spp + site + year.month + (1 | individual), data = rwc_pv_df)

m5 <- lmer(ttfg ~ mpa.scaled + dw.flam.sample + sample.wt + spp + site + year.month + (1 | individual), data = rwc_pv_df)

# Adding in mpa.scaled+ WW and mpa.scaled+ LFM despite their moderate collinearity to see how they affect model selection

m6 <- lmer(ttfg ~ mpa.scaled + ww.flam.sample + spp + site + year.month + (1 | individual), data = rwc_pv_df)

m7 <- lmer(ttfg ~ mpa.scaled + lfm + sample.wt + spp + site + year.month + (1 | individual), data = rwc_pv_df)
m2.5 <- lmer(ttfg ~ mpa.scaled+ spp + site + year.month + sample.wt + ldmc + (1 | individual), data = rwc_pv_df)

performance::multicollinearity(m2.5)

m3.5 <- lmer(ttfg ~ spp + site + year.month + sample.wt + ldmc + (1 | individual), data = rwc_pv_df)

performance::multicollinearity(m3.5)

m4.5 <- lmer(ttfg ~ spp + site + year.month + ldmc + (1 | individual), data = rwc_pv_df)

performance::multicollinearity(m4.5)

m5.5 <- lmer(ttfg ~ spp + mpa.scaled+ site + year.month + ldmc + (1 | individual), data = rwc_pv_df)

performance::multicollinearity(m5.5)
```


```{r}
model_list <- c(m1, m2, m3, m4, m5, m6, m7, m2.5, m3.5, m4.5, m5.5)

aictab(model_list)
```
conclusion: for TTFG, none made it in. 

```{r}
m3.5
```

#CONCLUSIONS: 

Models with LDMC were generally top-performing, but only when combined with water potential: for no flammability metric was LDMC alone the best predictor. LDMC functions in a similar manner to other covariates that incorporate aspects of dry matter content. Compared to these other metrics, LDMC performed equally well or very slightly better than LFM, sample dry weight, or sample fresh weight. AS LDMC requires full rehydration to calculate, the slight improvements in model selection will typically not outweigh the additional work of rehydrating leaves for LDMC determination. 

#3.0 SEGMENTED REGRESSION: 

#CEME

### PC1

n.s.
```{r}
o.mod.a<-lme(PC1~ldmc.unscaled + mpa.unscaled +year.month , random=~1|id, data=seg.ceme.subset)

os.mod.a<-segmented.default(o.mod.a, ~ldmc.unscaled, npsi=list(ldmc.unscaled=1))
#summarizing results (note the '.coef' argument)
#m <- slope(os.mod.a, .coef = fixef(os.mod.a))[[1]][[1]]
summary.segmented(os.mod.a)
```

###PC2
n.s.
```{r}
o.mod.a<-lme(PC2~ldmc.unscaled + mpa.unscaled +site + year.month, random=~1|id, 
             data=seg.ceme.subset)

os.mod.a<-segmented(o.mod.a, ~mpa.unscaled, npsi=list(mpa.unscaled=1))
#summarizing results (note the '.coef' argument)
#m <- slope(os.mod.a, .coef = fixef(os.mod.a))[[1]][[1]]
summary.segmented(os.mod.a)
```

###TTI
n.s.
```{r}
o.mod.a<-lme(tti~ldmc.unscaled + mpa.unscaled +year.month + site, random=~1|id, data=seg.ceme.subset)

os.mod.a<-segmented.default(o.mod.a, ~ mpa.unscaled, npsi=list(mpa.unscaled=1))
#summarizing results (note the '.coef' argument)
#m <- slope(os.mod.a, .coef = fixef(os.mod.a))[[1]][[1]]
summary.segmented(os.mod.a)
```

###FH
n.s.
```{r}
o.mod.a<-lme(fh~ldmc.unscaled + mpa.unscaled +year.month + site, random=~1|id, data=seg.ceme.subset)

os.mod.a<-segmented.default(o.mod.a, ~mpa.unscaled, npsi=list(mpa.unscaled=1))
#summarizing results (note the '.coef' argument)
#m <- slope(os.mod.a, .coef = fixef(os.mod.a))[[1]][[1]]
summary.segmented(os.mod.a)
```
###GD
n.s.
```{r}
o.mod.a<-lme(gd~ldmc.unscaled + mpa.unscaled +year.month + site, random=~1|id, data=seg.ceme.subset)

os.mod.a<-segmented.default(o.mod.a, ~mpa.unscaled, npsi=list(mpa.unscaled=1))
#summarizing results (note the '.coef' argument)
#m <- slope(os.mod.a, .coef = fixef(os.mod.a))[[1]][[1]]
summary.segmented(os.mod.a)
```

###Temp Max. 
2.407598e-01, Est = 0.4200656  (42%)

```{r}
o.mod.a<-lme(temp.max~ldmc.unscaled + mpa.unscaled +year.month + site, random=~1|id, data=seg.ceme.subset)

os.mod.a<-segmented.default(o.mod.a, ~mpa.unscaled, npsi=list(mpa.unscaled=1))
#summarizing results (note the '.coef' argument)
#m <- slope(os.mod.a, .coef = fixef(os.mod.a))[[1]][[1]]
summary.segmented(os.mod.a)
```

###Flame duration
n.s.
```{r}
o.mod.a<-lme(fd~ldmc.unscaled + mpa.unscaled +year.month + site, random=~1|id, data=seg.ceme.subset)

os.mod.a<-segmented.default(o.mod.a, ~mpa.unscaled, npsi=list(mpa.unscaled=1))
#summarizing results (note the '.coef' argument)
#m <- slope(os.mod.a, .coef = fixef(os.mod.a))[[1]][[1]]
summary.segmented(os.mod.a)
```
###PFG

n.s.
```{r}
o.mod.a<-lme(pfg~ldmc.unscaled + mpa.unscaled +year.month + site, random=~1|id, data=seg.ceme.subset)

os.mod.a<-segmented.default(o.mod.a, ~mpa.unscaled, npsi=list(mpa.unscaled=1))
#summarizing results (note the '.coef' argument)
#m <- slope(os.mod.a, .coef = fixef(os.mod.a))[[1]][[1]]
summary.segmented(os.mod.a)
```

#ADFA: 

### PC1

Not able to do version with random effects; also n.s. 
```{r}
o.mod.a<-lm(PC1~ldmc.unscaled + mpa.unscaled +site + year.month, #random=~1|id, 
             data=seg.adfa.subset)

os.mod.a<-segmented(o.mod.a, ~ldmc.unscaled, npsi=list(ldmc.unscaled=1))
#summarizing results (note the '.coef' argument)
#m <- slope(os.mod.a, .coef = fixef(os.mod.a))[[1]][[1]]
summary.segmented(os.mod.a)
```
###PC2
7.68e-05, Est = 0.698 (69.8%)
```{r}
o.mod.a<-lm(PC2~ldmc.unscaled + mpa.unscaled +site + year.month, random=~1|id, 
             data=seg.adfa.subset)

os.mod.a<-segmented(o.mod.a, ~ldmc.unscaled, npsi=list(ldmc.unscaled=1))
#summarizing results (note the '.coef' argument)
#m <- slope(os.mod.a, .coef = fixef(os.mod.a))[[1]][[1]]
summary.segmented(os.mod.a)
```

###TTI
n.s.
```{r}
o.mod.a<-lme(tti~ldmc.unscaled + mpa.unscaled +year.month + site, random=~1|id, data=seg.adfa.subset)

os.mod.a<-segmented.default(o.mod.a, ~ldmc.unscaled, npsi=list(ldmc.unscaled=1))
#summarizing results (note the '.coef' argument)
#m <- slope(os.mod.a, .coef = fixef(os.mod.a))[[1]][[1]]
summary.segmented(os.mod.a)
```

###FH
p = 2.239756e-07, Est. = 93%
```{r}
o.mod.a<-lme(fh~ldmc.unscaled + mpa.unscaled +year.month + site, random=~1|id, data=seg.adfa.subset)

os.mod.a<-segmented.default(o.mod.a, ~ldmc.unscaled, npsi=list(ldmc.unscaled=1))
#summarizing results (note the '.coef' argument)
#m <- slope(os.mod.a, .coef = fixef(os.mod.a))[[1]][[1]]
summary.segmented(os.mod.a)
```
###GD
p = 1.507566e-01, Est = 0.9752026 (97.5%)
```{r}
o.mod.a<-lme(gd~ldmc.unscaled + mpa.unscaled +year.month + site, random=~1|id, data=seg.adfa.subset)

os.mod.a<-segmented.default(o.mod.a, ~ldmc.unscaled, npsi=list(ldmc.unscaled=1))
#summarizing results (note the '.coef' argument)
#m <- slope(os.mod.a, .coef = fixef(os.mod.a))[[1]][[1]]
summary.segmented(os.mod.a)
```
###Temp Max. 
p = 1.848159e-02, Est = 0.6638564  (63%)
```{r}
o.mod.a<-lme(temp.max~ldmc.unscaled + mpa.unscaled +year.month + site, random=~1|id, data=seg.adfa.subset)

os.mod.a<-segmented.default(o.mod.a, ~ldmc.unscaled, npsi=list(ldmc.unscaled=1))
#summarizing results (note the '.coef' argument)
#m <- slope(os.mod.a, .coef = fixef(os.mod.a))[[1]][[1]]
summary.segmented(os.mod.a)
```

###Flame duration
n.s.
```{r}
o.mod.a<-lme(fd~ldmc.unscaled + mpa.unscaled +year.month + site, random=~1|id, data=seg.adfa.subset)

os.mod.a<-segmented.default(o.mod.a, ~ldmc.unscaled, npsi=list(ldmc.unscaled=1))
#summarizing results (note the '.coef' argument)
#m <- slope(os.mod.a, .coef = fixef(os.mod.a))[[1]][[1]]
summary.segmented(os.mod.a)
```
###PFG
6.712821e-04, Est = 0.9479989 
```{r}
o.mod.a<-lme(pfg~ldmc.unscaled + mpa.unscaled +year.month + site, random=~1|id, data=seg.adfa.subset)

os.mod.a<-segmented.default(o.mod.a, ~ldmc.unscaled, npsi=list(ldmc.unscaled=1))
#summarizing results (note the '.coef' argument)
#m <- slope(os.mod.a, .coef = fixef(os.mod.a))[[1]][[1]]
summary.segmented(os.mod.a)
```

#CONCLUSIONS: 

Not enough data for LDMC alone (LDMC only present for each individual), and inclusion of LDMC as a covariate with water potential also did not increase the likelihood of a significant/realistic breakpoint.
